% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordion.R
\name{accordion}
\alias{accordion}
\alias{accordionPanel}
\title{Accordion panels}
\usage{
accordion(...)

accordionPanel(header, body = NULL, collapsed = TRUE, ...)
}
\arguments{
\item{...}{Any number of \code{accordionPanel}s or named arguments passed as HTML
attributes to the parent element.}

\item{header}{A character string specifying the header of an accordion panel.}

\item{body}{A character string or tag element(s) specifying the body content
of the panel, defaults to \code{NULL}.}

\item{collapsed}{If \code{TRUE}, the accordion panel renders in the collapsed
state, defaults to \code{TRUE}.}
}
\description{
Similar to \code{collapse}, \code{accordion} is a way to hide content. An accordion
is made up of multiple panels and at most one panel is open. There is one
exception, when the page renders multiple panels may start in the open state,
however once the user interacts with accordion the default behavior returns.

\code{panel} is a helper function to facilitate building accordion panels and does
not have an explicit corresponding bootstrap class, see details for more
information.
}
\details{
\code{accordion} will generate an HTML id if an \code{id} argument is not passed in
\code{...} and child panels are assigned IDs based on the id of the accordion
parent.
}
\examples{
accordion(
  panel(
    heading = "Group 1",
    "Body text of the first group."
  ),
  panel(
    heading = "Group 2",
    "Body text of the second group."
  ),
  panel(
    heading = "Group 3",
    "Body text of the third group."
  )
)

if (interactive()) {
  shinyApp(
    ui = container(
      accordion(
        accordionPanel(
          "Option 1",
          "What do you think of option 1?",
          collapse = FALSE
        ),
        accordionPanel(
          "Option 2",
          "What do you think of option 2?"
        ),
        accordionPanel(
          "Option 3",
          "What do you think of option 3?"
        )
      )
    ),
    server = function(input, output) {

    }
  )
}

}
\seealso{
For more information on accordion's, please refer to
\url{https://v4-alpha.getbootstrap.com/components/collapse/}.
}
