% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R, R/collapse.R
\name{actions}
\alias{actions}
\alias{action}
\alias{toggleCollapseTarget}
\alias{showCollapseTarget}
\alias{hideCollapseTarget}
\title{Input actions}
\usage{
toggleCollapseTarget(id, ...)

showCollapseTarget(id, ...)

hideCollapseTarget(id, ...)
}
\arguments{
\item{id}{A character string specifying the id of a \code{\link[=navPane]{navPane()}} or
\code{\link[=collapsePane]{collapsePane()}}.}

\item{...}{Additional arguments, currently ignored.}
}
\description{
These functions are used in conjunction with yonder's reactive input functions'
\code{actions} argument.
}
