% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{align}
\alias{align}
\title{Align an element}
\usage{
align(tag, position)
}
\arguments{
\item{tag}{A tag object.}

\item{position}{One of \code{"top"}, \code{"middle"}, or \code{"bottom"} specifying how
the element is vertically aligned.}
}
\description{
The \code{align} utility function applies Bootstrap classes to change how an
\emph{inline} element or \emph{table cell} element is vertically aligned.
}
\examples{
tags$span() \%>\%
  align("bottom")

}
\seealso{
Other utilities: \code{\link{affix}},
  \code{\link{background}}, \code{\link{border}},
  \code{\link{display}}, \code{\link{float}},
  \code{\link{font}}, \code{\link{padding}},
  \code{\link{text}}, \code{\link{width}}
}
