% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{alignment}
\alias{alignment}
\title{Align element text}
\usage{
alignment(tag, default = NULL, sm = NULL, md = NULL, lg = NULL,
  xl = NULL)
}
\arguments{
\item{tag}{A tag object.}

\item{default}{One of \code{"left"}, \code{"right"}, \code{"center"}, or \code{"justified"}
specifying the default text alignment of the element.}

\item{sm}{Like \code{default}, but the text alignment is applied once the viewport
is 576 pixels wide, think phone in landscape mode.}

\item{md}{Like \code{default}, but the text alignment is applied once the
viewport is 768 pixels wide, think tablets.}

\item{lg}{Like \code{default}, but the text alignment is applied once the
viewport is 992 pixels wide, think desktop.}

\item{xl}{Like \code{default}, but the text alignment is applied once the
viewport is 1200 pixels wide, think large desktop.}
}
\description{
The \code{alignment} utility applies Bootstrap classes to change how an element's
text is aligned. Like with \link{display} or \link{padding} different text alignments
can be applied based on the viewport size.
}
\examples{


}
\seealso{
Other utilities: \code{\link{affix}},
  \code{\link{background}}, \code{\link{display}},
  \code{\link{float}}, \code{\link{font}},
  \code{\link{padding}}, \code{\link{rounded}},
  \code{\link{width}}
}
