% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design-background.R
\name{background}
\alias{background}
\title{Backgrounds}
\usage{
background(x, color)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{color}{One of \code{"primary"}, \code{"secondary"}, \code{"success"}, \code{"info"}, \code{"warning"}, \code{"danger"}, \code{"light"} or \code{"dark"} specifying the background color of the tag element, defaults to \code{NULL}.}
}
\description{
Use \code{background()} to modify the background color of a tag element.
}
\examples{

div() \%>\% background("primary")
# ->
div() \%>\% cascadess::background_color("primary")

}
\keyword{internal}
