% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge}
\alias{badge}
\title{Badges}
\usage{
badge(...)
}
\arguments{
\item{...}{Named arguments passed as HTML attributes to the parent
element or tag elements passed as children to the parent element.}
}
\description{
Small highlighted content which scales to its parent's size. A badge may
be dynamically updated with \code{\link[=replaceContent]{replaceContent()}}, in which case be sure to
pass an \code{id} argument as part of \code{...}.
}
\seealso{
Other components: 
\code{\link{alert}()},
\code{\link{blockquote}()},
\code{\link{collapsePane}()},
\code{\link{d1}()},
\code{\link{dropdown}()},
\code{\link{img}()},
\code{\link{jumbotron}()},
\code{\link{modal}()},
\code{\link{pre}()},
\code{\link{toast}()}
}
\concept{components}
