% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badgeElement}
\alias{badgeElement}
\title{Badges}
\usage{
badgeElement(...)
}
\arguments{
\item{...}{Named arguments passed as HTML attributes to the parent
element or tag elements passed as children.}
}
\description{
Small highlighted content which scales to its parent's size.
}
\details{
Use \code{\link[=replaceElement]{replaceElement()}} and \code{\link[=removeElement]{removeElement()}} to modify the contents of a
badge.
}
\examples{

## Possible colors

colors <- c(
  "red", "purple", "indigo", "blue", "cyan", "teal", "green",
  "yellow", "amber", "orange", "grey", "white"
)

div(
  lapply(colors, function(color) {
    badgeElement(color) \%>\%
      background(color) \%>\%
      margin(2)
  })
) \%>\%
  display("flex") \%>\%
  flex(wrap = TRUE)

}
\seealso{
Other rendering: \code{\link{outputElement}}
}
\concept{rendering}
