% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{border}
\alias{border}
\title{Element borders}
\usage{
border(tag, color)
}
\arguments{
\item{tag}{A tag element object.}

\item{color}{A character string specifying the border color, see details for
a list of possible colors.}
}
\description{
Apply a border color to an element.
}
\examples{
tags$h1("Hello, world!") \%>\%
  border("grey", sides = c("top", "bottom"))

tags$div() \%>\%
  border("orange")

if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          checkbarInput(
            id = NULL,
            choices = paste("Choice", 1:4)
          ) \%>\%
            background("cyan") \%>\%
            border("indigo")
        )
      )
    ),
    server = function(input, output) {

    }
  )
}

}
\seealso{
Other utilities: \code{\link{affix}}, \code{\link{align}},
  \code{\link{background}}, \code{\link{display}},
  \code{\link{float}}, \code{\link{padding}},
  \code{\link{text}}, \code{\link{width}}
}
