% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breadcrumb.R
\name{breadcrumbs}
\alias{breadcrumbs}
\title{Nav breadcrumbs}
\usage{
breadcrumbs(location = NULL)
}
\arguments{
\item{location}{Initial nav location, defaults to \code{NULL}.}
}
\description{
Build a navbar.
}
\examples{

}
