% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{bs}
\alias{bs}
\title{Bootstrap Builder Functions}
\description{
These functions are alternatives to the builder functions found in \link{tags}. In
the past, Bootstrap would automatically style a tag. Bootstrap version 4 has
taken a step back and does not automatically style tags. Instead they offer
classes like \code{.table} which will apply the previously standard styles. To
clue the user into possible, available styles and options new arguments have
been added to these rebooted builder functions.

The following rebooted builder functions are bundled into the \code{bs} object
and, include specific arguments in addition to \code{...},
\itemize{
\item \link[=table]{table, thead, tr, th, td}
\item \link[=lists]{ol, ul, dl, dt, dd}
\item \link[=img]{img, figure}
\item \link{blockquote}
}
}
