% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button-group.R
\name{button_group}
\alias{button_group}
\alias{button_toolbar}
\title{Group button inputs}
\usage{
button_group(...)

button_toolbar(...)
}
\arguments{
\item{...}{Button inputs to group.}

\item{vertical}{A boolean. If \code{TRUE}, vertically stack the buttons.}
}
\description{
A set of buttons with custom values.
}
