% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{buttonGroup}
\alias{buttonGroup}
\alias{buttonToolbar}
\title{Button groups and toolbars}
\usage{
buttonGroup(...)

buttonToolbar(..., margins = TRUE)
}
\arguments{
\item{...}{Button elements passed to \code{buttonGroup} or button groups passed to
\code{buttonToolbar}, or named arguments passed as HTML attributes to the parent
element.}

\item{margins}{If \code{TRUE}, margins are added around button groups for better
spacing, defaults to \code{TRUE}.}
}
\description{
Group together buttons.
}
\examples{
buttonGroup(
  button("Left"),
  button("Middle"),
  button("Right")
)

}
