% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{cardGroup}
\alias{cardGroup}
\alias{cardDeck}
\title{Card groups and decks}
\usage{
cardGroup(...)

cardDeck(...)
}
\arguments{
\item{...}{Any number of card elements or named arguments passed as HTML
attributes to the parent element.}
}
\description{
Group card objects together as groups or decks. Card groups placed cards
directly side by side. Card decks put space in between cards. When grouping
cards footers are aligned.
}
