% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{changeInput}
\alias{changeInput}
\title{Change input selection}
\usage{
changeInput(id, pattern, fixed = FALSE, invert = FALSE, reset = TRUE,
  session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{A character string specifying the id of an input.}

\item{pattern}{A character string specifying a regular expression, if \code{fixed}
is \code{FALSE}, values which match the expression will be selected. If \code{fixed}
is \code{TRUE}, a character vector of one or more values specifying exact values
to select.}

\item{fixed}{One of \code{TRUE} or \code{FALSE} specifying if \code{pattern} is interpreted
as a regular expression or a vector of character literals, defaults to
\code{FALSE}.}

\item{invert}{One of \code{TRUE} or \code{FALSE} specifying how values are matched,
if \code{TRUE} values which do \emph{not} match \code{pattern} will be selected, defaults
to \code{FALSE}.}

\item{reset}{One of \code{TRUE} or \code{FALSE} indicating how to handle any currently
selected values, if \code{TRUE} the current selection is dropped before
selecting new values, defaults to \code{TRUE}.}

\item{session}{A reactive context, defaults to \code{\link[=getDefaultReactiveDomain]{getDefaultReactiveDomain()}}.}
}
\description{
Use \code{changeInput()} to change an input's selected values. Values may be
selected using regular expression pattern matching or exact matches, see
\code{fixed}.
}
