% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartist.R
\name{chartistThruput}
\alias{chartistThruput}
\title{A chartist thruput}
\usage{
chartistThruput(id)
}
\arguments{
\item{id}{A character string specifying the id of the thruput control.}
}
\description{
\strong{Under development}. A reactive chart control.
}
\examples{

shinyApp(
  ui = container(
    row(
      col(
        chartistThruput("test")
      )
    )
  ),
  server = function(input, output) {

  }
)

}
