% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{checkbarInput}
\alias{checkbarInput}
\alias{radiobarInput}
\title{Checkbar and radiobar inputs}
\usage{
checkbarInput(id, choices, values = choices, selected = NULL)

radiobarInput(id, choices, values = choices, selected = NULL)
}
\arguments{
\item{id}{A character string specifying the id of the input.}

\item{choices}{A character vector or flat list of character strings
specifying the labels of the checkbar or radiobar options.}

\item{values}{A character vector, flat list of character strings, or object
to coerce to either, specifying the values of the checkbar or radiobar
options, defaults to \code{choices}.}

\item{selected}{One or more of \code{values} indicating which of the checkbar or
radiobar options are selected by default, defaults to \code{NULL}, in which case
there is no default option.}
}
\description{
These inputs behave similarly to their counter parts, checkbox and radio
inputs. However, yonder's checkbox input is a singleton value, thus the
checkbar input is more akin to shiny's checkbox group input.
}
\examples{

## An alternative to checkbox groups

checkbarInput(
  id = NULL,
  choices = c(
    "Check 1",
    "Check 2",
    "Check 3"
  ),
  selected = "Check 1"
) \%>\%
  background("blue") \%>\%
  margin(2)

## Radiobars in comparison

radiobarInput(
  id = NULL,
  choices = c(
    "fusce sagittis",
    "libero non molestie",
    "magna orci",
    "ultrices dolor"
  ),
  selected = "ultrices dolor"
) \%>\%
  background("grey")

}
\seealso{
Other inputs: \code{\link{addressInput}},
  \code{\link{buttonGroupInput}},
  \code{\link{buttonInput}}, \code{\link{checkboxInput}},
  \code{\link{dateInput}}, \code{\link{fileInput}},
  \code{\link{formInput}}, \code{\link{groupInput}},
  \code{\link{loginInput}}, \code{\link{radioInput}},
  \code{\link{rangeInput}}, \code{\link{selectInput}},
  \code{\link{textInput}}

Other inputs: \code{\link{addressInput}},
  \code{\link{buttonGroupInput}},
  \code{\link{buttonInput}}, \code{\link{checkboxInput}},
  \code{\link{dateInput}}, \code{\link{fileInput}},
  \code{\link{formInput}}, \code{\link{groupInput}},
  \code{\link{loginInput}}, \code{\link{radioInput}},
  \code{\link{rangeInput}}, \code{\link{selectInput}},
  \code{\link{textInput}}
}
\concept{inputs}
