% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkbox.R
\name{checkbox}
\alias{checkbox}
\alias{updateCheckbox}
\alias{checkboxBar}
\title{Checkbox, checkbox bars and groups}
\usage{
checkbox(label = NULL, value = NULL, inline = FALSE, check = FALSE,
  disable = FALSE, ...)

updateCheckbox(id, context, session = getDefaultReactiveDomain())

checkboxBar(labels = NULL, context = "secondary", ..., id = NULL)
}
\arguments{
\item{label, labels}{A character vector or list of tag elements used as the
checkbox labels, defaults to \code{NULL}.}

\item{value}{A character vector specifying the value of the checkbox,
defaults to \code{NULL}.}

\item{inline}{If \code{TRUE}, the button is rendered inline, defaults to \code{FALSE}.}

\item{check}{If \code{TRUE}, the button is in the checked state when initially
rendered, defaults to \code{FALSE}.}

\item{...}{Additional named arguments passed on as HTML attributes to the
parent element.}

\item{context}{Used to specify the visual context of the checkboxBar, one of
\code{"primary"}, \code{"secondary"}, \code{"success"}, \code{"info"}, \code{"warning"}, \code{"danger"},
or \code{"link"}, defaults to \code{"secondary"}.}
}
\description{
Checkbox inputs, be sure to include an \code{id} argument to register the checkbox
as a reactive input. Checkbox bars are checkboxes group and styled as a
button bar, but still act as a checkbox group.
}
\examples{

stub

}
