% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{col}
\alias{col}
\alias{row}
\alias{container}
\title{Grid system and responsive layout}
\usage{
col(..., default = NULL, small = NULL, medium = NULL, large = NULL,
  extralarge = NULL)

row(...)

container(..., fluid = TRUE)
}
\arguments{
\item{...}{Any number of child elements or named arguments passed as HTML
attributes to the parent element. \code{row}s need to be placed inside a
\code{container}. A \code{row} typically contains only \code{col}s. A \code{col} may contain
other \code{col}s or any number of other elements.}
}
\description{
Flexbox, grid layout, containers, rows, and columns (oh my). Arguments \code{width} and
\code{order} can be applied depending on the width of the page, see the details
section for more information.
}
\examples{

stub

}
