% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\title{Collapse Content}
\usage{
collapse(...)
}
\arguments{
\item{...}{Tag elements to build into a collapsible \code{div} or additional named
arguments passed as attributes to the top-level \code{div}.}
}
\description{
Build HTML content into a collapsible \code{<div>}.
}
