% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_panel_open}
\alias{collapse_panel_open}
\alias{collapse_panel_close}
\alias{collapse_panel_toggle}
\title{Collapse panel logic}
\usage{
collapse_panel_open(id, session = get_current_session())

collapse_panel_close(id, session = get_current_session())

collapse_panel_toggle(id, session = get_current_session())
}
\arguments{
\item{id}{A character string. The id of the reactive input.}

\item{session}{A shiny session object.}
}
\description{
Server-side functions for \link{collapse_panel}s.
}
\section{Functions}{
\itemize{
\item \code{collapse_panel_open()}: Open a collapse panel.

\item \code{collapse_panel_close()}: Close a collapse panel.

\item \code{collapse_panel_toggle()}: Toggle a collapse panel open or closed.

}}
