% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collate}
\alias{collate}
\title{Join HTML attribute values}
\usage{
collate(..., collapse = " ")
}
\arguments{
\item{...}{Any number of arguments to combine together as a single HTML
attribute value.}

\item{collapse}{A character string specifying how the arguments are pasted
together, passed to \code{paste}, defaults to \code{" "}.}
}
\description{
A function to join HTML attribute values with \code{htmltools} in mind. If all
arguments are \code{NULL}, \code{NULL} is returned so the attribute can be dropped. If
all arguments are the empty string, \code{NULL} is returned so the attribute can
be dropped. If all arguments are \code{NA}, then \code{NA} is returned to preserve a
boolean attribute. Otherwise, all arguments are collapsed together into a
single string suitable for an HTML attribute.
}
\examples{
collate(
  "container",
  "my-custom-class"
)

collate(
  "container",
  utils$border(round = "all")
)

}
\keyword{internal}
