% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{colors}
\alias{colors}
\title{Color text and backgrounds}
\usage{
utils$text(context)

utils$background(context)
}
\arguments{
\item{context}{A character string specifying the visual context of a block of
text or background color of an element.

For \strong{\code{utils$text}}, one of \code{"muted"}, \code{"primary"}, \code{"success"}, \code{"info"},
\code{"warning"}, \code{"danger"}, or \code{"white"}.

For \strong{\code{utils$background}}, one of \code{"muted"}, \code{"primary"}, \code{"success"},
\code{"info"}, \code{"warning"}, \code{"danger"}, \code{"inverse"} or \code{"faded"}.}
}
\description{
Utilities functions to change element text or background color. These
utilities are intended to give visual context to blocks of text or elements.
When using a dark background one may want to specify a lighter color for
any element text.
}
\seealso{
Other utilities: \code{\link{flexbox}},
  \code{\link{float}}, \code{\link{hidden}},
  \code{\link{position}}, \code{\link{sizing}},
  \code{\link{spacing}}, \code{\link{viewport}}
}
