% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{column}
\alias{column}
\alias{row}
\alias{container}
\title{Grid layout}
\usage{
column(..., width = NULL)

row(..., gutters = TRUE)

container(..., center = FALSE)
}
\arguments{
\item{...}{Any number of tags elements passed as child elements or named
arguments passed as HTML attributes to the parent element.}

\item{width}{A \link{responsive} argument. One of \code{1:12} or \code{"auto"}, defaults to
\code{NULL}.}

\item{gutters}{One of \code{TRUE} or \code{FALSE} specifying if columns inside the row
are padded, defaults to \code{TRUE}. If \code{FALSE} column content renders flush
against the border of the column. Most often you will want to leave this
\code{gutters} as \code{TRUE}.}

\item{center}{One of \code{TRUE} or \code{FALSE} specifying if the container is
responsively centered or if the container occupies the entire width of the
viewport, defaults to \code{FALSE}.}
}
\description{
These functions are the foundation of any application. Grid elements are
nested as follows: \code{container > row > column ~ column}. Columns may be nested
within columns. Columns may be created with an explicit width, 1 through 12.
To fit a column automatically to its content use \code{width = "auto"}. To divide
the space in a row evenly amongst all columns leave \code{width} as \code{NULL}. For
examples and usage tips see the sections below.
}
\section{Equal width columns}{
\preformatted{container(
  row(
    column(
      "Aliquam erat volutpat."
    ),
    column(
      "Mauris mollis tincidunt felis."
    ),
    column(
      "Cum sociis natoque penatibus et magnis dis parturient montes,",
      "nascetur ridiculus mus."
    )
  )
)
}
}

\section{Shiny's panel with sidebar layout}{
\preformatted{container(
  row(
    column(
      width = 4
    ),
    column()
  )
)
}
}

\section{Mobile friendly grids}{


Use \code{column()}s \link{responsive} \code{width} argument to make mobile friendly
applications.\preformatted{container(
  row(
    column(
      width = c(sm = 4),
      "Mauris ac felis vel velit tristique imperdiet."
    ),
    column(
      width = c(sm = 4),
      "Nam vestibulum accumsan nisl."
    ),
    column(
      width = c(sm = 4),
      "Proin neque massa, cursus ut, gravida ut, lobortis eget, lacus."
    )
  )
)
}\preformatted{container(
  row(
    column(
      width = c(sm = 4)
    ),
    column(
      width = c(sm = 8)
    )
  )
)
}
}

\section{Fit columns to their content}{
\preformatted{container(
  row(
    column(),
    column(
      width = "auto",
      "Cras placerat accumsan nulla.  Aenean in sem ac leo mollis blandit."
    ),
    column()
  )
)
}
}

\examples{

}
\seealso{
Other layout: \code{\link{fieldset}},
  \code{\link{formGroup}}, \code{\link{navbar}},
  \code{\link{tabTabs}}

Other layout: \code{\link{fieldset}},
  \code{\link{formGroup}}, \code{\link{navbar}},
  \code{\link{tabTabs}}

Other layout: \code{\link{fieldset}},
  \code{\link{formGroup}}, \code{\link{navbar}},
  \code{\link{tabTabs}}
}
