% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex.R
\name{content}
\alias{content}
\title{Align flex items}
\usage{
content(tag, default = NULL, sm = NULL, md = NULL, lg = NULL,
  xl = NULL)
}
\arguments{
\item{default}{One of \code{"start"}, \code{"end"}, \code{"center"}, \code{"between"} or
\code{"around"} specifying the default alignment of the element's flex items.}

\item{sm}{Like \code{default}, but the alignment is applied once the viewport is
576 pixels wide, think phone in landscape mode.}

\item{md}{Like \code{default}, but the alignment is applied once the viewport is
768 pixels wide, think tablets.}

\item{lg}{Like \code{default}, but the alignment is applied once the viewport is
992 pixels wide, think desktop.}

\item{xl}{Like \code{default}, but the alignment is applied once the viewport is
1200 pixels wide, think large desktop.}
}
\description{
The \code{content} function adds Bootstrap classes to a tag element to change how
flex item's align on the x-axis (when the flex direction is row) or on the
y-axis (when the flex is direction is column). For more on flex directions
see \link{direction}.
}
\examples{
lapply(1:5, tags$div) \%>\%
  tags$div() \%>\%
  display("flex") \%>\%
  content("center")

}
