% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{d1}
\alias{d1}
\alias{d2}
\alias{d3}
\alias{d4}
\alias{d5}
\alias{d6}
\title{Display headings}
\usage{
d1(...)

d2(...)

d3(...)

d4(...)

d5(...)

d6(...)
}
\arguments{
\item{...}{Any number of child elements passed to the parent element. Named
values are passed as HTML attributes to the parent element.}
}
\description{
Display headings are not meant to replace the standard HTML heading tags,
they are a stand out alternative for eye-catching titles.
}
\seealso{
Other components: 
\code{\link{alert}()},
\code{\link{badge}()},
\code{\link{collapse_panel}()},
\code{\link{modal_dialog}()},
\code{\link{toast}()}
}
\concept{components}
