% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{dateInput}
\alias{dateInput}
\title{Date time input}
\usage{
dateInput(id, value = NULL, placeholder = NULL, min = NULL, max = NULL,
  multiple = FALSE)
}
\arguments{
\item{id}{A character specifying the id of the datetime input.}

\item{value}{The default value of the input, defaults to \code{NULL}, in which
case the initial value is \code{NULL}.}

\item{min, max}{Date objects or character strings in the format \code{YYY-mm-dd}
specifying the minimum and maximum date that can be selecetd, both
default to \code{NULL} in which case there is no minimum or maximum value
respectively.}

\item{multiple}{\code{TRUE} or \code{FALSE} specifying whether multiple dates may be
selected, if \code{TRUE} the user may select multiple dates and a vector of
one or more dates is returned as the reactive value.}
}
\description{
A date time picker.
}
\examples{

if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          dateInput(
            id = "datetime",
            min = Sys.Date() + 10,
            multiple = TRUE
          )
        ),
        col(
          verbatimTextOutput("value")
        )
      )
    ),
    server = function(input, output) {
      output$value <- renderPrint({
        input$datetime
      })
    }
  )
}
}
