% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive-domain.R
\name{default_reactive_domain}
\alias{default_reactive_domain}
\title{Default reactive domain}
\usage{
default_reactive_domain()
}
\description{
A simple wrapper around \link[shiny:domains]{shiny::getDefaultReactiveDomain}.
}
