% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex.R
\name{direction}
\alias{direction}
\alias{reverse}
\title{Flex direction, rows and columns}
\usage{
direction(tag, default = NULL, sm = NULL, md = NULL, lg = NULL,
  xl = NULL)

reverse(tag, default = NULL, sm = NULL, md = NULL, lg = NULL,
  xl = NULL)
}
\arguments{
\item{tag}{A tag element.}
}
\description{
Change the direction of a flex boxes items. \code{reverse} is similar to
\code{direction}, but for \code{"row"} flex items start from the right of the parent
flex box and for \code{"column"} flex items start from the bottom of the parent
flex box.
}
\examples{

}
