% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{dots}
\alias{dots}
\title{Dots argument}
\usage{
dots(...)
}
\arguments{
\item{...}{Any number of child elements passed to the parent element. Named
values are passed as HTML attributes to the parent element.}
}
\description{
Dots argument
}
