% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{enableInput}
\alias{enableInput}
\alias{disableInput}
\title{Enable, disable input}
\usage{
enableInput(id, values = NULL, invert = FALSE,
  session = getDefaultReactiveDomain())

disableInput(id, values = NULL, invert = FALSE, reset = FALSE,
  session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{A character string specifying the reactive id of an input.}

\item{values}{A vector specifying values to enable or disable.}

\item{invert}{One of \code{TRUE} or \code{FALSE}, if \code{TRUE} choices which do \strong{not}
have a matching value are enabled or disabled, defaults to \code{FALSE}.}

\item{session}{A reactive context, defaults to \code{\link[=getDefaultReactiveDomain]{getDefaultReactiveDomain()}}.}

\item{reset}{One of \code{TRUE} or \code{FALSE}, if \code{TRUE} choices are all enabled
prior to disabling choices, defaults to \code{FALSE}.}
}
\description{
Prevent interacting with input choices.
}
