% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-awesome.R
\docType{data}
\name{fa}
\alias{fa}
\title{Font Awesome icons}
\format{An object of class \code{list} of length 875.}
\usage{
fa$<icon name>()
}
\arguments{
\item{...}{Additional named arguments passed as HTML attributes to the parent
element.}
}
\description{
A list of font awesome icons.
}
\examples{

if (interactive()) {
  shinyApp(
    ui = container(
      fluid = FALSE,
      selectInput(
        id = "name",
        choices = names(fa)
      ) \%>\%
        margins(3),
      div(
        htmlOutput("icon"),
        verbatimTextOutput("call")
      ) \%>\%
        margins(3) \%>\%
        display("flex") \%>\%
        direction("column") \%>\%
        items("center")
    ),
    server = function(input, output) {
      output$icon <- renderUI({
        fa[[input$name]](class = "fa-10x")
      })

      output$call <- renderText({
        paste0("fa$", input$name, "()")
      })
    }
  )
}

}
\keyword{datasets}
