% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design-float.R
\name{float}
\alias{float}
\title{Floats}
\usage{
float(x, side)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{side}{A \link{responsive} argument. One of \code{"left"} or \code{"right"} specifying
the side to float the element.}
}
\description{
Use \code{float()} to float an element to the left or right side of its parent
element. A newspaper layout is a classic example where an image is floated
with text wrapped around.
}
\examples{

div() \%>\% float("left")
# ->
div() \%>\% cascadess::float("left")

}
\keyword{internal}
