% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{font}
\alias{font}
\title{Text font}
\usage{
font(tag, weight = NULL)
}
\arguments{
\item{tag}{A tag element.}

\item{weight}{One of \code{"bold"}, \code{"normal"}, or \code{"light"} specifying the font
weight of the element's text, defaults to \code{NULL}, in which case the font
weight is unchanged.}
}
\description{
The \code{font} utility changes the weight and size of an element's text font.
Bold fonts are darker and heavier whereas light fonts are thinner.
}
\examples{

span("This and other news") \%>\%
  font(weight = "light")

}
\seealso{
Other utilities: \code{\link{affix}},
  \code{\link{alignment}}, \code{\link{background}},
  \code{\link{display}}, \code{\link{float}},
  \code{\link{padding}}, \code{\link{rounded}},
  \code{\link{width}}
}
\concept{utilities}
