% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{fontAwesome}
\alias{fontAwesome}
\title{Font Awesome icons}
\usage{
fontAwesome(name, solid = TRUE, size = "md", border = FALSE,
  float = NULL, fixed = FALSE, spin = FALSE, ...)
}
\arguments{
\item{name}{A character string specifying the icon name, the \code{fa-} prefix
is not necessary.}

\item{solid}{One of \code{TRUE} or \code{FALSE} specifying if the icon's solid variant
is used, defaults to \code{TRUE}. If the icon does not have a solid variant a
warning is raised.}

\item{size}{One of \code{"xs"}, \code{"sm"}, \code{"md"}, \code{"lg"}, or \code{"1x"} through \code{"10x"}
specifying the size of the icon, defaults to \code{"md"}.}

\item{border}{One of \code{TRUE} or \code{FALSE} specifying if a border is added to the
icon, defaults to \code{FALSE}.}

\item{float}{One of \code{"left"} or \code{"right"} specifying a direction in which to
float the element. This is particularly useful when wrapping text around an
icon. Defaults to \code{NULL} in which case a float is not applied.}

\item{fixed}{One of \code{TRUE} or \code{FALSE}, if \code{TRUE} the icon has a fixed width,
defaults to \code{FALSE}. This feature is particularly useful or necessary when
vertically aligning icons.}

\item{spin}{One of \code{TRUE} or \code{FALSE}, if \code{TRUE} the icon is animated and
spins, defaults to \code{FALSE}. This functionality works best with circular
icons like \code{"spinner"}, \code{"circle-notch"}, \code{"sync"}, or \code{"cog"}.}

\item{...}{Additional named arguments passed as HTML attributes to the parent
element.}
}
\description{
A set of icons developed by Fort Awesome. Icons have regular and solid
variants. The regular variant of an icon is often the inverse or outline
of the solid variant.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      fluid = FALSE,
      blockquote(
        fontAwesome("quote-left", size = "2x", border = TRUE, pull = "left"),
        "It's the questions we can't answer that teach us the most. They teach
        us how to think. If you give a man an answer, all he gains is a little
        fact. But give him a question and he'll look for his own answers.",
        source = "Patrick Rothfuss"
      )
    ),
    server = function(input, output) {

    }
  )
}

}
\seealso{
For a complete list of possible icons please refer to the Font Awesome
website, \url{https://fontawesome.com/icons?d=listing&m=free}.
}
