% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-form.R
\name{form_submit}
\alias{form_submit}
\title{Update form}
\usage{
form_submit(
  id,
  value = NULL,
  disable = NULL,
  session = default_reactive_domain()
)
}
\arguments{
\item{id}{A character string.}

\item{session}{A reactive domain.}

\item{submit}{A boolean or character string.
\itemize{
\item If \code{TRUE}, the form is submitted.
\item If \code{FALSE}, nothing happens.
\item If character, the form is submitted and the reactive value passed is the
character string specified.
}}
}
\description{
Update form
}
