% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{formGroup}
\alias{formGroup}
\alias{formRow}
\title{Input labels, help text, and formatting to inputs}
\usage{
formGroup(label, input, ..., help = NULL)

formRow(...)
}
\arguments{
\item{label}{A character string specifying a label for the input or \code{NULL}
in which case a label is not added.}

\item{input}{A tag element specifying the input to label.}

\item{...}{For \strong{formGroup}, additional named arguments passed as HTML
attributes to the parent element.

For \strong{formRow}, any number of \code{formGroup}s or additional named arguments
passed as HTML attributes to the parent element.}

\item{help}{A character string specifying help text for the input, defaults
to \code{NULL}, in which case help text is not added.}
}
\description{
Form groups are a way of labeling an input. Form rows arrange form groups
into a row and include additional styles intended for forms. The
flexibility provided by form rows and groups means you can confidently
develop shiny applications for devices and screens of varying sizes.
}
\concept{layout}
