% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{formInput}
\alias{formInput}
\title{Form inputs}
\usage{
formInput(id, ...)
}
\arguments{
\item{id}{A character string specifying an id for the form input, defaults to
\code{NULL}.}

\item{...}{Any number of inputs, other elements, or additional named
arguments passed as HTML attributes to the parent element.}
}
\description{
Form inputs are a new reactive input. A form input's reactive value is a list
of all the reactive inputs within it. Additionally, form inputs suppress
the reactive nature of their child inputs.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          formInput(
            id = "form",
            tags$label("Email"),
            emailInput(
              id = "email",
              placeholder = "Email"
            ),
            tags$label("Password"),
            passwordInput(
              id = "password",
              placeholder = "Password"
            ),
            tags$label("Radio"),
            radioInput(
              id = "options",
              choices = c(
                "Option one",
                "Option two",
                "Option three"
              )
            ),
            tags$label("Checkbox"),
            checkboxInput(
              id = "checkbox",
              choice = "Simple checkbox"
            ) \%>\%
              margins(c(0, 0, 2, 0))
          )
        ),
        col(
          verbatimTextOutput("value")
        )
      )
    ),
    server = function(input, output) {
      output$value <- renderPrint({
        str(input$form)
      })
    }
  )
}

if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          formInput(
            id = "real",
            tags$label("Real name"),
            textInput(
              id = "first",
              label = "First name"
            ),
            textInput(
              id = "last",
              label = "Last name"
            )
          )
        ),
        col(
          verbatimTextOutput("realvalue")
        ),
        col(
          formInput(
            id = "nickname",
            tags$label("Nickname"),
            textInput(
              id = "first",
              label = "First name"
            ),
            textInput(
              id = "last",
              label = "Last name"
            )
          )
        ),
        col(
          verbatimTextOutput("nickvalue")
        )
      )
    ),
    server = function(input, output) {
      output$realvalue <- renderPrint({
        str(input$real)
      })

      output$nickvalue <- renderPrint({
        str(input$nickname)
      })
    }
  )
}

}
\seealso{
Bootstrap 4 forms documentation:
\url{https://getbootstrap.com/docs/4.0/components/forms/}

Other inputs: \code{\link{buttonInput}},
  \code{\link{checkboxInput}}, \code{\link{groupInput}},
  \code{\link{radioInput}}, \code{\link{selectInput}},
  \code{\link{textInput}}
}
