% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{group}
\alias{group}
\alias{inputs$group}
\title{Input group, buttons and text}
\arguments{
\item{placeholder}{A character string specifying placeholder text for the
input group, defaults to \code{NULL}.}

\item{value}{A character string specifying an initial value for the input
group, defaults to \code{NULL}.}

\item{left, right}{A character string, \link{inputs$button} element, or \link{dropdown}
element, used as the left or right addon, respectively, of the input group,
both default to \code{NULL}. Addon's affect the reactive events and value of the
input group, see the details section below for more information.}
}
\description{
An input group is composite reactive input which may consist of one or two
buttons, dropdowns, character addons, or any combination of these elements.
Character addons, specified with \code{left} and \code{right} may be used to ensure an
input group's value always has a certain prefix or suffix and render a
visual cue to indicate this behavior. Buttons and dropdowns may be included
to control when the input group's reactive value updates. See below for more
information and examples.
}
\details{
\strong{reactive event}

If either \code{left} or \code{right} is a button or a dropdown the reactive value of
the input group is changed when either button or a dropdown item is clicked.
If either \code{left} and \code{right} are character strings or \code{NULL} the reactive
value of the input group will update each time the text input is changed by
the user.

\strong{input value}

An input group's value is \code{NULL} when its text input is empty and neither,
if any, buttons or dropdown item's have been clicked. Otherwise, a input
group's value is list of two named elements. The first element is \code{value}
which is the concatenation of \code{left}, \emph{if a character string}, the text input
value, and \code{right}, \emph{if a character string}. The second element is \code{click}
and is indicates which button or dropdown item was clicked. The value of
\code{click} depends on the value of the button or dropdown item, see
\link{inputs$button} or \link{dropdownItem}.
}
\examples{
if (interactive()) {
  library(shiny)

  shinyApp(
    ui = container(
      inputs$group(
        id = "nearlytext"
      ),
      inputs$group(
        id = "username",
        placeholder = "username",
        left = "@",
        right = inputs$button(
          label = "Go!",
          value = "go"
        )
      ),
      inputs$group(
        id = "github",
        left = "https://github.com/"
      ),
      inputs$group(
        id = "preference",
        left = inputs$button(
          label = icons$fa("thumbs-up"),
          value = "up"
        ),
        right = inputs$button(
          label = "^", #icons$fa("thumbs-down"),
          value = "down"
        )
      ),
      inputs$group(
        id = "options",
        placeholder = "",
        left = dropdown(
          label = "Action",
          dropdownItem("Action", "action"),
          dropdownItem("Another action", "another"),
          dropdownItem("Something else", "something"),
          dropdownDivider(),
          dropdownItem("Final", "final")
        )
      )
    ),
    server = function(input, output) {
      observe({
        print(input$nearlytext)
      })

      observe({
        print(input$username)
      })

      observe({
        print(input$github)
      })

      observe({
        print(input$preference)
      })

      observe({
        print(input$options)
      })
    }
  )
}

}
