% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{hidden}
\alias{hidden}
\title{Responsive hidden elements}
\usage{
utils$hidden(above = NULL, below = NULL)
}
\arguments{
\item{above}{One of \code{"small"}, \code{"medium"}, \code{"large"}, or \code{"extra-large"}
specifying the minimum viewport size at which to start hiding the element,
e.g. if \code{above = "medium"} an element is hidden on medium, large, and
extra large viewports.}

\item{below}{One of \code{"small"}, \code{"medium"}, \code{"large"}, or \code{"extra-large"}
specifying the maximum viewport size at which an element is hidden,
e.g. if \code{below = "medium"} an element is hidden on extra small, small,
and medium viewports.}
}
\description{
These utility functions allow elements to hide responsively depending on the
viewport size.
}
\seealso{
Other utilities: \code{\link{colors}},
  \code{\link{flexbox}}, \code{\link{float}},
  \code{\link{position}}, \code{\link{sizing}},
  \code{\link{spacing}}, \code{\link{viewport}}
}
