% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\docType{data}
\name{icons}
\alias{icons}
\title{Icon data}
\format{A data frame with 3563 rows and 4 columns:

\describe{
\item{set}{the icon family name}
\item{keyword}{a keyword to identify the icon by}
\item{name}{the name of the icon}
\item{prefix}{used by the font awesome family}
}}
\usage{
icons
}
\description{
Data frame of icon sets, keywords, names, and additional meta information.
}
\examples{
icons
}
\keyword{internal}
