% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{img}
\alias{img}
\alias{figure}
\alias{image}
\title{Responsive images and figures}
\usage{
img(src, ...)

figure(image, caption = NULL, ...)
}
\arguments{
\item{src}{A character string specifying the source of the image.}

\item{...}{Additional tag elements or named arguments passed as HTML attributes
to the parent element.}

\item{image}{An \code{<img>} tag, typically a call to \code{img}.}

\item{caption}{A character string specifying the image caption, defaults to
\code{NULL}.}
}
\description{
A small update to \code{tags$img} and \code{tags$figure}. Create responsive images with
\code{img}. \code{figure} has specific arguments for an image child element and image
caption.
}
\examples{

if (interactive()) {
  shinyApp(
    ui = container(
      img("https://upload.wikimedia.org/wikipedia/commons/1/18/Grey_Square.svg") \%>\%
        float("left") \%>\%
        rounded(),
      img("https://upload.wikimedia.org/wikipedia/commons/1/18/Grey_Square.svg") \%>\%
        float("right") \%>\%
        rounded()
    ),
    server = function(input, output) {

    }
  )
}

# Thank you to Wikimedia Commons
# Grey square provided by Johannes Rössel (Own work) [Public domain]

if (interactive()) {
  shinyApp(
    ui = container(
      figure(
        image = rounded(img("http://bit.ly/2qchbEB")),
        caption = "Stock cat photo."
      )
    ),
    server = function(input, output) {

    }
  )
}

}
\seealso{
Other content: \code{\link{blockquote}},
  \code{\link{card}}, \code{\link{d1}},
  \code{\link{dropdown}}, \code{\link{icon}},
  \code{\link{jumbotron}}, \code{\link{pre}},
  \code{\link{showPopover}}, \code{\link{tooltip}}

Other content: \code{\link{blockquote}},
  \code{\link{card}}, \code{\link{d1}},
  \code{\link{dropdown}}, \code{\link{icon}},
  \code{\link{jumbotron}}, \code{\link{pre}},
  \code{\link{showPopover}}, \code{\link{tooltip}}
}
\concept{content}
