% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-button.R
\name{input_button}
\alias{input_button}
\alias{update_button}
\title{Button inputs}
\usage{
input_button(id, label, ...)

update_button(
  id,
  label = NULL,
  disable = NULL,
  session = get_current_session()
)
}
\arguments{
\item{id}{A character string. The id of the reactive input.}

\item{label}{A character string. The text content of the button.}

\item{...}{Optional named arguments specifying HTML attributes for the input
element.}

\item{disable}{A boolean. The checkbox starts disabled if \code{TRUE}.}

\item{session}{A shiny session object.}
}
\value{
A \link[htmltools:tag]{htmltools::tag} object.
}
\description{
Button inputs are useful as triggers for reactive or observer expressions.
The reactive value of a button input begins as \code{NULL}, but subsequently is
the number of clicks.
}
\seealso{
Other inputs:
\code{\link[=floating_label]{floating_label()}},
\code{\link[=input_checkbox]{input_checkbox()}},
\code{\link[=input_checkbox_group]{input_checkbox_group()}},
\code{\link[=input_chip_group]{input_chip_group()}},
\code{\link[=input_form]{input_form()}},
\code{\link[=input_list_group]{input_list_group()}},
\code{\link[=input_menu]{input_menu()}},
\code{\link[=input_multi_select]{input_multi_select()}},
\code{\link[=input_numeric]{input_numeric()}},
\code{\link[=input_radio_group]{input_radio_group()}},
\code{\link[=input_range]{input_range()}},
\code{\link[=input_select]{input_select()}},
\code{\link[=input_text]{input_text()}},
\code{\link[=input_text_group]{input_text_group()}}
}
\concept{inputs}
