% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-checkbox-button.R
\name{input_checkbox_button}
\alias{input_checkbox_button}
\alias{update_checkbox_button}
\title{Checkbox button}
\usage{
input_checkbox_button(
  id,
  choices,
  ...,
  values = choices,
  select = NULL,
  disable = NULL,
  layout = "row"
)

update_checkbox_button(
  id,
  choices = NULL,
  values = choices,
  select = NULL,
  disable = NULL,
  session = get_current_session()
)
}
\arguments{
\item{id}{A character string. The id of the reactive input.}

\item{choices}{A character vector or list. The labels for the input choices.}

\item{...}{Optional named arguments specifying HTML attributes for the input
element.}

\item{values}{A character vector. The values for the input, defaults to
\code{choices}.}

\item{select}{A character vector. The values selected by default, one or more
of \code{values}.}

\item{disable}{A character vector. The values disabled by default.}

\item{layout}{A character string. The layout of the input choices, one of
\code{"column"} or \code{"row"}.}

\item{session}{A shiny session object.}
}
\value{
A shiny tag object.
}
\description{
A checkbox appearing as a button.
}
\seealso{
Other inputs: 
\code{\link{checkbarInput}()},
\code{\link{fileInput}()},
\code{\link{input_button}()},
\code{\link{input_checkbox}()},
\code{\link{input_chip}()},
\code{\link{input_form}()},
\code{\link{listGroupInput}()},
\code{\link{menuInput}()},
\code{\link{radioInput}()},
\code{\link{radiobarInput}()},
\code{\link{rangeInput}()},
\code{\link{selectInput}()},
\code{\link{textInput}()}
}
\concept{inputs}
