% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-checkbutton.R
\name{input_checkbutton}
\alias{input_checkbutton}
\title{Checkbox button}
\usage{
input_checkbutton(id, choices, values = choices, selected = NULL, ...)
}
\arguments{
\item{id}{A string.}

\item{choices}{A list.}

\item{values}{A list. Defaults to \code{choices}.}

\item{selected}{The default selected value or \code{NULL}.}
}
\description{
A checkbox appearing as a button.
}
\seealso{
Other inputs: 
\code{\link{checkbarInput}()},
\code{\link{fileInput}()},
\code{\link{input_button}()},
\code{\link{input_checkbox}()},
\code{\link{input_chip}()},
\code{\link{input_form}()},
\code{\link{listGroupInput}()},
\code{\link{menuInput}()},
\code{\link{radioInput}()},
\code{\link{radiobarInput}()},
\code{\link{rangeInput}()},
\code{\link{selectInput}()},
\code{\link{textInput}()}
}
\concept{inputs}
