% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-list-group.R
\name{input_list_group}
\alias{input_list_group}
\alias{update_list_group}
\title{List group inputs}
\usage{
input_list_group(
  id,
  choices,
  ...,
  label = NULL,
  values = choices,
  select = NULL,
  disable = NULL,
  appearance = c("default", "flush"),
  layout = c("column", "row")
)

update_list_group(
  id,
  choices = NULL,
  values = choices,
  select = NULL,
  disable = NULL,
  session = get_current_session()
)
}
\arguments{
\item{id}{A character string. The id of the reactive input.}

\item{choices}{A character vector or list. The labels for the input choices.}

\item{...}{Optional named arguments specifying HTML attributes for the input
element.}

\item{label}{A character string. The label of the input, rendered ahead
of the input. Floating labels are not supported.}

\item{values}{A character vector. The values for the input, defaults to
\code{choices}.}

\item{select}{A character vector. The values selected by default, one or more
of \code{values}.}

\item{disable}{A character vector. The values disabled by default.}

\item{appearance}{A character string. The appearance of the input's list
items.

Flush list items have their outer borders and rounded corners removed.}

\item{layout}{A character vector. The layout of the choices.}

\item{session}{A shiny session object.}
}
\description{
List group inputs are an actionable list of items. They behave similarly to
checkboxes or radios, that is, users may select one or more items from the
list. However, list group items may include highly variable content.
}
\seealso{
Other inputs:
\code{\link[=floating_label]{floating_label()}},
\code{\link[=input_button]{input_button()}},
\code{\link[=input_checkbox]{input_checkbox()}},
\code{\link[=input_checkbox_group]{input_checkbox_group()}},
\code{\link[=input_chip_group]{input_chip_group()}},
\code{\link[=input_form]{input_form()}},
\code{\link[=input_menu]{input_menu()}},
\code{\link[=input_multi_select]{input_multi_select()}},
\code{\link[=input_numeric]{input_numeric()}},
\code{\link[=input_radio_group]{input_radio_group()}},
\code{\link[=input_range]{input_range()}},
\code{\link[=input_select]{input_select()}},
\code{\link[=input_text]{input_text()}},
\code{\link[=input_text_group]{input_text_group()}}
}
\concept{inputs}
