% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{invalidateInput}
\alias{invalidateInput}
\alias{validateInput}
\title{Validate, invalidate input}
\usage{
invalidateInput(id, message, session = getDefaultReactiveDomain())

validateInput(id, session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{A character string.}

\item{message}{A character string specifying the message.}
}
\description{
\code{invalidateInput()} and \code{validateInput()} are new utilities for conveying
information to users. \code{invalidateInput()} adds text to an input letting a
user know why an input is valid or invalid. Additionally, \code{invalidateInput()}
will immediately freeze an input. As a result, further observers or reactives
using the input are not triggered. \code{validateInput()} will immediately thaw
the reactive input, thus allowing subsequent observers and reactives to
trigger.
}
