% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex.R
\name{items}
\alias{items}
\title{Flex items, cross axis alignment}
\usage{
items(tag, default = NULL, sm = NULL, md = NULL, lg = NULL, xl = NULL)
}
\arguments{
\item{default}{One of \code{"start"}, \code{"end"}, \code{"center"}, \code{"baseline"} or
\code{"stretch"} specifying the default cross axis alignment of the element's
flex items.}

\item{sm}{Like \code{default}, but the alignment is applied once the viewport is
576 pixels wide, think phone in landscape mode.}

\item{md}{Like \code{default}, but the alignment is applied once the viewport is
768 pixels wide, think tablets.}

\item{lg}{Like \code{default}, but the alignment is applied once the viewport is
992 pixels wide, think desktop.}

\item{xl}{Like \code{default}, but the alignment is applied once the viewport is
1200 pixels wide, think large desktop.}
}
\description{
The \code{items} utility function applies Bootstrap classes to a tag element in
order to change the cross axis alignment of its flex items. The element must
must use a flex display. To change the display property of a tag, see
\link{display} for more information.
}
\details{
When the flex direction is row, see \link{direction}, the cross axis is the
y-axis. When the flex direction is column, see \link{direction}, the cross axis is
the x-axis. The default direction is row. In this case, use \code{items} to
control where flex items are placed vertically within the tag element.
}
\section{Alignments}{


\strong{\code{"start"}}, flex items are aligned at the top of the parent element.\preformatted{| Item 1 | Item 2 | Item 3 | ============= |
|        |        |        |               |
|        |        |        |               |
}

\strong{\code{"end"}}, flex items are aligned at the bottom of the parent element.\preformatted{|        |        |        |               |
|        |        |        |               |
| Item 1 | Item 2 | Item 3 | ============= |
}

\strong{\code{"center"}}, flex items are aligned at the center of the parent element.\preformatted{|        |        |        |               |
| Item 1 | Item 2 | Item 3 | ============= |
|        |        |        |               |
}

\strong{\code{"baseline"}}, flex items are aligned by font size.\preformatted{| Item 1 | Item 2 | Item 3 | ============= |
|        |        |        |               |
|        |        |        |               |
}

\strong{\code{"stretch"}}, flex items stretch to fill the height of their parent
element. This is the browser defalut.\preformatted{| It     | It     | It     | ============= |
|   em   |   em   |   em   |               |
|      1 |      2 |      3 |               |
}
}

\examples{

}
\seealso{
Other flex: \code{\link{content}}, \code{\link{direction}},
  \code{\link{wrap}}
}
\concept{flex}
