% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex.R
\name{justify}
\alias{justify}
\title{Flex layout main axis}
\usage{
justify(tag, default = NULL, sm = NULL, md = NULL, lg = NULL,
  xl = NULL)
}
\arguments{
\item{tag}{A tag element.}

\item{default}{One of \code{"start"}, \code{"end"}, \code{"center"}, \code{"between"} or
\code{"around"} specifying the default alignment of the element's flex items.}

\item{sm}{Like \code{default}, but the alignment is applied once the viewport is
576 pixels wide, think phone in landscape mode.}

\item{md}{Like \code{default}, but the alignment is applied once the viewport is
768 pixels wide, think tablets.}

\item{lg}{Like \code{default}, but the alignment is applied once the viewport is
992 pixels wide, think desktop.}

\item{xl}{Like \code{default}, but the alignment is applied once the viewport is
1200 pixels wide, think large desktop.}
}
\description{
The \code{justify()} function allows you to control how elements or flex items
inside a flex element are aligned along the main axis. For more on the main
and cross axis see below.
}
\details{
The diagrams below demonstrate how \code{justify()} changes the alignment of flex
items.

\code{"start"}
\itemize{
\item \code{| Item 1 | Item 2 | Item 3 | ============= |}
}

\code{"end"}
\itemize{
\item \code{| ============= | Item 1 | Item 2 | Item 3 |}
}

\code{"center"}
\itemize{
\item \code{| ===== | Item 1 | Item 2 | Item 3 | ===== |}
}

\code{"between"}
\itemize{
\item \code{| Item 1 | ===== | Item 2 | ===== | Item 3 |}
}

\code{"around"}
\itemize{
\item \code{| = | Item 1 | = | Item 2 | = | Item 3 | = |}
}
}
\section{Main and cross axis}{

}

\examples{
lapply(1:5, tags$div) \%>\%
  tags$div() \%>\%
  display("flex") \%>\%
  content("center")

}
\seealso{
Other flex: \code{\link{align}}, \code{\link{direction}},
  \code{\link{wrap}}
}
\concept{flex}
