% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav.R
\name{linkNav}
\alias{linkNav}
\title{Page Navigation}
\usage{
linkNav(labels, panes = labels, ids = NULL, align = "right",
  block = FALSE, ...)
}
\arguments{
\item{align}{One of \code{"left"}, \code{"right"}, \code{"center"}, or \code{"horizontal"}
specifying the alignment of the nav, defaults to \code{"left"}.}

\item{...}{\code{navItem}s or \code{navDropdown}s passed to \code{nav}, \code{dropdownItem}s
passed to \code{navDropdown}, or named arguments passed as HTML attributes to
the respective parent element.}

\item{width}{If specified, one of \code{"fill"} or \code{"justify"} specifying how the
nav will fill its parent element, defaults to \code{NULL}. If \code{"justify"} all
nav links have equal width, when \code{"fill"} nav links are only equally
spaced.}

\item{vertical}{If \code{TRUE}, the nav is rendered vertically, defaults to
\code{FALSE}.}

\item{tabs}{If \code{TRUE}, the nav is rendered as a set of tabs for a tabbed
interface, defaults to \code{FALSE}.}

\item{pills}{If \code{TRUE}, \code{navItem}s are rendered as pills, defaults to
\code{FALSE}.}

\item{label}{A character vector specifying the label of a nav item,
defaults to \code{NULL}.}

\item{active}{If \code{TRUE}, the nav item renders in a an active state, defaults
to \code{FALSE}.}

\item{disabled}{If \code{TRUE}, the nav item renders in a disabled state and
will not receive click events, defaults to \code{FALSE}.}

\item{align}{One of \code{"left"}, \code{"right"}, or \code{"center"} specifying how the
nav is aligned on the page, defaults to \code{"left"}.}
}
\description{
Create a nav.
}
\details{
To center a horizontal nav specify \code{align = "horiztonal"} and \code{width = "fill"} or \code{width = "justify"}.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = linkNav(
      labels = c("Home", "Profile", "Messages", "Settings"),
      panes = list(
        tags$p(
          "To write, to rhyme, to put a slew of words to verse, these ",
          "powers I do not claim to fully grasp, but like a hem catching ",
          "the dew in the grey of early dawn, I catch a whisper ",
          "each chance I stumble from the dark."
        ),
        tags$p(
          "Stumbling down the mountain side I chanced to turn and look ",
          "upon the path I'd chanced to take."
        ),
        tags$p(
          ""
        ),
        tags$p(
          "A stone set in diamond."
        )
      )
    ),
    server = function(input, output) {

    }
  )
}

}
\seealso{
For more information about the Bootstrap nav please follow this
\href{http://v4-alpha.getbootstrap.com/components/navs/}{link}.

Other navs: \code{\link{listGroupNav}},
  \code{\link{tabsNav}}
}
