% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-group.R
\name{listGroup}
\alias{listGroup}
\alias{listGroupItem}
\alias{updateListGroupItem}
\title{List Group}
\usage{
listGroup(...)

listGroupItem(label = NULL, value = NULL, context = NULL,
  active = FALSE, disabled = FALSE, hover = FALSE, badge = NULL, ...)

updateListGroupItem(id, label = NULL, value = NULL, context = NULL,
  active = NULL, disabled = NULL, session = getDefaultReactiveDomain())
}
\arguments{
\item{...}{List items, built with \code{listGroupItem}, character strings, custom
HTML content, or named arguments passed as HTML attributes to the parent
\code{div}.}

\item{label}{List item label, defaults to \code{NULL}.}

\item{value}{The value of the list item, defaults to \code{NULL}.}

\item{context}{A character string specifying the visual context of the list
item, one of \code{"success"}, \code{"info"}, \code{"warning"}, or \code{"danger"}, defaults to
\code{NULL}.}

\item{active}{If \code{TRUE} the list item is rendered in the active state,
defaults to \code{FALSE}.}

\item{disabled}{If \code{TRUE} the list item is rendered in the disabled state,
disabled items do not receive click events, defaults to \code{FALSE}.}

\item{hover}{If \code{TRUE} the list item receives additional hover animation,
under the hood the list item is a \code{<a>} element instead of \code{<li>}, defaults
to \code{FALSE}.}

\item{badge}{A \code{\link{badge}} which is right aligned inside the list item, see
\code{\link{renderBadge}} for more on how to work with badges, defaults to \code{NULL}.}
}
\description{
A way of handling and outlining content as a list.
}
\details{
Character strings are converted into list items. Custom tags are left as is
and must include Bootstrap classes ahead of time. To specify a value for a
list item please include a \code{data-value} attribute in the \code{.list-group-item}
element.
}
\examples{
listGroup(
  "Item 1",
  "Item 2",
  "Item 3"
)

# highlight list item
listGroup(
  "Item 1",
  listGroupItem("Item 2", active = TRUE),
  "Item 3"
)

# disable list items
listGroup(
  listGroupItem("Item 1", disabled = TRUE),
  "Item 2",
  listGroupItem("Item 3", disabled = TRUE)
)

if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          width = 4,
          tags$h4("Any preferences?"),
          listGroup(
            id = "preferences",
            listGroupItem("No heights", "noheights"),
            listGroupItem("No animals", "noanimals")
          )
        ),
        col(
          tags$h4("Possible outings"),
          listGroup(
            listGroupItem(
              id = "optZoo",
              "Zoo visit!"
            ),
            listGroupItem(
              id = "optSkydive",
              "Sky diving!"
            ),
            listGroupItem(
              id = "optPern",
              "Visit Pern"
            )
          )
        )
      )
    ),
    server = function(input, output) {
      observe({
        if ("noheights" \%in\% input$preferences) {
          updateListGroupItem("optSkydive", context = "warning")
        }
      })

    }
  )
}

}
\seealso{
For more information on Bootstrap list groups please refer to the
\href{https://v4-alpha.getbootstrap.com/components/list-group/}{reference page}.
}
