% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav.R
\name{listGroupNav}
\alias{listGroupNav}
\title{List group nav}
\usage{
listGroupNav(labels, panes, selected = NULL, ids = NULL,
  context = "default", ...)
}
\arguments{
\item{labels}{A character vector or flast list of character strings
specifying the nav tab labels.}

\item{panes}{A list of character strings or custom elements specifying the
nav panes.}

\item{selected}{One of \code{labels} indicating which tab is open by default,
defaults to \code{NULL}, in which case the first tab is the default.}

\item{ids}{}

\item{context}{One of \code{"primary"}, \code{"secondary"}, \code{"success"}, \code{"info"},
\code{"warning"}. \code{"danger"}, \code{"light"}, \code{"dark"}, or \code{"default"}, specifying
the visual context of the list group nav, defaults to \code{"default"}.}
}
\description{
A list group nav.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      listGroupNav(
        labels = c("Home", "Profile", "Messages", "Settings"),
        panes = list(
          tags$p(
            "When all is said and done, the sun's brow dipped below the ",
            "horizon, then shall we rest ourselves, and then shall we ",
            "share a smile, for our hardwork is behind us for a while, ",
            "and for a while can we admire the life we build together."
          ),
          tags$p(
            "How, when the field is laid low, how, when the snow begins to ",
            "fall, shall I find you once again? I have let this wonderful ",
            "life slip through my fingers and so I am left with none, but ",
            "empty fists clenched against my sides"
          ),
          tags$p(

          ),
          tags$p(

          )
        )
      )
    ),
    server = function(input, output) {

    }
  )
}

}
\seealso{
Other navs: \code{\link{linksNav}}, \code{\link{tabsNav}}
}
