% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{lists}
\alias{lists}
\alias{ul}
\alias{ol}
\alias{dl}
\alias{dt}
\alias{dd}
\title{List Styling and Unstyling}
\usage{
bs$ul(..., styled = TRUE, inline = FALSE)

bs$ol(..., styled = TRUE, inline = FALSE)

bs$dl(...)

bs$dt(..., width = NULL, truncate = FALSE)

bs$dd(..., width = NULL)
}
\arguments{
\item{styled}{When \code{FALSE} the default list style is removed and left margins
are removed, but this applies only to the \emph{direct children elements},
defaults to \code{TRUE}.}

\item{inline}{When \code{TRUE}, list bullets are removed and the list is rendered
inline, defaults to \code{FALSE}.}

\item{width}{Description lists can make use of the row, column grid system. 1
through 12, defaults to \code{NULL}, alternate widths for different viewports
may be specified using a named multi-length vector or list, see details
section.}

\item{truncate}{When \code{TRUE} long description titles are shortened and a "..."
is added, defaults to \code{FALSE}.}
}
\description{
Rebooted builder functions for \code{<ul>}, \code{<ol>}, \code{<dl>}, \code{<dt>}, and \code{<dd>}.
Ordered and unordered lists can now be unstyled, removing bullets and some
margins. Description lists make use of the container system.
}
\details{
One can specify alternate column offsets or widths for different viewports by
using the following naming convention. A specific offset or width will be
applied when the viewport is \strong{at least} the corresponding size, see below.
\itemize{
\item \strong{unnamed}, applies to all viewports.
\item \strong{\code{sm}}, at least 576 pixels wide.
\item \strong{\code{md}}, at least 768 pixels wide.
\item \strong{\code{lg}}, at least 992 pixels wide.
\item \strong{\code{xl}}, at least 1200 pixels wide.
}

See examples for more on how to make use of this functionality.
}
\examples{

}
