% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textual.R
\name{loginInput}
\alias{loginInput}
\title{Login input}
\usage{
loginInput(id, ...)
}
\arguments{
\item{id}{A character string specifying the HTML id of the login input.}

\item{...}{Additional named arguments passed as HTML attibutes to the login
input.}
}
\description{
A composite input which consists of a username field and a password field.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      row(
        column(
          loginInput(
            id = "login"
          )
        ),
        column(
          verbatimTextOutput("value")
        )
      )
    ),
    server = function(input, output) {
      output$value <- renderPrint({
        input$login
      })

    }
  )
}

}
\seealso{
Other inputs: \code{\link{addressInput}},
  \code{\link{buttonGroupInput}},
  \code{\link{buttonInput}}, \code{\link{checkbarInput}},
  \code{\link{checkboxInput}}, \code{\link{dateInput}},
  \code{\link{fileInput}}, \code{\link{formInput}},
  \code{\link{groupInput}}, \code{\link{radioInput}},
  \code{\link{rangeInput}}, \code{\link{selectInput}},
  \code{\link{textInput}}
}
\concept{inputs}
