% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{markInvalid}
\alias{markInvalid}
\alias{markValid}
\title{Input validation}
\usage{
markInvalid(id, msg)

markValid(id, msg = NULL)
}
\arguments{
\item{id}{A character string specifying the id of an input to mark as invalid
or valid.}

\item{msg}{A character string letting the user know why an input is invalid
or valid. For \code{markInvalid} this argument is required, for \code{markValid} the
argument is optional.}
}
\description{
\code{markInvalid} and \code{markValid} are new utilities for conveying information to
users. Both functions can add text to an input letting a user know why an
input is valid or invalid. Additionally, \code{markInvalid} will immediately
freeze an input. As a result, further observers or reactives using the input
are not triggered. \code{markValid} will immediately thaw the reactive input, thus
allowing subsequent observers and reactives to trigger.
}
\examples{

}
\seealso{
Other utilities: \code{\link{collapse}},
  \code{\link{downloadEvent}}, \code{\link{updateChoices}}
}
\concept{utilities}
