% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-menu.R
\name{menu_options}
\alias{menu_options}
\title{Menu options.}
\usage{
menu_options(
  direction = c("down", "up", "left", "right"),
  align = c("left", "right"),
  close = c("auto", "inside", "outside", "manual")
)
}
\arguments{
\item{direction}{A character string.}

\item{close}{}

\item{<[responsive]>}{A character string.}
}
\value{
A list of menu options.
}
\description{
Construct an options object for a menu input.
}
