% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modal_body}
\alias{modal_body}
\alias{modal_title}
\alias{modal_header}
\alias{modal_close_icon}
\alias{modal_footer}
\alias{modal_close_button}
\title{Modal items}
\usage{
modal_body(...)

modal_title(...)

modal_header(..., close = modal_close_icon())

modal_close_icon(...)

modal_footer(..., close = modal_close_button())

modal_close_button(..., label = "Close")
}
\arguments{
\item{...}{Any number of child elements passed to the parent element. Named
values are passed as HTML attributes to the parent element.}

\item{close}{A button tag element used to close the modal, defaults to
\code{\link[=modal_close_icon]{modal_close_icon()}} for headers and \code{\link[=modal_close_button]{modal_close_button()}} for footers.}

\item{label}{A string. The text of the close button.}
}
\description{
Components of a modal dialog.
}
\section{Functions}{
\itemize{
\item \code{modal_body()}: The main content of a modal.

\item \code{modal_title()}: A title suitable for use in modal dialog.

\item \code{modal_header()}: Add a header to a modal dialog.

\item \code{modal_close_icon()}: An icon-only button used to close a modal dialog,
typically found in the modal header.

\item \code{modal_footer()}: Add a footer to a modal dialog.

\item \code{modal_close_button()}: A button with text used to close a modal dialog,
typically found in the modal footer.

}}
