% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modal_show}
\alias{modal_show}
\alias{modal_hide}
\title{Modal server functions}
\usage{
modal_show(modal, session = get_current_session())

modal_hide(session = get_current_session())
}
\arguments{
\item{modal}{A \link{modal_dialog} or string. A new modal to show or the id of an
existing modal.}

\item{session}{A shiny session object.}
}
\description{
Modal server functions.
}
