% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modalEvent}
\alias{modalEvent}
\alias{onInitialized}
\alias{modal}
\title{Modal dialogs}
\usage{
modalEvent(event, modal, domain = getDefaultReactiveDomain())

onInitialized()

modal(title, body, footer = NULL)
}
\arguments{
\item{title}{A character string specifying the modal's title.}

\item{body}{A character string specifying the body of the modal or
custom element to use as the body of the modal, defaults to \code{NULL}.}

\item{footer}{Custom tags to include at the bottom of the modal, defaults to
\code{NULL}.}
}
\description{
Modals are a flexible alert window, which disable interaction with the page
behind them. Modals may include inputs or buttons or simply include text.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(

    ),
    server = function(input, output, session) {
      modalEvent(onInitialized(), {
        modal("Initialize", "Session initialized", size = "large")
      })
    }
  )
}

}
