% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav.R
\name{nav}
\alias{nav}
\title{Page Navigation}
\usage{
nav(labels, hrefs = NULL, active = NULL, style = "links", ...)
}
\arguments{
\item{labels}{A character vector specifying the labels of the nav links.}

\item{hrefs}{A character vector specifying the hrefs of the nav links.}

\item{active}{One of \code{labels} specifying which nav link is active by default,
defaults to \code{NULL}, in which case there is no active tab.}

\item{...}{Additional named arguments to pass as HTML attributes to the
parent element.}
}
\description{
Create a nav. The nav may be styled as the default links, tabs, or pills. To
create dynamic tabbed content see \link{tabContent}.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      fluid = FALSE,
      nav(c("Link", "Link", "Link"), character(3)) \%>\%
        div() \%>\%
        border() \%>\%
        margins(2),
      nav(c("Link", "Link", "Link")) \%>\%
        direction("column") \%>\%
        div() \%>\%
        border() \%>\%
        margins(2),
      nav(c("Link", "Link", "Link"), style = "tabs") \%>\%
        div() \%>\%
        border() \%>\%
        padding(2) \%>\%
        margins(2),
      nav(c("Link", "Link", "Link"), style = "pills") \%>\%
        div() \%>\%
        border() \%>\%
        padding(2) \%>\%
        margins(2),
      nav(c("Link", "Link", "Link"), style = "pills") \%>\%
        content("center") \%>\%
        div() \%>\%
        border() \%>\%
        padding(2) \%>\%
        margins(2),
      nav(c("Link", "Link", "Link"), style = "pills") \%>\%
        content("end") \%>\%
        div() \%>\%
        border() \%>\%
        padding(2) \%>\%
        margins(2),
      nav(c("Link", "Link", "Link"), style = "pills") \%>\%
        direction("column") \%>\%
        div() \%>\%
        display("flex") \%>\%
        content("end") \%>\%
        border() \%>\%
        padding(2) \%>\%
        margins(2),
      nav(c("Link", "Link", "Link"), style = "fill") \%>\%
        div() \%>\%
        border() \%>\%
        padding(2) \%>\%
        margins(2)
    ),
    server = function(input, output) {

    }
  )
}


}
\seealso{
Bootstrap 4 nav documentation:
\url{https://getbootstrap.com/docs/4.0/components/navs/}
}
