% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar.R
\name{navbar}
\alias{navbar}
\title{Navigation bar}
\usage{
navbar(..., brand = NULL, collapse = NULL)
}
\arguments{
\item{...}{A tab toggle, inline forms, or text to add to include as part of
the navigation bar.}

\item{brand}{A tag element or text placed on the left end of the navbar,
defaults to \code{NULL}, in which case nothing is added.}

\item{collapse}{One of \code{"sm"}, \code{"md"}, \code{"lg"}, \code{"xl"}, or \code{NULL} specifying
the breakpoint at which the navbar collaspes, defaults to \code{NULL}, in which
case the navbar is always expanded.}
}
\description{
Add a navigation bar to your application with \code{navbar()}. Navigation bars may
include a tab toggle (useful for multi-page applications), inline forms
(perhaps a search feature or login item), or character strings to add simple
text. Navbars are responsive and will collapse on small screens, think mobile
device. A button is automatically added to toggle between the collapsed and
expanded states.
}
\details{
\subsection{Navbar with tabs}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{div(
  navbar(
    brand = "Navbar",
    navInput(
      .style \%>\%
        margin(right = "auto"),
      id = NULL,
      choices = c("Home", "About", "Our process")
    ),
    formInput(
      inline = TRUE,
      id = NULL,
      textInput(
        .style \%>\%
          margin(right = c(sm = 2)),
        type = "search",
        id = NULL,
        placeholder = "Search"
      ),
      formSubmit(
        .style \%>\%
          background("primary"),
        label = "Search",
        value = "search"
      )
    )
  ),
  container(
    p(". . .")
  )
)
}\if{html}{\out{</div>}}\if{html}{\out{
<div>
<nav class="navbar navbar-expand navbar-light">
<a class="navbar-brand" href="#">Navbar</a>
<button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#nav-content-433-345" aria-controls="nav-content-433-345" aria-expanded="false" aria-label="Toggle navigation">
<span class="navbar-toggler-icon"></span>
</button>
<div class="collapse navbar-collapse" id="nav-content-433-345">
<ul class="yonder-nav nav navbar-nav mr-auto">
<li class="nav-item">
<button class="nav-link btn btn-link active" value="Home">Home</button>
</li>
<li class="nav-item">
<button class="nav-link btn btn-link" value="About">About</button>
</li>
<li class="nav-item">
<button class="nav-link btn btn-link" value="Our process">Our process</button>
</li>
</ul>
</div>
</nav>
<div class="container-fluid">
<p>. . .</p>
</div>
</div>
}}

}
}
\seealso{
Other layout functions: 
\code{\link{column}()},
\code{\link{fieldset}()},
\code{\link{webpage}()}
}
\concept{layout functions}
