% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number.R
\name{numberInput}
\alias{numberInput}
\title{Number input}
\usage{
numberInput(id, value = NULL, placeholder = NULL, ..., min = NULL,
  max = NULL, step = 1)
}
\arguments{
\item{id}{A character string specifying the id of the reactive input.}

\item{value}{A number specifying the default value of the number input,
defaults to \code{NULL}.}

\item{placeholder}{A character string specifying placeholder text for the
input, defaults to \code{NULL}, in which case there is no placeholder text.}

\item{...}{Additional named arguments passed as HTML attributes to the
parent element or tag elements passed as child elements to the parent
element.}

\item{min}{A number specifying the minimum allowed value of the number input,
defaults to \code{NULL}.}

\item{max}{A number specifying the maximum allowed value of the number input,
defaults to \code{NULL}.}

\item{step}{A number specifying the increment step of the number input,
defaults to 1.}
}
\description{
A number input. Specify a minimum or maximum allowed value with \code{min} or
\code{max}. Specifying \code{step} controls how the incremement and decrement buttons
modify the input value.
}
\examples{

### Default number input

numberInput(id = "num1")

### Specify `min`, `max`, and `step`

numberInput(
  id = "num2",
  min = 1,
  max = 10,
  step = 2
)

}
