% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{open_collapse_panel}
\alias{open_collapse_panel}
\alias{close_collapse_panel}
\alias{toggle_collapse_panel}
\title{Collapse panel logic}
\usage{
open_collapse_panel(id, session = get_current_session())

close_collapse_panel(id, session = get_current_session())

toggle_collapse_panel(id, session = get_current_session())
}
\arguments{
\item{id}{A character string. The id of the reactive input.}

\item{session}{A shiny session object.}
}
\description{
Server-side functions for \link{collapse_panel}s.
}
\section{Functions}{
\itemize{
\item \code{open_collapse_panel()}: Open a collapse panel.

\item \code{close_collapse_panel()}: Close a collapse panel.

\item \code{toggle_collapse_panel()}: Toggle a collapse panel open or closed.

}}
