% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element.R
\name{outputElement}
\alias{outputElement}
\alias{replaceElement}
\alias{removeElement}
\title{Dynamic elements}
\usage{
outputElement(id, ...)

replaceElement(id, ..., session = getDefaultReactiveDomain())

removeElement(id, session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{A character string specifying a reactive id.}

\item{...}{Additional named arguments passed as HTML attributes to the parent
element or unnamed arguments passed as the new contents of the output
element.}

\item{session}{A reactive context, defaults to \code{\link[=getDefaultReactiveDomain]{getDefaultReactiveDomain()}}.}
}
\description{
An application may require dynamic content. This content may be quite simple.
The content could also be quite variable. These tools are an alternative to
the standard \code{output} related \code{render*()} functions.
}
\details{
These functions are experimental and are subject to change. Additionally,
they may be moved from this package entirely.
}
\seealso{
Other rendering: \code{\link{badgeElement}}
}
\concept{rendering}
