% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{padding}
\alias{padding}
\alias{margins}
\title{Element margins and padding}
\usage{
padding(tag, default = NULL, sm = NULL, md = NULL, lg = NULL,
  xl = NULL)

margins(tag, default = NULL, sm = NULL, md = NULL, lg = NULL,
  xl = NULL)
}
\arguments{
\item{tag}{A tag element.}

\item{default}{One of 0, 1, 2, 3, 4, 5 specifying the default margins or
padding to apply. If the margins and padding remain the same across all
viewports then only \code{default} needs to be specified.

For \strong{margins}, specifying \code{"auto"} leaves the spacing up to the browser.
For example, you could horizontally center an element for all viewports
by specifying \code{default = c(1, "auto", 1, "auto")}.}

\item{sm}{Like \code{default}, but the margins or padding are applied once the
viewport is 576 pixels wide, think phone in landscape mode.}

\item{md}{Like \code{default}, but the margins or padding are applied once the
viewport is 768 pixels wide, think tablets.}

\item{lg}{Like \code{default}, but the margins or padding are applied once the
viewport is 992 pixels wide, think desktop.}

\item{xl}{Like \code{default}, but the margins or padding are applied once the
viewport is 1200 pixels wide, think large desktop.}
}
\description{
These utilities change the padding or margins of a tag. Each argument value
may be a single value or a vector of four values. Margins and padding are
specified as 0, 1, 2, 3, 4, or 5, where 0 removes all space and 5 adds the most
space.

Specifying a single value changes the margins or padding along all four sides
of \code{tag}. To apply different margins or padding to each side pass a vector of
four values. In this case, the first value adjusts the top, second the right
side, third the bottom, and the fourth value adjusts the left side. As a wise
help page once said, think "\strong{tr}ou\strong{bl}e" to help remember the order.
}
\details{
Padding refers to the space between an element and its child element(s).

Margins refer to the space outside and around an element.
}
\examples{

if (interactive()) {
  shinyApp(
    ui = container(
      lapply(
        1:5,
        padding,
        tag = div("Nunc aliquet, augue nec") \%>\%
          width(25) \%>\%
          margins(1) \%>\%
          border("blue") \%>\%
          rounded() \%>\%
          text("center")
      )
    ) \%>\%
      display(flex = TRUE) \%>\%
      flex(wrap = TRUE),
    server = function(input, output) {

    }
  )
}

}
\seealso{
Other utilities: \code{\link{affix}},
  \code{\link{background}}, \code{\link{border}},
  \code{\link{display}}, \code{\link{float}},
  \code{\link{font}}, \code{\link{rounded}},
  \code{\link{text}}, \code{\link{width}}

Other utilities: \code{\link{affix}},
  \code{\link{background}}, \code{\link{border}},
  \code{\link{display}}, \code{\link{float}},
  \code{\link{font}}, \code{\link{rounded}},
  \code{\link{text}}, \code{\link{width}}
}
\concept{utilities}
