% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popover.R
\name{popover}
\alias{popover}
\alias{showPopover}
\alias{closePopover}
\title{Display a popover}
\usage{
popover(..., title = NULL)

showPopover(target, popover, placement = "right", duration = NULL,
  session = getDefaultReactiveDomain())

closePopover(id, session = getDefaultReactiveDomain())
}
\arguments{
\item{title}{A character string specifying a title for the popover, defaults
to \code{NULL}, in which case a title is not added.}

\item{target}{A character string specifying the id of the element where the
popover is shown.}

\item{popover}{The popover element to show, typically a call to \code{popover()}.}

\item{placement}{One of \code{"top"}, \code{"left"}, \code{"bottom"}, or \code{"right"}
specifying where the popover is positioned relative to the target tag
element indicated by \code{id}.}

\item{duration}{A positive integer specifying the duration of the popover
in seconds or \code{NULL}, in which case the popover is not automatically
removed. When \code{NULL} is specified the popover can be removed with
\code{closePopover()}.}

\item{session}{A reactive context, defaults to \code{\link[=getDefaultReactiveDomain]{getDefaultReactiveDomain()}}.}

\item{id}{A character string specifying the HTML id of a popover's target tag
element.}

\item{content}{A character string or tag element specifying the content of
the popover.}
}
\description{
Popovers are small windows of content associated with a tag element. Use
\code{popover()} to construct the element and \code{showPopover()} to add it to any tag
element with an HTML id. Popovers are great for explaining inputs and giving
hints to the users. Popovers are hidden with \code{closePopover()}.
}
\section{Example application}{
\preformatted{ui <- container(
  buttonInput("showHelp", "Help!"),
  div(
    id = "textBlock1",
    "Sociis natoque penatibus et magnis"
  ) \%>\%
    padding(3)
) \%>\%
  display("flex") \%>\%
  flex(justify = "around")

server <- function(input, output) {
  observeEvent(input$showHelp, ignoreInit = TRUE, {
    showPopover(
      target = "textBlock1",
      popover(title = "Hint", "I am a <div> element!"),
      placement = "bottom",
      duration = 4
    )
  })
}

shinyApp(ui, server)
}
}

\examples{

### Popover preview

popover(
  title = "Mr. Popover",
  "... and his penguins"
)

### No title

popover(
  "Pellentesque dapibus suscipit ligula."
)

}
\seealso{
Other content: \code{\link{alert}},
  \code{\link{blockquote}}, \code{\link{card}},
  \code{\link{d1}}, \code{\link{dropdown}},
  \code{\link{icon}}, \code{\link{img}},
  \code{\link{jumbotron}}, \code{\link{modal}},
  \code{\link{pre}}, \code{\link{progressOutlet}},
  \code{\link{tooltip}}
}
\concept{content}
