% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{position}
\alias{position}
\title{Affix elements}
\usage{
utils$position(fixed)
}
\arguments{
\item{fixed}{One of \code{"top"}, \code{"bottom"}, or \code{"sticky"} positioning and
fixing an element to the top or bottom of the viewport. If \code{"sticky"},
the element is only fixed to the top of the viewport after scrolling past
it.}
}
\description{
These utility functions are used to affix elements to the top or bottom of
a page. When the page is scrolled affixed elements will remain on the top
or bottom. A sticky element is affixed to the top of the page once the page
scrolls past it. Note that the CSS used in the sticky functionality is not
fully supported by all browsers.
}
\seealso{
Other utilities: \code{\link{colors}},
  \code{\link{flexbox}}, \code{\link{float}},
  \code{\link{hidden}}, \code{\link{sizing}},
  \code{\link{spacing}}, \code{\link{viewport}}
}
