% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{pre}
\alias{pre}
\title{Scrollable code snippets}
\usage{
pre(...)
}
\arguments{
\item{...}{Text, tag elements, or named arguments passed as HTML attributes
to the tag.}
}
\description{
The \code{pre} function adds a maximum height and scroll bar to the standard
\code{<pre>} element.
}
\examples{

}
\seealso{
Other content: \code{\link{blockquote}},
  \code{\link{card}}, \code{\link{dropdown}},
  \code{\link{icon}}, \code{\link{img}},
  \code{\link{jumbotron}}, \code{\link{typography}}
}
\concept{content}
