% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{print.targeter}
\alias{print.targeter}
\title{Explain a targetting action}
\usage{
\method{print}{targeter}(x, ...)
}
\arguments{
\item{x}{An object with class \code{"targeter"}.}

\item{...}{Additional arguments passed on to other methods.}
}
\description{
Printing a targetting action displays a message about the action performed.
}
\keyword{internal}
