% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress}
\alias{progress}
\alias{progressbar}
\alias{bar}
\alias{renderBar}
\title{Progress bars}
\usage{
progress(...)

bar(value = 0, label = NULL, context = NULL, striped = FALSE,
  animated = FALSE, ...)

renderBar(value, label = NULL, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{...}{\code{bar} elements passed to progress or named arguments passed as
HTML attributes to the respective parent element.}

\item{value}{An integer(s), between 0 and 100, specifying the initial
value(s) of a progress bar or components of a progress bar, defaults to 0.}

\item{label}{Label(s) for a progress bar or components of a progress bar,
defaults to \code{NULL}.}

\item{context}{Used to specify the visual context of the progress bar,
\code{"success"}, \code{"info"}, \code{"warning"} or \code{"danger"}.}

\item{striped}{If \code{TRUE}, the progress bar has a striped gradient, defaults
to \code{FALSE}.}

\item{animated}{If \code{TRUE}, the progress bar is animated and \code{striped} is set
to \code{TRUE}, defaults to \code{FALSE}.}
}
\description{
Create progress bars. A single progress bar may be compososed of multiple
progress bar components. These components may have individual values, labels,
and contexts. To create a composite progress bar pass multiple calls to \code{bar}
to \code{progress}. Each \code{bar} component may have its own HTML id.
}
\examples{

if (interactive()) {
  library(shiny)

  shinyApp(
    ui = container(
      progress(
        bar(id = "clicks", value = 0)
      ),
      button(id = "inc", "inc progress")
    ),
    server = function(input, output, session) {
      output$clicks <- renderBar(
        min(input$inc$count, 10) / 10 * 100,
        min(input$inc$count, 10) / 10 * 100
      )
    }
  )
}

}
