% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progressOutlet}
\alias{progressOutlet}
\alias{bar}
\alias{showBar}
\title{Progress bars}
\usage{
progressOutlet(id, ...)

bar(id, value, label = NULL, striped = FALSE, ...)

showBar(id, bar, session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{A character string specifying the HTML id of a progress output.}

\item{...}{One or more \code{bar} elements passed to a progress output or named
arguments passed as HTML attributes to the parent element.}

\item{value}{An integer between 0 and 100 specifying the initial value
of a bar.}

\item{label}{A character string specifying the label of a bar, defaults to
\code{NULL}, in which case a label is not added.}

\item{striped}{If \code{TRUE}, the progress bar has a striped gradient, defaults
to \code{FALSE}.}

\item{session}{A reactive context, defaults to \code{\link[=getDefaultReactiveDomain]{getDefaultReactiveDomain()}}.}
}
\description{
Create simple or composite progress bars. To create a composite progress bar
pass multiple calls to \code{bar} to a progress output. Each \code{bar} component has
its own id, value, label, and attributes. Furthermore, utility functions may
be applied to individual bars for added customization.
}
\examples{

}
\seealso{
Other content: \code{\link{blockquote}},
  \code{\link{card}}, \code{\link{d1}},
  \code{\link{dropdown}}, \code{\link{icon}},
  \code{\link{img}}, \code{\link{jumbotron}},
  \code{\link{modal}}, \code{\link{pre}},
  \code{\link{tooltip}}
}
\concept{content}
