% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial.R
\name{radialOutput}
\alias{radialOutput}
\alias{renderRadial}
\title{Radial output}
\usage{
radialOutput(id)

renderRadial(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{id}{A character string specifying the id of the output.}

\item{expr}{An expression.}

\item{env}{The evaluation environment.}

\item{quoted}{Is \code{expr} quoted?}
}
\description{
A radial output, suitable for precentages.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      radialOutput("radial")
    ),
    server = function(input, output) {
      output$radial <- renderRadial({

      })
    }
  )
}

}
