% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radio.R
\name{radio}
\alias{radio}
\alias{radioBar}
\title{Radio, radio bars and groups}
\usage{
radio(content = NULL, value = NULL, name = NULL, inline = FALSE,
  disable = FALSE, ...)

radioBar(content = NULL, context = "secondary", name = NULL, ...)
}
\arguments{
\item{content}{A character vector specifying the label of the radio control,
defaults to \code{NULL}.}

\item{name}{A character vector specifying the HTML name attribute of the
input tag, defaults to \code{NULL}.}

\item{disable}{If \code{TRUE}, the radio control is initialized with in a disabled
state, defaults to \code{FALSE}.}

\item{...}{Optional, additional named arguments passed as HTML attributes.}

\item{context}{Used to specify the visual context of the radioBar, one of
\code{"primary"}, \code{"secondary"}, \code{"success"}, \code{"info"}, \code{"warning"}, \code{"danger"},
or \code{"link"}, defaults to \code{"secondary"}.}
}
\description{
Create radio controls. Radio bars are stylized as button bars, but act as
radio groups.
}
