% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radio.R
\name{radioInput}
\alias{radioInput}
\title{Radios}
\usage{
radioInput(labels = NULL, values = NULL, selected = NULL,
  inline = FALSE, ..., id = NULL)
}
\arguments{
\item{labels}{A list or character vectors of labels for each of the
individual radio elements, defaults to \code{NULL}.}

\item{values}{A list or vector of values for each of the individual radio
elements, defaults to \code{NULL}.}

\item{selected}{A logical vector specifying which of the radio choices
renders in the selected state, defaults to \code{NULL}. If unspecified the first
value is selected by default.}

\item{inline}{If \code{TRUE}, the radio options render inline, otherwise the
options appear on individual lines, defaults to \code{FALSE}.}

\item{...}{Additional named arguments passed on to the parent element as
HTML attributes.}

\item{id}{A character string specifying the id of the radio input,}
}
\description{
Create an input of one or more radio inputs. If an \code{id} parameter is
specified the \code{name} attribute of each child radio element is given this
value.
}
\examples{

stub

}
\seealso{
Other inputs: \code{\link{button}},
  \code{\link{groupInput}}, \code{\link{selectInput}},
  \code{\link{textInput}}
}
