% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radio.R
\name{radiosInput}
\alias{radiosInput}
\title{Radios}
\usage{
radiosInput(labels = NULL, values = NULL, stacked = FALSE, ...)
}
\arguments{
\item{labels}{A list or character vectors of labels for each of the
individual radio elements, defaults to \code{NULL}.}

\item{values}{A list or vector of values for each of the individual radio
elements, one value may be named \code{checked} to indicate a default value for
the element, defaults to \code{NULL}.}

\item{stacked}{If \code{TRUE} the radio elements appear on sepearate lines,
defaults to \code{FALSE}, in which case the radio elements are rendered inline.}

\item{...}{Additional named arguments passed on to the parent element as
HTML attributes.}
}
\description{
Create an input of one or more radio inputs. If an \code{id} parameter is
specified the \code{name} attribute of each child radio element is given this
value.
}
\examples{

stub

}
\seealso{
Other inputs: \code{\link{button}},
  \code{\link{groupInput}}, \code{\link{selectInput}},
  \code{\link{textInput}}
}
