% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{replaceContent}
\alias{replaceContent}
\alias{removeContent}
\title{Dynamic content}
\usage{
replaceContent(id, ..., session = getDefaultReactiveDomain())

removeContent(id, session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{A character string specifying a reactive id.}

\item{...}{Additional named arguments passed as HTML attributes to the parent
element or unnamed arguments passed as the new contents of the output
element.}

\item{session}{A reactive context, defaults to \code{\link[=getDefaultReactiveDomain]{getDefaultReactiveDomain()}}.}
}
\description{
An application may require dynamic content. This content may be quite simple.
The content could also be quite variable. These tools are an alternative to
the standard \code{output} related \code{render*()} functions.
}
\details{
These functions are experimental and are subject to change. Additionally,
they may be moved from this package entirely.
}
