% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rounded}
\alias{rounded}
\title{Round tag element corners}
\usage{
rounded(tag, sides = "all")
}
\arguments{
\item{tag}{A tag element.}

\item{sides}{One of \code{"top"}, \code{"right"}, \code{"bottom"}, \code{"left"}, \code{"circle"}, \code{"all"} or
\code{"none"}, defaults to \code{"all"}.}
}
\description{
The \code{rounded} utility function applies Bootstrap classes to an element.
The styles are applied by sides, e.g. \code{"left"} or \code{"bottom"}. The \code{"circle"}
value, as it suggests, rounds the corners such that the element appears
as a circle.
}
\examples{

}
\seealso{
Other utilities: \code{\link{affix}},
  \code{\link{alignment}}, \code{\link{background}},
  \code{\link{display}}, \code{\link{float}},
  \code{\link{font}}, \code{\link{padding}},
  \code{\link{width}}
}
