% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{scroll}
\alias{scroll}
\title{Vertical and horizontal scroll}
\usage{
scroll(.tag, direction)
}
\arguments{
\item{.tag}{A tag element.}

\item{direction}{One of \code{"x"} or \code{"y"} specifying which direction to scroll
the tag's content, defaults to \code{"y"}, in which case horizontal scroll is
applied.}
}
\description{
Many of the applications you build depsite a complex layout will still fit
onto a single page. To help scroll long content along side shorter content
use the \code{scroll()} utility function.
}
\examples{

div(
  rep(c("Integer placerat tristique nisl.", br()), 20),
) \%>\%
  height(25) \%>\%
  scroll()

}
