% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{selectInput}
\alias{selectInput}
\alias{option}
\title{Select input}
\usage{
selectInput(id, labels, choices = labels, selected = NULL, ...)

option(label = NULL, value = NULL, selected = FALSE)
}
\arguments{
\item{id}{A character string specifying the id of the select input.}

\item{labels}{A character vector specifying the labels of the select input
options.}

\item{choices}{A character vector specifying the values of the select input
options, defaults to \code{labels}.}

\item{selected}{One of \code{choices} indicating the default value of the select
input, defaults to \code{NULL}. If \code{NULL} the first value is selected by
default.}

\item{...}{Additional named arguments passed as HTML attributes to the parent
element.}
}
\description{
A select input.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          selectInput(
            id = "select",
            c("Choose one", "One", "Two", "Three"),
            list(NULL, 1, 2, 3)
          )
        ),
        col(
          display4(
            textOutput("value")
          )
        )
      )
    ),
    server = function(input, output) {
      output$value <- renderText({
        input$select
      })
    }
  )
}


}
\seealso{
Other inputs: \code{\link{buttonInput}},
  \code{\link{checkboxInput}}, \code{\link{formInput}},
  \code{\link{groupInput}}, \code{\link{radioInput}},
  \code{\link{textInput}}
}
