% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{sendAlert}
\alias{sendAlert}
\title{Send and add alerts}
\usage{
sendAlert(id, content, context = "secondary",
  session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{A character vector of HTML id(s) where the alerts are inserted.}

\item{content}{Text and elements to include}

\item{context}{One of \code{"success"}, \code{"info"}, \code{"warning"}, or \code{"danger"}
specifying the visual context of the alert, defaults to \code{"secondary"}.}

\item{session}{A \code{session} object passed to the shiny server function,
defaults to \code{\link[=getDefaultReactiveDomain]{getDefaultReactiveDomain()}}.}
}
\description{
\code{sendAlert}
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      `font-awesome`(),
      buttonInput(id = "button", "A button")
    ),
    server = function(input, output) {
      observeEvent(input$button, {
        sendAlert(
          id = "button",
          content = "You clicked the button!",
          context = "warning"
        )
      })
    }
  )
}

}
\seealso{
For more information on bootstrap alerts please refer to the
\href{https://v4-alpha.getbootstrap.com/components/alerts/}{reference page}.
}
