% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{sendModal}
\alias{sendModal}
\alias{closeModal}
\title{Modal dialogs}
\usage{
sendModal(title, body, footer = NULL,
  session = getDefaultReactiveDomain())

closeModal(session = getDefaultReactiveDomain())
}
\arguments{
\item{title}{A character string specifying the modal's title.}

\item{body}{A character string specifying the body of the modal or
custom element to use as the body of the modal, defaults to \code{NULL}.}

\item{footer}{Custom tags to include at the bottom of the modal, defaults to
\code{NULL}.}

\item{session}{A reactive context, defaults to \code{\link[=getDefaultReactiveDomain]{getDefaultReactiveDomain()}}.}
}
\description{
Modals are a flexible alert window, which disable interaction with the page
behind them. Modals may include inputs or buttons or simply include text.
}
\section{Displaying a modal}{


sendModal(
title = "A simple modal",
body = paste(
"Cras mattis consectetur purus sit amet fermentum.",
"Cras justo odio, dapibus ac facilisis in, egestas",
"eget quam. Morbi leo risus, porta ac consectetur",
"ac, vestibulum at eros."
)
)
}

\seealso{
Other server: \code{\link{showPopover}}
}
\concept{server}
