% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{show_modal}
\alias{show_modal}
\alias{hide_modal}
\title{Modal server functions}
\usage{
show_modal(modal, session = get_current_session())

hide_modal(session = get_current_session())
}
\arguments{
\item{modal}{A \link{modal_dialog} or string. A new modal to show or the id of an
existing modal.}

\item{session}{A shiny session object.}
}
\description{
Modal server functions.
}
