% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{showAlert}
\alias{showAlert}
\title{Show an alert}
\usage{
showAlert(text, ..., color = NULL, duration = 4)
}
\arguments{
\item{text}{A character string specifying the message text of the alert.}

\item{...}{Additional named arguments passed as HTML attributes to the
alert element.}

\item{color}{A character string specifying the color of the alert,
for possible colors see \link{background}.}

\item{duration}{A positive integer specifying the duration of the alert,
by default the alert is removed after 4 seconds.}
}
\description{
Use \code{showAlert} to let the user know of successes or to call attention to
problems.
}
\examples{

if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          buttonInput("show", "Alert!")
        )
      )
    ),
    server = function(input, output) {
      observeEvent(input$show, {
        color <- sample(c("teal", "red", "orange", "blue"), 1)
        showAlert("Alert", color = color)
      })
    }
  )
}

}
\seealso{
Boostrap 4 alert documentation:
\url{https://getbootstrap.com/docs/4.0/components/alerts/}
}
