% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sizing}
\alias{sizing}
\title{Element sizing}
\usage{
utils$width(percentage = NULL, max = NULL)

utils$height(percentage = NULL, max = NULL)
}
\arguments{
\item{percentage}{One of 25, 50, 75, or 100 specifying an element's width or
height of an element as a percentage of its parent's width or height,
respectively, defaults to \code{NULL}.}

\item{max}{One of 25, 50, 75, or 100 specifying an element's maximum width or
height as a percentage of its parent's width or height, respectively,
defaults to \code{NULL}.}
}
\description{
Construct utility classes to modify an element's width or height. The width
and height of a child element are controlled relative to their parent's
width and height, respectively.
}
\seealso{
Other utilities: \code{\link{colors}},
  \code{\link{flexbox}}, \code{\link{float}},
  \code{\link{hidden}}, \code{\link{position}},
  \code{\link{spacing}}, \code{\link{viewport}}
}
