% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{sliderInput}
\alias{sliderInput}
\title{Slider with custom choices}
\usage{
sliderInput(id, choices, values = choices, selected = values[[1]],
  ticks = TRUE, fill = FALSE, prefix = NULL, suffix = NULL, ...)
}
\arguments{
\item{id}{A character string specifying the id of the reactive input.}

\item{choices}{A character vector specifying the slider labels.}

\item{values}{A character vector specifying the values of the input,
defaults to \code{choices}.}

\item{selected}{One of \code{values} specifying the initial value of the slider
input, defaults to \code{values[[1]]}.}

\item{ticks}{One of \code{TRUE} or \code{FALSE} specifying if tick marks are added to
the input, defaults to \code{TRUE}.}

\item{fill}{One of \code{TRUE} or \code{FALSE} specifying whether the filled portion of
the slider input is shown, defaults to \code{FALSE}.}

\item{prefix}{A character string specifying a prefix for the input value,
defaults to \code{NULL}, in which case a prefix is not prepended.}

\item{suffix}{A character string specifying a suffix for the input value,
defaults to \code{NULL}, in which case a prefix is not appended.}

\item{...}{Additional named arguments passed as HTML attributes to the parent
element.}
}
\description{
A slider input with custom choices and values. Unlike \code{\link[=rangeInput]{rangeInput()}} or
\code{\link[=intervalInput]{intervalInput()}} the slider input does range over a set of numeric values.
Instead the slider input may be used similarly to a radio input.
}
\examples{

### Custom chocies

# Select a value from a set of choices using a slider.

sliderInput(
  id = "slider1",
  choices = c("Closest", "Close", "Far", "Farthest")
)

}
\seealso{
Other inputs: \code{\link{buttonGroupInput}},
  \code{\link{buttonInput}}, \code{\link{checkboxInput}},
  \code{\link{fileInput}}, \code{\link{formInput}},
  \code{\link{groupInput}}, \code{\link{listGroupInput}},
  \code{\link{menuInput}}, \code{\link{navInput}},
  \code{\link{radioInput}}, \code{\link{rangeInput}},
  \code{\link{selectInput}}, \code{\link{textInput}}
}
\concept{inputs}
