% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{spacing}
\alias{spacing}
\title{Element Spacing}
\usage{
utils$margin(size, side = "all", viewport = NULL)

utils$padding(size, side = "all", viewport = NULL)
}
\arguments{
\item{size}{An integer, 0 through 5, specifying the relative adjustment of an
element's margin or padding, where 3 is the default amount of spacing, 1 is
0.25 times the default, and 5 is 3 times the default. 0 removes margins or
padding.}

\item{side}{One of \code{"top"}, \code{"left"}, \code{"bottom"}, \code{"right"}, or \code{"all"}
specifying which side(s) to apply spacing to, defaults to \code{"all"}.}

\item{viewport}{One of \code{"small"}, \code{"medium"}, \code{"large"}, or \code{"extra-large"},
specifying the screen size at which the style is applied, defaults
to \code{NULL}.}
}
\description{
Use these utilities to add responsive padding or margins to a tag element.
Both \code{margin} and \code{padding} are viewport responsive. Margin and padding sizes
are specified relative to the font size of a page's root \code{<html>} element.
}
\seealso{
The complete list of \link{utilities}. For more on viewport
responsiveness see \link{viewport}.

Other utilities: \code{\link{colors}},
  \code{\link{flexbox}}, \code{\link{float}},
  \code{\link{hidden}}, \code{\link{position}},
  \code{\link{sizing}}, \code{\link{viewport}}
}
