% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkline.R
\name{sparklineOutput}
\alias{sparklineOutput}
\alias{renderSparkline}
\title{Sparkline output}
\usage{
sparklineOutput(id, type = "bar", labels = TRUE, ...)

renderSparkline(values, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{id}{A character string specifying the id of the sparkline output.}

\item{type}{One of \code{"bar"}, \code{"dot"}, or \code{"line"} specifying the type of
chart to render, defaults to \code{"bar"}.}

\item{labels}{If \code{TRUE}, the first and last value of the sparkline data
are used as labels, defaults to \code{TRUE}.}

\item{values}{A numeric vector of values specifying the sparkline data
points. Bar and dot sparklines expect values to be between 0 and 100 and
line sparklines expect values to be between 0 and 10.}
}
\description{
Display concise, reactive inline bar, line, or dot charts.
}
\examples{

}
