% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{spinnerOutput}
\alias{spinnerOutput}
\alias{startSpinner}
\alias{stopSpinner}
\title{A spinner}
\usage{
spinnerOutput(id, type = "circle", pulse = FALSE, ...)

startSpinner(id, session = getDefaultReactiveDomain())

stopSpinner(id, session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{A character specifying the id of the spinner output.}

\item{type}{One of \code{"circle"}, \code{"cog"}, \code{"dots"}, or \code{"sync"} specifying the
type of spinner, defaults to \code{"circle"}.}

\item{pulse}{One of \code{TRUE} or \code{FALSE}, if \code{TRUE} the spinner rotates in 8
discrete steps, defaults to \code{FALSE}.}
}
\description{
Start or stop a spinner based on process progress.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      col(
        spinnerOutput("spin"),
        buttonInput("trigger", "Start/stop")
      ) \%>\%
        display("flex") \%>\%
        content("around")
    ),
    server = function(input, output) {
      observeEvent(input$trigger, {
        if (input$trigger \%\% 2 == 1) {
          startSpinner("spin")
        } else {
          stopSpinner("spin")
        }
      })
    }
  )
}

}
