% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{streamOutput}
\alias{streamOutput}
\alias{sendStreamMessage}
\title{Stream notifications}
\usage{
streamOutput(id, ...)

sendStreamMessage(id, message, context = NULL,
  session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{A character string specifying the id of the stream output.}

\item{message}{A character string specifying an update message.}

\item{context}{One of \code{"secondary"}, \code{"success"}, \code{"info"}, \code{"warning"}, or
\code{"danger"} specifying the visual context of the button input, defaults to
\code{"secondary"}.}

\item{session}{A \code{session} object passed to the shiny server function,
defaults to \code{\link[=getDefaultReactiveDomain]{getDefaultReactiveDomain()}}.}
}
\description{
The stream output is used to send updates to the user during long-running
processes. Unlike conventional reactive outputs, the stream output does not
have a render function instead messages are sent with \code{sendStreamMessage}.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          buttonInput(id = "trigger", "Go!")
        ),
        col(
          streamOutput(
            id = "stream"
          )
        )
      )
    ),
    server = function(input, output, session) {
      observeEvent(input$trigger, {
        for (i in seq_len(5)) {
          updateStream(
            id = "stream",
            message = paste("Update:", i),
            context = if (i \%\% 2) "warning" else "info"
          )
          Sys.sleep(1)
        }
      })
    }
  )
}

}
