% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav.R
\name{tabbableContent}
\alias{tabbableContent}
\alias{tabPane}
\title{Create dynamic tabbed content}
\usage{
tabbableContent(nav, ...)

tabPane(..., fade = TRUE)
}
\arguments{
\item{nav}{A \code{nav()} element.}

\item{...}{Any number of \code{tabPane}s or named attributes passed as HTML
attributes to the parent element.}

\item{fade}{One of \code{TRUE} or \code{FALSE}, if \code{TRUE} the tab pane fades in,
defaults to \code{TRUE}.}
}
\description{
Create tabbable content stylized as tabs, horizontal pills, or vertical
pills. Tab sets or tab panels are distinct from navs.
}
\details{
To know which tab is active check the reactive value of the nav.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          tabbableContent(
            nav = nav(
              id = "tabs",
              labels = c("One", "Two", "Three"),
              style = "tabs"
            ) \%>\%
              margins(c(0, 0, 3, 0)),
            tabPane(
              "A tab"
            ),
            tabPane(
              "Another tab"
            ),
            tabPane(
              "A final tab"
            )
          )
        ),
        col(
          verbatimTextOutput("which")
        )
      )
    ),
    server = function(input, output) {
      output$which <- renderPrint({
        input$tabs
      })
    }
  )
}

}
