% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{tabContent}
\alias{tabContent}
\alias{tabPane}
\title{Tab content, panes}
\usage{
tabContent(...)

tabPane(..., fade = TRUE)
}
\arguments{
\item{...}{For \strong{tabContent}*, calls to \code{tabPane} or named arguments passed
as HTML attributes to the parent element.

For \strong{tabPane}, any number of tag elements or named arguments passed as
HTML attributes to the parent element.}

\item{id}{A character string specifying the id of the tab pane, this id
must be passed to a tab list, see \strong{panes} in \link{tabList}.}
}
\description{
Create tabbed content.
}
\examples{

}
\seealso{
Other tabs: \code{\link{tabList}}

Other tabs: \code{\link{tabList}}
}
