% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{tabList}
\alias{tabList}
\title{Tab list}
\usage{
tabList(id, labels, values = labels, active = values[1], ...)
}
\arguments{
\item{id}{A character string specifying the id of the tab list, this
value must also be passed to \code{\link[=tabContent]{tabContent()}} as the \code{list} argument.

Tabs have a reactive value which is by default is the label of the open
tab. For custom values see \code{values}.}

\item{labels}{A character vector specifying the labels of the tabs.}

\item{values}{A character vector specifying a reactive value for each tab,
defaults to \code{labels}.}

\item{active}{One of \code{values} specifying which tab is initially shown,
defaults to \code{values[1]}.`}

\item{...}{Additional named arguments passed as HTML attributes to the
parent element.}
}
\description{
Create a set of tabs for controlling tab content. Tab lists are separated
from tab content or tab panes for flexible placement, see the example below
combining cards and tabs. Because of this separation it is important to
remember a tab list does very little alone. Be sure to add corresponding
\link[=tabPane]{tab panes}.
}
\examples{

tabList(
  labels = c("Home", "About", "Posts"),
  values = c("home", "about", "blog")
)

if (interactive()) {
  shinyApp(
    ui = container(
      card(
        header = tabList(
          id = "tabs",
          labels = c("Home", "Profile", "Contact"),
        ),
        tabContent(
          list = "tabs",
          tabPane(
            "Vestibulum id ligula porta felis euismod semper. Cras justo
             odio, dapibus ac facilisis in, egestas eget quam."
          ),
          tabPane(
            "Duis mollis, est non commodo luctus, nisi erat porttitor ligula,
             eget lacinia odio sem nec elit. Aenean eu leo quam. Pellentesque
             ornare sem lacinia quam venenatis vestibulum."
          ),
          tabPane(
            "Donec ullamcorper nulla non metus auctor fringilla. Nullam id
             dolor id nibh ultricies vehicula ut id elit."
          )
        )
      ) \%>\%
        margins(4) \%>\%
        width(50)
    ),
    server = function(input, output) {
      observe({
        showAlert(input$tabs, duration = 2, color = "grey")
      })
    }
  )
}
}
\seealso{
Other tabs: \code{\link{tabContent}}
}
