% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav.R
\name{tabsNav}
\alias{tabsNav}
\alias{pillsNav}
\title{Navigation tabs}
\usage{
tabsNav(labels, panes, ids = NULL, align = "right")

pillsNav(labels, panes = labels, ids = NULL, align = "right",
  block = FALSE)
}
\arguments{
\item{labels}{A character vector or list of character vectors specifying
the labels of the nav's tabs, named items are converted into dropdown tabs.}

\item{panes}{A list of custom tags specifying the panes controlled by the
nav. Unlike \code{labels} or \code{ids} this is a flat list of items.}

\item{ids}{A character vector or list of character vectors mimicing the
structure of \code{labels}, but specifying the HTML ids to use for each tab and
pane pair, defaults to \code{NULL}, in which case ids are automatically
generated.}
}
\description{
A set of navigation tabs.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      tabsNav(
        labels = list(
          "Home",
          "Profile",
          "Other" = c("Action", "Another action")
        ),
        panes = list(
          row(
            col(
              formInput(
                id = "form",
                numberInput(
                  id = "columns",
                  label = "Number of columns",
                  value = 1,
                  min = 0,
                ),
                numberInput(
                  id = "rows",
                  label = "Number of rows",
                  value = 5,
                  min = 0
                )
              )
            ),
            col(
              tableThruput(id = "table")
            )
          ),
          tags$p(
            "The window, the window, the second story window."
          ),
          tags$p(
            "We must away, ere break of day, ",
            "To find our long-forgotten gold."
          ),
          tags$div(
            lead("This text is stylized"),
            tags$p(
              "This is the body of the whole text and nothing but the ",
              "text."
            )
          )
        )
      )
    ),
    server = function(input, output) {
      output$table <- renderTable({
        req(input$form)

        form <- input$form

        df <- as.data.frame(
          matrix(
            data = rep.int("Hello, world!", form$columns * form$rows),
            nrow = form$rows,
            ncol = form$columns
          )
        )

        colnames(df) <- paste("Column", seq_len(form$columns))

        df
      })
    }
  )
}

}
\seealso{
Other navs: \code{\link{linkNav}},
  \code{\link{listGroupNav}}
}
