% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav.R
\name{tabsTabs}
\alias{tabsTabs}
\alias{pillsTabs}
\alias{listTabs}
\title{Tabbable content}
\usage{
tabsTabs(labels, panes, ids = NULL, ...)

pillsTabs(labels, panes, ids = NULL, ...)

listTabs(labels, panes, ids = NULL, selected = NULL, context = NULL, ...)
}
\arguments{
\item{labels}{A character vector or list of character vectors specifying
the labels of the nav's tabs. \code{tabsTabs} and \code{pillsTabs} understand named
items which are converted into dropdown tabs.}

\item{panes}{A list of custom tags specifying the panes controlled by the
nav. Unlike \code{labels} or \code{ids} this is a flat list of items.}

\item{ids}{A character vector or list of character vectors mimicing the
structure of \code{labels}, but specifying the HTML ids to use for each tab and
pane pair, defaults to \code{NULL}, in which case ids are automatically
generated.}

\item{...}{Additional named arguments passed as HTML attributes to the
parent element.}

\item{selected}{One of \code{labels} indicating which tab is open by default,
defaults to \code{NULL}, in which case the first tab is the default.}

\item{context}{One of \code{"primary"}, \code{"secondary"}, \code{"success"}, \code{"info"},
\code{"warning"}. \code{"danger"}, \code{"light"}, \code{"dark"}, or \code{NULL}, specifying
the visual context of the list nav, defaults to \code{NULL}, in which case a
visual context is not applied.}
}
\description{
Create tabbable content with headers stylized as tabs, pills, or a list
group. Tab sets or tab panels are distinct from navs. \code{tabsTabs} and
\code{pillsTabs} may include nested tabs as dropdowns by passing named character
vectors or lists as elements of the \code{labels} argument. However, \code{listTabs}
does not allow nested tabs.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      tabsTabs(
        labels = list(
          "Home",
          "Profile",
          "Other" = c("Action", "Another action")
        ),
        panes = list(
          row(
            col(
              formInput(
                id = "form",
                numberInput(
                  id = "columns",
                  label = "Number of columns",
                  value = 1,
                  min = 0,
                ),
                numberInput(
                  id = "rows",
                  label = "Number of rows",
                  value = 5,
                  min = 0
                )
              )
            ),
            col(
              tableThruput(id = "table")
            )
          ),
          tags$p(
            "The window, the window, the second story window."
          ),
          tags$p(
            "We must away, ere break of day, ",
            "To find our long-forgotten gold."
          ),
          tags$div(
            tags$p(
              class = "lead",
              "This text is stylized"
            ),
            tags$p(
              "This is the body of the whole text and nothing but the ",
              "text."
            )
          )
        )
      )
    ),
    server = function(input, output) {
      output$table <- renderTable({
        req(input$form)

        form <- input$form

        df <- as.data.frame(
          matrix(
            data = rep.int("Hello, world!", form$columns * form$rows),
            nrow = form$rows,
            ncol = form$columns
          )
        )

        colnames(df) <- paste("Column", seq_len(form$columns))

        df
      })
    }
  )
}

if (interactive()) {
  shinyApp(
    ui = container(
      pillsTabs(
        labels = c("Home", "Profile", "Messages", "Settings"),
        panes = list(
          lapply(
            c(
              "To write, to rhyme, to put a slew of words to verse",
              "These powers I cannot claim to fully grasp",
              "But like my hem catches the dew in the grey of early dawn",
              "I catch a whisper each chance I stumble from the darkness."
            ),
            tags$p
          ),
          tags$p(
            "Stumbling down the mountain side I chanced to turn and look ",
            "upon the path I'd chanced to take."
          ),
          tags$p(
            ""
          ),
          tags$p(
            "A stone set in diamond."
          )
        )
      )
    ),
    server = function(input, output) {

    }
  )
}

if (interactive()) {
  shinyApp(
    ui = container(
      listTabs(
        labels = c("Home", "Profile", "Messages", "Settings"),
        panes = list(
          tags$p(
            "When all is said and done, the sun's brow dipped below the ",
            "horizon, then shall we rest ourselves, and then shall we ",
            "share a smile, for our hardwork is behind us for a while, ",
            "and for a while can we admire the life we build together."
          ),
          tags$p(
            "How, when the field is laid low, how, when the snow begins to ",
            "fall, shall I find you once again? I have let this wonderful ",
            "life slip through my fingers and so I am left with none, but ",
            "empty fists clenched against my sides"
          ),
          tags$p(

          ),
          tags$p(

          )
        )
      )
    ),
    server = function(input, output) {

    }
  )
}

}
\seealso{
Bootstrap 4 nav documentation:
\url{https://getbootstrap.com/docs/4.0/components/navs/}
}
