% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tagReduce}
\alias{tagReduce}
\title{Apply styles to a tag}
\usage{
tagReduce(...)
}
\arguments{
\item{...}{Any number of utility function calls \strong{and} one tag object.}
}
\description{
This function applies any number of styles generated by utility function
calls.
}
\examples{
tagReduce(
  width(25),
  padding(3),
  tags$div()
)

tagReduce(
  margins(sm = 1, xl = 5),
  tags$span()
)

}
