% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{text}
\alias{text}
\title{Tag element text alignment}
\usage{
text(tag, default = NULL, sm = NULL, md = NULL, lg = NULL, xl = NULL)
}
\arguments{
\item{tag}{A tag object.}

\item{default}{One of \code{"left"}, \code{"right"}, \code{"center"}, or \code{"justified"}
specifying the default text alignment of the element.}

\item{sm}{Like \code{default}, but the text alignment is applied once the viewport
is 576 pixels wide, think phone in landscape mode.}

\item{md}{Like \code{default}, but the text alignment is applied once the
viewport is 768 pixels wide, think tablets.}

\item{lg}{Like \code{default}, but the text alignment is applied once the
viewport is 992 pixels wide, think desktop.}

\item{xl}{Like \code{default}, but the text alignment is applied once the
viewport is 1200 pixels wide, think large desktop.}
}
\description{
The \code{text()} utility applies Bootstrap classes to change how an element's
text is aligned. Like with \code{\link[=display]{display()}} or \code{\link[=padding]{padding()}} different text alignments
can be applied based on the viewport size.
}
\examples{

if (interactive()) {
  shinyApp(
    ui = container(
      p("Text always aligned center. Resize your browser window to see these",
        "examples in action") \%>\%
        text("center"),
      p("Text centered on screens >= large, left aligned by default") \%>\%
        text("left", lg = "center"),
      p("Text aligned left on screens >= medium, right aligned by default") \%>\%
        text("right", md = "left"),
      p("Text aligned left on screens >= medium, centered for >= small, justified",
        "for mobile") \%>\%
        text("justify", sm = "center", md = "left")
    ),
    server = function(input, output) {

    }
  )
}

}
\seealso{
Other utilities: \code{\link{affix}},
  \code{\link{background}}, \code{\link{border}},
  \code{\link{display}}, \code{\link{float}},
  \code{\link{font}}, \code{\link{padding}},
  \code{\link{rounded}}, \code{\link{width}}
}
\concept{utilities}
