% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textual.R
\name{textInput}
\alias{textInput}
\alias{updateTextInput}
\alias{searchInput}
\alias{updateSearchInput}
\alias{emailInput}
\alias{updateEmailInput}
\alias{urlInput}
\alias{updateUrlInput}
\alias{telephoneInput}
\alias{updateTelephoneInput}
\alias{passwordInput}
\alias{updatePassowrdInput}
\alias{numberInput}
\alias{updateNumberInput}
\title{Textual inputs}
\usage{
textInput(id, value = NULL, placeholder = NULL, ...)

updateTextInput(id, value = NULL, enable = NULL, disable = NULL,
  valid = NULL, invalid = NULL, session = getDefaultReactiveDomain())

searchInput(id, value = NULL, placeholder = NULL, ...)

updateSearchInput(id, value = NULL, enable = NULL, disable = NULL,
  valid = NULL, invalid = NULL, session = getDefaultReactiveDomain())

emailInput(id, value = NULL, placeholder = NULL, ...)

updateEmailInput(id, value = NULL, enable = NULL, disable = NULL,
  valid = NULL, invalid = NULL, session = getDefaultReactiveDomain())

urlInput(id, value = NULL, placeholder = NULL, ...)

updateUrlInput(id, value = NULL, enable = NULL, disable = NULL,
  valid = NULL, invalid = NULL, session = getDefaultReactiveDomain())

telephoneInput(id, value = NULL, placeholder = NULL, ...)

updateTelephoneInput(id, value = NULL, enable = NULL, disable = NULL,
  valid = NULL, invalid = NULL, session = getDefaultReactiveDomain())

passwordInput(id, value = NULL, placeholder = NULL, ...)

updatePassowrdInput(id, value = NULL, enable = NULL, disable = NULL,
  valid = NULL, invalid = NULL, session = getDefaultReactiveDomain())

numberInput(id, value = NULL, placeholder = NULL, ...)

updateNumberInput(id, value = NULL, enable = NULL, disable = NULL,
  valid = NULL, invalid = NULL, session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{A character string specifying the id of the reactive input.}

\item{value}{A character string or a value coerced to a character string
specifying the default value of the textual input.}

\item{placeholder}{A character string specifying placeholder text for the
input, defaults to \code{NULL}, in which case there is no placeholder text.}

\item{...}{Additional named arguments passed as HTML attributes to the parent
element.}
}
\description{
Different types of textual inputs are provided to best support mobile
keyboards and assistive technologies. A password input will mask its
contents. Email inputs offer client-side validation depending on the browser.
}
\examples{

### Basic text

textInput(id = "text")

### Search

searchInput(id = "search")

### Email

emailInput(id = "email")

### URLs

urlInput(id = "url")

### Telephone numbers

telephoneInput(id = "tele")

### Passwords

passwordInput(id = "password")

### Numbers

numberInput(id = "num")

}
\seealso{
Other inputs: \code{\link{buttonGroupInput}},
  \code{\link{buttonInput}}, \code{\link{checkboxInput}},
  \code{\link{chipInput}}, \code{\link{fileInput}},
  \code{\link{formInput}}, \code{\link{groupTextInput}},
  \code{\link{listGroupInput}}, \code{\link{menuInput}},
  \code{\link{navInput}}, \code{\link{radioInput}},
  \code{\link{rangeInput}}, \code{\link{selectInput}},
  \code{\link{sliderInput}}
}
\concept{inputs}
