% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toast.R
\name{toast}
\alias{toast}
\alias{toast_container}
\title{Toasts}
\usage{
toast(
  ...,
  id = NULL,
  visibility = c("hide", "show"),
  duration = NULL,
  wrapper = toast_body
)

toast_container(..., position = c("top", "bottom"), padding = 3)
}
\arguments{
\item{...}{Any number of child elements passed to the parent element. Named
values are passed as HTML attributes to the parent element.}

\item{id}{A string. An optional reactive id to track the state of
the toast via \verb{input$<id>}. When visible this value is \code{"shown"} and when
the toast is dismissed the value is \code{"hidden"}.}

\item{visibility}{A string. The default visibility of the toast.}

\item{duration}{A number. The number of seconds to show the toast for.}

\item{wrapper}{A \link[htmltools:builder]{htmltools::tag} function. The function used to wrap the
components passed in \code{...}.}
}
\description{
Toasts relay notification-level information to the user.
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

library(shiny)
library(bslib)

ui <-
  page_fluid(
    toast_container(
      toast(
        id = "notify_begin",
        duration = 5,
        toast_header("Starting task"),
        "This task may take a while."
      )
    ),
    card(
      input_button(
        id = "begin",
        text = "Begin task"
      )
    )
  )

server <-
  function(input, output) {
    observeEvent(input$begin, {
      toast_show("notify_begin", 2)
    })
  }

shinyApp(ui, server)

\dontshow{\}) # examplesIf}
}
\seealso{
Other components: 
\code{\link{alert}()},
\code{\link{badge}()},
\code{\link{collapse_panel}()},
\code{\link{d1}()},
\code{\link{modal_dialog}()}
}
\concept{components}
