% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toast.R
\name{toast_add}
\alias{toast_add}
\alias{toast_show}
\alias{toast_hide}
\title{Toast actions}
\usage{
toast_add(target, toast, session = get_current_session())

toast_show(id, duration = NULL, session = get_current_session())

toast_hide(id, session = get_current_session())
}
\arguments{
\item{target}{A string. The id of a toast container.}

\item{toast}{A \code{\link[=toast]{toast()}}. A toast component to add to a toast container. Once
added, the toast may be shown or hidden.}

\item{session}{A shiny session object.}

\item{id}{A string. The id of a \code{\link[=toast_container]{toast_container()}}.}

\item{duration}{A number. The number of seconds to show a toast before
automatically hiding it. If \code{NULL}, the toast remains shown until manually
hidden.}
}
\description{
Do things with toasts.
}
\section{Functions}{
\itemize{
\item \code{toast_add()}: Add a toast to a toast container.

\item \code{toast_show()}: Show a toast.

\item \code{toast_hide()}: Hide a toast.

}}
