% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toast.R
\name{toast_body}
\alias{toast_body}
\alias{toast_header}
\alias{toast_button}
\title{Toast components}
\usage{
toast_body(...)

toast_header(title, ..., icon = NULL, dismiss = toast_button())

toast_button()
}
\arguments{
\item{...}{Any number of child elements passed to the parent element. Named
values are passed as HTML attributes to the parent element.}

\item{title}{A string. The title of the toast, displayed in the toast
header.}

\item{icon}{A tag element. An icon displayed before the toast title.}

\item{dismiss}{A button tag element used to dismiss the toast, defaults to
\code{\link[=toast_button]{toast_button()}}.}
}
\value{
A \code{shiny.tag} object.
}
\description{
Components to include inside a \link{toast}.
}
\seealso{
\code{\link[=toast]{toast()}} for creating toasts.
}
