% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip.R
\name{tooltip}
\alias{tooltip}
\title{Button or link tooltips}
\usage{
tooltip(..., placement = "top", fade = TRUE)
}
\arguments{
\item{...}{Character strings or tag elements (such as \code{em} or \code{b}) specifying
the contents of the tooltip.}

\item{placement}{One of \code{"top"}, \code{"right"}, \code{"bottom"}, or \code{"left"}
specifying what side of the tag element the tooltip appears on.}

\item{fade}{One of \code{TRUE} or \code{FALSE} specifying if the tooltip fades in when
shown and fades out when hidden, defaults to \code{TRUE}.}
}
\description{
Use \code{tooltip()} to contruct a tooltip for a button or link input.
}
\examples{

### Link with tooltip

linkInput(
  id = "link1",
  "A link",
  tooltip = tooltip("But, with a tooltip!")
)

}
\seealso{
Other content: \code{\link{alert}},
  \code{\link{blockquote}}, \code{\link{card}},
  \code{\link{collapsePane}}, \code{\link{d1}},
  \code{\link{dropdown}}, \code{\link{img}},
  \code{\link{jumbotron}}, \code{\link{modal}},
  \code{\link{navContent}}, \code{\link{popover}},
  \code{\link{pre}}, \code{\link{toast}}
}
\concept{content}
