% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{typography}
\alias{typography}
\alias{d1}
\alias{d2}
\alias{d3}
\alias{d4}
\title{Text and Heading Styles}
\usage{
d1(...)

d2(...)

d3(...)

d4(...)
}
\arguments{
\item{...}{Child elements or named arguments passed as HTML attributes
to the parent element.}
}
\description{
These function encapsulate Bootstrap classes for styling text and headings.
Display headings are not meant to replace the standard HTML heading tags,
they are a stand out alternative for eye-catching titles. Use \code{lead} to
emphasize a paragraph or snippet of text.

\code{mark} and \code{small} are HTML class-based alternatives to \code{<mark>} and
\code{<small>}. Typically, \code{<mark>} is used to indicate relevance. By using
Bootstrap's class alternatives one can utilize the highlighting aspect
of \code{<mark>} while avoiding the usage and purpose restrictions.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      d1("Eye-catching!"),
      d3("Wowza."),
      d4("You'll never guess what happens next."),
      h1("This is an <h1> heading for comparision")
    ),
    server = function(input, output) {

    }
  )
}

}
\seealso{
Other content: \code{\link{blockquote}},
  \code{\link{card}}, \code{\link{dropdown}},
  \code{\link{icon}}, \code{\link{img}},
  \code{\link{jumbotron}}, \code{\link{pre}}
}
\concept{content}
