% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-checkbox.R
\name{update_checkbox}
\alias{update_checkbox}
\title{Update checkbox}
\usage{
update_checkbox(
  id,
  choices = NULL,
  values = choices,
  selected = NULL,
  disable = NULL,
  session = default_reactive_domain()
)
}
\arguments{
\item{choices}{choices}

\item{values}{values}

\item{selected}{selected}

\item{disable}{disable}

\item{session}{A reactive session, \code{\link[=default_reactive_domain]{default_reactive_domain()}}.}
}
\description{
Update checkbox
}
