% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-numeric.R
\name{update_numeric}
\alias{update_numeric}
\title{Update a numeric input}
\usage{
update_numeric(
  id,
  ...,
  value = NULL,
  min = NULL,
  max = NULL,
  step = NULL,
  disable = NULL,
  session = get_current_session()
)
}
\arguments{
\item{id}{A character string. The id of the reactive input.}

\item{...}{These dots are for future extensions and must be empty.}

\item{value}{A number. The new value of the input.}

\item{min}{A number. The new minimum value of the input.}

\item{max}{A number. The new maximum value of the input.}

\item{step}{A number. The new interval when stepping between values.}

\item{disable}{A boolean. If \code{TRUE}, disable the input.}

\item{session}{A shiny session object.}
}
\value{
No return value, called for side effects.
}
\description{
Update the value, bounds, or state of an \code{\link[=input_numeric]{input_numeric()}}.
}
\seealso{
\code{\link[=input_numeric]{input_numeric()}}
}
