% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{updateChoices}
\alias{updateChoices}
\alias{updateValues}
\alias{updateSelected}
\title{Update choices, values, selected choices}
\usage{
updateChoices(id, ...)

updateValues(id, ...)

updateSelected(id, ...)
}
\description{
Functions to update choices, values, and selected choices.
}
