% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{updateInput}
\alias{updateInput}
\title{Update input}
\usage{
updateInput(id, choices = NULL, values = NULL, selected = NULL,
  session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{A character string specifying the reactive id of an input.}

\item{choices}{A character vector, tag element, or list specifying choices
for the input.}

\item{values}{An object specifying values for the input, this object is
coerced to a character vector.}

\item{selected}{One of \code{values} specifying which choice to select, defaults
to \code{NULL}, in which case a choice is not selected. Note that browsers may
automatically select a choice if not specified.}

\item{session}{A reactive context, defaults to \code{\link[=getDefaultReactiveDomain]{getDefaultReactiveDomain()}}.}
}
\description{
Input utilities.
}
\section{Combinations of \code{choices} and \code{values}}{


\strong{choices} and \strong{values} specified: the input's choices and values are
updated.

\strong{choices} is specified and \strong{values} is \code{NULL}: \code{values} defaults to
\code{choices}.

\strong{choices} is \code{NULL} and \strong{values} is specified: use this case for inputs
without choices, e.g. a text input, to update only the current value.

\strong{choices} and \strong{values} both \code{NULL}: used to preserve choices and values
while changing the selected choice with \code{selected}.
}

