% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{updateStream}
\alias{updateStream}
\title{Stream notifications}
\usage{
updateStream(id, msg, session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{A character string specifying the id of the stream output.}

\item{msg}{A character string specifying an update message.}

\item{session}{The shiny session, defaults to the default reactive domain.}
}
\description{
These functionalities allow an application to convey text progress updates
to the user during long running processes.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      button(id = "trigger", "Go!"),
      tags$div(
        id = "myStream"
      )
    ),
    server = function(input, output, session) {
      observeEvent(input$trigger, {
        for (i in seq_len(5)) {
          updateStream("myStream", paste("Update:", i))
          Sys.sleep(1);
        }
      })
    }
  )
}

}
