% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{viewport}
\alias{viewport}
\title{Viewport}
\description{
Many styles are viewport responsive. These styles change depending on the
size of the viewing application. This allows the programmer to build content
which can adapt to laptop screens, mobile devices, and more.

Responsive styles are applied depending on the width of the viewport. Extra
small screens are 0 px and up. Small screens are 576 px and up. Medium
screens are 768 px and up. Large screens are 992 px and up. Extra large
screens are 1200 px and up.
}
\examples{
if (interactive()) {
  library(shiny)

  shinyApp(
    ui = container(
      listGroup(
        class = collate(
          utils$margin(0),
          utils$margin(5, "left", "medium"),
          utils$margin(5, "right", "medium")
        ),
        listGroupItem(
          class = utils$bg("info"),
          label = "Hello, world!"
        ),
        listGroupItem(
          class = utils$bg("info"),
          label = "Goodnight, moon!"
        )
      )

    ),
    server = function(input, output) {

    }
  )
}

}
