% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{webpage}
\alias{webpage}
\title{User interface}
\usage{
webpage(..., nav = NULL)
}
\arguments{
\item{...}{Any number of tag elements or named values added as children and
attributes to the main section of the page.}

\item{nav}{A navigation element, typically a call to \code{\link[=navbar]{navbar()}}, added at the
top of the page, defaults to \code{NULL}.}
}
\description{
Begin creating a user interface. The \code{webpage()} function properly lays out a
navigation bar and main section of elements.
}
\seealso{
Other layout functions: 
\code{\link{column}()},
\code{\link{fieldset}()},
\code{\link{navbar}()}
}
\concept{layout functions}
