% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{width}
\alias{width}
\title{Tag element width}
\usage{
width(.tag, size)
}
\arguments{
\item{.tag}{A tag element.}

\item{size}{A character string or number specifying the width of the tag
element. Possible values:

An integer between 1 and 20, in which case the width of the element is
relative to the font size of the page.

One of "1/2", "1/3", "2/3", "1/4", "3/4", "1/5", "2/5", "3/5", "4/5", or
"full", in which case the element's width is a percentage of its parent's
width. The height of the parent element must be specified for percentage
widths to work. Percentages do not account for margins or padding and may
cause an element to extend beyond its parent.

Or "auto", in which case the element's width is determined by the browser.
The browser will take into account the width, padding, margins, and border
of the tag element's parent to keep the element from extending beyond its
parent.}
}
\description{
Utility function to change a tag element's width. Widths are specified
relative to the font size of page (browser default is 16px), relative to
their parent element (i.e. 1/2 the width of their parent), or relative to the
element's content.
}
\examples{

### Numeric values

# When specifying a numeric value the width of the element is relative to the
# default font size of the page.

lapply(
  1:20,
  width,
  .tag = div() \%>\%
    border("black") \%>\%
    height(4)
)

### Fractional values

# When specifying a fraction the element's width is a percentage of its
# parent's width.

div(
  div() \%>\%
    margin(b = 3) \%>\%
    border("red") \%>\%
    width("1/3")
) \%>\%
  width(20)

}
\seealso{
Other design: \code{\link{affix}},
  \code{\link{background}}, \code{\link{border}},
  \code{\link{display}}, \code{\link{float}},
  \code{\link{font}}, \code{\link{height}},
  \code{\link{padding}}, \code{\link{scroll}},
  \code{\link{shadow}}
}
\concept{design}
