% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{wrap}
\alias{wrap}
\title{Wrapping flex items}
\usage{
wrap(tag, default = NULL, sm = NULL, md = NULL, lg = NULL, xl = NULL)
}
\arguments{
\item{tag}{A tag element.}

\item{default}{One of \code{"none"}, \code{"wrap"}, or \code{"reverse} specifying how the
flex items of an element wrap.}

\item{sm}{Like \code{default}, but the wrapping behavior is applied once the
viewport is 576 pixels wide, think phone in landscape mode.}

\item{md}{Like \code{default}, but the wrapping behavior is applied once the
viewport is 768 pixels wide, think tablets.}

\item{lg}{Like \code{default}, but the wrapping behavior is applied once the
viewport is 992 pixels wide, think desktop.}

\item{xl}{Like \code{default}, but the wrapping behavior is applied once the
viewport is 1200 pixels wide, think large desktop.}
}
\description{
This function applies bootstrap classes to a tag element to change how the
element's flex items wrap or do not wrap.
}
\examples{

}
