% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordion.R
\name{accordion}
\alias{accordion}
\alias{accordionPanel}
\title{Accordion panels}
\usage{
accordion(...)

accordionPanel(label, ..., collapsed = TRUE)
}
\arguments{
\item{...}{For \strong{accordion}, any number of \code{accordionPanel}s or named
arguments passed as HTML attributes to the parent element.

For \strong{accordionPanel}, any number of elements to include in the panel or
named arguments passed as HTML attributes to the parent element.}

\item{label}{A character string specifying the accordion panel's header.}

\item{collapsed}{If \code{TRUE}, the accordion panel renders in the collapsed
state, defaults to \code{TRUE}.}
}
\description{
Similar to \code{collapse}, \code{accordion} is a way to hide content. An accordion
is made up of multiple panels. As one panel is toggled open other panels
are closed. A page may render with multiple panels open, but once toggled
only one panel will remain open. \code{accordionPanel} is a helper function to
facilitate building the accordion's panels.
}
\details{
\code{accordion} will generate an HTML id if an \code{id} argument is not passed in
\code{...} and child panels are assigned IDs based on the id of the accordion
parent.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      accordion(
        accordionPanel(
          "Option 1",
          "What do you think of option 1?",
          collapse = FALSE
        ),
        accordionPanel(
          "Option 2",
          "What do you think of option 2?"
        ),
        accordionPanel(
          "Option 3",
          "What do you think of option 3?"
        )
      )
    ),
    server = function(input, output) {

    }
  )
}

}
\seealso{
For more information on accordion's, please refer to
\url{https://v4-alpha.getbootstrap.com/components/collapse/}.
}
