% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordion.R
\name{accordion}
\alias{accordion}
\alias{panel}
\title{Accordions}
\usage{
accordion(...)

panel(heading, body)
}
\arguments{
\item{...}{\code{panel}s converted into accordion panels or named arguments passed
as HTML attributes to the accordion parent element.

\strong{Note}, a \code{panel} or custom panel named \code{show} will render in an open
state rather than the default collapsed state.}

\item{heading}{A panel heading, character string or tag element(s).}

\item{body}{Body content for the panel, character string or tag element(s).}
}
\description{
Accordion collapsible content. \code{panel} is a helper function to facilitate
building accordion panels, see details for more information.
}
\details{
\code{panel} is little more than a thin wrapper around \code{list} to ensure each
accordion panel has a heading and a body. If you want more customization over
accordion panels you may ignore \code{panel} and pass unnamed tag elements. When
creating custom panels be sure to follow the \code{data-target}, \code{data-toggle}
conventions, for more on how to build custom accordions check out the
Bootstrap
\href{https://v4-alpha.getbootstrap.com/components/collapse/}{collapse reference}.

\code{accordion} will generate an HTML id if one is not specified and child panels
are assigned IDs based on the id of the accordion parent.
}
\examples{
accordion(
  panel(
    "Heading 1",
    "Body text"
  ),
  panel(
    "Heading 2",
    "Body text"
  ),
  panel(
    "Heading 3",
    "Body text"
  )
)

}
