% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordion.R
\name{accordion}
\alias{accordion}
\title{Accordion panels}
\usage{
accordion(labels, contents, ids = NULL, ...)
}
\arguments{
\item{labels}{A character vector or flat list of character strings specifying
the link labels of the accordion panels.}

\item{contents}{A list of custom tags specifying the content of the panels.}

\item{ids}{A character vector or flat list of character strings specifying
the HTML ids of the accordion panels, defaults to \code{NULL}, in which case
ids are automatically generated.}

\item{...}{Additional named arguments passed as HTML attributes to the
parent element.}
}
\description{
Similar to \code{collapse}, \code{accordion} is a way to condense content. An accordion
is made up of multiple panels. As one panel is toggled open other panels
are closed. A page may render with multiple panels open, but once toggled
only one panel will remain open.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          accordion(
            labels = c("Option 1", "Option 2", "Option 3"),
            contents = list(
              tags$p("What do you think of option 1?"),
              tags$p("What do you think of option 2?"),
              tags$p("What do you think of option 3?")
            )
          )
        )
      )
    ),
    server = function(input, output) {

    }
  )
}

}
\seealso{
Bootstrap 4 accordion documentation:
\url{https://getbootstrap.com/docs/4.0/components/collapse/#accordion-example}
}
