% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{actions}
\alias{actions}
\alias{showTarget}
\alias{hideTarget}
\title{Input actions}
\usage{
showTarget(id, ...)

hideTarget(id, ...)
}
\arguments{
\item{id}{A character string specifying the id of a \code{\link[=navPane]{navPane()}} or
\code{\link[=collapsePane]{collapsePane()}}.}

\item{...}{Additional arguments, currently ignored.}
}
\description{
These functions are used in conjunction with the \code{targets} argument found
in all of yonder's reactive input functions.
}
