% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{active}
\alias{active}
\title{Color selected choices}
\usage{
active(tag, color)
}
\arguments{
\item{tag}{A tag element.}

\item{color}{One of \code{"red"}, \code{"purple"}, \code{"indigo"}, \code{"blue"}, \code{"cyan"},
\code{"teal"}, \code{"green"}, \code{"yellow"}, \code{"amber"}, \code{"orange"}, \code{"grey"}, \code{"white"}
specifying the active color of selected choices.}
}
\description{
Use \code{active()} to change the color of active selected choices. This utility
will only have an effect on elements with selectable choices, e.g.
\code{checkbarInput()}.
}
\seealso{
Other design: \code{\link{affix}},
  \code{\link{background}}, \code{\link{border}},
  \code{\link{display}}, \code{\link{float}},
  \code{\link{font}}, \code{\link{height}},
  \code{\link{padding}}, \code{\link{scroll}},
  \code{\link{shadow}}, \code{\link{width}}
}
\concept{design}
