% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{active}
\alias{active}
\title{Selected choice color}
\usage{
active(tag, color)
}
\arguments{
\item{tag}{A tag element.}

\item{color}{One of "\code{primary}", "\code{secondary}", "\code{success}", "\code{info}", "\code{warning}", "\code{danger}", "\code{light}" or "\code{dark}" specifying the active color of the tag element, defaults to \code{NULL}}
}
\description{
As part of an effort to revert yonder's default bootstrap styles, the
\code{active()} utility has been deprecated. In future versions of the application
the function will be removed entirely.

Previously, \code{active()} would change the
highlight color of an input's selected choices.
}
\seealso{
Other design utilities: 
\code{\link{affix}()},
\code{\link{background}()},
\code{\link{border}()},
\code{\link{display}()},
\code{\link{float}()},
\code{\link{font}()},
\code{\link{height}()},
\code{\link{padding}()},
\code{\link{scroll}()},
\code{\link{shadow}()},
\code{\link{width}()}
}
\concept{design utilities}
