% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{addressInput}
\alias{addressInput}
\title{Address input}
\usage{
addressInput(id)
}
\arguments{
\item{id}{A character string specifying the id of the address input.}

\item{placeholders}{If \code{TRUE}, placeholder text is added to all the address
input fields, defaults to \code{TRUE}.}

\item{abbreviate}{If \code{TRUE}, state abbreviations are used instead of state
names, defaults to \code{TRUE}.}
}
\description{
An address input which includes a street field, apartment or unit field, city
field, state field, and a zip code field.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          tags$form(
            addressInput("address")
          )
        ),
        col(
          verbatimTextOutput("value")
        )
      )
    ),
    server = function(input, output) {
      output$value <- renderPrint({
        input$address
      })
    }
  )
}

}
