% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textual.R
\name{addressInput}
\alias{addressInput}
\title{Address input}
\usage{
addressInput(id, ...)
}
\arguments{
\item{id}{A character string specifying the id of the address input.}

\item{...}{Additional named arguments passed as HTML attributes to the
top-level element.}
}
\description{
A composite input which includes a street field, apartment or unit field,
city field, state field, and a zip code field.
}
\examples{

### Default fields

addressInput(id = NULL)

}
\seealso{
Other inputs: \code{\link{buttonGroupInput}},
  \code{\link{buttonInput}}, \code{\link{checkboxInput}},
  \code{\link{dateInput}}, \code{\link{fileInput}},
  \code{\link{formInput}}, \code{\link{groupInput}},
  \code{\link{loginInput}}, \code{\link{radioInput}},
  \code{\link{rangeInput}}, \code{\link{selectInput}},
  \code{\link{textInput}}
}
\concept{inputs}
