% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design-affix.R
\name{affix}
\alias{affix}
\title{Fixed position elements}
\usage{
affix(x, position)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{position}{One of \code{"top"}, \code{"bottom"}, or \code{"sticky"} specifying the
fixed behavior of an element.}
}
\description{
The \code{affix} utility function applies Bootstrap classes to fix elements to the
top or bottom of a page. Use \code{"sticky"} to cause an element to fix to the top
of a page after the element is scrolled past. \emph{Important}, the IE11 and Edge
browsers do not support the sticky behavior.
}
\examples{

div() \%>\% affix("sticky")
# ->
div() \%>\% cascadess::position_sticky()

}
