% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{affix}
\alias{affix}
\title{Affix elements to top or bottom of page}
\usage{
affix(tag, position)
}
\arguments{
\item{position}{One of \code{"top"}, \code{"bottom"}, or \code{"sticky"} specifying the
fixed behavior of an element.}
}
\description{
The \code{affix} utility function applies Bootstrap classes to fix elements to the
top or bottom of a page. Use \code{"sticky"} to cause an element to fix to the top
of a page \emph{after} the element is scrolled past. \emph{Important}, the IE11 and
Edge browsers do not support the sticky behavior.
}
\examples{
tags$div("A simple banner") \%>\%
  width(100) \%>\%
  affix("top")

}
\seealso{
Other utilities: \code{\link{align}}, \code{\link{border}},
  \code{\link{float}}, \code{\link{font}},
  \code{\link{padding}}, \code{\link{text}},
  \code{\link{width}}
}
