% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert.R
\name{alert}
\alias{alert}
\alias{alert_heading}
\alias{alert_button}
\title{Alerts}
\usage{
alert(
  ...,
  type = NULL,
  dismiss = c("animate", "instant", "none"),
  dismiss_button = alert_button()
)

alert_heading(..., container = htmltools::h5)

alert_button()
}
\arguments{
\item{...}{Components to include. Named arguments are passed as HTML
attributes to the parent element.}

\item{type}{Optional semantic type.}

\item{dismiss}{The method for removing the alert. One of \code{"animate"},
\code{"instant"}, or \code{"none"}.}

\item{dismiss_button}{The dismiss button element. A \link[htmltools:tag]{htmltools::tag} object.}

\item{container}{An \link[htmltools:tag]{htmltools::tag} function.}
}
\value{
A \link[htmltools:tag]{htmltools::tag} object.
}
\description{
Alerts highlight information for users.
}
\examples{
\dontshow{if (rlang::is_interactive()) withAutoprint(\{ # examplesIf}

alert(
  type = "info",
  dismiss = "none",
  alert_heading("Heads up!"),
  "Make sure you know the thing is happening",
  htmltools::hr(),
  "Use alert classes to change the color of an alert"
)

alert(
  type = "warning",
  "Double check those figures."
)
\dontshow{\}) # examplesIf}
}
\seealso{
Other components:
\code{\link[=badge]{badge()}},
\code{\link[=collapse_panel]{collapse_panel()}},
\code{\link[=d1]{d1()}},
\code{\link[=modal_dialog]{modal_dialog()}}
}
\concept{components}
