% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert.R
\name{alert}
\alias{alert}
\alias{alert_heading}
\alias{alert_button}
\title{Alerts}
\usage{
alert(..., class = NULL)

alert_heading(..., container = htmltools::h6)

alert_button()
}
\arguments{
\item{...}{Components to include. Named arguments are passed as HTML
attributes to the parent element.}

\item{class}{Additional CSS classes for the alert element.}

\item{container}{An \link[htmltools:builder]{htmltools::tag} function.}
}
\description{
Alerts highlight information for users.
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

alert(
  class = "alert-info",
  alert_heading("Heads up!"),
  "Make sure you know the thing is happening",
  htmltools::hr(),
  "Use alert classes to change the color of an alert"
)

alert(
  class = "alert-warning",
  "Double check those figures.",
  alert_button()
)
\dontshow{\}) # examplesIf}
}
\seealso{
Other components: 
\code{\link{badge}()},
\code{\link{blockquote}()},
\code{\link{collapsePane}()},
\code{\link{d1}()},
\code{\link{dropdown}()},
\code{\link{img}()},
\code{\link{jumbotron}()},
\code{\link{modal_dialog}()},
\code{\link{pre}()},
\code{\link{toast}()}
}
\concept{components}
