% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert.R
\name{alert}
\alias{alert}
\title{Alert boxes}
\usage{
alert(..., title = NULL)
}
\arguments{
\item{...}{Character strings specifying the text of the alert or additional
named arguments passed as HTML attributes to the alert element.}

\item{title}{A character string or tag element specifying a heading for the
alert, defaults to \code{NULL}, in which case a title is not added.}
}
\description{
Use an alert element to let the user know of successes or to call attention
to problems.
}
\examples{

### Default alert

alert("Donec at pede.")

### A more complex alert

alert(
  p("Etiam vel tortor sodales"),
  hr(),
  p("Fusce commodo.")
) \%>\%
  background("amber")

}
\seealso{
Other content: \code{\link{blockquote}},
  \code{\link{card}}, \code{\link{collapsiblePane}},
  \code{\link{d1}}, \code{\link{dropdown}},
  \code{\link{img}}, \code{\link{jumbotron}},
  \code{\link{modal}}, \code{\link{navContent}},
  \code{\link{popover}}, \code{\link{pre}},
  \code{\link{progressOutlet}}, \code{\link{toast}},
  \code{\link{tooltip}}
}
\concept{content}
