% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design-background.R
\name{background}
\alias{background}
\title{Backgrounds}
\usage{
background(x, color)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{color}{One of \code{"primary"}, \code{"secondary"}, \code{"success"}, \code{"info"}, \code{"warning"}, \code{"danger"}, \code{"light"} or \code{"dark"} specifying the background color of the tag element, defaults to \code{NULL}.}
}
\description{
Use \code{background()} to modify the background color of a tag element.
}
\examples{

### Modifying input elements

checkbarInput(
  .style \%>\% background("info"),
  id = "bar1",
  choices = c(
    "Nunc rutrum turpis sed pede.",
    "Etiam vel neque.",
    "Lorem ipsum dolor sit amet."
  )
)

}
\seealso{
Other design utilities: 
\code{\link{active}()},
\code{\link{affix}()},
\code{\link{border}()},
\code{\link{display}()},
\code{\link{flex}()},
\code{\link{float}()},
\code{\link{font}()},
\code{\link{height}()},
\code{\link{margin}()},
\code{\link{padding}()},
\code{\link{scroll}()},
\code{\link{shadow}()},
\code{\link{width}()}
}
\concept{design utilities}
\section{Changing input backgrounds}{

Input backgrounds will change differently according to the particular
input. Changing the background of a \code{buttonGroupInput()} will change the
background colors of the buttons in the group.\if{html}{\out{<div class="r">}}\preformatted{buttonGroupInput(
  .style \%>\%
    background("info"),
  id = "flavors",
  choices = c("Mint", "Strawberry", "Swirl"),
  values = c("mint", "sberry", "swirl")
)
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve-->
}}
\if{html}{\out{
<div id="flavors" class="yonder-button-group btn-group btn-group-info" role="group">
}}
\if{html}{\out{
<button type="button" class="btn" value="mint">
}}


Mint\if{html}{\out{
</button>
}}
\if{html}{\out{
<button type="button" class="btn" value="sberry">
}}


Strawberry\if{html}{\out{
</button>
}}
\if{html}{\out{
<button type="button" class="btn" value="swirl">
}}


Swirl\if{html}{\out{
</button>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<!--/html_preserve-->
}}


Another one.
}
