% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{background}
\alias{background}
\title{Background color}
\usage{
background(tag, context)
}
\arguments{
\item{context}{One of \code{"primary"}, \code{"secondary"}, \code{"success"}, \code{"info"},
\code{"warning"}, \code{"danger"}, \code{"light"}, \code{"dark"}, or \code{"white"}.}
}
\description{
The \code{background} utility function applies Bootstrap classes to a tag element
to change the element's background color.
}
\examples{
tags$div("light text, dark background") \%>\%
  text("white") \%>\%
  background("dark")

}
