% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{background}
\alias{background}
\alias{text}
\alias{border}
\title{Chance text, background, or border color}
\usage{
background(tag, color, tone = 0)

text(tag, color, tone = 0)

border(tag, color, tone = 0)
}
\arguments{
\item{tag}{A tag element.}

\item{color}{A character string specifying the background color, see details
for all possible values.}

\item{tone}{An integer -4 through 4 specifying how much to darken or lighten
a color, defaults to 0, in which case the base color is not modified.
Negative values result in darker tones, whereas positive values create a
lighter tone.}
}
\description{
The \code{text}, \code{background}, and \code{border} utility functions may be used to
change the text, background, or border color of a tag element, respectively.
}
\examples{
tags$div("light text, dark background") \%>\%
  background("grey", -4) \%>\%
  text("yellow", 3)

if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          div("hello hello") \%>\%
            text("amber", 4) \%>\%
            background("blue-grey", -4)
        )
      )
    ),
    server = function(input, output) {

    }
  )
}

tags$h1("Hello, world!") \%>\%
  border("grey", sides = c("top", "bottom"))

tags$div() \%>\%
  border("orange")

if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          checkbarInput(
            id = NULL,
            choices = paste("Choice", 1:4)
          ) \%>\%
            background("cyan") \%>\%
            border("indigo")
        )
      )
    ),
    server = function(input, output) {

    }
  )
}

}
\seealso{
Other utilities: \code{\link{affix}}, \code{\link{align}},
  \code{\link{display}}, \code{\link{float}},
  \code{\link{font}}, \code{\link{padding}},
  \code{\link{width}}

Other utilities: \code{\link{affix}}, \code{\link{align}},
  \code{\link{display}}, \code{\link{float}},
  \code{\link{font}}, \code{\link{padding}},
  \code{\link{width}}

Other utilities: \code{\link{affix}}, \code{\link{align}},
  \code{\link{display}}, \code{\link{float}},
  \code{\link{font}}, \code{\link{padding}},
  \code{\link{width}}
}
