% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{background}
\alias{background}
\title{Tag element background color}
\usage{
background(tag, color)
}
\arguments{
\item{tag}{A tag element.}

\item{color}{A character string specifying the background color, see below
for all possible values.}
}
\description{
Use \code{background()} to change the background color of a tag element.
}
\details{
The possible background colors are,
\itemize{
\item red
\item purple
\item indigo
\item blue
\item cyan
\item teal
\item green
\item yellow
\item amber
\item orange
\item grey
\item white
\item transparent
}
}
\examples{

div("Nullam eu ante vel est convallis dignissim.") \%>\%
  background("grey")


checkbarInput(
  id = NULL,
  choices = c(
    "Nunc rutrum turpis sed pede.",
    "Etiam vel neque.",
    "Lorem ipsum dolor sit amet."
  )
) \%>\%
  background("cyan")


if (interactive()) {
  colors <- c(
    "red", "purple", "indigo", "blue", "cyan", "teal", "green",
    "yellow", "amber", "orange", "grey", "white"
  )

  shinyApp(
    ui = container(
      fluid = FALSE,
      lapply(
        colors,
        background,
        tag = div() \%>\%
          padding(5) \%>\%
          margins(2)
      )
    ) \%>\%
      display(flex = TRUE) \%>\%
      flex(wrap = TRUE),
    server = function(input, output) {

    }
  )
}

}
\seealso{
Other utilities: \code{\link{affix}}, \code{\link{border}},
  \code{\link{display}}, \code{\link{float}},
  \code{\link{font}}, \code{\link{padding}},
  \code{\link{rounded}}, \code{\link{text}},
  \code{\link{width}}
}
\concept{utilities}
