% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge}
\alias{badge}
\title{Badges}
\usage{
badge(..., pill = FALSE)
}
\arguments{
\item{...}{Additional named argument passed as HTML attributes to the parent
element.}

\item{pill}{One of \code{TRUE} or \code{FALSE} specifying if the badge has more rounded
corners, defaults to \code{FALSE}.}
}
\description{
Small highlighted content which scales to its parent's size.
}
\examples{

## Buttons with badges

buttonInput(
  id = "button1",
  label = "Process",
  badge(7) \%>\%
    background("cyan")
)

## Possible colors

colors <- c(
  "red", "purple", "indigo", "blue", "cyan", "teal", "green",
  "yellow", "amber", "orange", "grey", "white"
)

div(
  lapply(colors, function(color) {
    badge(color) \%>\%
      background(color) \%>\%
      margin(2)
  })
) \%>\%
  display("flex") \%>\%
  flex(wrap = TRUE)

}
\seealso{
Other content: \code{\link{alert}},
  \code{\link{blockquote}}, \code{\link{card}},
  \code{\link{collapsiblePane}}, \code{\link{d1}},
  \code{\link{dropdown}}, \code{\link{img}},
  \code{\link{jumbotron}}, \code{\link{modal}},
  \code{\link{navContent}}, \code{\link{popover}},
  \code{\link{pre}}, \code{\link{progressOutlet}},
  \code{\link{toast}}, \code{\link{tooltip}}
}
\concept{content}
