% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge}
\alias{badge}
\alias{renderBadge}
\title{Badges, tags}
\usage{
badge(content = NULL, context = NULL, pill = FALSE, ...)

renderBadge(value, context = NULL, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{content}{The text or label of the badge, defaults to \code{NULL}.}

\item{context}{A character string specifying the visual context of the badge,
one of \code{"primary"}, \code{"success"}, \code{"info"}, \code{"warning"}, or \code{"danger"},
defaults to \code{NULL}.}

\item{pill}{If \code{TRUE} the badge has rounder corners, defaults to \code{FALSE}.}

\item{...}{Additional named argument passed on as HTML attributes to the parent
element.}
}
\description{
A badge is a useful means of highlighting text.
}
\examples{

if (interactive()) {
  library(shiny)

  shinyApp(
    ui = container(
      listGroup(
        listItem(
          label = "Button clicks",
          badge = badge(
            id = "buttonClicks",
            content = 0
          )
        )
      ),
      button(
        id = "myButton",
        "Click here!"
      )
    ),
    server = function(input, output) {
      numClicks <- 0
      output$buttonClicks <- renderBadge({
        req(input$myButton)
        numClicks <<- numClicks + 1
        numClicks
      }, {
        if (numClicks > 19) {
          "danger"
        } else if (numClicks > 9) {
          "warning"
        } else {
          "info"
        }
      })
    }
  )
}

}
