% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge}
\alias{badge}
\title{Badges}
\usage{
badge(...)
}
\arguments{
\item{...}{Named arguments passed as HTML attributes to the parent
element or tag elements passed as children.}
}
\description{
Small highlighted content which scales to its parent's size.
}
\examples{

## Possible colors

colors <- c(
  "red", "purple", "indigo", "blue", "cyan", "teal", "green",
  "yellow", "amber", "orange", "grey", "white"
)

div(
  lapply(colors, function(color) {
    badge(color) \%>\%
      background(color) \%>\%
      margin(2)
  })
) \%>\%
  display("flex") \%>\%
  flex(wrap = TRUE)

}
\seealso{
Other content: \code{\link{alert}},
  \code{\link{blockquote}}, \code{\link{card}},
  \code{\link{collapsiblePane}}, \code{\link{d1}},
  \code{\link{dropdown}}, \code{\link{img}},
  \code{\link{jumbotron}}, \code{\link{modal}},
  \code{\link{navContent}}, \code{\link{popover}},
  \code{\link{pre}}, \code{\link{progressOutlet}},
  \code{\link{toast}}, \code{\link{tooltip}}
}
\concept{content}
