% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge}
\alias{badge}
\title{Badges}
\usage{
badge(..., appearance = c("default", "pill"))
}
\arguments{
\item{...}{Components to include inside the badge. Named arguments are passed
as HTML attributes to the parent element.}

\item{appearance}{A string. The appearance of the badge.}
}
\description{
Highlight small pieces of content with badges. A badge's size scales with
their parent's size.
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

badge(99)

badge(
  class = "text-bg-primary",
   "New"
)
\dontshow{\}) # examplesIf}
}
\seealso{
Other components: 
\code{\link{alert}()},
\code{\link{blockquote}()},
\code{\link{collapsePane}()},
\code{\link{d1}()},
\code{\link{dropdown}()},
\code{\link{img}()},
\code{\link{jumbotron}()},
\code{\link{modal_dialog}()},
\code{\link{pre}()},
\code{\link{toast}()}
}
\concept{components}
