% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badgeElement}
\alias{badgeElement}
\title{Badges}
\usage{
badgeElement(id, ...)
}
\arguments{
\item{id}{A character string specifying a reactive id.}

\item{...}{Named arguments passed as HTML attributes to the parent
element or tag elements passed as children to the parent element.}
}
\description{
Small highlighted content which scales to its parent's size. A badge may
be dynamically updated with \code{\link[=replaceElement]{replaceElement()}}.
}
\details{
Use \code{\link[=replaceElement]{replaceElement()}} and \code{\link[=removeElement]{removeElement()}} to modify the contents of a
badge.
}
\section{Example application}{
\preformatted{ui <- container(
  buttonInput(
    id = "clicker",
    label = list(
      "Clicks",
      badgeElement("counter") \%>\%
        margin(left = 2) \%>\%
        background("teal")
    )
  )
)

server <- function(input, output) {
  observe({
    clicks <- if (is.null(input$clicker)) 0 else input$clicker
    replaceElement("counter", clicks)
  })
}

shinyApp(ui, server)
}
}

\examples{

### Possible colors

colors <- c(
  "red", "purple", "indigo", "blue", "cyan", "teal", "green",
  "yellow", "amber", "orange", "grey", "white"
)

div(
  lapply(colors, function(color) {
    badgeElement(color, color) \%>\%
      background(color) \%>\%
      margin(2) \%>\%
      padding(1)
  })
) \%>\%
  display("flex") \%>\%
  flex(wrap = TRUE)

}
\seealso{
Other rendering: \code{\link{outputElement}}
}
\concept{rendering}
