% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badgeOutput}
\alias{badgeOutput}
\alias{renderBadge}
\title{Badge outputs}
\usage{
badgeOutput(id, ...)

renderBadge(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{id}{A character string specifying the id of the badge output.}

\item{...}{Additional named argument passed as HTML attributes to the parent
element.}

\item{expr}{An expression which returns a character string specifying the
label of the badge.}

\item{env}{The environment in which to evaluate \code{expr}, defaults to the
calling environment.}

\item{quoted}{One of \code{TRUE} or \code{FALSE} specifying if \code{expr} is a quoted
expression.}
}
\description{
Small highlighted content which scales to its parent's size. Useful for
displaying dynamically changing counts or tickers, drawing attention to new
options, or tagging content.
}
\examples{

if (interactive()) {
  shinyApp(
    ui = container(
      row(
        column(
          buttonInput(
            id = "mybutton",
            label = list(
              "Number of clicks: ",
              badgeOutput("clicks") \%>\%
                background("red")
            )
          ),
          badgeOutput("another")
        )
      )
    ),
    server = function(input, output) {
      output$another <- renderBadge({
        200
      })

      output$clicks <- renderBadge({
        input$mybutton
      })
    }
  )
}

}
