% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badgeOutput}
\alias{badgeOutput}
\alias{renderBadge}
\title{Badge outputs}
\usage{
badgeOutput(id, content, context = "secondary", ...)

renderBadge(content, context = "secondary", env = parent.frame(),
  quoted = FALSE)
}
\arguments{
\item{id}{A character string specifying the id of the badge output.}

\item{content}{A character string specifying the content of the badge or an
expression which returns a character string.}

\item{context}{An expression which returns one of \code{"primary"}, \code{"secondary"},
\code{"success"}, \code{"info"}, \code{"warning"}, \code{"danger"}, \code{"light"}, or \code{"dark"}. The
expression may contain reactive values.}

\item{...}{Additional named argument passed as HTML attributes to the parent
element.}
}
\description{
Small highlighted content which scales to its parent's size. Useful for
displaying dynamically changing counts or tickers, drawing attention to new
options, or tagging content.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          listGroupInput(
            id = NULL,
            items = "Button clicks",
            badges = list(
              badgeOutput(
                id = "clicks",
                content = 0
              )
            )
          )
        ),
        col(
          buttonInput(
            id = "clicker",
            "Click here!"
          )
        )
      )
    ),
    server = function(input, output) {
      output$clicks <- renderBadge(
        content = {
          input$clicker
        },
        context = {
          req(input$clicker)
          clicks <- input$clicker

          if (clicks > 19) {
            "danger"
          } else if (clicks > 9) {
            "warning"
          } else {
            "info"
          }
        }
      )
    }
  )
}

}
