% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badgeOutput}
\alias{badgeOutput}
\alias{renderBadge}
\title{Badge outputs}
\usage{
badgeOutput(id, content, ...)

renderBadge(content, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{id}{A character string specifying the id of the badge output.}

\item{content}{A character string specifying the content of the badge or an
expression which returns a character string.}

\item{...}{Additional named argument passed as HTML attributes to the parent
element.}

\item{env}{The environment in which to evaluate \code{content}.}

\item{quoted}{One of \code{TRUE} or \code{FALSE} specifying if \code{content} is a quoted
expression.}
}
\description{
Small highlighted content which scales to its parent's size. Useful for
displaying dynamically changing counts or tickers, drawing attention to new
options, or tagging content.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          buttonInput(
            id = "button",
            label = list(
              "Number of clicks: ",
              badgeOutput("clicks", 0) \%>\%
                background("red", -2)
            )
          )
        )
      )
    ),
    server = function(input, output) {
      output$clicks <- renderBadge({
        input$button
      })
    }
  )
}

}
