% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badgeOutput}
\alias{badgeOutput}
\alias{renderBadge}
\title{Badge outputs}
\usage{
badgeOutput(id, ...)

renderBadge(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{id}{A character string specifying the id of the output.}

\item{...}{Additional named argument passed as HTML attributes to the parent
element.}

\item{expr}{An expression which returns a character string specifying the
label of the badge.}

\item{env}{The environment in which to evaluate \code{expr}, defaults to the
calling environment.}

\item{quoted}{One of \code{TRUE} or \code{FALSE} specifying if \code{expr} is a quoted
expression.}
}
\description{
Small highlighted content which scales to its parent's size. Useful for
displaying dynamically changing counts or tickers, drawing attention to new
options, or tagging content.
}
\examples{

## Buttons with badges

# Typically, you would use `renderBadge()` to update a badge's
# value. Here we are hard-coding a default value of 7.

buttonInput(
  id = NULL,
  label = "Process",
  badgeOutput(
    id = NULL,
    7
  ) \%>\%
    background("cyan")
)

## Possible colors

colors <- c(
  "red", "purple", "indigo", "blue", "cyan", "teal", "green",
  "yellow", "amber", "orange", "grey", "white"
)

div(
  lapply(
    colors,
    function(color) {
      badgeOutput(
        id = NULL,
        color
      ) \%>\%
        background(color) \%>\%
        margin(2)
    }
  )
) \%>\%
  display("flex") \%>\%
  flex(wrap = TRUE)

}
\concept{outputs}
