% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{border}
\alias{border}
\title{Tag element borders}
\usage{
border(.tag, color = NULL, sides = "all")
}
\arguments{
\item{.tag}{A tag element.}

\item{color}{A character string specifying the border color, defaults to
\code{NULL}, in which case the browser default is used. See below for possible
border colors.}

\item{sides}{One or more of \code{"top"}, \code{"right"}, \code{"bottom"}, \code{"left"} or
\code{"all"} or \code{"none"} specifying which sides to add a border to, defaults to
\code{"all"}.}
}
\description{
Use \code{border()} to add borders to a tag element or change the color of a tag
element's border.
}
\details{
The following border colors are available,
\itemize{
\item red
\item purple
\item indigo
\item blue
\item cyan
\item teal
\item green
\item yellow
\item amber
\item orange
\item grey
\item white
}
}
\examples{

### Possible colors

colors <- c(
  "red", "purple", "indigo", "blue", "cyan", "teal", "green",
  "yellow", "amber", "orange", "grey", "white"
)

div(
  lapply(
    colors,
    border,
    .tag = div() \%>\%
      padding(5) \%>\%
      margin(2)
  )
) \%>\%
  display("flex") \%>\%
  flex(wrap = TRUE)

}
\seealso{
Other design: \code{\link{affix}},
  \code{\link{background}}, \code{\link{display}},
  \code{\link{float}}, \code{\link{font}},
  \code{\link{height}}, \code{\link{padding}},
  \code{\link{rounded}}, \code{\link{scroll}},
  \code{\link{shadow}}, \code{\link{width}}
}
\concept{design}
