% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{border}
\alias{border}
\title{Element borders}
\usage{
border(tag, context = NULL, sides = "all", rounded = "none")
}
\arguments{
\item{tag}{A tag element object.}

\item{context}{One of \code{NULL}, \code{"primary"}, \code{"secondary"}, \code{"success"},
\code{"info"}, \code{"warning"}, \code{"danger"}, \code{"light"}, \code{"dark"}, or \code{"white"}
specifying the visual context and border color. If \code{NULL}, the border is
the browser default.}

\item{sides}{One of \code{"all"}, \code{"none"}, or one or more of \code{"top"}, \code{"right"},
\code{"bottom"}, \code{"left"} specifying the sides to apply a border, defaults to
\code{"all"}.}

\item{rounded}{One of \code{"all"}, \code{"circle"}, \code{"none"}, or one or more of
\code{"top"}, \code{"right"}, \code{"bottom"}, \code{"left"} specifying the side whose corners
to round.}
}
\description{
Apply borders, border colors, and border radius to a tag element.
}
\examples{
tags$h1("Hello, world!") \%>\%
  border(sides = c("top", "bottom"))

tags$div() \%>\%
  border("warning")

}
\seealso{
Other utilities: \code{\link{affix}}, \code{\link{align}},
  \code{\link{background}}, \code{\link{display}},
  \code{\link{float}}, \code{\link{font}},
  \code{\link{padding}}, \code{\link{text}},
  \code{\link{width}}
}
