% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design-border.R
\name{border}
\alias{border}
\title{Border colors}
\usage{
border(x, color = NULL, sides = "all", round = NULL)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{color}{One of \code{"primary"}, \code{"secondary"}, \code{"success"}, \code{"info"}, \code{"warning"}, \code{"danger"}, \code{"light"} or \code{"dark"} specifying the border color of the tag element, defaults to \code{NULL}.}

\item{sides}{One or more of \code{"top"}, \code{"right"}, \code{"bottom"}, \code{"left"} or
\code{"all"} or \code{"none"} specifying which sides to add a border to, defaults to
\code{"all"}.}

\item{round}{One or more of \code{"top"}, \code{"right"}, \code{"bottom"}, \code{"left"},
\code{"circle"}, \code{"all"}, or \code{"none"} specifying how to round the border(s) of a
tag element, defaults to \code{NULL}, in which case the argument is ignored.}
}
\description{
Use \code{border()} to add or modify tag element borders.
}
\examples{

div() \%>\% border("primary")
# ->
div() \%>\% border_color("primary")

}
\keyword{internal}
