% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design-border.R
\name{border}
\alias{border}
\title{Border colors}
\usage{
border(x, color = NULL, sides = "all", round = NULL)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{color}{One of \code{"primary"}, \code{"secondary"}, \code{"success"}, \code{"info"}, \code{"warning"}, \code{"danger"}, \code{"light"} or \code{"dark"} specifying the border color of the tag element, defaults to \code{NULL}.}

\item{sides}{One or more of \code{"top"}, \code{"right"}, \code{"bottom"}, \code{"left"}, or
\code{"all"} or \code{"none"} specifying which sides to add a border to, defaults to
\code{"all"}. Specifying \code{"none"} will remove the element's borders.}

\item{round}{One or more of \code{"top"}, \code{"right"}, \code{"bottom"}, \code{"left"},
\code{"all"}, or \code{"none"} specifying how to round the border(s) of a tag
element, defaults to \code{NULL}, in which case the argument is ignored.}
}
\description{
Use \code{border()} to add or modify tag element borders.
}
\details{
\subsection{Cards}{

Card borders and font colors.\if{html}{\out{<div class="r">}}\preformatted{card(
  .style \%>\% 
    border("primary") \%>\% 
    font("primary"),
  header = "Primary",
  p("Cras justo odio, dapibus ac facilisis in, egestas eget quam."),
  p("Etiam porta sem malesuada magna mollis euismod.")
)
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve-->
}}
\if{html}{\out{
<div class="r-output">
}}
\if{html}{\out{
<div class="card border border-primary text-primary">
}}
\if{html}{\out{
<div class="card-header">
}}


Primary\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="card-body">
}}
\if{html}{\out{
<p class="card-text">
}}


Cras justo odio, dapibus ac facilisis in, egestas eget quam.\if{html}{\out{
</p>
}}
\if{html}{\out{
<p class="card-text">
}}


Etiam porta sem malesuada magna mollis euismod.\if{html}{\out{
</p>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<!--/html_preserve-->
}}

}
}
\seealso{
Other design utilities: 
\code{\link{active}()},
\code{\link{affix}()},
\code{\link{background}()},
\code{\link{display}()},
\code{\link{flex}()},
\code{\link{float}()},
\code{\link{font}()},
\code{\link{height}()},
\code{\link{margin}()},
\code{\link{padding}()},
\code{\link{scroll}()},
\code{\link{shadow}()},
\code{\link{width}()}
}
\concept{design utilities}
