% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{button}
\alias{button}
\alias{submit}
\alias{updateButton}
\title{Buttons and button groups}
\usage{
button(label = NULL, value = NULL, context = "secondary",
  outline = FALSE, block = FALSE, disabled = FALSE, ..., id = NULL)

submit(label = NULL, outline = FALSE, block = FALSE, ...)

updateButton(id, count = NULL, context = NULL,
  session = getDefaultReactiveDomain())
}
\arguments{
\item{label}{A character string or tag elements to use as a button label or a
character vector specifying labels for a button group, defaults to \code{NULL}.}

\item{value}{A character string specifying a value for the button or a
character vector specifying values for a button group, defaults to
\code{NULL}.}

\item{context}{Used to specify the visual context of the button, one of
\code{"primary"}, \code{"secondary"}, \code{"success"}, \code{"info"}, \code{"warning"}, \code{"danger"},
or \code{"link"}, defaults to \code{"secondary"}.

Primary buttons are blue, secondary buttons are white and grey, buttons for
success are green, informative buttons are a lighter blue, warning buttons
are yellow, and buttons for danger are red. Specifying \code{"link"} makes a
button render with the appearance of a link.}

\item{outline}{If \code{TRUE}, the button's background is transparent, \code{context}
is preserved, defaults to \code{FALSE}.}

\item{block}{If \code{TRUE}, the button is block-level instead of inline, defaults
to \code{FALSE}. A block-level element will occupy the entire space of its
parent element, thereby creating a "block."}

\item{disabled}{If \code{TRUE}, the button renders in a disabled state, defaults
to \code{FALSE}.}

\item{...}{Named arguments passed as HTML attributes to the parent element.}

\item{id}{A character string specifying the id of the button, button group,
or submit button, defaults to \code{NULL}. If specified, a reactive value is
available to the shiny server function.}
}
\description{
Buttons, submit and reset buttons. A button's reactive value is a list of two
items. The first item is \code{count}, the number of clicks on the button. The
second item is \code{value}, the HTML data-value attribute of the button which may
be set with the \code{value} argument.
}
\examples{
button("Primary", context = "primary")

button("Secondary")

button("Success", context = "success")

button("Info", context = "info", outline = TRUE)

button("\\u2715", context = "warning")

button("Danger!", context = "danger", disable = TRUE)

if (interactive()) {
  shinyApp(
    ui = container(
      listGroup(
        listGroupItem(
          inlineForm(
            button(
              id = "clicker1",
              "Simple button"
            )
          ),
          badge = badge(id = "badge1", 0)
        ),
        listGroupItem(
          inlineForm(
            button(
              id = "clicker2",
              label = "Click me!",
            )
          ),
          badge = badge(id = "badge2", 0)
        ),
        listGroupItem(
          button(
            id = "reset",
            label = "Reset",
            context = "primary"
          )
        )
      )
    ),
    server = function(input, output) {
      output$badge1 <- renderBadge(input$clicker1$count)

      output$badge2 <- renderBadge(
        value = {
          input$clicker2$count
        },
        context = {
          if (input$clicker2$count > 5) {
            "warning"
          } else {
            "default"
          }
        }
      )

      observeEvent(input$reset, {
        updateButton("clicker1", count = 0)
        updateButton("clicker2", count = 0)
      })
    }
  )

  shinyApp(
    ui = container(
      button(
        id = "group",
        label = c("First", "Second", "Third"),
        value = c("first", "second", "third")
      )
    ),
    server = function(input, output) {
      observe({
        print(input$group)
      })
    }
  )
}

}
\seealso{
For more about block-level elements please refer to the block-level elements
MDN
\href{https://developer.mozilla.org/en-US/docs/Web/HTML/Block-level_elements}{reference section}.

For more about buttons and button groups please refer to the bootstrap
\href{https://v4-alpha.getbootstrap.com/components/input-group/#button-addons}{reference section}.

Other inputs: \code{\link{groupInput}},
  \code{\link{radioInput}}, \code{\link{selectInput}},
  \code{\link{textInput}}
}
