% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{buttonGroupInput}
\alias{buttonGroupInput}
\title{Button group inputs}
\usage{
buttonGroupInput(id, labels, values = labels, ...)
}
\arguments{
\item{id}{A character string specifying the id of the reactive input.}

\item{labels}{A character vector of labels, a button is added to the group
for each label specified.}

\item{values}{A character vector of values, one for each button specified,
defaults to \code{labels}.}

\item{...}{Additional named arguments passed as HTML attributes to the parent
element.}
}
\description{
A set of buttons with custom values.
}
\examples{

### Default input

buttonGroupInput(
  id = "group1",
  labels = c("Once", "Twice", "Thrice"),
  values = c(1, 2, 3)
)

### Styling the button group

buttonGroupInput(
  id = "group2",
  labels = c("Button 1", "Button 2", "Button 3")
) \%>\%
  background("blue") \%>\%
  width("1/3")

}
\seealso{
Other inputs: \code{\link{buttonInput}},
  \code{\link{checkboxInput}}, \code{\link{chipInput}},
  \code{\link{fileInput}}, \code{\link{formInput}},
  \code{\link{groupInput}}, \code{\link{listGroupInput}},
  \code{\link{menuInput}}, \code{\link{navInput}},
  \code{\link{radioInput}}, \code{\link{rangeInput}},
  \code{\link{selectInput}}, \code{\link{sliderInput}},
  \code{\link{textInput}}
}
\concept{inputs}
