% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{buttonGroupInput}
\alias{buttonGroupInput}
\title{Button groups}
\usage{
buttonGroupInput(id, labels, values = labels, context = "secondary",
  outline = FALSE)
}
\arguments{
\item{id}{A character string specifying the id of the button group input.}

\item{labels}{A character vector of labels, a button is added to the group
for each label specified.}

\item{values}{A character vector of values, one for each button specified,
defaults to \code{labels}.}

\item{context}{One of \code{"primary"}, \code{"secondary"}, \code{"success"}, \code{"info"},
\code{"warning"}, or \code{"danger"} specifying the visual context of the button
group, defaults to \code{"secondary"}.}

\item{outline}{One of \code{TRUE} or \code{FALSE}, if \code{TRUE} the button group appears
as an outline, defaults to \code{FALSE}.`}
}
\description{
Groups of buttons with persisting value.
}
\examples{
buttonGroup("group", c("Once", "Twice"), c(1, 2))

if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          buttonGroup(
            id = "group",
            labels = c("Once", "Twice", "Thrice"),
            values = c(1, 2, 3)
          )
        ),
        col(
          verbatimTextOutput("value")
        )
      )
    ),
    server = function(input, output) {
      output$value <- renderPrint({
        input$group
      })
    }
  )
}

}
