% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{buttonGroupInput}
\alias{buttonGroupInput}
\title{Button group inputs}
\usage{
buttonGroupInput(id, labels, values = labels)
}
\arguments{
\item{id}{A character string specifying the id of the button group input.}

\item{labels}{A character vector of labels, a button is added to the group
for each label specified.}

\item{values}{A character vector of values, one for each button specified,
defaults to \code{labels}.}
}
\description{
Groups of buttons with a persisting value.
}
\examples{
buttonGroupInput("group", c("Once", "Twice"), c(1, 2))

if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          buttonGroupInput(
            id = "group",
            labels = c("Once", "Twice", "Thrice"),
            values = c(1, 2, 3)
          )
        ),
        col(
          verbatimTextOutput("value")
        )
      )
    ),
    server = function(input, output) {
      output$value <- renderPrint({
        input$group
      })
    }
  )
}

if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          buttonGroupInput(
            id = "bg1",
            labels = c("Button 1", "Button 2", "Button 3")
          ) \%>\%
            background("blue", -2) \%>\%
            margins(3)
        ),
        col(
          buttonGroupInput(
            id = "bg2",
            labels = c("Groupee 1", "Groupee 2", "Groupee 3")
          ) \%>\%
            background("yellow", +1) \%>\%
            margins(3)
        )
      )
    ),
    server = function(input, output) {
      observe({
        print(input$bg1)
      })

      observe({
        print(input$bg2)
      })
    }
  )
}

}
