% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-button-group.R
\name{buttonGroupInput}
\alias{buttonGroupInput}
\alias{updateButtonGroupInput}
\title{Button group inputs}
\usage{
buttonGroupInput(
  ...,
  id,
  choices = NULL,
  values = choices,
  direction = "horizontal"
)

updateButtonGroupInput(
  id,
  choices = NULL,
  values = choices,
  enable = NULL,
  disable = NULL,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{...}{Additional named arguments passed as HTML attributes to the parent
element.}

\item{id}{A character string. specifying the id of the reactive input.}

\item{choices}{A character vector specifying the labels for each button in
the group.}

\item{values}{A character vector specifying the values of each button in the
group, defaults to \code{choices}.}

\item{direction}{One of \code{"horizontal"} or \code{"vertical"} specifying the layout
of the buttons, defaults to \code{"horizontal"}.}

\item{enable}{One of \code{values} indicating individual buttons to enable or
\code{TRUE} to enable the entire input, defaults to \code{NULL}.}

\item{disable}{One of \code{values} indicating individual buttons to disable or
\code{TRUE} to disable the entire input, defaults to \code{NULL}.}
}
\description{
A set of buttons with custom values.
}
\seealso{
Other inputs: 
\code{\link{checkbarInput}()},
\code{\link{checkboxInput}()},
\code{\link{chipInput}()},
\code{\link{fileInput}()},
\code{\link{formInput}()},
\code{\link{input_button}()},
\code{\link{listGroupInput}()},
\code{\link{menuInput}()},
\code{\link{radioInput}()},
\code{\link{radiobarInput}()},
\code{\link{rangeInput}()},
\code{\link{selectInput}()},
\code{\link{textInput}()}
}
\concept{inputs}
