% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{buttonInput}
\alias{buttonInput}
\alias{submitInput}
\alias{linkInput}
\title{Button and submit inputs}
\usage{
buttonInput(id, label, ..., block = FALSE)

submitInput(label = "Submit", value = label, block = FALSE, ...)

linkInput(id, text, ...)
}
\arguments{
\item{id}{A character string specifying the id of the reactive input.}

\item{label}{A character string specifying the label text on the button
input.}

\item{...}{Additional named arguments passed as HTML attributes to the parent
element.}

\item{block}{If \code{TRUE}, the input is block-level instead of inline, defaults
to \code{FALSE}. A block-level element will occupy the entire width of its
parent element.}

\item{value}{A character string specifying the value of a submit input,
defaults to \code{label}. This value is used to distinguish form submission
types in the case where a form input has multiple submit inputs.}

\item{text}{A character string specifying the text displayed as part of the
link input.}
}
\description{
Button inputs are useful as triggers for reactive or observer expressions.
The reactive value of a button input begins as \code{NULL}, but subsequently is
the number of clicks.

A submit input is a special type of button used to control form input
submission. Because of their specific usage, submit inputs do not require an
\code{id}, but may have a specified \code{value}. Submit inputs will \emph{not} freeze all
reactive inputs, see \code{\link[=formInput]{formInput()}}.
}
\examples{
### Simple vs block button

buttonInput(
  id = NULL,
  label = "Simple"
)

# Block buttons will fill the width of their parent element
buttonInput(
  id = NULL,
  label = "Block",
  block = TRUE
) \%>\%
  background("red")

### A submit button

submitInput()

# Or use custom text to clarify the action taken when clicked by the user.

submitInput("Place order")

### Possible colors

colors <- c(
  "red", "purple", "indigo", "blue", "cyan", "teal", "green",
  "yellow", "amber", "orange", "grey"
)

div(
  lapply(
    colors,
    function(color) {
      buttonInput(
        id = NULL,
        label = color
      ) \%>\%
        background(color) \%>\%
        margin(2)
    }
  )
) \%>\%
  display("flex") \%>\%
  flex(wrap = TRUE)

### Reactive links

div("Curabitur ", linkInput("inline", "vulputate"), " vestibulum lorem.")

}
\seealso{
Other inputs: \code{\link{buttonGroupInput}},
  \code{\link{checkboxInput}}, \code{\link{chipInput}},
  \code{\link{fileInput}}, \code{\link{formInput}},
  \code{\link{groupInput}}, \code{\link{listGroupInput}},
  \code{\link{menuInput}}, \code{\link{navInput}},
  \code{\link{radioInput}}, \code{\link{rangeInput}},
  \code{\link{selectInput}}, \code{\link{sliderInput}},
  \code{\link{textInput}}
}
\concept{inputs}
