% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{buttonInput}
\alias{buttonInput}
\alias{submitInput}
\alias{linkInput}
\title{Button and submit inputs}
\usage{
buttonInput(id, label, block = FALSE, ...)

submitInput(label = "Submit", block = FALSE, ...)

linkInput(id, text, ...)
}
\arguments{
\item{id}{A character string specifying the id of the button or link input.}

\item{label}{A character string specifying the label text on the button
input.}

\item{block}{If \code{TRUE}, the input is block-level instead of inline, defaults
to \code{FALSE}. A block-level element will occupy the entire width of its
parent element.}

\item{...}{Additional named arguments passed as HTML attributes to the parent
element.}

\item{text}{A character string specifying the text displayed as part of the
link input.}
}
\description{
Button inputs are useful as triggers for reactive or observer expressions.
The reactive value of a button input is the number of times it has been
clicked.

A submit input is a special type of button used to control HTML form
submission. Unlike shiny's \code{submitButton}, \code{submitInput} will not freeze all
reactive inputs on the page.
}
\details{
A submit input is automatically included as part of a \code{\link{formInput}}.
}
\examples{

if (interactive()) {
  shinyApp(
    ui = container(
      row(
        column(
          buttonInput(
            id = "button",
            label = "Click me!"
          ) \%>\%
            background("green")
        ),
        column(
          d4(
            textOutput("clicks")
          )
        )
      )
    ),
    server = function(input, output) {
      output$clicks <- renderText({
        input$button
      })
    }
  )
}


if (interactive()) {
  shinyApp(
    ui = container(
      center = TRUE,
      buttonInput(
        id = "button",
        label = list(
          "Increment badge",
          badgeOutput(
            id = "badge"
          ) \%>\%
            background("green")
        )
      ) \%>\%
        background("amber")
    ),
    server = function(input, output) {
      output$badge <- renderBadge({
        input$button
      })
    }
  )
}


if (interactive()) {
  shinyApp(
    ui = container(
      "Curabitur ", linkInput("inline", "vulputate"), " vestibulum lorem."
    ),
    server = function(input, output) {
      observeEvent(input$inline, {
        showPopover(
          input$inline,
          content = "This means beef?",
          placement = "bottom"
        )
      })
    }
  )
}

}
\seealso{
Bootstrap 4 button documentation:
\url{https://getbootstrap.com/docs/4.0/components/buttons/}

Other inputs: \code{\link{buttonGroupInput}},
  \code{\link{checkbarInput}}, \code{\link{checkboxInput}},
  \code{\link{dateInput}}, \code{\link{dropdownInput}},
  \code{\link{fileInput}}, \code{\link{formInput}},
  \code{\link{groupInput}}, \code{\link{listGroupThruput}},
  \code{\link{radioInput}}, \code{\link{rangeInput}},
  \code{\link{selectInput}}, \code{\link{textInput}}

Other inputs: \code{\link{buttonGroupInput}},
  \code{\link{checkbarInput}}, \code{\link{checkboxInput}},
  \code{\link{dateInput}}, \code{\link{dropdownInput}},
  \code{\link{fileInput}}, \code{\link{formInput}},
  \code{\link{groupInput}}, \code{\link{listGroupThruput}},
  \code{\link{radioInput}}, \code{\link{rangeInput}},
  \code{\link{selectInput}}, \code{\link{textInput}}

Other inputs: \code{\link{buttonGroupInput}},
  \code{\link{checkbarInput}}, \code{\link{checkboxInput}},
  \code{\link{dateInput}}, \code{\link{dropdownInput}},
  \code{\link{fileInput}}, \code{\link{formInput}},
  \code{\link{groupInput}}, \code{\link{listGroupThruput}},
  \code{\link{radioInput}}, \code{\link{rangeInput}},
  \code{\link{selectInput}}, \code{\link{textInput}}
}
\concept{inputs}
