% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{card}
\alias{card}
\title{Cards}
\usage{
card(header = NULL, block = NULL, footer = NULL, text = "left",
  context = NULL, outline = FALSE, inverse = FALSE)
}
\arguments{
\item{header}{Card header, a character vector or tag, defaults to \code{NULL}.}

\item{block}{The card body, a character vector or tag(s), defaults to \code{NULL}.}

\item{footer}{Card footer, a character vector or tag, defaults to \code{NULL}.}

\item{text}{One of \code{"left"}, \code{"right"}, or \code{"center"}, specifying the text
alignment of the card, defaults to \code{"left"}.}

\item{context}{One of \code{"primary"}, \code{"success"}, \code{"info"}, \code{"warning"}, or
\code{"danger"}, used to add a visual context to the card, defaults to \code{NULL}.}

\item{outline}{If \code{TRUE}, the card context is styled as the card outline
rather than the card background, defaults to \code{FALSE}.}

\item{inverse}{If \code{TRUE}, the card has a dark background and light text as
opposed to the default dark text on light background, defaults to \code{FALSE}.}
}
\description{
Handy card tools.
}
