% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{card}
\alias{card}
\alias{deck}
\title{Cards, blocks of content}
\usage{
card(..., header = NULL, title = NULL, subtitle = NULL,
  image = NULL, footer = NULL)

deck(...)
}
\arguments{
\item{...}{For \strong{card}, character strings, tag elements, or list groups to
include as the body of a card or additional named arguments passed as HTML
attributes to the parent element.

For \strong{deck}, any number of cards or additional named arguments passed as
HTML attributes to the parent element.}

\item{header}{A character string or tag element specifying the header of the
card, defaults to \code{NULL}, in which case a header is not added.}

\item{title}{A character string or tag element specifying the title of the
card, defaults to \code{NULL}, in which case a title is not added.}

\item{subtitle}{A character string or tag element specifying the subtitle of
the card, defaults to \code{NULL}, in which case a subtitle is not added.}

\item{image}{An \code{\link[=img]{img()}} element specifying an image to add to the card,
defaults to \code{NULL}, in which case an image is not added.}

\item{footer}{A character string or tag element specifying the footer of the
card, defaults to \code{NULL}, in which case a footer is not added.}
}
\description{
Create blocks of content with \code{card}. \code{deck} is used to group and add padding
is placed around any number of cards. Additionally, grouping cards with
\code{deck} has the benefit of aligning the footer of each card.
}
\examples{

## A simple card

column(
  width = 4,
  card("Praesent fermentum tempor tellus.")
)

## Adding a title, subtitle

column(
  width = 4,
  card(
    title = "Mauris mollis tincidunt felis.",
    subtitle = "Phasellus at dui in ligula mollis ultricies.",
    "Nullam tempus. Mauris mollis tincidunt felis.",
    "Nullam libero mauris, consequat quis, varius et, dictum id, arcu."
  )
)

## Styling cards

deck(
  card(
    header = div("Donec pretium posuere tellus") \%>\%
      background("teal"),
    "Donec hendrerit tempor tellus.",
    "Cras placerat accumsan nulla."
  ),
  card(
    "Aliquam posuere.",
    "Phasellus neque orci, porta a, aliquet quis, semper a, massa.",
    "Pellentesque dapibus suscipit ligula."
  ) \%>\%
    border("orange"),
  card(
    header = div("Phasellus lacus") \%>\%
      background("indigo"),
    "Etiam laoreet quam sed arcu.",
    "Etiam vel tortor sodales tellus ultricies commodo.",
    footer = "Nam euismod tellus id erat."
  ) \%>\%
      background("grey")
)

## Cards with list groups

column(
  width = 4,
  card(
    listGroupThruput(
      id = "important",
      flush = TRUE,
      listGroupItem(
        "Pellentesque tristique imperdiet tortor."
      ),
      listGroupItem(
        "Lorem ipsum dolor sit amet, consectetuer adipiscing elit."
      ),
      listGroupItem(
        "Phasellus purus."
      )
    )
  )
)

## Tabbed content in cards

card(
  header = tabTabs(
    id = "example-card-tabs",
    labels = c("Phasellus", "Donec", "Fusce")
  ),
  tabContent(
    tabs = "example-card-tabs",  # same as the id in `tabTabs`
    tabPane(
      "Phasellus purus.",
      "Proin neque massa, cursus ut, gravida ut, lobortis eget, lacus."
    ),
    tabPane(
      "Donec at pede. Praesent augue."
    ),
    tabPanel(
      "Fusce suscipit, wisi nec facilisis facilisis,",
      "est dui fermentum leo, quis tempor ligula erat quis odio."
    )
  )
)


deck(
  card(
    title = "Nullam tristique",
    "Fusce sagittis, libero non molestie mollis, magna orci ultrices dolor, at vulputate neque nulla lacinia eros.",
    "Nunc rutrum turpis sed pede.",
    footer = "Cras placerat accumsan nulla."
  ),
  card(
    title = "Integer placerat",
    "Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Donec hendrerit tempor tellus.",
    footer = "Cras placerat accumsan nulla."
  ),
  card(
    title = "Phasellus neque",
    "Donec at pede. Etiam vel neque nec dui dignissim bibendum.",
    footer = "Cras placerat accumsan nulla."
  )
)

}
\seealso{
Other content: \code{\link{blockquote}}, \code{\link{d1}},
  \code{\link{dropdown}}, \code{\link{icon}},
  \code{\link{img}}, \code{\link{jumbotron}},
  \code{\link{pre}}, \code{\link{showPopover}},
  \code{\link{tooltip}}
}
\concept{content}
