% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{card}
\alias{card}
\alias{deck}
\title{Cards, blocks of content}
\usage{
card(..., header = NULL, title = NULL, subtitle = NULL, image = NULL,
  footer = NULL)

deck(...)
}
\arguments{
\item{...}{For \strong{card}, character strings, tag elements, or list groups to
include as the body of a card or additional named arguments passed as HTML
attributes to the parent element.

For \strong{deck}, any number of cards or additional named arguments passed as
HTML attributes to the parent element.}

\item{header}{A character string or tag element specifying the header of the
card, defaults to \code{NULL}, in which case a header is not added.}

\item{title}{A character string or tag element specifying the title of the
card, defaults to \code{NULL}, in which case a title is not added.}

\item{subtitle}{A character string or tag element specifying the subtitle of
the card, defaults to \code{NULL}, in which case a subtitle is not added.}

\item{image}{An \code{\link[=img]{img()}} element specifying an image to add to the card,
defaults to \code{NULL}, in which case an image is not added.}

\item{footer}{A character string or tag element specifying the footer of the
card, defaults to \code{NULL}, in which case a footer is not added.}
}
\description{
Create blocks of content with \code{card}. \code{deck} is used to group and add padding
is placed around any number of cards. Additionally, grouping cards with
\code{deck} has the benefit of aligning the footer of each card.
}
\examples{

card("Praesent fermentum tempor tellus.")


card(
  title = "Mauris mollis tincidunt felis.",
  subtitle = "Phasellus at dui in ligula mollis ultricies.",
  "Nullam tempus. Mauris mollis tincidunt felis.",
  "Nullam libero mauris, consequat quis, varius et, dictum id, arcu."
)


card(
  listGroupThruput(
    id = NULL,
    listGroupItem(
      "Pellentesque tristique imperdiet tortor."
    ),
    listGroupItem(
      "Lorem ipsum dolor sit amet, consectetuer adipiscing elit."
    ),
    listGroupItem(
      "Phasellus purus."
    )
  )
)


card(
  header = "Nunc rutrum turpis sed pede.",
  title = "Sed bibendum.",
  "Etiam vel neque nec dui dignissim bibendum. Etiam vel neque nec dui dignissim bibendum.",
  buttonInput(id = NULL, label = "Phasellus purus")
)


card(
  header = tabTabs(
    id = "myCardTabs",
    labels = c("Phasellus", "Donec", "Fusce")
  ),
  tabContent(
    tabs = "myCardTabs",
    tabPane(
      "Phasellus purus. Proin neque massa, cursus ut, gravida ut, lobortis eget, lacus."
    ),
    tabPane(
      "Donec at pede. Praesent augue."
    ),
    tabPanel(
      "Fusce suscipit, wisi nec facilisis facilisis, est dui fermentum leo, quis tempor ligula erat quis odio."
    )
  )
)


card(
  header = div("Donec pretium posuere tellus.") \%>\%
    background("teal"),
  "Donec hendrerit tempor tellus.",
  "Cras placerat accumsan nulla."
)


deck(
  card(
    title = "Nullam tristique",
    "Fusce sagittis, libero non molestie mollis, magna orci ultrices dolor, at vulputate neque nulla lacinia eros.",
    "Nunc rutrum turpis sed pede.",
    footer = "Cras placerat accumsan nulla."
  ),
  card(
    title = "Integer placerat",
    "Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Donec hendrerit tempor tellus.",
    footer = "Cras placerat accumsan nulla."
  ),
  card(
    title = "Phasellus neque",
    "Donec at pede. Etiam vel neque nec dui dignissim bibendum.",
    footer = "Cras placerat accumsan nulla."
  )
)

}
