% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{card}
\alias{card}
\alias{deck}
\title{Cards, blocks of content}
\usage{
card(..., header = NULL, title = NULL, subtitle = NULL, footer = NULL)

deck(...)
}
\arguments{
\item{...}{For \strong{card}, character strings, tag elements, or list groups to
include as the body of a card or additional named arguments passed as HTML
attributes to the parent element.

For \strong{deck}, any number of cards or additional named arguments passed as
HTML attributes to the parent element.}

\item{header}{A character string or tag element specifying the header of the
card, defaults to \code{NULL}, in which case a header is not added.}

\item{title}{A character string or tag element specifying the title of the
card, defaults to \code{NULL}, in which case a title is not added.}

\item{subtitle}{A character string or tag element specifying the subtitle of
the card, defaults to \code{NULL}, in which case a subtitle is not added.}

\item{footer}{A character string or tag element specifying the footer of the
card, defaults to \code{NULL}, in which case a footer is not added.}
}
\description{
Create blocks of content with \code{card}. \code{deck} is used to group and add padding
is placed around any number of cards. Additionally, grouping cards with
\code{deck} has the benefit of aligning the footer of each card.
}
\examples{

}
