% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carousel.R
\name{carousel}
\alias{carousel}
\title{Carousel}
\usage{
carousel(..., id = NULL, controls = TRUE, fade = FALSE)
}
\arguments{
\item{...}{Any number of tag elements.}

\item{id}{A character string specifying an optional id for the carousel,
defaults to \code{NULL}.}

\item{controls}{One of \code{TRUE} or \code{FALSE} specifying if next and previous
slide controls are added to the carousel, defaults to \code{TRUE}.}

\item{fade}{One of \code{TRUE} or \code{FALSE} specifying if slides fade in instead of
sliding in, defaults to \code{FALSE}.}
}
\description{
A carousel of tag elements.
}
