% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-checkbar.R
\name{checkbarInput}
\alias{checkbarInput}
\alias{updateCheckbarInput}
\title{Checkbar input}
\usage{
checkbarInput(..., id, choices = NULL, values = choices, selected = NULL)

updateCheckbarInput(
  id,
  choices = NULL,
  values = choices,
  selected = NULL,
  enable = NULL,
  disable = NULL,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{...}{Optional named arguments specifying HTML attributes for the input
element.}

\item{id}{A character string. The id of the reactive input.}

\item{choices}{A character vector or list of tag element specifying the
input's choices, defaults to \code{NULL}.}

\item{values}{A vector of values specifying the values of the input's
choices, defaults to \code{choices}.}

\item{selected}{One or more of \code{values} specifying the input's default
selected values, defaults to \code{NULL}.}

\item{disable}{A character vector. The values disabled by default.}

\item{session}{A shiny session object.}
}
\description{
A stylized checkbox input. The checkbar input appears similar to a group of
buttons, but with a checked or highlighted state.
}
\seealso{
Other inputs: 
\code{\link{fileInput}()},
\code{\link{input_button}()},
\code{\link{input_checkbox}()},
\code{\link{input_checkbox_button}()},
\code{\link{input_chip}()},
\code{\link{input_form}()},
\code{\link{listGroupInput}()},
\code{\link{menuInput}()},
\code{\link{radioInput}()},
\code{\link{radiobarInput}()},
\code{\link{rangeInput}()},
\code{\link{selectInput}()},
\code{\link{textInput}()}
}
\concept{inputs}
