% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{checkbarInput}
\alias{checkbarInput}
\title{Checkbar inputs}
\usage{
checkbarInput(id, choices, values = choices, selected = NULL,
  label = NULL, context = "secondary")
}
\description{
Similar to a checkbox input.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          checkbarInput(
            id = "checkbar",
            choices = c(
              "Check 1 (pre-selected)",
              "Check 2",
              "Check 3"
            ),
            values = paste0("check", 1:3)
          )
        ),
        col(
          verbatimTextOutput("value")
        )
      )

    ),
    server = function(input, output) {
      output$value <- renderPrint({
        input$checkbar
      })
    }
  )
}
}
