% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkbox.R
\name{checkboxInput}
\alias{checkboxInput}
\alias{updateCheckboxInput}
\alias{validateCheckboxInput}
\alias{disableCheckboxInput}
\alias{enableCheckboxInput}
\title{Checkbox inputs and checkbox bars}
\usage{
checkboxInput(id, label, value, checked = FALSE, ...)

updateCheckboxInput(id, label, session = getDefaultReactiveDomain())

validateCheckboxInput(id, state, session = getDefaultReactiveDomain())

disableCheckboxInput(id, session = getDefaultReactiveDomain())

enableCheckboxInput(id, session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{A character string specifying the id of the checkbox input,
defaults to \code{NULL}, in which case the checkbox element is rendered, but a
reactive value is not passed to the shiny server function.}

\item{label}{A character string specifying a label for the checkbox, defaults
to \code{NULL}, in which case a label is not added.}

\item{checked}{One of \code{TRUE} or \code{FALSE} specifying if the checkbox renders in
the checked state, defaults to \code{FALSE}.}

\item{...}{Checkboxes passed to a checkbox input or additional named
arguments passed on as HTML attributes to the respective parent element.}

\item{values}{A character string specifying a value for the checkbox,
defaults to \code{NULL}, in which case the value of the checkbox is \code{NULL}.}

\item{inline}{If \code{TRUE}, the checkbox input is rendered inline, otherwise
individual checkboxes are put on separate lines, defaults to \code{FALSE}.}

\item{context}{One of \code{"success"}, \code{"warning"}, or \code{"danger"} specifying a
visual context to add to a checkbox input, specifying \code{"none"} will remove
any context on the checkbox, defaults to \code{NULL}. Whenever \code{NULL} the
current checkbox context remains unchanged.}

\item{disable}{One of \code{TRUE}, \code{FALSE}, or a value coercible to either
specifying whether to disable or enable a checkbox input, by default
checkbox inputs render in an enabled state, defaults to \code{NULL}. Whenever
\code{NULL} the current state of the checkbox remains unchanged.}
}
\description{
A reactive input for selecting one or more values. A checkbox bar is a
visually re-styling of a group of checkboxes. A checkbox bar appears as a set
of buttons, but still acts as a checkbox group. Because checkbox bars are
pseudo-buttons a visual context may be specified.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          checkboxInput(
            id = "choices",
            checkbox("Choice 1", 1),
            checkbox("Choice 2", 2),
            checkbox("Choice 3", 3)
          ),
          buttonInput(
            id = "disable",
            label = "Disable / Enable"
          )
        ),
        col(
          checkboxInput(
            id = "options",
            checkbox("Option 1", 1),
            checkbox("Option 2", 2),
            checkbox("Option 3", 3)
          ),
          dropdownInput(
            id = "context",
            label = "Apply a context",
            dropdownItem("success", "success"),
            dropdownItem("warning", "warning"),
            dropdownItem("danger", "danger"),
            dropdownItem("none", "none")
          )
        )
      )
    ),
    server = function(input, output) {
      observe({
        req(input$disable)

        updateCheckbox("choices", disable = input$disable$count \%\% 2)
      })

      observe({
        req(input$context)

        updateCheckbox("options", context = input$context)
      })
    }
  )
}

}
