% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkbox.R
\name{checkboxInput}
\alias{checkboxInput}
\alias{updateCheckboxInput}
\alias{validateCheckboxInput}
\alias{disableCheckboxInput}
\alias{enableCheckboxInput}
\title{Checkbox inputs}
\usage{
checkboxInput(id, choice, value = choice, label = NULL, checked = FALSE,
  ...)

updateCheckboxInput(id, choice, value = choice, checked = FALSE,
  session = getDefaultReactiveDomain())

validateCheckboxInput(id, state, session = getDefaultReactiveDomain())

disableCheckboxInput(id, session = getDefaultReactiveDomain())

enableCheckboxInput(id, session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{A character string specifying the id of the checkbox input, the
reactive value of the checkbox input is available to the shiny server
function as part of the \code{input} object.}

\item{choice}{A character string specifying a label for the checkbox.}

\item{value}{A character string, object to coerce to a character string, or
\code{NULL} specifying the value of the checkbox or a new value for the
checkbox, defaults to \code{choice}.}

\item{label}{A character string specifying a label for the checkbox input,
defaults to \code{NULL}, in which case a label is not added.}

\item{checked}{If \code{TRUE} the checkbox renders in a checked state, defaults
to \code{FALSE}.}

\item{...}{Additional named arguments passed as HTML attributes to the parent
element.}

\item{state}{One of \code{"valid"}, \code{"warning"}, or \code{"danger"} indicating the
state of the checkbox input. If the return value is \code{"valid"} any visual
context is removed.}
}
\description{
A reactive checkbox input. When a checkbox input is unchecked the reactive
value is \code{NULL}. When checked the checkbox input reactive value is \code{value}.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          checkboxInput(
            id = "checkbox",
            label = "A checkbox",
            choice = "Are you there?",
            value = "yes"
          )
        ),
        col(
          display1(
            textOutput("value")
          )
        )
      )
    ),
    server = function(input, output) {
      output$value <- renderText({
        input$checkbox
      })
    }
  )
}

if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          checkboxInput(
            id = "caps",
            choice = "Capslock?",
            value = "capslock"
          )
        ),
        col(
          checkboxInput(
            id = "question",
            choice = "Please check this box",
            value = "checked"
          )
        )
      )
    ),
    server = function(input, output) {
      observe({
        if (isTruthy(input$caps)) {
          updateCheckboxInput(
            id = "question",
            choice = "PLEASE CHECK THIS BOX",
            value = "checked"
          )
        } else {
          updateCheckboxInput(
            id = "question",
            choice = "Please check this box",
            value = "checked"
          )
        }
      })
    }
  )
}

}
\seealso{
Other inputs: \code{\link{buttonInput}},
  \code{\link{formInput}}, \code{\link{groupInput}},
  \code{\link{radioInput}}, \code{\link{selectInput}},
  \code{\link{textInput}}
}
