% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkbox.R
\name{checkboxInput}
\alias{checkboxInput}
\alias{switchInput}
\alias{checkbarInput}
\title{Checkbox inputs}
\usage{
checkboxInput(id, choices, values = choices, selected = NULL, ...,
  inline = FALSE)

switchInput(id, choices, values = choices, selected = NULL, ...)

checkbarInput(id, choices, values = choices, selected = NULL, ...)
}
\arguments{
\item{id}{A character string specifying the id of the reactive input.}

\item{choices}{A character string or vector specifying a label or labels for
the checkbox or checkbar.}

\item{values}{A character string or vector specifying values for the
checkbox or checkbar input, defaults to \code{choice} or \code{values}, respectively.}

\item{selected}{One or more of \code{values} specifying which choices are
selected by default, defaults to \code{NULL}, in which case no choices are
initially selected.}

\item{...}{Additional named arguments passed as HTML attributes to the parent
element.}

\item{inline}{One of \code{TRUE} or \code{FALSE} specifying if the checkbox input
choices render inline or stacked, defaults to \code{FALSE}, in which case the
choices are stacked.}
}
\description{
Reactive checkbox and checkbar inputs. Users may select one or more choices.
The checkbox input appears as a standard checkbox or set of checkboxes. The
checkbar input appears similar to a group of buttons, but with a checked or
highlighted state. When a checkbox or checkbar input has no selected choices
the reactive value is \code{NULL}. Switch inputs differ from checkboxes only in
appearance.
}
\examples{

### One option

checkboxInput(
  id = "checkbox1",
  choices = "Choice 1",
  selected = "Choice 1"
)

### Multiple options

checkboxInput(
  id = "checkbox2",
  choices = c("Choice 1", "Choice 2")
)

### Inline checkbox

checkboxInput(
  id = "checkbox3",
  choices = c("Choice 1", "Choice 2", "Choice 3"),
  inline = TRUE
)

### Checkbar

checkbarInput(
  id = "checks",
  choices = c(
    "Check 1",
    "Check 2",
    "Check 3"
  ),
  selected = "Check 1"
) \%>\%
  background("blue") \%>\%
  margin(2)

### Labeling a checkbar

formGroup(
  label = "Toppings",
  checkbarInput(
    id = "fixins",
    choices = c(
      "Sprinkles",
      "Nuts",
      "Fudge"
    )
  )
)

### Switches

switchInput(
  id = "switch1",
  choices = paste("Switch choice", 1:3),
  selected = "Switch choice 3"
) \%>\%
  active("indigo")

}
\seealso{
Other inputs: \code{\link{buttonGroupInput}},
  \code{\link{buttonInput}}, \code{\link{chipInput}},
  \code{\link{fileInput}}, \code{\link{formInput}},
  \code{\link{groupInput}}, \code{\link{listGroupInput}},
  \code{\link{menuInput}}, \code{\link{navInput}},
  \code{\link{radioInput}}, \code{\link{rangeInput}},
  \code{\link{selectInput}}, \code{\link{sliderInput}},
  \code{\link{textInput}}
}
\concept{inputs}
