% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkbox.R
\name{checkboxInput}
\alias{checkboxInput}
\alias{checkbarInput}
\title{Checkbox inputs}
\usage{
checkboxInput(id, choice, value = choice, checked = FALSE, ...)

checkbarInput(id, choices, values = choices, selected = NULL, ...)
}
\arguments{
\item{id}{A character string specifying the id of the checkbox input, the
reactive value of the checkbox input is available to the shiny server
function as part of the \code{input} object.}

\item{choice, choices}{A character string or vector specifying a label for the
checkbox or checkbar.}

\item{value, values}{A character string or vector specifying values for the
checkbox or checkbar input, defaults to \code{choice} or \code{values}, respectively.}

\item{checked}{If \code{TRUE} the checkbox renders in a checked state, defaults
to \code{FALSE}.}

\item{...}{Additional named arguments passed as HTML attributes to the parent
element.}

\item{selected}{One of \code{values} specifying the initial value of the checkbar
input, defaults to \code{NULL}, in which case no choice is selected.}
}
\description{
A reactive checkbox input. When a checkbox input is unchecked the reactive
value is \code{NULL}. When checked the checkbox input reactive value is \code{value}.
Unlike shiny, yonder's checkbox inputs are a singleton value.
}
\examples{

### Start checked

checkboxInput(
  id = "potenti",
  choice = "Suspendisse potenti",
  checked = TRUE
)

### Checkbar, checkbox group

checkbarInput(
  id = "checks",
  choices = c(
    "Check 1",
    "Check 2",
    "Check 3"
  ),
  selected = "Check 1"
) \%>\%
  background("blue") \%>\%
  margin(2)

### Labeling a checkbar

formGroup(
  label = "Toppings",
  checkbarInput(
    id = "fixins",
    choices = c(
      "Sprinkles",
      "Nuts",
      "Fudge"
    )
  )
)

}
\seealso{
Other inputs: \code{\link{addressInput}},
  \code{\link{buttonGroupInput}},
  \code{\link{buttonInput}}, \code{\link{dateInput}},
  \code{\link{fileInput}}, \code{\link{formInput}},
  \code{\link{groupInput}}, \code{\link{loginInput}},
  \code{\link{navInput}}, \code{\link{radioInput}},
  \code{\link{rangeInput}}, \code{\link{selectInput}},
  \code{\link{textInput}}
}
\concept{inputs}
