% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip.R
\name{chipInput}
\alias{chipInput}
\title{A chips input}
\usage{
chipInput(id, choices, values = choices, selected = NULL, ...,
  max = NULL, block = TRUE)
}
\arguments{
\item{id}{A character string specifying the id of the reactive input.}

\item{choices}{A character vector or list specifying the possible choices.}

\item{values}{A character vector or list of strings specifying the input's
values, defaults to \code{choices}.}

\item{selected}{One or more of \code{values} specifying which values are selected
by default.}

\item{...}{Additional named arguments passed as HTML attributes to the parent
element.}

\item{max}{A number specifying the maximum number of items a user may select.}

\item{block}{One of \code{TRUE} or \code{FALSE} specifying the layout of chips. If
\code{TRUE} chips fill the width of the parent element, otherwise if \code{FALSE} the
chips are rendered inline, defaults to \code{TRUE}. Rendering chips inline is
useful when choice text is consistently short.}
}
\description{
A selectize alternative.
}
\seealso{
Other inputs: \code{\link{buttonGroupInput}},
  \code{\link{buttonInput}}, \code{\link{checkboxInput}},
  \code{\link{fileInput}}, \code{\link{formInput}},
  \code{\link{groupInput}}, \code{\link{listGroupInput}},
  \code{\link{menuInput}}, \code{\link{navInput}},
  \code{\link{radioInput}}, \code{\link{rangeInput}},
  \code{\link{selectInput}}, \code{\link{sliderInput}},
  \code{\link{textInput}}
}
\concept{inputs}
