% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-chip.R
\name{chipInput}
\alias{chipInput}
\alias{updateChipInput}
\title{Chip inputs}
\usage{
chipInput(
  id,
  choices = NULL,
  values = choices,
  selected = NULL,
  ...,
  placeholder = NULL,
  max = Inf,
  inline = TRUE,
  sort = "stack"
)

updateChipInput(
  id,
  choices = NULL,
  values = choices,
  selected = NULL,
  max = NULL,
  enable = NULL,
  disable = NULL,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{choices}{A character vector or list specifying the possible choices.}

\item{values}{A character vector or list of strings specifying the input's
values, defaults to \code{choices}.}

\item{selected}{One or more of \code{values} specifying which values are selected
by default.}

\item{placeholder}{A character string specifying placeholder text of the
chip input, defaults to \code{NULL}.}

\item{max}{A number specifying the maximum number of items a user may select,
defaults to \code{Inf}.}

\item{inline}{One of \code{TRUE} or \code{FALSE} specifying if chips are rendered
inline. If \code{TRUE} multiple chips may fit onto a single row, otherwise, if
\code{FALSE}, chips expand to fill the width of their parent element, one chip
per row.}

\item{sort}{One of \code{"stack"}, \code{"queue"}, or \code{"fixed"} specifying how
selected chips are ordered, defaults to \code{"stack"}.

\code{"stack"}, selected chips are placed ahead of other selected chips.

\code{"queue"}, selected chips are placed behind other selected chips.

\code{"fixed"}, selected chips appear in the order specified by
\code{choices} and \code{values}. Use \code{"fixed"} and sort \code{choices} to keep selected
chips in the same sorted order.}
}
\description{
The chip input is a selectize alternative. Choices are selected from a
dropdown menu and appear as chips below the input's text box. Chips do not
appear in the order they are selected. Instead chips are shown in the order
specified by the \code{choices} argument. Use the \code{max} argument to limit the
number of choices a user may select.
}
\seealso{
Other inputs: 
\code{\link{checkbarInput}()},
\code{\link{fileInput}()},
\code{\link{formInput}()},
\code{\link{input_button}()},
\code{\link{input_checkbox}()},
\code{\link{listGroupInput}()},
\code{\link{menuInput}()},
\code{\link{radioInput}()},
\code{\link{radiobarInput}()},
\code{\link{rangeInput}()},
\code{\link{selectInput}()},
\code{\link{textInput}()}
}
\concept{inputs}
