% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\alias{showCollapse}
\alias{hideCollapse}
\alias{toggleCollapse}
\title{Collapsible sections}
\usage{
collapse(button, content, ...)

showCollapse(id, session = getDefaultReactiveDomain())

hideCollapse(id, session = getDefaultReactiveDomain())

toggleCollapse(id, session = getDefaultReactiveDomain())
}
\arguments{
\item{button}{A button input, when clicked the button will hide or show the
collapsable content.}

\item{content}{A tag element speicfying the content to collapse.}

\item{...}{Additional named arguments passed as attributes to the parent
element.}

\item{id}{A character string specifying the id of a collapse section.}

\item{session}{A \code{session} object passed to the shiny server function,
defaults to \code{\link[=getDefaultReactiveDomain]{getDefaultReactiveDomain()}}.}
}
\description{
\code{collapse} allows content to be hidden and toggled into a shown state.
The collapsed content is toggled by a button or may be toggled between hidden
and shown states using \code{toggleCollapse}. \code{showCollapse} and \code{hideCollapse}
are alternatives to \code{toggleCollapse} to change the state of collapsed content
to shown or hidden, respectively.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          buttonInput(
            id = "trigger",
            label = "The Time Machine"
          ) \%>\%
            background("purple", +1) \%>\%
            margins(2) \%>\%
            collapse(
              tags$div(
                class  = "card card-block",
                "\\"The Time Traveller (for so it will be convenient to speak
                of him) was expounding a recondite matter to us. His grey eyes
                shone and twinkled, and his usually pale face was flushed and
                animated. The fire burned brightly, and the soft radiance of
                the incandescent lights in the lilies of silver caught the
                bubbles that flashed and passed in our glasses.\\""
              )
            )
        ),
        col(
          buttonInput(id = "toggle", "Toggle"),
          buttonInput(id = "hide", "Hide"),
          buttonInput(id = "show", "Show")
        )
      )
    ),
    server = function(input, output) {
      observeEvent(input$toggle, {
        toggleCollapse("trigger")
      })

      observeEvent(input$hide, {
        hideCollapse("trigger")
      })

      observeEvent(input$show, {
        showCollapse("trigger")
      })
    }
  )
}

}
