% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\alias{showCollapse}
\alias{hideCollapse}
\alias{toggleCollapse}
\title{Collapsible sections}
\usage{
collapse(tag, id, ...)

showCollapse(id)

hideCollapse(id)

toggleCollapse(id)
}
\arguments{
\item{tag}{A tag element.}

\item{id}{A character string specifying an HTML id. Pass this id to the
\code{*Collapse()} functions to hide or show the collapsible element.}

\item{...}{Additional named arguments passed as HTML attributes to the
collapsible div.}
}
\description{
The \code{collapse()} function allows you to make a tag element collapsible. The
state of the element, shown or hidden, is toggled using \code{hideCollapse()},
\code{showCollapse()}, and \code{toggleCollapse()}.
}
\examples{

## Making an element collapsible

# On the server side you will need to call `hideCollapse` or
# `toggleCollapse`

card(
  '"The Time Traveller (for so it will be convenient to speak
    of him) was expounding a recondite matter to us. His grey eyes
    shone and twinkled, and his usually pale face was flushed and
    animated. The fire burned brightly, and the soft radiance of
    the incandescent lights in the lilies of silver caught the
    bubbles that flashed and passed in our glasses."'
) \%>\%
  background("grey") \%>\%
  collapse("an-html-id")  # pass this id to the `*Collapse` function

}
\seealso{
Other utilities: \code{\link{markInvalid}},
  \code{\link{showAlert}}, \code{\link{showPopover}},
  \code{\link{updateChoices}}
}
\concept{utilities}
