% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_panel}
\alias{collapse_panel}
\alias{collapse_panel_button}
\title{Collapsible panes}
\usage{
collapse_panel(
  id,
  ...,
  state = c("closed", "open"),
  direction = c("vertical", "horizontal")
)

collapse_panel_button(target, text)
}
\arguments{
\item{id}{A character string. The id of the reactive input.}

\item{...}{Any number of child elements passed to the parent element. Named
values are passed as HTML attributes to the parent element.}

\item{state}{A string. The default state of the panel.}

\item{direction}{A string. The direction to open the panel.}

\item{target}{A string. A panel id.}

\item{text}{A string. The content of the button.}
}
\value{
A \link[htmltools:builder]{htmltools::tag} object.
}
\description{
Create collapsible containers to hide and show content. Open and close a
panel with \code{collapse_panel_toggle()}.
}
\examples{

collapse_panel("panel1", "Hidden panel")

collapse_panel_button("panel1", "Toggle the hidden panel")

}
\seealso{
Other components: 
\code{\link{alert}()},
\code{\link{badge}()},
\code{\link{d1}()},
\code{\link{modal_dialog}()},
\code{\link{toast}()}
}
\concept{components}
