% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{column}
\alias{column}
\alias{columns}
\alias{container}
\title{Grid layout}
\usage{
column(..., width = "equal")

columns(...)

container(..., centered = FALSE)
}
\arguments{
\item{...}{Any number of tags elements passed as child elements or named
arguments passed as HTML attributes to the parent element.}

\item{width}{A \link{responsive} argument. One of \code{1:12}, \code{"content"}, or
\code{"equal"}, defaults to \code{"equal"}.}

\item{centered}{One of \code{TRUE} or \code{FALSE} specifying how a container fills the
browser or viewport window. If \code{TRUE} the container is responsively
centered, otherwise, if \code{FALSE}, the container occupies the entire width of
the viewport, defaults to \code{FALSE}.}
}
\description{
These functions are the foundation of any application. Grid elements are
nested as follows: \verb{container() > columns() > column() ~ column()}. A
\code{column()} may be created with an explicit width, 1 through 12. To fit a
column automatically to its content use \code{width = "content"}. To divide the
space in a row evenly amongst all columns use \code{width = "equal"}. For examples
and usage tips see the sections below.
}
\details{
\subsection{Equal width columns}{\if{html}{\out{<div class="r">}}\preformatted{container(
  columns(
    column(
      "Aliquam erat volutpat."
    ),
    column(
      "Mauris mollis tincidunt felis."
    ),
    column(
      "Cum sociis natoque penatibus et magnis dis parturient montes,",
      "nascetur ridiculus mus."
    )
  )
)
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve-->
}}
\if{html}{\out{
<div class="r-output">
}}
\if{html}{\out{
<div class="container-fluid">
}}
\if{html}{\out{
<div class="row">
}}
\if{html}{\out{
<div class="col">
}}


Aliquam erat volutpat.\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="col">
}}


Mauris mollis tincidunt felis.\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="col">
}}


Cum sociis natoque penatibus et magnis dis parturient montes, nascetur
ridiculus mus.\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<!--/html_preserve-->
}}

}

\subsection{Sidebar layout}{\if{html}{\out{<div class="r">}}\preformatted{container(
  columns(
    column(
      width = 4,  # <-
      card(
        h5("Sidebar panel")
      )
    ),
    column(
      h5("Main panel")
    )
  )
)
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve-->
}}
\if{html}{\out{
<div class="r-output">
}}
\if{html}{\out{
<div class="container-fluid">
}}
\if{html}{\out{
<div class="row">
}}
\if{html}{\out{
<div class="col-4">
}}
\if{html}{\out{
<div class="card">
}}
\if{html}{\out{
<div class="card-body">
}}
\if{html}{\out{
<h5 class="card-title">
}}


Sidebar panel\if{html}{\out{
</h5>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="col">
}}
\if{html}{\out{
<h5>
}}


Main panel\if{html}{\out{
</h5>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<!--/html_preserve-->
}}

}

\subsection{Mobile friendly}{

Use the \link{responsive} \code{width} argument to make mobile
friendly applications. On extra small screens each column spans the
entire width of the row. On larger screens the columns will instead
split the row width evenly.\if{html}{\out{<div class="r">}}\preformatted{container(
  columns(
    column(
      width = c(xs = 12, sm = 4),
      "Mauris ac felis vel velit tristique imperdiet."
    ),
    column(
      width = c(xs = 12, sm = 4),
      "Nam vestibulum accumsan nisl."
    ),
    column(
      width = c(xs = 12, sm = 4),
      "Proin neque massa, cursus ut, gravida ut, lobortis eget, lacus."
    )
  )
)
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve-->
}}
\if{html}{\out{
<div class="r-output">
}}
\if{html}{\out{
<div class="container-fluid">
}}
\if{html}{\out{
<div class="row">
}}
\if{html}{\out{
<div class="col-12 col-sm-4">
}}


Mauris ac felis vel velit tristique imperdiet.\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="col-12 col-sm-4">
}}


Nam vestibulum accumsan nisl.\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="col-12 col-sm-4">
}}


Proin neque massa, cursus ut, gravida ut, lobortis eget, lacus.\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<!--/html_preserve-->
}}

}

\subsection{Automatic column sizing}{

Columns may be fit to the width of their content. Other columns in this
example will split the remaining space equally.\if{html}{\out{<div class="r">}}\preformatted{container(
  columns(
    column(
      .style \%>\%
        background("warning")
    ),
    column(
      width = "content",
      "Cras placerat accumsan nulla. Aenean in sem ac leo mollis blandit.",
      "Nam euismod tellus id erat.",
      "Donec neque quam, dignissim in, mollis nec, sagittis eu, wisi."
    ),
    column(
      .style \%>\%
        background("warning")
    )
  )
)
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve-->
}}
\if{html}{\out{
<div class="r-output">
}}
\if{html}{\out{
<div class="container-fluid">
}}
\if{html}{\out{
<div class="row">
}}
\if{html}{\out{
<div class="bg-warning col">
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="col-auto">
}}


Cras placerat accumsan nulla. Aenean in sem ac leo mollis blandit. Nam
euismod tellus id erat. Donec neque quam, dignissim in, mollis nec,
sagittis eu, wisi.\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="bg-warning col">
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<!--/html_preserve-->
}}


Without specficiying widths the space in a row is evenly split between
columns.\if{html}{\out{<div class="r">}}\preformatted{container(
  columns(
    column(),
    column(),
    column(),
    column()
  )
)
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve-->
}}
\if{html}{\out{
<div class="r-output">
}}
\if{html}{\out{
<div class="container-fluid">
}}
\if{html}{\out{
<div class="row">
}}
\if{html}{\out{
<div class="col">
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="col">
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="col">
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="col">
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<!--/html_preserve-->
}}

}
}
\seealso{
Other layout functions: 
\code{\link{fieldset}()},
\code{\link{navbar}()},
\code{\link{webpage}()}
}
\concept{layout functions}
