% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex.R
\name{content}
\alias{content}
\title{Align flex items}
\usage{
content(tag, default = NULL, sm = NULL, md = NULL, lg = NULL,
  xl = NULL)
}
\arguments{
\item{tag}{A tag element.}

\item{default}{One of \code{"start"}, \code{"end"}, \code{"center"}, \code{"between"} or
\code{"around"} specifying the default alignment of the element's flex items.}

\item{sm}{Like \code{default}, but the alignment is applied once the viewport is
576 pixels wide, think phone in landscape mode.}

\item{md}{Like \code{default}, but the alignment is applied once the viewport is
768 pixels wide, think tablets.}

\item{lg}{Like \code{default}, but the alignment is applied once the viewport is
992 pixels wide, think desktop.}

\item{xl}{Like \code{default}, but the alignment is applied once the viewport is
1200 pixels wide, think large desktop.}
}
\description{
The \code{content} function adds Bootstrap classes to a tag element to change how
flex item's align on the x-axis (when the flex direction is row) or on the
y-axis (when the flex is direction is column). For more on flex directions
see \link{direction}.
}
\details{
The flex box mindset is no easy feat. Here is one way to think about the
effects of \code{content}. If you would, take a look over the diagrams below.

\code{"start"}
\itemize{
\item \code{| Item 1 | Item 2 | Item 3 | ============= |}
}

\code{"end"}
\itemize{
\item \code{| ============= | Item 1 | Item 2 | Item 3 |}
}

\code{"center"}
\itemize{
\item \code{| ===== | Item 1 | Item 2 | Item 3 | ===== |}
}

\code{"between"}
\itemize{
\item \code{| Item 1 | ===== | Item 2 | ===== | Item 3 |}
}

\code{"around"}
\itemize{
\item \code{| = | Item 1 | = | Item 2 | = | Item 3 | = |}
}

With \code{content} each possible value relates to where the content \strong{as a
whole} is placed along the main axis. \code{"between"} and \code{"around"} prove
troublesome.
}
\examples{
lapply(1:5, tags$div) \%>\%
  tags$div() \%>\%
  display("flex") \%>\%
  content("center")

}
\seealso{
Other flex: \code{\link{direction}}, \code{\link{items}},
  \code{\link{wrap}}
}
\concept{flex}
