% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{d1}
\alias{d1}
\alias{d2}
\alias{d3}
\alias{d4}
\title{Headings}
\usage{
d1(...)

d2(...)

d3(...)

d4(...)
}
\arguments{
\item{...}{Any number of character strings or tag elements or named arguments
passed as HTML attributes to the parent element.}
}
\description{
Display headings are not meant to replace the standard HTML heading tags,
they are a stand out alternative for eye-catching titles.
}
\seealso{
Other components: 
\code{\link{alert}()},
\code{\link{badge}()},
\code{\link{blockquote}()},
\code{\link{collapsePane}()},
\code{\link{dropdown}()},
\code{\link{img}()},
\code{\link{jumbotron}()},
\code{\link{modal}()},
\code{\link{pre}()},
\code{\link{toast}()}
}
\concept{components}
