% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{dateInput}
\alias{dateInput}
\alias{dateRangeInput}
\title{Date inputs}
\usage{
dateInput(id, choices = NULL, selected = NULL, min = NULL,
  max = NULL, multiple = FALSE, ...)

dateRangeInput(id, choices = NULL, selected = NULL, min = NULL,
  max = NULL, ...)
}
\arguments{
\item{id}{A character specifying the id of the datetime input.}

\item{choices}{Date objects or character strings specifying the set of
dates the user may choose from, defaults to \code{NULL} in which case the user
may choose any date.}

\item{selected}{Date objects or character strings specifying the dates
selected by default in the date input, defaults \code{NULL} in which case no
dates are selected by default.}

\item{min, max}{Date objects or character strings in the format \code{YYYY-mm-dd}
specifying the minimum and maximum date that can be selecetd, both
default to \code{NULL} in which case there is no minimum or maximum value
respectively.}

\item{multiple}{One of \code{TRUE} or \code{FALSE} specifying whether multiple dates
may be selected, if \code{TRUE} the user may select multiple dates and a vector
of one or more dates is returned as the reactive value.}

\item{...}{Additional named arguments passed as HTML attributes to the parent
element.}
}
\description{
A date time picker. Alternatively, use the date time range picker to select
a range of dates. The value of the date range picker is always two dates.
}
\examples{

### Preselect a date

dateInput(
  id = NULL,
  selected = Sys.Date() + 1
)

### Set a min and max

dateInput(
  id = NULL,
  min = Sys.Date() - 3,
  max = Sys.Date() + 3
)

### Select multiple dates

dateInput(
  id = NULL,
  choices = Sys.Date() + seq(-6, 6, by = 2),
  selected = Sys.Date() + 1,
  multiple = TRUE
)

### Date ranges

dateRangeInput(
  id = NULL,
  selected = c(Sys.Date(), Sys.Date() + 3)
)

}
\seealso{
Other inputs: \code{\link{addressInput}},
  \code{\link{buttonGroupInput}},
  \code{\link{buttonInput}}, \code{\link{checkboxInput}},
  \code{\link{fileInput}}, \code{\link{formInput}},
  \code{\link{groupInput}}, \code{\link{loginInput}},
  \code{\link{radioInput}}, \code{\link{rangeInput}},
  \code{\link{selectInput}}, \code{\link{textInput}}
}
\concept{inputs}
