% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex.R
\name{direction}
\alias{direction}
\alias{reverse}
\title{Flex box}
\usage{
direction(tag, default = NULL, sm = NULL, md = NULL, lg = NULL,
  xl = NULL)

reverse(tag, default = NULL, sm = NULL, md = NULL, lg = NULL,
  xl = NULL)
}
\arguments{
\item{tag}{A tag element.}
}
\description{
Use flexbox utilities.
}
\examples{
tags$div() \%>\%
  display("flex") \%>\%
  direction("row") \%>\%

Flex direction, rows and columns

Change the direction of a flex boxes items. `reverse` is similar to
`direction`, but for `"row"` flex items start from the right of the parent
flex box and for `"column"` flex items start from the bottom of the parent
flex box.


}
