% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex.R
\name{direction}
\alias{direction}
\alias{reverse}
\title{Flex direction, rows and columns}
\usage{
direction(tag, default = NULL, sm = NULL, md = NULL, lg = NULL,
  xl = NULL)

reverse(tag, default = NULL, sm = NULL, md = NULL, lg = NULL,
  xl = NULL)
}
\arguments{
\item{tag}{A tag element.}

\item{default}{One of \code{"row"} or \code{"column"} specifying the default direction
of an element's flex items.}

\item{sm}{Like \code{default}, but the direction is applied once the viewport is
576 pixels wide, think phone in landscape mode.}

\item{md}{Like \code{default}, but the direction is applied once the viewport is
768 pixels wide, think tablets.}

\item{lg}{Like \code{default}, but the direction is applied once the viewport is
992 pixels wide, think desktop.}

\item{xl}{Like \code{default}, but the direction is applied once the viewport is
1200 pixels wide, think large desktop.}
}
\description{
Change the direction of a flex boxes items. \code{reverse} is similar to
\code{direction}, but for \code{"row"} flex items start from the right of the parent
flex box and for \code{"column"} flex items start from the bottom of the parent
flex box.
}
\examples{

}
\seealso{
Other flex: \code{\link{content}}, \code{\link{items}},
  \code{\link{wrap}}

Other flex: \code{\link{content}}, \code{\link{items}},
  \code{\link{wrap}}
}
