% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{display}
\alias{display}
\title{Tag element display}
\usage{
display(tag, inline = NULL, block = NULL, flex = NULL, none = NULL)
}
\arguments{
\item{tag}{A tag element.}

\item{inline}{..}

\item{block}{..}

\item{flex}{..}

\item{none}{..}
}
\description{
Use the \code{display()} utility to adjust a tag element's display property. This
allows you to hide elements on small screens or convert elements from inline
to block on large screens.
}
\examples{

div() \%>\%
  display(none = TRUE, block = c(md = TRUE))

}
\seealso{
Other utilities: \code{\link{affix}},
  \code{\link{background}}, \code{\link{border}},
  \code{\link{float}}, \code{\link{font}},
  \code{\link{padding}}, \code{\link{rounded}},
  \code{\link{text}}, \code{\link{width}}
}
\concept{utilities}
