% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design-display.R
\name{display}
\alias{display}
\title{Display}
\usage{
display(x, type)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{type}{A \link{responsive} argument. One of \code{"inline"}, \code{"block"},
\code{"inline-block"}, \code{"flex"}, \code{"inline-flex"}, or \code{"none"}.}
}
\description{
Use the \code{display()} utility to adjust how a tag element is rendered. All
arguments are responsive allowing you to hide elements on small screens or
convert elements from inline to block on large screens.
}
\seealso{
Other design utilities: 
\code{\link{active}()},
\code{\link{affix}()},
\code{\link{background}()},
\code{\link{border}()},
\code{\link{flex}()},
\code{\link{float}()},
\code{\link{font}()},
\code{\link{height}()},
\code{\link{margin}()},
\code{\link{padding}()},
\code{\link{scroll}()},
\code{\link{shadow}()},
\code{\link{width}()}
}
\concept{design utilities}
