% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{display}
\alias{display}
\title{Tag element display}
\usage{
display(.tag, render = NULL, print = NULL)
}
\arguments{
\item{.tag}{A tag element.}

\item{render, print}{A \link{responsive} argument. One of \code{"inline"}, \code{"block"},
\code{"inline-block"}, \code{"flex"}, \code{"inline-flex"}, or \code{"none"}, defaults to
\code{NULL}.}
}
\description{
Use the \code{display()} utility to adjust how a tag element is rendered. All
arguments are responsive allowing you to hide elements on small screens or
convert elements from inline to block on large screens. Most of the time
you will use the \code{render} argument. However if you want to control how an
element appears (or does not appear) when the page is printed use the \code{print}
argument.
}
\examples{

display(div(), render = c(xs = "none", md = "block"))

display(div(), render = c(xs = "inline", sm = "block"))

display(div(), print = "none")

}
\seealso{
Other utilities: \code{\link{affix}},
  \code{\link{background}}, \code{\link{border}},
  \code{\link{flex}}, \code{\link{float}},
  \code{\link{font}}, \code{\link{markInvalid}},
  \code{\link{padding}}, \code{\link{rounded}},
  \code{\link{sendModal}}, \code{\link{showPopover}},
  \code{\link{tooltip}}, \code{\link{updateChoices}},
  \code{\link{width}}
}
\concept{utilities}
