% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pronoun-style.R
\docType{data}
\name{.style}
\alias{.style}
\title{Style pronoun}
\usage{
.style
}
\description{
The \code{.style} pronoun allows you to define styles for a tag element within the
context of the element. Prior to the introduction of the \code{.style} pronoun tag
styles were always applied outside or after constructing a tag element.\if{html}{\out{<div class="R">}}\preformatted{card() \%>\% background("primary") \%>\% display("flex")
}\if{html}{\out{</div>}}

However, once the content of a tag element grows to more than a few lines,
associating the element's styles with the element becomes increasingly
unintuitive. In these situations, make use of the \code{.style} pronoun.\if{html}{\out{<div class="R">}}\preformatted{card(
  .style \%>\%
    border("primary") \%>\%
    font("primary")
)
}\if{html}{\out{</div>}}
}
\keyword{datasets}
