% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown.R
\name{dropdown}
\alias{dropdown}
\title{Dropdown Menu}
\usage{
dropdown(button = NULL, links = NULL, ...)
}
\arguments{
\item{button}{A button to trigger the dropdown, see \code{button}.}

\item{links}{A list of \code{tags$a} elements.}

\item{...}{Additional named arguments passed on as HTML attributes.}
}
\description{
A Bootstrap dropdown menu.
}
\examples{
# A dropdown without `id` attribute
#   - standard `tags$a` utility
dropdown(
  button("Secondary"),
  list(
    tags$a(href = "#", "Option 1"),
    tags$a(href = "3", "Option 2")
  )
)

# A dropdown with `id` attribute
#   - `tags$a` now act similar `shiny::actionButton`
#   - when `tags$a` clicked `input$flavor` will have `href` value
#   - default `tags$a` behavior suppressed
dropdown(
  id = "flavor",
  button("Ice cream flavors"),
  list(
    tags$a(href = "vanilla", "Vanilla"),
    tags$a(href = "chocolate", "Chocolate"),
    tags$a(href = "mint", "Mint")
  )
)

}
