% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown.R
\name{dropdown}
\alias{dropdown}
\title{Dropdown Menu}
\usage{
dropdown(button = NULL, ...)
}
\arguments{
\item{button}{A button element used to trigger the dropdown, see \code{button}.}

\item{...}{\code{dropdownItem}s passed to \code{dropdown} or additional named arguments
passed as HTML attributes to the parent element.}
}
\description{
A Bootstrap dropdown menu.
}
\examples{
# A dropdown without `id` attribute
#   - standard `tags$a` utility
dropdown(
  button = button("Secondary"),
  dropdownItem("Option 1", href = "#"),
  dropdownItem("Option 2", href = "#")
)

# A dropdown with `id` attribute
#   - now `dropdownItem`s similar to `shiny::actionButton`
#   - when clicked `dropdownItem` will have `href` value
dropdown(
  id = "flavor",
  button = button("Ice cream flavors"),
  dropdownItem("Vanilla", href = "vanilla"),
  dropdownItem("Chocolate", href = "chocolate"),
  dropdownItem("Mint", href = "mint")
)

}
