% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown.R
\name{dropdown}
\alias{dropdown}
\title{Dropdown menus}
\usage{
dropdown(label, ..., direction = "down", split = FALSE)
}
\arguments{
\item{label}{A character string specifying the label of the dropdown's
button.}

\item{...}{Character strings or vectors, header tag elements, button inputs,
or form inputs specifying the elements of the dropdown. These elements may
be grouped into lists to create a menu with sections. \code{h6()} is the
recommended heading level for menu headers. Character vectors are converted
into paragraphs of text. To format menu text use \code{p()} and utility
functions.  Named arguments are passed HTML attributes to the parent
element.}

\item{direction}{One of \code{"up"}, \code{"right"}, \code{"down"}, or \code{"left"} specifying
the direction in which the menu opens, defaults to \code{"down"}.}

\item{split}{One of \code{TRUE} or \code{FALSE} specifying if the dropdown toggle button
is split into two distinct buttons, defaults to \code{FALSE}. This is a stylistic
modification which properly spaces the dropdown toggle icon and aligns the
dropdown menu to the toggle icon.}
}
\description{
Dropdown menus are a container for buttons, text, and form inputs. See
argument \code{...} for details on composing dropdown menus.
}
\examples{

if (interactive()) {
  shinyApp(
    ui = container(
      row(
        column(
          dropdown(
            label = "Dropdown",
            split = TRUE,
            formInput(
              id = "login",
              formGroup(
                label = "Email address",
                textInput(
                  id = "email",
                  placeholder = "email@example.com"
                )
              ),
              formGroup(
                label = "Password",
                passwordInput(
                  id = "password",
                  placeholder = "*****"
                )
              ),
              submit = submitInput(
                label = "Sign in"
              )
            ) \%>\%
              padding(3, 4, 3, 4),
            list(
              buttonInput(
                id = "signup",
                label = "New? Sign up"
              ),
              buttonInput(
                id = "forgot",
                label = "Forgot password?"
              )
            )
          )
        ),
        column(
          verbatimTextOutput("values")
        )
      )
    ),
    server = function(input, output) {
      output$values <- renderPrint(
        list(
          email = input$email,
          password = input$password,
          signup = input$signup,
          forgot = input$forgot
        )
      )
    }
  )
}

}
\seealso{
Other content: \code{\link{blockquote}},
  \code{\link{card}}, \code{\link{d1}}, \code{\link{icon}},
  \code{\link{img}}, \code{\link{jumbotron}},
  \code{\link{pre}}, \code{\link{showPopover}},
  \code{\link{tooltip}}
}
\concept{content}
