% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown.R
\name{dropdown}
\alias{dropdown}
\title{Dropdown menus}
\usage{
dropdown(label, ..., direction = "down", align = "left")
}
\arguments{
\item{label}{A character string specifying the label of the dropdown's
button.}

\item{...}{Character strings or vectors, header tag elements, button inputs,
or form inputs specifying the elements of the dropdown. These elements may
be grouped into lists to create a menu with sections. \code{h6()} is the
recommended heading level for menu headers. Character vectors are converted
into paragraphs of text. To format menu text use \code{p()} and utility
functions. Named arguments are passed as HTML attributes to the parent
element.}

\item{direction}{One of \code{"up"}, \code{"right"}, \code{"down"}, or \code{"left"} specifying
the direction in which the menu opens, defaults to \code{"down"}.}

\item{align}{One of \code{"left"} or \code{"right"} specifying which side of the button
to align the dropdown menu to, defaults to \code{"left"}.}
}
\description{
Dropdown menus are a container for buttons, text, and form inputs. See
argument \code{...} for details on composing dropdown menus.
}
\examples{

### Simple options w/ buttons

dropdown(
  label = "Choices",
  buttonInput("choice1", "Choice 1"),
  buttonInput("choice2", "Choice 2"),
  buttonInput("choice3", "Choice 3")
)

### Grouped sections

dropdown(
  label = "Sections",
  list(
    h6("Section 1"),
    buttonInput("addA", "Add A"),
    buttonInput("addB", "Add B")
  ),
  list(
    h6("Section 2"),
    buttonInput("calcC", "Calculate C"),
    buttonInput("calcD", "Calculate D")
  )
)

### Direction variations

div(
  lapply(
    c("up", "down", "left", "right"),
    function(d) {
      dropdown(
        label = d,
        direction = d,
        buttonInput(NULL, "Nam euismod"),
        buttonInput(NULL, "Nunc eleifend"),
        buttonInput(NULL, "Nullam eu")
      ) \%>\%
        margin(3)
    }
  )
) \%>\%
  display("flex")

### Include forms

dropdown(
  label = "Sign in",
  formInput(
    id = "login",
    formGroup(
      label = "Email address",
      textInput(
        id = "email",
        placeholder = "email@example.com"
      )
    ),
    formGroup(
      label = "Password",
      passwordInput(
        id = "password",
        placeholder = "*****"
      )
    ),
    submit = submitInput(
      label = "Sign in"
    )
  ) \%>\%
    padding(3, 4, 3, 4)
)

}
\seealso{
Other content: \code{\link{alert}},
  \code{\link{blockquote}}, \code{\link{card}},
  \code{\link{collapsiblePane}}, \code{\link{d1}},
  \code{\link{img}}, \code{\link{jumbotron}},
  \code{\link{modal}}, \code{\link{navContent}},
  \code{\link{popover}}, \code{\link{pre}},
  \code{\link{progressOutlet}}, \code{\link{toast}},
  \code{\link{tooltip}}
}
\concept{content}
