% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown.R
\name{dropdownInput}
\alias{dropdownInput}
\alias{disableDropdownInput}
\alias{enableDropdownInput}
\title{Dropdown input}
\usage{
dropdownInput(id, label, items, values = items, disabled = NULL,
  dividers = NULL, context = "secondary", dropup = FALSE, ...)

disableDropdownInput(id, disabled = NULL,
  session = getDefaultReactiveDomain())

enableDropdownInput(id, enabled = NULL,
  session = getDefaultReactiveDomain())
}
\arguments{
\item{items}{A character vector specifying the labels of the dropdown menu
items.}

\item{values}{A character vector, list of character strings, vector of values
to coerce to character strings, or list of values to coerce to character
strings specifying the values of the dropdown input's choices, defaults to
\code{items}.}

\item{disabled, enabled}{One or more of \code{values} indicating which dropdown
menu items to disable or enable, defaults to \code{NULL}. If \code{NULL} then
\code{disableDropdownInput} and \code{enableDropdownInput} will disable or enable all
the dropdown input's items, respectively.}

\item{dividers}{One or more of \code{values} indicating which dropdown menu items
are the start of a new section, defaults to \code{NULL}. Divider lines will be
placed above the indicated values separating the dropdown menu items into
sections.}

\item{context}{One of \code{"primary"}, \code{"secondary"}, \code{"success"}, \code{"info"},
\code{"warning"}, \code{"danger"}, \code{"light"}, \code{"dark"}, or \code{"link"}, specifying the
visual context of the dropdown input, defaults to \code{"secondary"}.}

\item{dropup}{If \code{TRUE}, the dropdown menu opens upwards instead of
downwards, defaults to \code{FALSE}.}

\item{...}{Additional named arguments passed as HTML attributes to the parent
element.}
}
\description{
Dropdown inputs, or dropdown menus, function similar to a set of buttons.
The initial click to open a dropdown menu does not trigger a reactive event,
but a click on one of the dropdown items triggers a reactive event. The
reactive value of a dropdown input is the value of the most recently clicked
dropdown menu item. Disabling a dropdown menu triggers a reactive event and
resets the reactive value to \code{NULL}.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          dropdownInput(
            id = "dropdown",
            label = "A dropdown"
            items = paste("Action", 1:5),
            dividers = "Action 4"
          )
        ),
        col(
          display4(
            textOutput("value")
          )
        )
      )
    ),
    server = function(input, output) {
      output$value <- renderText({
        input$dropdown
      })
    }
  )
}

if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          dropdownInput(
            id = "actions",
            label = "Actions",
            items = c("disable", "enable", "disable some", "enable some"),
            dividers = "disable some"
          )
        ),
        col(
          dropdownInput(
            id = "dropdown",
            label = "Other actions",
            items = paste("Action", 1:5),
            disabled = "Action 3"
          )
        ),
        col(
          display4(
            textOutput("value")
          )
        )
      )
    ),
    server = function(input, output) {
      observeEvent(input$actions, {
        if (input$actions == "disable") {
          disableDropdownInput("dropdown")
        } else if (input$actions == "enable") {
          enableDropdownInput("dropdown")
        } else if (input$actions == "disable some") {
          disableDropdownInput("dropdown", paste("Action", c(2, 3, 5)))
        } else if (input$actions == "enable some") {
          enableDropdownInput("dropdown", paste("Action", c(2, 3, 5)))
        }
      })

      output$value <- renderText({
        input$dropdown
      })
    }
  )
}

}
