% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dull.R
\docType{package}
\name{dull}
\alias{dull}
\alias{dull-package}
\alias{dull-package}
\title{A new approach to shiny applications}
\description{
More freedom to the programmer, more to come.
}
\section{Q and A}{

}

\author{
\strong{Maintainer}: Nathan Teetor \email{nathanteetor@gmail.com}

}
