% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{fieldset}
\alias{fieldset}
\title{Visually group and label inputs}
\usage{
inputs$fieldset(legend = NULL, ...)
}
\arguments{
\item{legend}{A label for the associated inputs or a custom tag, defaults to
\code{NULL}.}

\item{disabeld}{If \code{TRUE}, all inputs within the fieldset are rendered in a
disabled state, defaults to \code{FALSE}.}

\item{...}{Inputs to group or additional named arguments passed as HTML
attributes to the parent \code{<fieldset>} element.}
}
\description{
Use \code{input$fieldset} to associate inputs. Good for screen readers and other
assitive technologies.
}
