% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textual.R
\name{fieldset}
\alias{fieldset}
\title{Group and label multiple inputs}
\usage{
fieldset(legend, ...)
}
\arguments{
\item{legend}{A character string specifying the fieldset's legend.}

\item{...}{Any number of inputs to group or named arguments passed as HTML
attributes to the parent element.}
}
\description{
Use \code{fieldset} to associate and label inputs. Good for screen readers and
other assitive technologies.
}
\examples{

}
\seealso{
Other layout: \code{\link{column}},
  \code{\link{formGroup}}, \code{\link{navbar}},
  \code{\link{tabTabs}}
}
\concept{layout}
