% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{fileInput}
\alias{fileInput}
\title{Upload user files}
\usage{
fileInput(id, placeholder = "Choose file", left = NULL, right = "Browse",
  ..., multiple = TRUE, accept = NULL)
}
\arguments{
\item{id}{A character string specifying the HTML id of the file input.}

\item{placeholder}{A character string specifying the text inside the file
input, defaults to \code{"Choose file"}.}

\item{left, right}{A character string or button element placed prepended or
appended respectively to the file input. For more information refer to
\code{\link[=groupInput]{groupInput()}}.

Clicking on an element specified by \code{right} also opens the file input
dialog.}

\item{...}{Additional named arguments passed on as HTML attributes to the
parent element.}

\item{multiple}{One of \code{TRUE} or \code{FALSE} specifying whether or not the user
can upload multiple files at once, defaults to \code{TRUE}.}

\item{accept}{A character vector of possible MIME types or file extensions,
defaults to \code{NULL}, in which case any file type may be selected.}
}
\description{
Upload files to the server.
}
\details{
Be careful when adjusting the right or left margin of a file input. In the
current version of Bootstrap file inputs can be pushed off the side of a
page.
}
\examples{

if (interactive()) {
  shinyApp(
    ui = container(
      fileInput("upload") \%>\%
        margin(0, "auto", 0, "auto")
    ),
    server = function(input, output) {
      observe({
        req(input$upload)
        print(input$upload)
      })
    }
  )
}

if (interactive()) {
  shinyApp(
    ui = container(
      fileInput(
        id = "upload",
        left = buttonInput(NULL, "Upload") \%>\%
          background("white") \%>\%
          border("green", -1)
      ) \%>\%
        margin("auto", 0, "auto", 0)
    ),
    server = function(input, output) {
      observe({
        req(input$upload)

        for (path in input$upload$datapath) {
          cat(readLines(path), "\\n")
        }
      })
    }
  )
}

if (interactive()) {
  shinyApp(
    ui = container(
      fileInput("upload1") \%>\%
        margin(3),
      fileInput("upload2") \%>\%
        margin(3)
    ),
    server = function(input, output) {
      observeEvent(input$upload1, {
        cat("Upload 1:", input$upload1$name, "\\n")
      })

      observeEvent(input$upload2, {
        cat("Upload 2:", input$upload2$name, "\\n")
      })
    }
  )
}

}
