% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-file.R
\name{fileInput}
\alias{fileInput}
\title{File inputs}
\usage{
fileInput(
  ...,
  id,
  placeholder = "Choose file",
  browse = "Browse",
  multiple = TRUE,
  accept = NULL
)
}
\arguments{
\item{placeholder}{A character string specifying the text inside the file
input, defaults to \code{"Choose file"}.}

\item{browse}{A character string specifying the label of file input, defaults
to \code{"Browse"}.}

\item{multiple}{One of \code{TRUE} or \code{FALSE} specifying whether or not the user
can upload multiple files at once, defaults to \code{TRUE}.}

\item{accept}{A character vector of possible MIME types or file extensions,
defaults to \code{NULL}, in which case any file type may be selected.}
}
\description{
Upload files to the server.
}
\seealso{
Other inputs: 
\code{\link{checkbarInput}()},
\code{\link{input_button}()},
\code{\link{input_checkbox}()},
\code{\link{input_checkbutton}()},
\code{\link{input_chip}()},
\code{\link{input_form}()},
\code{\link{listGroupInput}()},
\code{\link{menuInput}()},
\code{\link{radioInput}()},
\code{\link{radiobarInput}()},
\code{\link{rangeInput}()},
\code{\link{selectInput}()},
\code{\link{textInput}()}
}
\concept{inputs}
