% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-file.R
\name{fileInput}
\alias{fileInput}
\title{File inputs}
\usage{
fileInput(
  id,
  ...,
  placeholder = "Choose file",
  browse = "Browse",
  multiple = TRUE,
  accept = NULL
)
}
\arguments{
\item{id}{A character string. The id of the reactive input.}

\item{...}{Optional named arguments specifying HTML attributes for the input
element.}

\item{placeholder}{A character string specifying the text inside the file
input, defaults to \code{"Choose file"}.}

\item{browse}{A character string specifying the label of file input, defaults
to \code{"Browse"}.}

\item{multiple}{One of \code{TRUE} or \code{FALSE} specifying whether or not the user
can upload multiple files at once, defaults to \code{TRUE}.}

\item{accept}{A character vector of possible MIME types or file extensions,
defaults to \code{NULL}, in which case any file type may be selected.}
}
\description{
Upload files to the server.
}
\seealso{
Other inputs: 
\code{\link{input_button}()},
\code{\link{input_checkbox}()},
\code{\link{input_checkbox_group}()},
\code{\link{input_form}()},
\code{\link{input_list_group}()},
\code{\link{input_menu}()},
\code{\link{input_multi_select}()},
\code{\link{input_numeric}()},
\code{\link{input_radio_group}()},
\code{\link{input_range}()},
\code{\link{input_select}()},
\code{\link{input_text}()},
\code{\link{input_text_group}()}
}
\concept{inputs}
