% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{fileInput}
\alias{fileInput}
\title{Upload user files}
\usage{
fileInput(id, placeholder = "Choose file", left = NULL,
  right = "Browse", ..., multiple = TRUE, accept = NULL)
}
\arguments{
\item{id}{A character string specifying the HTML id of the file input.}

\item{placeholder}{A character string specifying the text inside the file
input, defaults to \code{"Choose file"}.}

\item{left, right}{A character string or button element placed prepended or
appended respectively to the file input. For more information refer to
\code{\link[=groupInput]{groupInput()}}.

Clicking on an element specified by \code{right} also opens the file input
dialog.}

\item{...}{Additional named arguments passed on as HTML attributes to the
parent element.}

\item{multiple}{One of \code{TRUE} or \code{FALSE} specifying whether or not the user
can upload multiple files at once, defaults to \code{TRUE}.}

\item{accept}{A character vector of possible MIME types or file extensions,
defaults to \code{NULL}, in which case any file type may be selected.}
}
\description{
Upload files to the server.
}
\details{
Be careful when adjusting the right or left margin of a file input. In the
current version of Bootstrap file inputs can be pushed off the side of a
page.
}
\section{Uploading a file}{
\preformatted{shinyApp(
  ui = container(
    fileInput("upload") \%>\%
      margin(0, "auto", 0, "auto")
  ),
  server = function(input, output) {
    observe({
      req(input$upload)

      print(input$upload)
    })
  }
)
}
}

\examples{

## Standard file input

fileInput(id = NULL)

## Adding another button

fileInput(
  id = NULL,
  left = buttonInput(NULL, "Upload") \%>\%
    background("transparent") \%>\%
    border("green")
) \%>\%
  margin("auto", 0, "auto", 0)

}
\seealso{
Other inputs: \code{\link{addressInput}},
  \code{\link{buttonGroupInput}},
  \code{\link{buttonInput}}, \code{\link{checkboxInput}},
  \code{\link{dateInput}}, \code{\link{formInput}},
  \code{\link{groupInput}}, \code{\link{loginInput}},
  \code{\link{radioInput}}, \code{\link{rangeInput}},
  \code{\link{selectInput}}, \code{\link{textInput}}
}
\concept{inputs}
