% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{float}
\alias{float}
\title{Responsive floated elements}
\usage{
utils$float(side, viewport = NULL)
}
\arguments{
\item{side}{One of \code{"left"}, \code{"right"}, or \code{"none"}, Which side of the parent
element to float the text on or if \code{"none"} the element cannot be floated.}
}
\description{
These utility functions float an element to the left or right of its parent.
Alternatively, one can specify \code{"none"} to disable an element's float.
}
\seealso{
Other utilities: \code{\link{colors}},
  \code{\link{flexbox}}, \code{\link{hidden}},
  \code{\link{position}}, \code{\link{sizing}},
  \code{\link{spacing}}, \code{\link{viewport}}
}
