% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design-float.R, R/style-float.R
\name{float}
\alias{float}
\title{Floats}
\usage{
float(tag, side)

float(tag, side)
}
\arguments{
\item{tag}{A tag element.}

\item{side}{A \link{responsive} argument. One of \code{"left"} or \code{"right"} specifying
the side to float the element.}
}
\description{
Use \code{float()} to float an element to the left or right side of its parent
element. A newspaper layout is a classic example where an image is floated
with text wrapped around.

Use \code{float()} to float an element to the left or right side of its parent
element. A newspaper layout is a classic example where an image is floated
with text wrapped around.
}
\examples{

### Float an alert

div(
  .style \%>\%
    padding(2),
  alert("This just in!") \%>\%
    margin(3) \%>\%
    background("warning") \%>\%
    float("left"),
  p("Fusce commodo. Nullam tempus. Nunc rutrum turpis sed pede.",
    "Phasellus lacus.  Cras placerat accumsan nulla.",
    "Fusce sagittis, libero non molestie mollis, ",
    "magna orci ultrices dolor, at vulputate neque nulla lacinia eros."),
  p("Nulla facilisis, risus a rhoncus fermentum, tellus tellus",
    "lacinia purus, et dictum nunc justo sit amet elit."),
  p("Proin neque massa, cursus ut, gravida ut, lobortis eget, lacus.",
    "Aliquam posuere.",
    "Sed id ligula quis est convallis tempor."),
  p("Fusce dapibus, tellus ac cursus commodo, ",
    "tortor mauris condimentum nibh, ut fermentum massa justo sit",
    "amet risus.")
)


### Float an alert

div(
  .style \%>\%
    padding(2),
  alert("This just in!") \%>\%
    margin(3) \%>\%
    background("warning") \%>\%
    float("left"),
  p("Fusce commodo. Nullam tempus. Nunc rutrum turpis sed pede.",
    "Phasellus lacus.  Cras placerat accumsan nulla.",
    "Fusce sagittis, libero non molestie mollis, ",
    "magna orci ultrices dolor, at vulputate neque nulla lacinia eros."),
  p("Nulla facilisis, risus a rhoncus fermentum, tellus tellus",
    "lacinia purus, et dictum nunc justo sit amet elit."),
  p("Proin neque massa, cursus ut, gravida ut, lobortis eget, lacus.",
    "Aliquam posuere.",
    "Sed id ligula quis est convallis tempor."),
  p("Fusce dapibus, tellus ac cursus commodo, ",
    "tortor mauris condimentum nibh, ut fermentum massa justo sit",
    "amet risus.")
)

}
\seealso{
Other design utilities: 
\code{\link{active}()},
\code{\link{affix}()},
\code{\link{background}()},
\code{\link{border}()},
\code{\link{display}()},
\code{\link{flex}()},
\code{\link{font}()},
\code{\link{height}()},
\code{\link{margin}()},
\code{\link{padding}()},
\code{\link{scroll}()},
\code{\link{shadow}()},
\code{\link{width}()}

Other design utilities: 
\code{\link{active}()},
\code{\link{affix}()},
\code{\link{background}()},
\code{\link{border}()},
\code{\link{display}()},
\code{\link{flex}()},
\code{\link{font}()},
\code{\link{height}()},
\code{\link{margin}()},
\code{\link{padding}()},
\code{\link{scroll}()},
\code{\link{shadow}()},
\code{\link{width}()}
}
\concept{design utilities}
