% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{float}
\alias{float}
\title{Tag element float}
\usage{
float(.tag, side)
}
\arguments{
\item{.tag}{A tag element.}

\item{side}{A \link{responsive} argument. One of \code{"left"} or \code{"right"} specifying
the side to float the element.}
}
\description{
Use \code{float()} to float an element to the left or right side of its parent
element. A classic example using floats is a newspaper layout where text is
wrapped around a picture.
}
\section{Newspaper layout}{
\preformatted{div(
  icon("table-tennis") \%>\%
    font(size = "5x") \%>\%
    padding(2) \%>\%
    float("left"),
  p(
    "Fusce commodo. Nullam tempus. Nunc rutrum turpis sed pede.",
    "Phasellus lacus.  Cras placerat accumsan nulla.",
    "Fusce sagittis, libero non molestie mollis, ",
    "magna orci ultrices dolor, at vulputate neque nulla lacinia eros."
  ),
  p(
    "Nulla facilisis, risus a rhoncus fermentum, tellus tellus",
    "lacinia purus, et dictum nunc justo sit amet elit."
  ),
  p(
    "Proin neque massa, cursus ut, gravida ut, lobortis eget, lacus.",
    "Aliquam posuere.",
    "Sed id ligula quis est convallis tempor."
  )
)
}
}

\examples{

}
\seealso{
Other utilities: \code{\link{affix}},
  \code{\link{background}}, \code{\link{border}},
  \code{\link{collapse}}, \code{\link{display}},
  \code{\link{flex}}, \code{\link{font}},
  \code{\link{markInvalid}}, \code{\link{padding}},
  \code{\link{rounded}}, \code{\link{scroll}},
  \code{\link{sendModal}}, \code{\link{shadow}},
  \code{\link{showPopover}}, \code{\link{tooltip}},
  \code{\link{updateChoices}}, \code{\link{width}}
}
\concept{utilities}
