% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{float}
\alias{float}
\title{Float an element}
\usage{
float(tag, default = NULL, sm = NULL, md = NULL, lg = NULL, xl = NULL)
}
\arguments{
\item{tag}{A tag element.}

\item{default}{One of \code{"left"}, \code{"right"}, or \code{"none"} specifying the default
float of the element.}

\item{sm}{Like \code{default}, but the float is applied once the viewport is 576
pixels wide, think phone in landscape mode.}

\item{md}{Like \code{default}, but the float is applied once the viewport is 768
pixels wide, think tablets.}

\item{lg}{Like \code{default}, but the float is applied once the viewport is 992
pixels wide, think desktop.}

\item{xl}{Like \code{default}, but the float is applied once the viewport is 1200
pixels wide, think large desktop.}
}
\description{
The \code{float} utility function applies Bootstrap float classes to a tag
element. These classes cause a tag element to float to the left or right
in its parent element. Alternatively, specify \code{"none"} to remove the
element's float. The float utilities are viewport responsive.
}
\examples{

}
\seealso{
Other utilities: \code{\link{affix}}, \code{\link{align}},
  \code{\link{background}}, \code{\link{display}},
  \code{\link{font}}, \code{\link{padding}},
  \code{\link{width}}
}
