% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{font}
\alias{font}
\title{Font classes}
\usage{
font(.tag, color = NULL, size = NULL, weight = NULL, align = NULL)
}
\arguments{
\item{color}{One of \code{"red"}, \code{"purple"}, \code{"indigo"}, \code{"blue"}, \code{"cyan"},
\code{"teal"}, \code{"green"}, \code{"yellow"}, \code{"amber"}, \code{"orange"}, \code{"grey"}, or
\code{"white"} specifying the text color, defaults to \code{NULL}, in which case text
color is unmodified.}

\item{size}{One of \code{"xs"}, \code{"sm"}, \code{"base"}, \code{"lg"}, or \code{"xl"} specifying a
font size relative to the base font size of the page, defaults to \code{"base"}.}

\item{weight}{One of \code{"bold"}, \code{"normal"}, \code{"light"}, \code{"italic"}, or
\code{"monospace"} specifying the font weight of the element's text, defaults to
\code{"normal"}.}

\item{align}{A \link{responsive} argument. One of \code{"left"}, \code{"center"}, \code{"right"},
or \code{"justify"}.}

\item{transform}{One of \code{"upper"}, \code{"lower"}, or \code{"capitalize"} specifying a
text transformation, defaults to \code{NULL}, in which case a transformation is
not applied.}
}
\description{
The \code{font()} utility may be used to change the color, size, or weight of a
tag element's text font. Font size's are changed relative to the base font
size of the web page.
}
\details{
The possible font colors are,
}
\examples{

### Possible colors

colors <- c(
  "red", "purple", "indigo", "blue", "cyan", "teal", "green",
  "yellow", "amber", "orange", "body", "grey", "white"
)

div(
  lapply(
    head(colors, -1),
    font,
    .tag = div("Pellentesque tristique imperdiet tortor.") \%>\%
      padding(5)
  )
) \%>\%
  display("flex") \%>\%
  flex(wrap = TRUE)

}
\seealso{
Other design: \code{\link{affix}},
  \code{\link{background}}, \code{\link{border}},
  \code{\link{display}}, \code{\link{float}},
  \code{\link{height}}, \code{\link{padding}},
  \code{\link{rounded}}, \code{\link{scroll}},
  \code{\link{shadow}}, \code{\link{width}}
}
\concept{design}
