% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{font}
\alias{font}
\title{Text font}
\usage{
font(tag, size = NULL, weight = NULL)
}
\arguments{
\item{tag}{A tag element.}

\item{size}{One of \code{"2x"}, \code{"3x"}, ..., or \code{"10x"} specifying a factor to
increase a tag element's font size by (e.g. \code{"2x"} is double the base font
size), defaults to \code{NULL}, in which case the font size is unchanged.}

\item{weight}{One of \code{"bold"}, \code{"normal"}, \code{"light"}, \code{"italic"}, or
\code{"monospace"} specifying the font weight of the element's text, defaults to
\code{NULL}, in which case the font is unchanged.}
}
\description{
The \code{font} utility changes the weight and size of an element's text font.
Bold fonts are darker and heavier whereas light fonts are thinner. Font
size's are changed relative to the current font size.
}
\examples{

span("This and other news") \%>\%
  font(weight = "light")

icon("anchor") \%>\%
  font("5x")

p("Ipsum Consectetur Nibh Bibendum Ullamcorper") \%>\%
  font("2x", "monospace") \%>\%
  font(weight = "italic")

}
\seealso{
Other utilities: \code{\link{affix}},
  \code{\link{alignment}}, \code{\link{background}},
  \code{\link{display}}, \code{\link{float}},
  \code{\link{padding}}, \code{\link{rounded}},
  \code{\link{width}}
}
\concept{utilities}
