% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{font}
\alias{font}
\title{Text font and styles}
\usage{
font(tag, weight = NULL, style = NULL)
}
\arguments{
\item{tag}{A tag element.}

\item{weight}{One of \code{"bold"}, \code{"normal"}, or \code{"light"} specifying the font
weight of an element's text. Defaults to \code{NULL}, element font weight is
left as is.}

\item{style}{One of \code{"bold"} or \code{"italic"} specifying the font style of an
element. Defaults to \code{NULL}, element font style is left as is. Specifying
\code{"bold"} is equivalent to \code{weight = "bold"}.}
}
\description{
The \code{font} utility function applies Bootstrap classes to a tag element.
These classes apply CSS to change the font weight of an element's text or the
font style of an element text.
}
\examples{
tags$span("This and other news") \%>\%
  font(weight = "light")

}
\seealso{
Other utilities: \code{\link{affix}}, \code{\link{align}},
  \code{\link{background}}, \code{\link{border}},
  \code{\link{display}}, \code{\link{float}},
  \code{\link{padding}}, \code{\link{text}},
  \code{\link{width}}
}
