% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design-font.R
\name{font}
\alias{font}
\title{Font color, size, weight}
\usage{
font(x, color = NULL, size = NULL, weight = NULL, case = NULL, align = NULL)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{color}{One of \code{"primary"}, \code{"secondary"}, \code{"success"}, \code{"info"}, \code{"warning"}, \code{"danger"}, \code{"light"} or \code{"dark"} specifying the text color of the tag element, defaults to \code{NULL}.}

\item{size}{Deprecated, in future versions of bootstrap resonsive font sizing
will be enabled by default,
\url{https://github.com/twbs/bootstrap/pull/29152}.

One of \code{"xs"}, \code{"sm"}, \code{"base"}, \code{"lg"}, \code{"xl"} specifying a font size
relative to the default base page font size, defaults to \code{NULL}.}

\item{weight}{One of \code{"bold"}, \code{"bolder"}, \code{"normal"}, \code{"lighter"}, or
\code{"light"} specifying the font weight of the element's text, defaults to
\code{NULL}. \code{"bolder"} and \code{"lighter"} change the weight relative to the
current font weight of the element.}

\item{case}{One of \code{"upper"}, \code{"lower"}, or \code{"title"} specifying a
transformation of the tag element's text, default to \code{NULL}.}

\item{align}{A \link{responsive} argument. One of \code{"left"}, \code{"center"}, \code{"right"},
or \code{"justify"}, specifying the alignment of the tag element's text,
defaults to \code{NULL}.}
}
\description{
The \code{font()} utility modifies the color, size, weight, case, or alignment of
a tag element's text. All arguments default to \code{NULL}, in which case they are
ignored.  For example, \code{font(.., size = "lg")} increases font size without
affecting color, weight, case, or alignment.
}
\details{
\subsection{Getting started}{

The font utility is used to change text colors, font weights, and more.
}

\subsection{Changing text color}{\if{html}{\out{<div class="r">}}\preformatted{card(
  .style \%>\%
    font("primary") \%>\%
    border("primary"),
  header = h3("Header"),
  p("Nullam tristique diam non turpis.",
    "Pellentesque dapibus suscipit ligula.",
    "Nullam eu ante vel est convallis dignissim."),
  p("Aliquam posuere.")
)
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve-->
}}
\if{html}{\out{
<div class="r-output">
}}
\if{html}{\out{
<div class="card text-primary border border-primary">
}}
\if{html}{\out{
<div class="card-header">
}}
\if{html}{\out{
<h3>
}}


Header\if{html}{\out{
</h3>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="card-body">
}}
\if{html}{\out{
<p class="card-text">
}}


Nullam tristique diam non turpis. Pellentesque dapibus suscipit ligula.
Nullam eu ante vel est convallis dignissim.\if{html}{\out{
</p>
}}
\if{html}{\out{
<p class="card-text">
}}


Aliquam posuere.\if{html}{\out{
</p>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<!--/html_preserve-->
}}

}

\subsection{Changing font weight}{\if{html}{\out{<div class="r">}}\preformatted{p(
  .style \%>\%
    font(weight = "bold"),
  "Curabitur lacinia pulvinar nibh."
)
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve-->
}}
\if{html}{\out{
<div class="r-output">
}}
\if{html}{\out{
<p class="font-weight-bold">
}}


Curabitur lacinia pulvinar nibh.\if{html}{\out{
</p>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<!--/html_preserve-->
}}
\if{html}{\out{<div class="r">}}\preformatted{p(
  .style \%>\%
    font(weight = "light"),
  "Proin quam nisl, tincidunt et."
)
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve-->
}}
\if{html}{\out{
<div class="r-output">
}}
\if{html}{\out{
<p class="font-weight-light">
}}


Proin quam nisl, tincidunt et.\if{html}{\out{
</p>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<!--/html_preserve-->
}}

}
}
\seealso{
Other design utilities: 
\code{\link{active}()},
\code{\link{affix}()},
\code{\link{background}()},
\code{\link{border}()},
\code{\link{display}()},
\code{\link{flex}()},
\code{\link{float}()},
\code{\link{height}()},
\code{\link{margin}()},
\code{\link{padding}()},
\code{\link{scroll}()},
\code{\link{shadow}()},
\code{\link{width}()}
}
\concept{design utilities}
