% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{fontAwesome}
\alias{fontAwesome}
\title{Include icons in applications}
\usage{
fontAwesome(name, ...)
}
\arguments{
\item{name}{A character string specifying the icon name, leading prefixes
unique to the icon set are not required.}

\item{...}{Additional named arguments passed as HTML attributes to the parent
element.}
}
\description{
Currently only FontAwesome icons are available.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      fontAwesome("thumbs-up"),
      fontAwesome("bathtub"),
      fontAwesome("cog"),
      fontAwesome("clock-o"),
      fontAwesome("microphone"),
      fontAwesome("print")
      # and more!
    ),
    server = function(input, output) {

    }
  )
}

}
\seealso{
For a complete list of possible FontAwesome icons please refer to
\url{http://fontawesome.io/icons/}.
}
