% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{formGroup}
\alias{formGroup}
\alias{formRow}
\title{Add labels, help text, and formatting to inputs}
\usage{
formGroup(label, input, help = NULL, ..., default = NULL, sm = NULL,
  md = NULL, lg = NULL, xl = NULL)

formRow(...)
}
\arguments{
\item{label}{A character string specifying a label for the input or \code{NULL}
in which case a label is not added.}

\item{input}{A tag element specifying the input to label.}

\item{help}{A character string specifying help text for the input, defaults
to \code{NULL}, in which case help text is not added.}

\item{...}{For \strong{formGroup}, additional named arguments passed as HTML
attributes to the parent element.

For \strong{formRow}, any number of \code{formGroup}s or additional named arguments
passed as HTML attributes to the parent element.}

\item{default, sm, md, lg, xl}{These arguments are taken directly from \code{\link[=col]{col()}}
because \code{formGroup}s can replicate column behaviour in order to build
highly flexible forms. To best understand these arguments please refer to
the \code{col} help page.}
}
\description{
Form groups are a way of labelling an input. Form rows are similar to
\code{\link[=row]{row()}}s, but include additional styles intended for forms. The flexibility
provided by form rows and groups means you can confidently develop shiny
applications for devices and screens of varying sizes.
}
\examples{

# to see this example in action adjust your browser window
# from large to small, notice how the form elements expand?
if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          md = 6,
          card(
            formRow(
              formGroup(
                md = 6,
                label = "Email",
                emailInput(
                  id = "email",
                  placeholder = "e@mail.com"
                )
              ),
              formGroup(
                md = 6,
                label = "Password",
                passwordInput(
                  id = "password",
                  placeholder = "123456"
                ),
                help = "Please consider something
                  better than 123456"
              )
            ),
            formGroup(
              label = "Username",
              groupInput(
                id = "username",
                left = "@"
              )
            ),
            buttonInput(
              id = "go",
              label = "Go!"
            ) \%>\%
              background("blue")
          ) \%>\%
            margins(3) \%>\%
            background("grey", +2)
        ) \%>\%
          margins("auto")
      )
    ),
    server = function(input, output) {
    }
  )
}

}
