% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{formInput}
\alias{formInput}
\title{Form and form group inputs}
\usage{
formInput(id, ...)
}
\arguments{
\item{id}{A character string specifying an id for the form or form group,
defaults to \code{NULL}. If specified a form or form group becomes a reactive
input. A reactive form group will modify its parent form's input.}

\item{...}{Any number of inputs, other elements, or named arguments passed
as HTML attributes to the parent element.}
}
\description{
Form inputs are a new reactive input. A form input's reactive value is a list
of all the reactive inputs within it. Additionally, form inputs suppress
the reactive nature of
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          formInput(
            id = "form",
            emailInput(
              id = "email",
              label = "Email",
              placeholder = "Email"
            ),
            passwordInput(
              id = "password",
              label = "Password",
              placeholder = "Password"
            ),
            fieldset(
              legend = "Radios",
              radioInput(
                id = "options",
                labels = c(
                  "Option one",
                  "Option two",
                  "Option three"
                )
              )
            ),
            checkboxInput(
              id = "checkbox",
              title = "Checkbox",
              label = "Check me out"
            )
          )
        ),
        col(
          verbatimTextOutput("value")
        )
      )
    ),
    server = function(input, output) {
      output$value <- renderPrint({
        str(input$form)
      })
    }
  )
}

if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          formInput(
            id = "real",
            tags$label("Real name"),
            textInput(
              id = "first",
              label = "First name"
            ),
            textInput(
              id = "last",
              label = "Last name"
            )
          )
        ),
        col(
          verbatimTextOutput("realvalue")
        ),
        col(
          formInput(
            id = "nickname",
            tags$label("Nickname"),
            textInput(
              id = "first",
              label = "First name"
            ),
            textInput(
              id = "last",
              label = "Last name"
            )
          )
        ),
        col(
          verbatimTextOutput("nickvalue")
        )
      )
    ),
    server = function(input, output) {
      output$realvalue <- renderPrint({
        str(input$real)
      })

      output$nickvalue <- renderPrint({
        str(input$nickname)
      })
    }
  )
}


}
\seealso{
For more information on forms and form groups please refer to the online
bootstrap \href{https://v4-alpha.getbootstrap.com/components/forms/}{referencepage}.

Other inputs: \code{\link{buttonInput}},
  \code{\link{checkboxInput}}, \code{\link{groupInput}},
  \code{\link{radioInput}}, \code{\link{selectInput}},
  \code{\link{textInput}}
}
