% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\docType{data}
\name{forms}
\alias{forms}
\alias{form}
\alias{inline}
\alias{updateFormGroup}
\title{Forms}
\usage{
forms$form(...)

forms$inline(...)

forms$group(..., state = NULL, fieldset = FALSE, label = NULL)

updateFormGroup(id, state, session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{A character string specifying the HTML id of a button to update.}

\item{state}{One of \code{"success"}, \code{"warning"}, or \code{"danger"}, specifying the
state of the form group, defaults to \code{NULL} in \code{forms$group}, is required
for \code{updateFormGroup}.}

\item{session}{The \code{session} object passed to a shiny server function,
defaults to \code{\link[shiny:getDefaultReactiveDomain]{shiny::getDefaultReactiveDomain()}}.}

\item{fieldset}{If \code{TRUE}, the form group is rendered inside a \code{<fieldset>}
instead of a \code{<div>}, defaults to \code{FALSE}. When building a form group
multiple inputs it is recommended to set \code{fieldset} to \code{TRUE} so that
\code{label} properly renders as a \code{<legend>}.}

\item{label}{A character vector specifying a label for the form group, if
\code{fieldset} is \code{TRUE}, the label is rendered inside a \code{<legend>} element
instead of a \code{<label>} element, defaults to \code{NULL}.}

\item{...}{Input elements, labels, or other tag elements or named arguments
passed as HTML attributes to the parent element.}
}
\description{
A reactive \code{forms$form} is a new reactive input. A form's reactive value is a
list of all the reactive inputs within it. Reactive form groups within a
form, those form groups \emph{which have an ID}, cause the form's value to take on
a nested structure. See details for more information.

\code{forms$inline} is an inline equivalent of \code{forms$form}. Inline forms are
handy for rendering small, compact forms.

\code{forms$group} helps visually organization and structure a form. On the
server side, a \code{forms$group} with an \code{id} argument becomes a reactive list
value comprised of its underlying reactive inputs.
}
\details{
** The value of a form **

Below is a small sample form,\preformatted{forms$form(
  id = "register",
  forms$group(
    id = "names",
    inputs$text(
      id = "first",
      label = "First name"
    ),
    inputs$text(
      id = "last",
      label = "Last name"
    )
  )
)
}

The initial value of this form, as would be printed from a shiny server
function.\preformatted{input
input$register
input$register$names
input$register$names$first
NULL

input$register$names$last
NULL
}

Here is a longer form based on the one above,\preformatted{forms$form(
  id = "register",
  forms$group(
    id = "names",
    inputs$text(
      id = "first",
      label = "First name"
    ),
    inputs$text(
      id = "last",
      label = "Last name"
    )
  ),
  forms$group(
    id = "contact",
    inputs$telephone(
      id = "mobile",
      label = "Cell phone"
    ),
    inputs$telephone(
      id = "home",
      label = "Home phone"
    )
  )
)
}

Below is the initial value of this larger form,\preformatted{input
input$register
input$register$names
input$register$names$first
NULL

input$register$names$last
NULL

input$register$contact
input$register$contact$mobile
NULL

input$register$contact$home
NULL
}

** Inputs within a form **

Any reactive inputs within a \strong{reactive} form, a form with an \code{id}, will not
trigger an obesrver or reactive expression nor do they take a value other
than \code{NULL}.
}
\examples{
if (interactive()) {
  library(shiny)

  shinyApp(
    ui = container(
      forms$form(
        id = "user",
        submit = TRUE,
        forms$group(
          id = "names",
          label = "Form group",
          inputs$text(id = "first", placeholder = "first name"),
          inputs$text(id = "last", placeholder = "last name")
        )
      )
    ),
    server = function(input, output) {
      observe({
        print(input$user)
      })

      # Note the observer triggers on application startup,
      # value is NULL, and then does not trigger again.
      observe({
        print(inputs$first)
      })
    }
  )
}

}
\seealso{
For more information on forms and form groups please refer to the online
bootstrap
\href{https://v4-alpha.getbootstrap.com/components/forms/}{reference page}.
}
\keyword{datasets}
