% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{icon}
\alias{icon}
\title{Icon elements}
\usage{
icon(name, ..., set = NULL)
}
\arguments{
\item{name}{A character string specifying the name of the icon.}

\item{...}{Additional named arguments passed as HTML attributes to the parent
element.}

\item{set}{A character string specifying the icon set to choose from, defaults
to \code{"NULL"}, in which case all icon sets are searched.}
}
\description{
Include an icon in your application. For now only Font Awesome icons are
included.
}
\examples{

if (interactive()) {
  shinyApp(
    ui = container(
      fluid = FALSE,
      selectInput(
        id = "name",
        choices = unique(icons$name)
      ) \%>\%
        margins(3),
      div(
        htmlOutput("icon")
      ) \%>\%
        margins(3) \%>\%
        display("flex") \%>\%
        direction("column") \%>\%
        items("center")
    ),
    server = function(input, output) {
      output$icon <- renderUI({
        icon(input$name) \%>\%
          font("8x")
      })
    }
  )
}

if (interactive()) {
  shinyApp(
    ui = container(
      lapply(
        unique(icons$set),
        function(s) {
          div(
            lapply(
              unique(icons[icons$set == s, ]$name),
              function(nm) {
                icon(nm, set = s) \%>\%
                  margins(2)
              }
            )
          ) \%>\%
            display("flex") \%>\%
            wrap("wrap")
        }
      )
    ),
    server = function(input, output) {

    }
  )
}

}
\seealso{
For more on Font Awesome check out \url{https://fontawesome.com}.
}
