% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{icon}
\alias{icon}
\title{Icon elements}
\usage{
icon(name, set = NULL, ...)
}
\arguments{
\item{name}{A character string specifying the name of the icon.}

\item{set}{A character string specifying the icon set to choose from,
defaults to \code{"NULL"}, in which case all icon sets are searched.

Possibles values include \code{"font awesome"}, \code{"material design"}, and
\code{"feather"}.}

\item{...}{Additional named arguments passed as HTML attributes to the parent
element.}
}
\description{
Include icons in your application UIs.
}
\examples{

### Font awesome

icon("clone")

### Material design

icon("mail", "material design")

### Feather

icon("mail", "feather")

}
\seealso{
Other content: \code{\link{alert}},
  \code{\link{blockquote}}, \code{\link{card}},
  \code{\link{d1}}, \code{\link{dropdown}},
  \code{\link{img}}, \code{\link{jumbotron}},
  \code{\link{modal}}, \code{\link{popover}},
  \code{\link{pre}}, \code{\link{progressOutlet}},
  \code{\link{tooltip}}
}
\concept{content}
