% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\docType{data}
\name{icons}
\alias{icons}
\title{Icons}
\usage{
icons$na(name, ...)
}
\arguments{
\item{name}{A character string specifying the icon name, leading prefixes
unique to the icon set are not required.}

\item{...}{Named arguments passed to the icon element as HTML attributes.}
}
\description{
Currently only FontAwesome icons.
}
\examples{

icons$fa("thumbs-up")

}
\keyword{datasets}
