% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{img}
\alias{img}
\alias{figure}
\alias{image}
\alias{figure}
\title{Responsive Images and Figures}
\usage{
img(..., fluid = TRUE, align = NULL, thumbnail = FALSE)

figure(..., image = NULL, caption = NULL, align = "left")
}
\arguments{
\item{fluid}{If \code{TRUE}, the image will scale with its parent element,
defaults to \code{TRUE}.}

\item{align}{For \code{img}, one of \code{"left"} or \code{"right"} specifying which side of
the parent element the image should align to, defaults to \code{NULL} in which
case no alignment is applied.

For \code{figure}, one of \code{"left"}, \code{"right"}, or \code{"center"} specifying which
side of the figure the figure caption is aligned to, defaults to \code{"left"}.}

\item{thumbnail}{If \code{TRUE}, the image is rendered with a small border,
defaults to \code{FALSE}.}

\item{image}{An \code{img} element, it is best to leave the image as fluid,
defaults to \code{NULL}.}

\item{caption}{A character vector specifying a caption for the image,
defaults to \code{NULL}.}
}
\description{
A small update to \code{<img>} and \code{<figure>}. Additional arguments has been added
to the \code{img} builder function. Figures have specific arguments for an \code{img}
child element and caption.
}
\examples{
if (interactive()) {
  library(shiny)

  shinyApp(
    ui = container(
      # A figure with a,
      #   - fluid image
      #   - left-aligned caption
      figure(
        image = img(src = "http://bit.ly/2qchbEB"),
        caption = "Stock cat photo."
      )
    ),
    server = function(input, output) {

    }
  )
}

}
\seealso{
Other content: \code{\link{blockquote}}
}
