% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{img}
\alias{img}
\alias{figure}
\title{Responsive images and figures}
\usage{
img(src, ...)

figure(image, caption = NULL, ...)
}
\arguments{
\item{src}{A character string specifying the source of the image.}

\item{...}{Additional tag elements or named arguments passed as HTML attributes
to the parent element.}

\item{image}{An \code{<img>} tag, typically a call to \code{img}.}

\item{caption}{A character string specifying the image caption, defaults to
\code{NULL}.}
}
\description{
A small update to \code{tags$img} and \code{tags$figure}. Create responsive images with
\code{img}. \code{figure} has specific arguments for an image child element and image
caption.
}
\seealso{
Other components: \code{\link{alert}}, \code{\link{badge}},
  \code{\link{blockquote}}, \code{\link{card}},
  \code{\link{collapsePane}}, \code{\link{d1}},
  \code{\link{dropdown}}, \code{\link{jumbotron}},
  \code{\link{modal}}, \code{\link{navContent}},
  \code{\link{popover}}, \code{\link{pre}},
  \code{\link{toast}}
}
\concept{components}
