% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-button.R
\name{input_button}
\alias{input_button}
\alias{update_button}
\title{Button inputs}
\usage{
input_button(id, text, ...)

update_button(id, text = NULL, disable = NULL, session = get_current_session())
}
\arguments{
\item{id}{A character string. The id of the reactive input.}

\item{text}{A character string. The text content of the button.}

\item{...}{Optional named arguments specifying HTML attributes for the input
element.}

\item{disable}{A boolean. The checkbox starts disabled if \code{TRUE}.}

\item{session}{A shiny session object.}
}
\value{
A \link[htmltools:builder]{htmltools::tag} object.
}
\description{
Button inputs are useful as triggers for reactive or observer expressions.
The reactive value of a button input begins as \code{NULL}, but subsequently is
the number of clicks.
}
\seealso{
Other inputs: 
\code{\link{fileInput}()},
\code{\link{input_checkbox}()},
\code{\link{input_checkbox_group}()},
\code{\link{input_chip}()},
\code{\link{input_form}()},
\code{\link{input_list_group}()},
\code{\link{input_menu}()},
\code{\link{input_radio_group}()},
\code{\link{input_range}()},
\code{\link{selectInput}()},
\code{\link{textInput}()}
}
\concept{inputs}
