% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-button.R
\name{input_button}
\alias{input_button}
\alias{input_link}
\title{Button and link inputs}
\usage{
input_button(id, label, ...)

input_link(id, label, stretch = FALSE, icon = NULL, ...)
}
\arguments{
\item{id}{A character string. specifying the id of the reactive input.}

\item{label}{A character string specifying the label text on the button or
link input.}

\item{...}{Additional named arguments passed as HTML attributes to the parent
element.}

\item{icon}{An icon, see \link[bsicons:bs_icon]{bsicons::bs_icon}.}
}
\description{
Button inputs are useful as triggers for reactive or observer expressions.
The reactive value of a button input begins as \code{NULL}, but subsequently is
the number of clicks.
}
\seealso{
Other inputs: 
\code{\link{checkbarInput}()},
\code{\link{fileInput}()},
\code{\link{input_checkbox}()},
\code{\link{input_checkbox_button}()},
\code{\link{input_chip}()},
\code{\link{input_form}()},
\code{\link{listGroupInput}()},
\code{\link{menuInput}()},
\code{\link{radioInput}()},
\code{\link{radiobarInput}()},
\code{\link{rangeInput}()},
\code{\link{selectInput}()},
\code{\link{textInput}()}
}
\concept{inputs}
