% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-button.R
\name{input_button}
\alias{input_button}
\alias{update_button}
\title{Button and link inputs}
\usage{
input_button(id, label, ...)

update_button(
  id,
  label = NULL,
  value = NULL,
  disable = NULL,
  session = get_current_session()
)
}
\arguments{
\item{id}{A character string. specifying the id of the reactive input.}

\item{label}{A character string specifying the label text on the button or
link input.}

\item{...}{Additional named arguments passed as HTML attributes to the parent
element.}

\item{icon}{An icon, see \link[bsicons:bs_icon]{bsicons::bs_icon}.}
}
\description{
Button inputs are useful as triggers for reactive or observer expressions.
The reactive value of a button input begins as \code{NULL}, but subsequently is
the number of clicks.
}
\seealso{
Other inputs: 
\code{\link{fileInput}()},
\code{\link{input_checkbox}()},
\code{\link{input_checkbox_group}()},
\code{\link{input_chip}()},
\code{\link{input_form}()},
\code{\link{input_list_group}()},
\code{\link{menuInput}()},
\code{\link{radioInput}()},
\code{\link{radiobarInput}()},
\code{\link{rangeInput}()},
\code{\link{selectInput}()},
\code{\link{textInput}()}
}
\concept{inputs}
