% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-checkbox.R
\name{input_checkbox}
\alias{input_checkbox}
\alias{update_checkbox}
\title{Checkbox input}
\usage{
input_checkbox(
  id,
  choice,
  ...,
  label = NULL,
  value = FALSE,
  disable = NULL,
  choice_placement = c("after", "before")
)

update_checkbox(
  id,
  choice = NULL,
  value = NULL,
  disable = NULL,
  session = get_current_session()
)
}
\arguments{
\item{id}{A character string. The id of the reactive input.}

\item{choice}{A character string. The text for the checkbox.}

\item{...}{Optional named arguments specifying HTML attributes for the input
element.}

\item{label}{A character string. The label of the input, rendered ahead
of the input. Floating labels are not supported.}

\item{value}{A boolean. The default value for the checkbox.}

\item{disable}{A boolean. The checkbox starts disabled if \code{TRUE}.}

\item{choice_placement}{A character string. The placement of the choice's
text relative to its checkbox.}

\item{session}{A shiny session object.}
}
\value{
A \link[htmltools:tag]{htmltools::tag} object.
}
\description{
A reactive checkbox input. When a checkbox input has no selected choices the
reactive value is \code{NULL}.
}
\details{
\subsection{Server value}{

A named logical vector.
}
}
\seealso{
Other inputs:
\code{\link[=floating_label]{floating_label()}},
\code{\link[=input_button]{input_button()}},
\code{\link[=input_checkbox_group]{input_checkbox_group()}},
\code{\link[=input_chip_group]{input_chip_group()}},
\code{\link[=input_form]{input_form()}},
\code{\link[=input_list_group]{input_list_group()}},
\code{\link[=input_menu]{input_menu()}},
\code{\link[=input_multi_select]{input_multi_select()}},
\code{\link[=input_numeric]{input_numeric()}},
\code{\link[=input_radio_group]{input_radio_group()}},
\code{\link[=input_range]{input_range()}},
\code{\link[=input_select]{input_select()}},
\code{\link[=input_text]{input_text()}},
\code{\link[=input_text_group]{input_text_group()}}
}
\concept{inputs}
