% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-checkbox.R
\name{input_checkbox}
\alias{input_checkbox}
\alias{input_checkbox_button}
\title{Checkbox and switch inputs}
\usage{
input_checkbox(
  id,
  choices = NULL,
  values = choices,
  selected = NULL,
  inline = FALSE,
  ...
)

input_checkbox_button(id, choices, values = choices, selected = NULL, ...)
}
\arguments{
\item{id}{A character string specifying the id of the reactive input.}

\item{choices}{A character string or vector specifying a label or labels for
the checkbox or checkbar.}

\item{values}{A character string or vector specifying values for the
checkbox or checkbar input, defaults to \code{choice} or \code{values}, respectively.}

\item{selected}{One or more of \code{values} specifying which choices are
selected by default, defaults to \code{NULL}, in which case no choices are
initially selected.}

\item{inline}{One of \code{TRUE} or \code{FALSE} specifying if the checkbox input
choices render inline or stacked, defaults to \code{FALSE}, in which case the
choices are stacked.}

\item{...}{Additional named arguments passed as HTML attributes to the
parent element or tag elements passed as child elements to the parent
element.}

\item{disable}{One of \code{values} specifying particular choices to disable or
\code{TRUE} specifying the entire input is disabled, defaults to \code{NULL}.}

\item{session}{A reactive context, defaults to \code{\link[=getDefaultReactiveDomain]{getDefaultReactiveDomain()}}.}
}
\description{
Reactive checkbox and checkbar inputs. Users may select one or more choices.
The checkbox input appears as a standard checkbox or set of checkboxes. When
a checkbox input has no selected choices the reactive value is \code{NULL}. Switch
inputs differ from checkboxes only in appearance.
}
\seealso{
Other inputs: 
\code{\link{checkbarInput}()},
\code{\link{chipInput}()},
\code{\link{fileInput}()},
\code{\link{formInput}()},
\code{\link{input_button}()},
\code{\link{listGroupInput}()},
\code{\link{menuInput}()},
\code{\link{radioInput}()},
\code{\link{radiobarInput}()},
\code{\link{rangeInput}()},
\code{\link{selectInput}()},
\code{\link{textInput}()}
}
\concept{inputs}
