% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-checkbox.R
\name{input_checkbox}
\alias{input_checkbox}
\alias{update_checkbox}
\title{Checkbox input}
\usage{
input_checkbox(
  id,
  choices,
  ...,
  values = choices,
  select = NULL,
  disable = NULL,
  layout = "column",
  label = "after"
)

update_checkbox(
  id,
  choices = NULL,
  values = choices,
  select = NULL,
  disable = NULL,
  layout = "column",
  label = "after",
  session = default_reactive_domain()
)
}
\arguments{
\item{id}{A character string. The id of the reactive input.}

\item{choices}{A character vector or list. The labels for the input choices.}

\item{...}{Optional named arguments specifying HTML attributes for the input
element.}

\item{values}{A character vector. The values for the input, defaults to
\code{choices}.}

\item{select}{A character vector. The values selected by default, one or more
of \code{values}.}

\item{disable}{A character vector. The values disabled by default.}

\item{layout}{A character string. The layout of the input choices, one of
\code{"column"} or \code{"row"}.}

\item{label}{A character string. The placement of the choice label relative
to the checkboxes, one of \code{"after"} or "\code{before}".}

\item{session}{A shiny session object.}
}
\value{
A shiny tag object.
}
\description{
A reactive checkbox input. Users may select one or more choices. When a
checkbox input has no selected choices the reactive value is \code{NULL}.
}
\details{
\subsection{Server value}{

A named logical vector.
}
}
\seealso{
Other inputs: 
\code{\link{checkbarInput}()},
\code{\link{fileInput}()},
\code{\link{input_button}()},
\code{\link{input_checkbutton}()},
\code{\link{input_chip}()},
\code{\link{input_form}()},
\code{\link{listGroupInput}()},
\code{\link{menuInput}()},
\code{\link{radioInput}()},
\code{\link{radiobarInput}()},
\code{\link{rangeInput}()},
\code{\link{selectInput}()},
\code{\link{textInput}()}
}
\concept{inputs}
