% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-checkbox.R, R/input-checkbox-group.R
\name{input_checkbox}
\alias{input_checkbox}
\alias{update_checkbox}
\alias{update_checkbox_group}
\title{Checkbox input}
\usage{
input_checkbox(
  id,
  choice,
  ...,
  value = FALSE,
  disable = NULL,
  label = c("after", "before")
)

update_checkbox(
  id,
  choice = NULL,
  value = NULL,
  disable = NULL,
  session = get_current_session()
)

update_checkbox_group(
  id,
  choices = NULL,
  values = choices,
  select = NULL,
  disable = NULL,
  layout = "column",
  label = "after",
  session = get_current_session()
)
}
\arguments{
\item{id}{A character string. The id of the reactive input.}

\item{choice}{A character string. The text for the checkbox.}

\item{...}{Optional named arguments specifying HTML attributes for the input
element.}

\item{value}{A boolean. The default value for the checkbox.}

\item{disable}{A boolean. The checkbox starts disabled if \code{TRUE}.}

\item{label}{A character string. The placement of a label relative to its
checkbox.}

\item{session}{A shiny session object.}
}
\value{
A \link[htmltools:builder]{htmltools::tag} object.
}
\description{
A reactive checkbox input. Users may select one or more choices. When a
checkbox input has no selected choices the reactive value is \code{NULL}.
}
\details{
\subsection{Server value}{

A named logical vector.
}
}
\seealso{
Other inputs: 
\code{\link{fileInput}()},
\code{\link{input_button}()},
\code{\link{input_checkbox_group}()},
\code{\link{input_chip}()},
\code{\link{input_form}()},
\code{\link{input_list_group}()},
\code{\link{input_menu}()},
\code{\link{input_radio_group}()},
\code{\link{input_range}()},
\code{\link{selectInput}()},
\code{\link{textInput}()}
}
\concept{inputs}
