% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-checkbox-group.R
\name{input_checkbox_group}
\alias{input_checkbox_group}
\title{Checkbox input}
\usage{
input_checkbox_group(
  id,
  choices,
  ...,
  values = choices,
  select = NULL,
  disable = NULL,
  appearance = c("default", "buttons", "switches", "list"),
  layout = c("column", "row"),
  label = c("after", "before")
)
}
\arguments{
\item{id}{A character string. The id of the reactive input.}

\item{choices}{A character vector or list. The labels for the input choices.}

\item{...}{Optional named arguments specifying HTML attributes for the input
element.}

\item{values}{A character vector. The values for the input, defaults to
\code{choices}.}

\item{select}{A character vector. The values selected by default, one or more
of \code{values}.}

\item{disable}{A character vector. The values disabled by default.}

\item{appearance}{A character string. The appearance of the input's
checkboxes.}

\item{layout}{A character string. The layout of the choices.}

\item{label}{A character string. The placement of a label relative to its
checkbox.}

\item{session}{A shiny session object.}
}
\value{
A shiny tag object.
}
\description{
A reactive checkbox input. Users may select one or more choices. When a
checkbox input has no selected choices the reactive value is \code{NULL}.
}
\details{
\subsection{Server value}{

A named logical vector.
}
}
\seealso{
Other inputs: 
\code{\link{fileInput}()},
\code{\link{input_button}()},
\code{\link{input_checkbox}()},
\code{\link{input_chip}()},
\code{\link{input_form}()},
\code{\link{input_list_group}()},
\code{\link{menuInput}()},
\code{\link{radioInput}()},
\code{\link{radiobarInput}()},
\code{\link{rangeInput}()},
\code{\link{selectInput}()},
\code{\link{textInput}()}
}
\concept{inputs}
