% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-form.R
\name{input_form}
\alias{input_form}
\alias{form_submit_button}
\alias{update_form}
\alias{submit_form}
\title{Form inputs}
\usage{
input_form(id, ..., label = NULL, layout = NULL)

form_submit_button(label, ..., value = label)

update_form(id, label = NULL, session = get_current_session())

submit_form(id, value, session = get_current_session())
}
\arguments{
\item{id}{A character string. The id of the reactive input.}

\item{...}{Reactive inputs.}

\item{label}{A character string. The label of the input.}

\item{layout}{<\link{responsive}> A character string.}

\item{value}{A character string. The input's value when the submit button is
clicked.}

\item{session}{A shiny session object.}

\item{gap}{A number. The space between inputs, \code{0} through \code{5}.}
}
\description{
Form inputs are a new reactive input. Form inputs are an alternative to
\link[shiny:submitButton]{shiny::submitButton}. A form input is comprised of any number of inputs. The
values of these inputs do not reactively update. The inputs will reactively
update when a form submit button is clicked.

A form input's reactive value depends on the clicked form submit button. This
allows server logic to distinguish between different form submission types,
think "login" versus "register".
}
\seealso{
Other inputs: 
\code{\link{fileInput}()},
\code{\link{input_button}()},
\code{\link{input_checkbox}()},
\code{\link{input_checkbox_group}()},
\code{\link{input_chip}()},
\code{\link{input_list_group}()},
\code{\link{menuInput}()},
\code{\link{radioInput}()},
\code{\link{radiobarInput}()},
\code{\link{rangeInput}()},
\code{\link{selectInput}()},
\code{\link{textInput}()}
}
\concept{inputs}
