% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-link.R
\name{input_link}
\alias{input_link}
\alias{update_link}
\title{Link input}
\usage{
input_link(id, label, ..., stretch = FALSE, icon = NULL)

update_link(id, label = NULL, session = get_current_session())
}
\arguments{
\item{id}{A character string. The id of the reactive input.}

\item{label}{A character string. The placement of a label relative to its
checkbox.}

\item{...}{Optional named arguments specifying HTML attributes for the input
element.}

\item{stretch}{A boolean.}

\item{icon}{A \verb{[bsicons::bs_icon]}.}

\item{session}{A shiny session object.}
}
\description{
A button-like input, clickable text. Link inputs include a unique stretch
feature. With \code{stretch = TRUE} a link input will detect clicks on its parent
container. For example, a \verb{[bslib::card]} may become clickable.
}
