% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-list-group.R
\name{input_list_group}
\alias{input_list_group}
\alias{update_list_group}
\title{List group inputs}
\usage{
input_list_group(
  id,
  choices,
  ...,
  values = choices,
  selected = NULL,
  layout = "vertical",
  flush = FALSE
)

update_list_group(
  id,
  choices = NULL,
  values = choices,
  selected = NULL,
  enable = NULL,
  disable = NULL,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{A character string. The id of the reactive input.}

\item{choices}{A character vector or list. The labels for the input choices.}

\item{...}{Optional named arguments specifying HTML attributes for the input
element.}

\item{values}{A character vector. The values for the input, defaults to
\code{choices}.}

\item{layout}{<\link{responsive}> A character vector. The layout of the list
group, \code{"column"} or \code{"row"}.}

\item{flush}{One of \code{TRUE} or \code{FALSE} specifying if the list group is
rendered without an outside border, defaults to \code{FALSE}. Removing the list
group border is useful when rendering a list group inside a custom parent
container, e.g. inside a \code{\link[=card]{card()}}.}

\item{disable}{A character vector. The values disabled by default.}

\item{session}{A shiny session object.}
}
\description{
List group inputs are an actionable list of items. They behave similarly to
checkboxes or radios, that is, users may select one or more items from the
list. However, list group items may include highly variable content.
}
\seealso{
Other inputs: 
\code{\link{fileInput}()},
\code{\link{input_button}()},
\code{\link{input_checkbox}()},
\code{\link{input_checkbox_group}()},
\code{\link{input_chip}()},
\code{\link{input_form}()},
\code{\link{menuInput}()},
\code{\link{radioInput}()},
\code{\link{radiobarInput}()},
\code{\link{rangeInput}()},
\code{\link{selectInput}()},
\code{\link{textInput}()}
}
\concept{inputs}
