% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-menu.R
\name{input_menu}
\alias{input_menu}
\alias{update_menu_input}
\title{Menu inputs}
\usage{
input_menu(
  id,
  label,
  choices,
  ...,
  values = choices,
  select = NULL,
  disable = NULL,
  menu_options = NULL
)

update_menu_input(
  id,
  label = NULL,
  choices = NULL,
  values = choices,
  selected = NULL,
  enable = NULL,
  disable = NULL,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{A character string. The id of the reactive input.}

\item{label}{A character string or tag element specifying the label of the
menu's toggle button.}

\item{choices}{A character vector specifying the choice text of the menu's
items.}

\item{...}{Optional named arguments specifying HTML attributes for the input
element.}

\item{values}{A character vector specifying the values of the menu's items,
defaults to \code{choices}.}

\item{select}{A character vector. The values selected by default, one or more
of \code{values}.}

\item{disable}{A character vector. The values disabled by default.}

\item{session}{A shiny session object.}

\item{direction}{One of \code{"up"}, \code{"right"}, \code{"down"}, or \code{"left"} specifying
which direction the menu opens, defaults to \code{"down"}.}

\item{align}{One or \code{"right"} or \code{"left"} specifying which side of the
toggle button the menu aligns to, defaults to \code{"left"}.`}
}
\description{
A dropdown menu.
}
\seealso{
Other inputs: 
\code{\link{fileInput}()},
\code{\link{input_button}()},
\code{\link{input_checkbox}()},
\code{\link{input_checkbox_group}()},
\code{\link{input_chip}()},
\code{\link{input_form}()},
\code{\link{input_list_group}()},
\code{\link{input_radio_group}()},
\code{\link{radiobarInput}()},
\code{\link{rangeInput}()},
\code{\link{selectInput}()},
\code{\link{textInput}()}
}
\concept{inputs}
