% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-menu.R
\name{input_menu}
\alias{input_menu}
\alias{update_menu_input}
\title{Menu inputs}
\usage{
input_menu(
  id,
  label,
  choices,
  ...,
  values = choices,
  disable = NULL,
  direction = c("down", "up", "left", "right"),
  align = c("left", "right"),
  close = c("auto", "inside", "outside", "manual")
)

update_menu_input(
  id,
  label = NULL,
  choices = NULL,
  values = choices,
  select = NULL,
  disable = NULL,
  session = get_current_session()
)
}
\arguments{
\item{id}{A character string. The id of the reactive input.}

\item{label}{A character string or tag element specifying the label of the
menu's toggle button.}

\item{choices}{A character vector or list. The labels for the input choices.}

\item{...}{Optional named arguments specifying HTML attributes for the input
element.}

\item{values}{A character vector. The values for the input, defaults to
\code{choices}.}

\item{disable}{A character vector. The values disabled by default.}

\item{direction}{A character string.}

\item{align}{<\link{responsive}> A character vector.}

\item{close}{A character string.}

\item{select}{A character vector. The values selected by default, one or more
of \code{values}.}

\item{session}{A shiny session object.}
}
\description{
A dropdown menu.
}
\seealso{
Other inputs: 
\code{\link{fileInput}()},
\code{\link{input_button}()},
\code{\link{input_checkbox}()},
\code{\link{input_checkbox_group}()},
\code{\link{input_chip}()},
\code{\link{input_form}()},
\code{\link{input_list_group}()},
\code{\link{input_radio_group}()},
\code{\link{radiobarInput}()},
\code{\link{rangeInput}()},
\code{\link{selectInput}()},
\code{\link{textInput}()}
}
\concept{inputs}
