% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-multi-select.R
\name{input_multi_select}
\alias{input_multi_select}
\alias{update_multi_select}
\title{Multi select input}
\usage{
input_multi_select(id, ...)

update_multi_select(
  id,
  choices = NULL,
  values = choices,
  selected = NULL,
  max = NULL,
  enable = NULL,
  disable = NULL,
  session = get_current_session()
)
}
\arguments{
\item{id}{A character string. The id of the reactive input.}

\item{...}{Optional named arguments specifying HTML attributes for the input
element.}

\item{disable}{A boolean. The checkbox starts disabled if \code{TRUE}.}

\item{session}{A shiny session object.}
}
\value{
A \link[htmltools:builder]{htmltools::tag} object.
}
\description{
The multi select input is a selectize alternative.
}
\seealso{
Other inputs: 
\code{\link{fileInput}()},
\code{\link{input_button}()},
\code{\link{input_checkbox}()},
\code{\link{input_checkbox_group}()},
\code{\link{input_form}()},
\code{\link{input_list_group}()},
\code{\link{input_menu}()},
\code{\link{input_numeric}()},
\code{\link{input_radio_group}()},
\code{\link{input_range}()},
\code{\link{input_select}()},
\code{\link{input_text}()},
\code{\link{input_text_group}()}
}
\concept{inputs}
