% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-numeric.R
\name{input_numeric}
\alias{input_numeric}
\title{Numeric input}
\usage{
input_numeric(
  id,
  ...,
  label = NULL,
  value = NULL,
  min = NULL,
  max = NULL,
  step = 1,
  placeholder = NULL
)
}
\arguments{
\item{id}{A character string. The id of the reactive input.}

\item{...}{Optional named arguments specifying HTML attributes for the input
element.}

\item{label}{A character string or \code{\link[=floating_label]{floating_label()}}. The label of the
input. A character string renders a standard label ahead of the input,
a \code{\link[=floating_label]{floating_label()}} renders a floating label inside the input.}

\item{value}{A number. The default value of the input.}

\item{min}{A number. A possible minimum value for the input.}

\item{max}{A number. A possible maximum value for the input.}

\item{step}{A number. The interval when stepping between values.}

\item{placeholder}{A string. The placeholder text displayed inside the input.}
}
\value{
A shiny tag object.
}
\description{
A numeric input.
}
\seealso{
\code{\link[=update_numeric]{update_numeric()}}

Other inputs:
\code{\link[=fileInput]{fileInput()}},
\code{\link[=floating_label]{floating_label()}},
\code{\link[=input_button]{input_button()}},
\code{\link[=input_checkbox]{input_checkbox()}},
\code{\link[=input_checkbox_group]{input_checkbox_group()}},
\code{\link[=input_chip_group]{input_chip_group()}},
\code{\link[=input_form]{input_form()}},
\code{\link[=input_list_group]{input_list_group()}},
\code{\link[=input_menu]{input_menu()}},
\code{\link[=input_multi_select]{input_multi_select()}},
\code{\link[=input_radio_group]{input_radio_group()}},
\code{\link[=input_range]{input_range()}},
\code{\link[=input_select]{input_select()}},
\code{\link[=input_text]{input_text()}},
\code{\link[=input_text_group]{input_text_group()}}
}
\concept{inputs}
