% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-numeric.R
\name{input_numeric}
\alias{input_numeric}
\title{Numeric input}
\usage{
input_numeric(
  id,
  ...,
  value = NULL,
  min = NULL,
  max = NULL,
  step = 1,
  placeholder = NULL
)
}
\arguments{
\item{id}{A character string. The id of the reactive input.}

\item{...}{Optional named arguments specifying HTML attributes for the input
element.}

\item{min}{A number. A possible minimum value for the input.}

\item{max}{A number. A possible maximum value for the input.}

\item{step}{A number. The interval when stepping between values.}

\item{placeholder}{A string. The placeholder text displayed inside the input.}
}
\value{
A shiny tag object.
}
\description{
A numeric input.
}
\seealso{
Other inputs: 
\code{\link{fileInput}()},
\code{\link{input_button}()},
\code{\link{input_checkbox}()},
\code{\link{input_checkbox_group}()},
\code{\link{input_form}()},
\code{\link{input_list_group}()},
\code{\link{input_menu}()},
\code{\link{input_multi_select}()},
\code{\link{input_radio_group}()},
\code{\link{input_range}()},
\code{\link{input_select}()},
\code{\link{input_text}()},
\code{\link{input_text_group}()}
}
\concept{inputs}
