% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-range.R
\name{input_range}
\alias{input_range}
\alias{update_range}
\title{Range input}
\usage{
input_range(id, ..., label = NULL, min = 0, max = 100, value = min, step = 1)

update_range(id, value = NULL, disable = NULL, session = get_current_session())
}
\arguments{
\item{id}{A character string. The id of the reactive input.}

\item{...}{Optional named arguments specifying HTML attributes for the input
element.}

\item{label}{A character string. The label of the input, rendered ahead
of the input. Floating labels are not supported.}

\item{min}{A number. The input's minimum value, defaults to \code{0}.}

\item{max}{A number. The input's maximum value, defaults to \code{100}.}

\item{value}{A number. The value of the input, between \code{min} and
\code{max}.}

\item{step}{A number. The input's step interval.}

\item{disable}{A boolean. The checkbox starts disabled if \code{TRUE}.}

\item{session}{A shiny session object.}
}
\value{
A \link[htmltools:tag]{htmltools::tag} object.
}
\description{
A simple numeric range input.
}
\details{
The sophistication of this input will improve as browsers adopt the latest
HTML standards.
}
\seealso{
Other inputs:
\code{\link[=floating_label]{floating_label()}},
\code{\link[=input_button]{input_button()}},
\code{\link[=input_checkbox]{input_checkbox()}},
\code{\link[=input_checkbox_group]{input_checkbox_group()}},
\code{\link[=input_chip_group]{input_chip_group()}},
\code{\link[=input_form]{input_form()}},
\code{\link[=input_list_group]{input_list_group()}},
\code{\link[=input_menu]{input_menu()}},
\code{\link[=input_multi_select]{input_multi_select()}},
\code{\link[=input_numeric]{input_numeric()}},
\code{\link[=input_radio_group]{input_radio_group()}},
\code{\link[=input_select]{input_select()}},
\code{\link[=input_text]{input_text()}},
\code{\link[=input_text_group]{input_text_group()}}
}
\concept{inputs}
