% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-select.R
\name{input_select}
\alias{input_select}
\alias{update_select}
\title{Select inputs}
\usage{
input_select(
  id,
  choices,
  ...,
  label = NULL,
  values = choices,
  select = NULL,
  disable = NULL,
  placeholder = NULL
)

update_select(
  id,
  choices = NULL,
  values = choices,
  select = NULL,
  disable = NULL,
  session = get_current_session()
)
}
\arguments{
\item{id}{A character string. The id of the reactive input.}

\item{choices}{A character vector specifying the input's choices.}

\item{...}{Optional named arguments specifying HTML attributes for the input
element.}

\item{label}{A character string or \code{\link[=floating_label]{floating_label()}}. The label of the
input. A character string renders a standard label ahead of the input,
a \code{\link[=floating_label]{floating_label()}} renders a floating label inside the input.}

\item{values}{A character vector specifying the values of the input's
choices, defaults to \code{choices}.}

\item{select}{One of \code{values} indicating the default value of the input,
defaults to \code{NULL}.}

\item{disable}{A boolean. The checkbox starts disabled if \code{TRUE}.}

\item{placeholder}{A character string specifying the placeholder text of
the select input, defaults to \code{NULL}.}

\item{session}{A shiny session object.}
}
\description{
Create a select input. Select elements typically appear as a simple menu of
choices and may have one selected choice. A group select input is a select
input with one or two additional components. These addon components are used
to change the reactivity or value of the input, see Details for more
information.
}
\seealso{
Other inputs:
\code{\link[=floating_label]{floating_label()}},
\code{\link[=input_button]{input_button()}},
\code{\link[=input_checkbox]{input_checkbox()}},
\code{\link[=input_checkbox_group]{input_checkbox_group()}},
\code{\link[=input_chip_group]{input_chip_group()}},
\code{\link[=input_form]{input_form()}},
\code{\link[=input_list_group]{input_list_group()}},
\code{\link[=input_menu]{input_menu()}},
\code{\link[=input_multi_select]{input_multi_select()}},
\code{\link[=input_numeric]{input_numeric()}},
\code{\link[=input_radio_group]{input_radio_group()}},
\code{\link[=input_range]{input_range()}},
\code{\link[=input_text]{input_text()}},
\code{\link[=input_text_group]{input_text_group()}}
}
\concept{inputs}
