% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-select.R
\name{input_select}
\alias{input_select}
\alias{update_select}
\title{Select inputs}
\usage{
input_select(
  id,
  choices,
  ...,
  values = choices,
  select = NULL,
  disable = NULL,
  placeholder = NULL
)

update_select(
  id,
  choices = NULL,
  values = choices,
  select = NULL,
  disable = NULL,
  session = get_current_session()
)
}
\arguments{
\item{id}{A character string. The id of the reactive input.}

\item{choices}{A character vector specifying the input's choices.}

\item{...}{Optional named arguments specifying HTML attributes for the input
element.}

\item{values}{A character vector specifying the values of the input's
choices, defaults to \code{choices}.}

\item{disable}{A boolean. The checkbox starts disabled if \code{TRUE}.}

\item{placeholder}{A character string specifying the placeholder text of
the select input, defaults to \code{NULL}.}

\item{session}{A shiny session object.}

\item{selected}{One of \code{values} indicating the default value of the input,
defaults to \code{values[[1]]}.}
}
\description{
Create a select input. Select elements typically appear as a simple menu of
choices and may have one selected choice. A group select input is a select
input with one or two additional components. These addon components are used
to change the reactivity or value of the input, see Details for more
information.
}
\seealso{
Other inputs: 
\code{\link{fileInput}()},
\code{\link{input_button}()},
\code{\link{input_checkbox}()},
\code{\link{input_checkbox_group}()},
\code{\link{input_chip}()},
\code{\link{input_form}()},
\code{\link{input_list_group}()},
\code{\link{input_menu}()},
\code{\link{input_radio_group}()},
\code{\link{input_range}()},
\code{\link{textInput}()}
}
\concept{inputs}
