% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-text.R
\name{input_text}
\alias{input_text}
\alias{update_text}
\title{Text inputs}
\usage{
input_text(id, ..., label = NULL, value = NULL, placeholder = NULL)

update_text(id, value = NULL, disable = NULL, session = get_current_session())
}
\arguments{
\item{id}{A character string. The id of the reactive input.}

\item{...}{Optional named arguments specifying HTML attributes for the input
element.}

\item{label}{A character string or \code{\link[=floating_label]{floating_label()}}. The label of the
input. A character string renders a standard label ahead of the input,
a \code{\link[=floating_label]{floating_label()}} renders a floating label inside the input.}

\item{value}{A character string. The value of the input.}

\item{placeholder}{A string. The placeholder text displayed inside the input.}

\item{disable}{A boolean. If \code{TRUE}, disable the input.}

\item{session}{A shiny session object.}
}
\value{
A shiny tag object.
}
\description{
A text input.
}
\seealso{
Other inputs:
\code{\link[=fileInput]{fileInput()}},
\code{\link[=floating_label]{floating_label()}},
\code{\link[=input_button]{input_button()}},
\code{\link[=input_checkbox]{input_checkbox()}},
\code{\link[=input_checkbox_group]{input_checkbox_group()}},
\code{\link[=input_chip_group]{input_chip_group()}},
\code{\link[=input_form]{input_form()}},
\code{\link[=input_list_group]{input_list_group()}},
\code{\link[=input_menu]{input_menu()}},
\code{\link[=input_multi_select]{input_multi_select()}},
\code{\link[=input_numeric]{input_numeric()}},
\code{\link[=input_radio_group]{input_radio_group()}},
\code{\link[=input_range]{input_range()}},
\code{\link[=input_select]{input_select()}},
\code{\link[=input_text_group]{input_text_group()}}
}
\concept{inputs}
