% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-text.R
\name{input_text}
\alias{input_text}
\alias{update_text}
\title{Text inputs}
\usage{
input_text(id, ..., value = NULL, placeholder = NULL)

update_text(id, value = NULL, disable = NULL, session = get_current_session())
}
\arguments{
\item{id}{A character string. The id of the reactive input.}

\item{...}{Optional named arguments specifying HTML attributes for the input
element.}

\item{placeholder}{A string. The placeholder text displayed inside the input.}
}
\value{
A shiny tag object.
}
\description{
A text input.
}
\seealso{
Other inputs: 
\code{\link{fileInput}()},
\code{\link{input_button}()},
\code{\link{input_checkbox}()},
\code{\link{input_checkbox_group}()},
\code{\link{input_chip}()},
\code{\link{input_form}()},
\code{\link{input_list_group}()},
\code{\link{input_menu}()},
\code{\link{input_numeric}()},
\code{\link{input_radio_group}()},
\code{\link{input_range}()},
\code{\link{input_select}()},
\code{\link{input_text_group}()}
}
\concept{inputs}
