% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-text-group.R
\name{input_text_group}
\alias{input_text_group}
\alias{update_text_group}
\title{A text group input}
\usage{
input_text_group(
  id,
  ...,
  left = NULL,
  right = NULL,
  value = NULL,
  placeholder = NULL
)

update_text_group(
  id,
  value = NULL,
  disable = NULL,
  session = get_current_session()
)
}
\arguments{
\item{id}{A character string. The id of the reactive input.}

\item{...}{Optional named arguments specifying HTML attributes for the input
element.}

\item{left}{A character vector. One or more character strings prepended to
the reactive value of the input.}

\item{right}{A character vector. One or more character strings appended to
the reactive value of the input.}

\item{placeholder}{A string. The placeholder text displayed inside the input.}
}
\value{
A shiny tag object.
}
\description{
A text input with a possible static prefix or suffix.
}
\seealso{
Other inputs: 
\code{\link{fileInput}()},
\code{\link{input_button}()},
\code{\link{input_checkbox}()},
\code{\link{input_checkbox_group}()},
\code{\link{input_form}()},
\code{\link{input_list_group}()},
\code{\link{input_menu}()},
\code{\link{input_multi_select}()},
\code{\link{input_numeric}()},
\code{\link{input_radio_group}()},
\code{\link{input_range}()},
\code{\link{input_select}()},
\code{\link{input_text}()}
}
\concept{inputs}
