% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{jumbotron}
\alias{jumbotron}
\title{Jumbotron}
\usage{
jumbotron(title, subtitle, ..., fluid = FALSE)
}
\arguments{
\item{title}{A character string specifying the jumbotron's title.}

\item{subtitle}{A character string specifying the jumbotron's subtitle.}

\item{...}{Additional tag elements or named arguments passed as HTML
attributes to the parent element.}

\item{fluid}{One of \code{TRUE} or \code{FALSE} specifying if the jumbotron fills the
width of its parent container, defaults to \code{FALSE}.}
}
\description{
Highlight messages.
}
\examples{

### Landing page welcome

jumbotron(
  title = "Welcome, welcome!",
  subtitle = "This simple jumbotron-style component calls attention to a new feature",
  tags$p(
    "Here we can talk more about this excellently superb new feature.",
    "The best."
  )
)

}
\seealso{
Other content: \code{\link{alert}},
  \code{\link{blockquote}}, \code{\link{card}},
  \code{\link{collapsiblePane}}, \code{\link{d1}},
  \code{\link{dropdown}}, \code{\link{img}},
  \code{\link{modal}}, \code{\link{navContent}},
  \code{\link{popover}}, \code{\link{pre}},
  \code{\link{progressOutlet}}, \code{\link{tooltip}}
}
\concept{content}
