% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{jumbotron}
\alias{jumbotron}
\title{Jumbotron}
\usage{
jumbotron(title, subtitle, ...)
}
\arguments{
\item{title}{A character string specifying the jumbotron's title.}

\item{subtitle}{A character string specifying the jumbotron's subtitle.}

\item{...}{Additional elements or named arguments passed as HTML attributes
to the parent element.}
}
\description{
Highlight messages.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      jumbotron(
        title = "Hello, world!",
        subtitle = "This simple jumbotron-style component calls attention to a new feature",
        tags$p(
          "Here we can talk more about this excellently superb new feature.",
          "The best."
        )
      )
    ),
    server = function(input, output) {

    }
  )
}

}
