% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{jumbotron}
\alias{jumbotron}
\title{Jumbotron}
\usage{
jumbotron(..., title = NULL, subtitle = NULL)
}
\arguments{
\item{...}{Tag elements passed as child elements or named arguments passed as
HTML attributes to the parent element.}

\item{title}{A character string specifying a title for the jumbotron,
defaults to \code{NULL}, in which case a title is not added.}

\item{subtitle}{A character string specifying a subtitle for the jumbotron,
defaults to \code{NULL}, in which case a subtitle is not added.}
}
\description{
A showcase banner, good for front or splash pages.
}
\examples{

### Landing page welcome

jumbotron(
  title = "Welcome, welcome!",
  subtitle = "Here we are showcasing the very showcase itself.",
  tags$p(
    "Now let's talk more about that superb new feature."
  )
)

}
\seealso{
Other components: 
\code{\link{alert}()},
\code{\link{badge}()},
\code{\link{blockquote}()},
\code{\link{card}()},
\code{\link{collapsePane}()},
\code{\link{d1}()},
\code{\link{dropdown}()},
\code{\link{img}()},
\code{\link{modal}()},
\code{\link{navContent}()},
\code{\link{popover}()},
\code{\link{pre}()},
\code{\link{toast}()}
}
\concept{components}
