% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav.R
\name{linksNav}
\alias{linksNav}
\title{Page Navigation}
\usage{
linksNav(labels, ids, align = "right", block = FALSE, ...)
}
\arguments{
\item{labels}{A character vector or flat list of character strings specifying
the labels of the nav.}

\item{align}{One of \code{"left"}, \code{"right"}, \code{"center"}, or \code{"horizontal"}
specifying the alignment of the nav, defaults to \code{"left"}.}

\item{hrefs}{A character vector or flat list of character strings specifying
the HTML href attribute for each nav link.}

\item{vertical}{If \code{TRUE}, the nav is rendered vertically, defaults to
\code{FALSE}.}

\item{align}{One of \code{"left"}, \code{"right"}, or \code{"center"} specifying how the nav
is aligned on the page, defaults to \code{"left"}.}
}
\description{
Create a nav element. This is not a set of tabs, for creating a tabbed interface
of dynamic content see \link{tabsNav} or \link{pillsNav}.
}
\examples{

stub

}
\seealso{
For more information about the Bootstrap nav please follow this
\href{http://v4-alpha.getbootstrap.com/components/navs/}{link}.

Other navs: \code{\link{listGroupNav}},
  \code{\link{tabsNav}}
}
