% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav.R
\name{linksNav}
\alias{linksNav}
\title{Page Navigation}
\usage{
linksNav(labels, hrefs, ...)
}
\arguments{
\item{labels}{A character vector or flat list of character strings specifying
the labels of the nav.}

\item{hrefs}{A character vector or flat list of character strings specifying
the HTML href attributes for each nav link.}

\item{...}{Additional named arguments to pass as HTML attributes to the
parent element.}
}
\description{
Create a nav element. This is not a set of tabs, for creating a tabbed interface
of dynamic content see \link{tabsNav} or \link{pillsNav}.
}
\examples{

stub

}
\seealso{
Bootstrap 4 nav documentation:
\url{https://getbootstrap.com/docs/4.0/components/navs/}
}
