% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-group.R
\name{listGroupInput}
\alias{listGroupInput}
\alias{listGroupItem}
\alias{updateListGroupItem}
\title{List group inputs}
\usage{
listGroupInput(..., id = NULL)

listGroupItem(label = NULL, value = NULL, context = NULL,
  active = FALSE, disabled = FALSE, badge = NULL, ..., id = NULL)

updateListGroupItem(id, context = NULL, active = NULL, disable = NULL,
  session = getDefaultReactiveDomain())
}
\arguments{
\item{...}{List group items or named arguments passed as HTML attributes to
the parent element.}

\item{id}{A character string specifying the id of the list group or list
group item, defaults to \code{NULL}.

If a \strong{list group} id is specified then a reactive value is available to
the shiny server.

Specifying a \strong{list group item} id is only necessary to use
\code{updateListGroupItem}.}

\item{label}{A character string specifying the label of the list group item,
defaults to \code{NULL}.}

\item{value}{A character string specifying the value of the list group item,
defaults to \code{NULL}.}

\item{context}{A character string specifying the visual context of the list
group item, one of \code{"success"}, \code{"info"}, \code{"warning"}, or \code{"danger"},
defaults to \code{NULL}. Use \code{"none"} to remove a list group item's context with
\code{updateListGroupItem}.}

\item{active}{If \code{TRUE} the list group item is rendered in an active state,
defaults to \code{FALSE}.}

\item{disabled}{If \code{TRUE} the list group item is rendered in a disabled
state, a disabled list group item will not receive click events, defaults
to \code{FALSE}. A list group item may be enabled using \code{updateListGroupItem}.}

\item{badge}{A \code{badgeOutput} which is right aligned inside the list group
item, see \code{\link{renderBadge}} for more on how to work with badges, defaults to
\code{NULL}.}
}
\description{
A way of handling and outlining content as a list. List groups function
similarly to checkbox groups. A list group returns a reactive vector of the
values from its active (selected) list group items. List group items are
selected or unselected by clicking on them. While list groups may be used as
reactive inputs they may also be used to simply display content, in which
case do not specify a list group id.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      row(
        tags$h3("Find something to do!"),
        br()
      ),
      row(
        col(
          width = 4,
          tags$h4("Any preferences?"),
          listGroupInput(
            id = "preferences",
            listGroupItem(
              label = "No heights",
              value = "noheights"
            ),
            listGroupItem(
              label = "No animals",
              value = "noanimals"
            )
          )
        ),
        col(
          tags$h4("Possible outings"),
          listGroupInput(
            hover = FALSE,
            listGroupItem(
              id = "optZoo",
              "Zoo visit!"
            ),
            listGroupItem(
              id = "optSkydive",
              "Sky diving!"
            ),
            listGroupItem(
              id = "optPern",
              "Visit Pern"
            )
          )
        )
      )
    ),
    server = function(input, output) {
      observe({
        prefs <- input$preferences

        updateListGroupItem(
          id = "optSkydive",
          context = if ("noheights" \%in\% prefs) "danger" else "none"
        )

        updateListGroupItem(
          id = "optZoo",
          context = if ("noanimals" \%in\% prefs) "danger" else "none"
        )

        updateListGroupItem(
          id = "optPern",
          context = if ("noanimals" \%in\% prefs) "danger" else "none"
        )
      })
    }
  )
}

}
\seealso{
For more information on Bootstrap list groups please refer to the
\href{https://v4-alpha.getbootstrap.com/components/list-group/}{reference page}.
}
