% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-group.R
\name{listGroupInput}
\alias{listGroupInput}
\alias{updateListGroupInput}
\alias{validateListGroupInput}
\title{List group inputs}
\usage{
listGroupInput(id, labels, values = labels, selected = NULL,
  disabled = NULL, ...)

updateListGroupInput(id, labels, values = labels, selected = NULL,
  disabled = NULL, session = getDefaultReactiveDomain())

validateListGroupInput(id, state, validate = NULL,
  session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{A character string specifying the id of the list group input, the
reactive value of the list group input is available to the shiny server
function as part of the \code{input} object. When creating a list group to
only display information pass \code{NULL} as the \code{id}, in this case an id is not
added and a reactive input is not created.}

\item{labels}{A character vector or list of character strings specifying the
labels of the list group items, defaults to \code{labels}.}

\item{values}{A character string or list of character strings specifying the
values of the list group items.}

\item{selected}{One or more of \code{values} indicating the default list group
items selected, defaults to \code{NULL}, in which case no items are selected and
the default value of the list group input is \code{NULL}.}

\item{disabled, enabled}{One or more of \code{values} indicating which list group
items to disable or enable, defaults to \code{NULL}. If \code{NULL} then
\code{disableListGroup} and \code{enableListGroup} will disable or enable all the
list group items, respectively.}

\item{...}{Additional named arguments passed as HTML attributes to the parent
element.}

\item{state}{One of \code{"success"}, \code{"info"}, \code{"warning"}, or \code{"danger"}
indicating the state of the list group items. If the return value is \code{NULL}
any visual context is removed.}

\item{validate}{One or more of \code{values} indicating which list group items
to mark with \code{state}, defaults to \code{NULL}. If \code{NULL} then the all list group
items are updated.}
}
\description{
A way of handling and outlining content as a list. List groups function
similarly to checkbox groups. A list group returns a reactive vector of the
values from its active (selected) list group items. List group items are
selected or unselected by clicking on them. While list groups may be used as
reactive inputs they may also be used to simply display content, in which
case do not specify a list group id.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          listGroupInput(
            id = "listgroup",
            labels = paste("Item", 1:5),
            values = 1:5,
            selected = 2
          )
        ),
        col(
          textOutput("selected")
        )
      )
    ),
    server = function(input, output) {
      output$selected <- renderText({
        paste0(input$listgroup, collapse = ", ")
      })
    }
  )
}

if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          listGroupInput(
            id = "sets",
            labels = c(
              "red, blue, yellow",
              "silver, gold, crystal",
              "sapphire, ruby, emerald"
            )
          )
        ),
        col(
          listGroupInput(
            id = "stub",
            labels = NULL
          )
        )
      )
    ),
    server = function(input, output) {
      observe({
        updateListGroupInput("stub", input$sets)
      })
    }
  )
}

if (interactive()) {
  teams <- c(
    `avengers` = "iron man, thor, hulk, ant-man, wasp",
    `west coast avengers` = "hawkeye, mockingbird, wonder man, tigra, iron man",
    `new avengers` = "luke cage, captain america, iron man, spider-man, spider-woman"
  )

  shinyApp(
    ui = container(
      row(
        col(
          listGroupInput(
            id = "state",
            labels = c("none", "success", "info", "warning", "danger"),
            values = list(NULL, "success", "info", "warning", "danger")
          )
        ),
        col(
          listGroupInput(
            id = "filter",
            labels = c("iron man", "wonder man", "spider-woman")
          )
        ),
        col(
          listGroupInput(
            id = "teams",
            labels = names(teams),
            values = teams
          )
        )
      )
    ),
    server = function(input, output) {
     observe({
       req(input$filter)

       validateListGroupInput(
         id = "teams",
         state = input$state,
         validate = grep(
           paste0(input$filter, collapse = "|"), teams, value = TRUE
         )
       )
     })
    }
  )
}

}
\seealso{
For more information on Bootstrap list groups please refer to the
\href{https://v4-alpha.getbootstrap.com/components/list-group/}{reference page}.
}
