% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-group.R
\name{listGroupThruput}
\alias{listGroupThruput}
\alias{listGroupItem}
\alias{renderListGroup}
\title{List group thruputs}
\usage{
listGroupThruput(id, ..., multiple = TRUE, flush = FALSE)

listGroupItem(..., value = NULL, selected = FALSE, disabled = FALSE)

renderListGroup(..., env = parent.frame())
}
\arguments{
\item{id}{A character vector specifying the reactive id of the list group
thruput.}

\item{...}{For \code{listGroupThruput()}, additional named arguments passed on as
HTML attributes to the parent list group element.

For \code{listGroupItem()}, the text or HTML content of the list group item.

For \code{renderListGroup()}, any number of expressions which return a
\code{listGroupItem()} or calls to \code{listGroupItem()}.}

\item{multiple}{One of \code{TRUE} or \code{FALSE} specifyng if multiple list group
items may be selected, defaults to \code{TRUE}.}

\item{flush}{One of \code{TRUE} or \code{FALSE} specifying if the list group is
rendered without a border, defaults to \code{FALSE}. Removing the list group
border is useful when rendering a list group inside a custom parent
container, e.g. inside a \code{card()}.}

\item{value}{A character string specifying the value of the list group item,
defaults to \code{NULL}, in which case the list group item has no value. List
group items without a value are not actionable, i.e. they cannot be
selected.}

\item{selected}{\code{TRUE} or \code{FALSE} specifying if the list group item is
selected by default, defaults to \code{FALSE}.}

\item{disabled}{\code{TRUE} or \code{FALSE} specifying if the list group item can be
selected, defaults to \code{FALSE}.}

\item{env}{The environment in which to evalute the expressions based to
\code{renderListGroup()}.}
}
\description{
A way of handling and outlining content as a list. List groups function
similarly to checkbox groups. A list group returns a reactive vector of
values from its active (selected) list group items. List group items are
selected or unselected by clicking on them.
}
\examples{

### Getting started

listGroupThruput(
  id = NULL,
  listGroupItem(
    rangeInput(NULL)
  )
)

### Fancier list items

lessons <- list(
  stars = c(
    "The stars and moon are far too bright",
    "Their beam and smile splashing o'er all",
    "To illuminate while turning my sight",
    "From the shadows wherein deeper shadows fall"
  ),
  joy = c(
    "A single step, her hand aloft",
    "More than a step, a joyful bound",
    "The moment, precious, small, soft",
    "And within a truth was found"
  )
)

listGroupThruput(
  id = NULL,
  multiple = FALSE,
  listGroupItem(
    value = "stars",
    h5("Stars"),
    lessons[["stars"]][1]
  ),
  listGroupItem(
    value = "joy",
    h5("Joy"),
    lessons[["joy"]][1]
  )
)

}
\seealso{
Other thruputs: \code{\link{tableThruput}}
}
\concept{thruputs}
