% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{loginInput}
\alias{loginInput}
\title{Login input}
\usage{
loginInput(id, ...)
}
\description{
Username
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          loginInput(
            id = "login"
          )
        ),
        col(
          verbatimTextOutput("value")
        )
      )
    ),
    server = function(input, output) {
      output$value <- renderPrint({
        input$login
      })
    }
  )
}
}
