% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{loginInput}
\alias{loginInput}
\title{Login input}
\usage{
loginInput(id, ...)
}
\arguments{
\item{...}{Additional named arguments passed as HTML attibutes to the login
input.}
}
\description{
A complex input which consists of a username field and a password field.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          loginInput(
            id = "login"
          )
        ),
        col(
          verbatimTextOutput("value")
        )
      )
    ),
    server = function(input, output) {
      output$value <- renderPrint({
        input$login
      })
    }
  )
}

}
