% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu.R
\name{menuInput}
\alias{menuInput}
\title{Menu input}
\usage{
menuInput(id, label, choices, values = choices, ...,
  direction = "down", align = "left")
}
\arguments{
\item{id}{A character string specifying the reactive id of the input.}

\item{label}{A character string or tag element specifying the label of the
menu's toggle button.}

\item{choices}{A character vector specifying the choice text of the menu's
items.}

\item{values}{A character vector specifying the values of the menu's items,
defaults to \code{choices}.}

\item{...}{Additional named arguments passed as HTML attributes to the parent element.}

\item{direction}{One of \code{"up"}, \code{"right"}, \code{"down"}, or \code{"left"} specifying
which direction the menu opens, defaults to \code{"down"}.}

\item{align}{One or \code{"right"} or \code{"left"} specifying which side of the
toggle button the menu aligns to, defaults to \code{"left"}.`}
}
\description{
A togglable dropdown menu input. Menu inputs may be used as standalone
reactive inputs or within a \code{\link[=navInput]{navInput()}}. When used within a nav input the
\code{id} argument is ignored and may be set to \code{NULL}. For building custom,
more complex dropdown elements please see \code{\link[=dropdown]{dropdown()}}.
}
\examples{

### A simple menu

menuInput(
  id = "menu1",
  label = "Menu",
  choices = c(
    "Choice 1",
    "Choice 2",
    "Choice 3"
  )
)

### Use in navigation

navInput(
  id = "nav1",
  choice = list(
    "Tab 1",
    menuInput(
      id = NULL,  # <- ignored
      label = "Tab 2",
      choices = c(
        "Option 1",
        "Option 2",
        "Option 3"
      )
    ),
    "Tab 3",
    "Tab 4"
  )
)

}
\seealso{
Other inputs: \code{\link{buttonGroupInput}},
  \code{\link{buttonInput}}, \code{\link{checkboxInput}},
  \code{\link{dateInput}}, \code{\link{fileInput}},
  \code{\link{formInput}}, \code{\link{groupInput}},
  \code{\link{listGroupInput}}, \code{\link{navInput}},
  \code{\link{radioInput}}, \code{\link{rangeInput}},
  \code{\link{selectInput}}, \code{\link{textInput}}
}
\concept{inputs}
