% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-menu.R
\name{menuInput}
\alias{menuInput}
\alias{updateMenuInput}
\title{Menu inputs}
\usage{
menuInput(
  id,
  label,
  choices = NULL,
  values = choices,
  selected = NULL,
  ...,
  direction = "down",
  align = "left"
)

updateMenuInput(
  id,
  label = NULL,
  choices = NULL,
  values = choices,
  selected = NULL,
  enable = NULL,
  disable = NULL,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{A character string. The id of the reactive input.}

\item{label}{A character string or tag element specifying the label of the
menu's toggle button.}

\item{choices}{A character vector specifying the choice text of the menu's
items.}

\item{values}{A character vector specifying the values of the menu's items,
defaults to \code{choices}.}

\item{...}{Optional named arguments specifying HTML attributes for the input
element.}

\item{direction}{One of \code{"up"}, \code{"right"}, \code{"down"}, or \code{"left"} specifying
which direction the menu opens, defaults to \code{"down"}.}

\item{align}{One or \code{"right"} or \code{"left"} specifying which side of the
toggle button the menu aligns to, defaults to \code{"left"}.`}

\item{disable}{A character vector. The values disabled by default.}

\item{session}{A shiny session object.}
}
\description{
A toggleable dropdown menu input. Menu inputs may be used as standalone
reactive inputs or within a \code{\link[=navInput]{navInput()}}. For building custom, more complex
dropdown elements please see \code{\link[=dropdown]{dropdown()}}.
}
\seealso{
Other inputs: 
\code{\link{fileInput}()},
\code{\link{input_button}()},
\code{\link{input_checkbox}()},
\code{\link{input_checkbox_button}()},
\code{\link{input_chip}()},
\code{\link{input_form}()},
\code{\link{listGroupInput}()},
\code{\link{radioInput}()},
\code{\link{radiobarInput}()},
\code{\link{rangeInput}()},
\code{\link{selectInput}()},
\code{\link{textInput}()}
}
\concept{inputs}
