% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modal}
\alias{modal}
\title{Modal dialogs}
\usage{
modal(header = NULL, body = NULL, footer = NULL, ...)
}
\arguments{
\item{header}{A character vector specifying a title for the modal or custom
tags to use as a header, defaults to \code{NULL}. When creating a custom header
title elements may need to include the \code{"modal-title"} HTML class.}

\item{body}{A character vectory specifying the main text of the modal or
custom tags to use as for the body of the modal, defaults to \code{NULL}.}

\item{footer}{Custom tags to include at the bottom of the modal, defaults to
\code{NULL}.}

\item{...}{Additional named arguments passed as HTML attributes to the parent
element.}
}
\description{
Create a modal that can by shown or hidden by \code{toggleModal}. Modals may be
built and included anywhere on the page. Unlike shiny modals, dull modals
must be defined ahead of time, but may be dynamically updated during the life
cycle of the shiny app by passing, for example, a \code{textOutput} as part of the
modal body.
}
\examples{
if (interactive()) {
  library(shiny)

  shinyApp(
    ui = container(
      button(id = "toggle", "Click to show modal"),
      modal(
        id = "hello",
        header = "Hello, world!",
        body = "A simple modal."
      )
    ),
    server = function(input, output) {
      observeEvent(input$toggle, {
        toggleModal("hello")
      })
    }
  )
}
}
