% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modal}
\alias{modal}
\alias{showModal}
\alias{closeModal}
\title{Modal dialogs}
\usage{
modal(
  id,
  ...,
  header = NULL,
  footer = NULL,
  center = FALSE,
  size = "md",
  fade = TRUE
)

showModal(modal, session = getDefaultReactiveDomain())

closeModal(id = NULL, session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{A character string specifying the id of the modal, when closed
\code{input[[id]]} is set to \code{TRUE}.}

\item{...}{Unnamed values passed as tag elements to the body of the modal.
or named values passed as HTML attributes to the body element of the
modal.}

\item{header}{A character string or tag element specifying the header of the
modal.}

\item{footer}{A character string or tag element specifying the footer of the
modal.}

\item{center}{One of \code{TRUE} or \code{FALSE} specifying whether the modal is
vertically centered on the page, defaults to \code{FALSE}.}

\item{size}{One of \code{"sm"} (small), \code{"md"} (medium), \code{"lg"} (large), or \code{"xl"}
(extra large) specifying the relative width of the modal, defaults to
\code{"md"}.}

\item{fade}{One of \code{TRUE} or \code{FALSE} specifying if the modal fades in when
shown and fades out when closed, defaults to \code{TRUE}.}

\item{modal}{A modal tag element created using \code{modal()}.}

\item{session}{A reactive context, defaults to \code{\link[=getDefaultReactiveDomain]{getDefaultReactiveDomain()}}.}
}
\description{
Modals are a flexible alert window, which disable interaction with the page
behind them. Modals may include inputs, buttons, or simply text. Each modal
may be assigned an \code{id}. By default \code{hideModal()} will hide all modals, but
you may instead specify a modal's \code{id} in which case only that modal is
closed. Additionally, when \code{id} is not \code{NULL} observers and reactives may
watch for the modal's close event.
}
\details{
\subsection{Getting started}{\if{html}{\out{<div class="r">}}\preformatted{modal(
  id = "modal1",
  header = h5("Header"),
  p("Cras placerat accumsan nulla.")
)
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve-->
}}
\if{html}{\out{
<div class="r-output">
}}
\if{html}{\out{
<div id="modal1" class="yonder-modal modal fade" tabindex="-1" role="dialog">
}}
\if{html}{\out{
<div class="modal-dialog" role="document">
}}
\if{html}{\out{
<div class="modal-content">
}}
\if{html}{\out{
<div class="modal-header">
}}
\if{html}{\out{
<h5 class="modal-title">
}}


Header\if{html}{\out{
</h5>
}}
\if{html}{\out{
<button type="button" class="close" data-dismiss="modal" aria-label="Close">
}}


\if{html}{\out{<span aria-hidden="true">}}×\if{html}{\out{</span>}}\if{html}{\out{
</button>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="modal-body">
}}
\if{html}{\out{
<p>
}}


Cras placerat accumsan nulla.\if{html}{\out{
</p>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<!--/html_preserve-->
}}
\if{html}{\out{
<!--html_preserve-->
}}
\if{html}{\out{
<div class="r-output">
}}
\if{html}{\out{
<button data-toggle="modal" data-target="#modal1">
}}


Open modal\if{html}{\out{
</button>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<!--/html_preserve-->
}}

}

\subsection{Modals with containers}{\if{html}{\out{<div class="r">}}\preformatted{modal(
  id = "modal2",
  size = "lg",
  header = h5("More complex"),
  container(
    columns(
      column("Cras placerat accumsan nulla."),
      column("Curabitur lacinia pulvinar nibh."),
      column(
        "Aliquam posuere.",
        "Praesent fermentum tempor tellus."
      )
    )
  )
)
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve-->
}}
\if{html}{\out{
<div class="r-output">
}}
\if{html}{\out{
<div id="modal2" class="yonder-modal modal fade" tabindex="-1" role="dialog">
}}
\if{html}{\out{
<div class="modal-dialog modal-lg" role="document">
}}
\if{html}{\out{
<div class="modal-content">
}}
\if{html}{\out{
<div class="modal-header">
}}
\if{html}{\out{
<h5 class="modal-title">
}}


More complex\if{html}{\out{
</h5>
}}
\if{html}{\out{
<button type="button" class="close" data-dismiss="modal" aria-label="Close">
}}


\if{html}{\out{<span aria-hidden="true">}}×\if{html}{\out{</span>}}\if{html}{\out{
</button>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="modal-body">
}}
\if{html}{\out{
<div class="container-fluid">
}}
\if{html}{\out{
<div class="row">
}}
\if{html}{\out{
<div class="col">
}}


Cras placerat accumsan nulla.\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="col">
}}


Curabitur lacinia pulvinar nibh.\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="col">
}}


Aliquam posuere. Praesent fermentum tempor tellus.\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<!--/html_preserve-->
}}
\if{html}{\out{
<!--html_preserve-->
}}
\if{html}{\out{
<div class="r-output">
}}
\if{html}{\out{
<button data-toggle="modal" data-target="#modal2">
}}


Open modal\if{html}{\out{
</button>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<!--/html_preserve-->
}}

}
}
\seealso{
Other components: 
\code{\link{alert}()},
\code{\link{badge}()},
\code{\link{blockquote}()},
\code{\link{card}()},
\code{\link{collapsePane}()},
\code{\link{d1}()},
\code{\link{dropdown}()},
\code{\link{img}()},
\code{\link{jumbotron}()},
\code{\link{navContent}()},
\code{\link{popover}()},
\code{\link{pre}()},
\code{\link{toast}()}
}
\concept{components}
