% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modal_body}
\alias{modal_body}
\alias{modal_title}
\alias{modal_header}
\alias{modal_close}
\alias{modal_footer}
\alias{modal_dismiss}
\title{Modal items}
\usage{
modal_body(...)

modal_title(...)

modal_header(..., close = modal_close())

modal_close()

modal_footer(..., close = modal_dismiss())

modal_dismiss(..., text = "Close")
}
\description{
Components of a modal dialog.
}
\section{Functions}{
\itemize{
\item \code{modal_body()}: The main content of a modal.

\item \code{modal_title()}: A title suitable for use in modal dialog.

\item \code{modal_header()}: Add a header to a modal dialog.

\item \code{modal_close()}: A button stylized as a large "X" used to close a modal
dialog, typically found in the modal header.

\item \code{modal_footer()}: Add a footer to a modal dialog.

\item \code{modal_dismiss()}: A button with text used to close a modal dialog,
typically found the modal footer.

}}
