% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modal_dialog}
\alias{modal_dialog}
\alias{modal_toggle}
\title{Modal dialogs}
\usage{
modal_dialog(
  id,
  ...,
  position = c("top", "center"),
  size = c("md", "sm", "lg", "xl", "fullscreen"),
  scroll = c("modal", "body"),
  backdrop = c("default", "static", "none"),
  wrapper = modal_body
)

modal_toggle(id, text, ...)
}
\arguments{
\item{id}{A string. The id of a modal. Use \verb{input$<id>} to query the state of
the modal.}

\item{...}{Any number of child elements passed to the parent element. Named
values are passed as HTML attributes to the parent element.}

\item{position}{A string. The position of the modal on the screen.}

\item{size}{A string. The size of the modal.}

\item{modal}{A modal tag element created using \code{modal()}.}
}
\value{
A \link[htmltools:builder]{htmltools::tag} object.
}
\description{
Modals are flexible alert windows, which disable interaction with the page
behind them. Modals may include inputs, buttons, or simply text.
}
\section{Functions}{
\itemize{
\item \code{modal_dialog()}: Construct a modal.

\item \code{modal_toggle()}: Open a modal.

}}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

modal_dialog(
  id = "example1",
  modal_header(modal_title("Example 1")),
  "Our first modal example.",
  "It's simple, but it gets the job done.",
  modal_footer()
)
\dontshow{\}) # examplesIf}
}
\seealso{
Other components: 
\code{\link{alert}()},
\code{\link{badge}()},
\code{\link{collapse_panel}()},
\code{\link{d1}()},
\code{\link{toast}()}
}
\concept{components}
