% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav.R
\name{nav}
\alias{nav}
\alias{navItem}
\alias{navDropdown}
\title{Page Navigation}
\usage{
nav(..., align = "left", width = NULL, vertical = FALSE, tabs = FALSE,
  pills = FALSE)

navItem(label = NULL, active = FALSE, disabled = FALSE, ...)

navDropdown(label = NULL, ...)
}
\arguments{
\item{...}{\code{navItem}s or \code{navDropdown}s passed to \code{nav}, \code{dropdownItem}s
passed to \code{navDropdown}, or named arguments passed as HTML attributes to
the respective parent element.}

\item{align}{One of \code{"left"}, \code{"right"}, \code{"center"}, or \code{"horizontal"}
specifying the alignment of the nav, defaults to \code{"left"}.}

\item{width}{If specified, one of \code{"fill"} or \code{"justify"} specifying how the
nav will fill its parent element, defaults to \code{NULL}. If \code{"justify"} all
nav links have equal width, when \code{"fill"} nav links are only equally
spaced.}

\item{vertical}{If \code{TRUE}, the nav is rendered vertically, defaults to
\code{FALSE}.}

\item{tabs}{If \code{TRUE}, the nav is rendered as a set of tabs for a tabbed
interface, defaults to \code{FALSE}.}

\item{pills}{If \code{TRUE}, \code{navItem}s are rendered as pills, defaults to
\code{FALSE}.}

\item{label}{A character vector specifying the label of a nav item,
defaults to \code{NULL}.}

\item{active}{If \code{TRUE}, the nav item renders in a an active state, defaults
to \code{FALSE}.}

\item{disabled}{If \code{TRUE}, the nav item renders in a disabled state and
will not receive click events, defaults to \code{FALSE}.}
}
\description{
Create a nav.
}
\details{
To center a horizontal nav specify \code{align = "horiztonal"} and \code{width = "fill"} or \code{width = "justify"}.
}
\examples{
nav(
  navItem("Home", href = "#home"),
  navItem("About", href = "#about")
)

# active link
nav(
  navItem("Home", href = "#home", active = TRUE),
  navItem("About", href = "#about")
)

# active tab
nav(
  tabs = TRUE,
  navItem("Home", href = "#home", active = TRUE),
  navItem("About", href = "#about"),
  navItem("Blog", href = "#blog")
)

if (interactive()) {
  shinyApp(
    ui = container(
      nav(
        navItem(
          label = "Home",
          active = TRUE
        ),
        navItem(
          label = "Posts"
        ),
        navItem(
          label = "About"
        ),
        navDropdown(
          dropdownItem()
        )
      )
    ),
    server = function(input, output) {

    }
  )
}

}
\seealso{
For more information about the Bootstrap nav please follow this
\href{http://v4-alpha.getbootstrap.com/components/navs/}{link}.
}
