% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav.R
\name{nav}
\alias{nav}
\title{Page Navigation}
\usage{
nav(content = NULL, align = "left", width = NULL, column = FALSE,
  tabs = FALSE, pills = FALSE)
}
\arguments{
\item{content}{A list of links, where links are constructed using \code{tags$a}.

To render a link in a disabled state name the corresponding list element(s)
\code{disabled}, see examples below.

When building a tabbed interface with \code{nav} a default tab may be specified
by naming an element \code{active}, see examples below.}

\item{align}{Alignment of the nav, one of \code{"left"}, \code{"right"}, \code{"center"}, or
\code{"horizontal"}, defaults to \code{"left"}.}

\item{width}{Used to indicate the nav should fill the width of its parent
element, one of \code{"fill"} or \code{"justify"}, defaults to \code{NULL}. If \code{"justify"}
all nav links have equal width, when \code{"fill"} nav links are only equally
spaced.}

\item{column}{If \code{TRUE}, the nav is rendered as a horizontal stack instead
of vertically, defaults to \code{FALSE}.}

\item{tabs}{If \code{TRUE}, the nav is rendered as a set of tabs for a tabbed
interface, defaults to \code{FALSE}. Specifying \code{TRUE} will also add a
\code{data-toggle} HTML attribute to the nav links allowing them to toggle tabs.}

\item{pills}{If \code{TRUE}, the nav is rendered as pills, defaults to \code{FALSE}.}
}
\description{
Create a nav.
}
\details{
To center a horizontal nav specify \code{align = "horiztonal"} and \code{width = "fill"} or \code{width = "justify"}.
}
\examples{
nav(
  list(
    tags$a(href = "#home", "Home"),
    tags$a(href = "#about", "About")
  )
)

# active link
nav(
  list(
    active = tags$a(href = "#home", "Home"),
    tags$a(href = "#about", "About")
  )
)

# active tab
nav(
  tabs = TRUE,
  list(
    active = tags$a(href = "#home", "Home"),
    tags$a(href = "#about", "About"),
    tags$a(href = "#blog", "Blog")
  )
)

}
\seealso{
For more information about the Bootstrap nav please follow this
\href{http://v4-alpha.getbootstrap.com/components/navs/}{link}.
}
