% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar.R
\name{navbar}
\alias{navbar}
\title{Page and content navigation}
\usage{
navbar(..., brand = NULL)
}
\arguments{
\item{...}{A tab toggle, inline forms, or text to add to include as part of
the navigation bar.}

\item{brand}{A tag element or text placed on the left end of the navbar,
defaults to \code{NULL}, in which case nothing is added.}
}
\description{
Add a navigation bar to your application with \code{navbar()}. Navigation bars may
include a tab toggle (useful for multi-page applications), inline forms
(perhaps a search feature or login item), or character strings to add simple
text. Navbars are responsive and will collapse on small screens, think mobile
device. A button is automatically added to toggle between the collapsed and
expanded states.
}
\examples{

if (interactive()) {
  shinyApp(
    ui = tagList(
      navbar(
        brand = "Navbar",
        tabTabs("myTabs", c("Home", "About", "Our process")) \%>\%
          margins(c(0, "auto", 0, 0)),
        formInput(
          inline = TRUE,
          id = "navForm",
          searchInput("search", placeholder = "Search") \%>\%
            margins(sm = c(0, 2, 0, 0)),
          submit = submitInput("Search") \%>\%
            background("amber")
        )
      ) \%>\%
        background("teal"),
      container(
        tabContent(
          tabs = "myTabs",
          tabPane(
            h3("Home")
          ),
          tabPane(
            h3("About")
          ),
          tabPane(
            h3("The process")
          )
        )
      )
    ),
    server = function(input, output) {

    }
  )
}

}
