% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar.R
\name{navbar}
\alias{navbar}
\title{Navigation bar}
\usage{
navbar(..., brand = NULL, collapse = NULL)
}
\arguments{
\item{...}{A tab toggle, inline forms, or text to add to include as part of
the navigation bar.}

\item{brand}{A tag element or text placed on the left end of the navbar,
defaults to \code{NULL}, in which case nothing is added.}

\item{collapse}{One of \code{"sm"}, \code{"md"}, \code{"lg"}, \code{"xl"}, or \code{NULL} specifying
the breakpoint at which the navbar collaspes, defaults to \code{NULL}, in which
case the navbar is always expanded.}
}
\description{
Add a navigation bar to your application with \code{navbar()}. Navigation bars may
include a tab toggle (useful for multi-page applications), inline forms
(perhaps a search feature or login item), or character strings to add simple
text. Navbars are responsive and will collapse on small screens, think mobile
device. A button is automatically added to toggle between the collapsed and
expanded states.
}
\details{
\subsection{Navbar with tabs}{\if{html}{\out{<div class="r">}}\preformatted{div(
  navbar(
    brand = "Navbar",
    navInput(
      .style \%>\%
        margin(right = "auto"),
      id = NULL,
      choices = c("Home", "About", "Our process")
    ),
    formInput(
      inline = TRUE,
      id = NULL,
      textInput(
        .style \%>\%
          margin(right = c(sm = 2)),
        type = "search",
        id = NULL,
        placeholder = "Search"
      ),
      formSubmit(
        .style \%>\%
          background("primary"),
        label = "Search",
        value = "search"
      )
    )
  ),
  container(
    p(". . .")
  )
)
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve-->
}}
\if{html}{\out{
<div class="r-output">
}}
\if{html}{\out{
<div>
}}
\if{html}{\out{
<nav class="navbar navbar-expand navbar-light">
}}


\if{html}{\out{<a class="navbar-brand" href="#">}}Navbar\if{html}{\out{</a>}}
\if{html}{\out{<button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#nav-content-433-345" aria-controls="nav-content-433-345" aria-expanded="false" aria-label="Toggle navigation">}}
\if{html}{\out{<span class="navbar-toggler-icon">}}\if{html}{\out{</span>}} \if{html}{\out{</button>}}\if{html}{\out{
<div id="nav-content-433-345" class="collapse navbar-collapse">
}}
\if{html}{\out{
<ul class="yonder-nav nav navbar-nav mr-auto">
}}
\if{html}{\out{
<li class="nav-item">
}}
\if{html}{\out{
<button class="nav-link btn btn-link active" value="Home">
}}


Home\if{html}{\out{
</button>
}}
\if{html}{\out{
</li>
}}
\if{html}{\out{
<li class="nav-item">
}}
\if{html}{\out{
<button class="nav-link btn btn-link" value="About">
}}


About\if{html}{\out{
</button>
}}
\if{html}{\out{
</li>
}}
\if{html}{\out{
<li class="nav-item">
}}
\if{html}{\out{
<button class="nav-link btn btn-link" value="Our process">
}}


Our process\if{html}{\out{
</button>
}}
\if{html}{\out{
</li>
}}
\if{html}{\out{
</ul>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</nav>
}}
\if{html}{\out{
<div class="container-fluid">
}}
\if{html}{\out{
<p>
}}


. . .\if{html}{\out{
</p>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<!--/html_preserve-->
}}

}
}
\seealso{
Other layout functions: 
\code{\link{column}()},
\code{\link{fieldset}()},
\code{\link{webpage}()}
}
\concept{layout functions}
