% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{navInput}
\alias{navInput}
\alias{navContent}
\alias{navPane}
\alias{showPane}
\title{Page navigation}
\usage{
navInput(id, items, values = items, ..., fill = FALSE,
  appearance = NULL)

navContent(...)

navPane(id, ...)

showPane(id, session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{A character string specifying the id of a pane.}

\item{values}{A character vector specifying custom values for each navigation
item, defaults to \code{labels}.}

\item{...}{For \strong{navContent}, calls to \code{navPane()} or named arguments passed
as HTML attributes to the parent element.

For \strong{navInput} and \strong{navPane}, any number of tag elements or named
arguments passed as HTML attributes to the parent element.}

\item{fill}{One of \code{TRUE} or \code{FALSE} specifying if the nav input fills the
width of its parent element. If \code{TRUE}, the space is divided evenly among
the nav items.}

\item{appearance}{One of \code{"pills"} or \code{"tabs"} specifying the appearance of
the nav input.}

\item{labels}{A character vector specifying the labels of the navigation
items.}
}
\description{
Create a set of tabs for controlling tab content. Tabs are separated from
content and panes for flexible placement. In the examples below you can see
how this allows tabs to be used inside \code{\link[=card]{card()}}s. The flexibility also allows
tabs to be added to navbars. When building tabbed content be sure to create a
pane for each label in your tabs and link tabs to content by the \code{id} and
\code{tabs} arguments, see below.
}
\section{App with pills}{
\preformatted{ui <- container(
  navInput(
    id = "tabs",
    items = paste("Tab", 1:3),
    values = paste0("pane", 1:3),
    appearance = "pills"
  ),
  navContent(
    navPane(
      id = "pane1",
      "Nullam tristique diam non turpis.",
      "Cum sociis natoque penatibus et magnis dis parturient montes, ",
      "nascetur ridiculus mus.",
      "Etiam laoreet quam sed arcu.",
      "Curabitur vulputate vestibulum lorem."
    ),
    navPane(
      id = "pane2",
      "Praesent fermentum tempor tellus.",
      "Proin neque massa, cursus ut, gravida ut, lobortis eget, lacus.",
      "Phasellus lacus.",
      "Nam euismod tellus id erat."
    ),
    navPane(
      id = "pane3",
      "Nullam eu ante vel est convallis dignissim.",
      "Phasellus at dui in ligula mollis ultricies.",
      "Fusce suscipit, wisi nec facilisis facilisis, est dui ",
      "fermentum leo, quis tempor ligula erat quis odio.",
      "Donec hendrerit tempor tellus."
    )
  )
)

server <- function(input, output) {
  observeEvent(input$tabs, {
    showPane(input$tabs)
  })
}

shinyApp(ui, server)
}
}

\section{App with dropdown}{
\preformatted{ui <- container(
  navInput(
    id = "tabs",
    items = list(
      "Tab 1",
      dropdown(
        label = "Tab 2",
        buttonInput("action", "Action"),
        buttonInput("another", "Another action")
      ),
      "Tab 3"
    ),
    values = paste0("pane", 1:3),
    appearance = "tabs"
  ),
  navContent(
    navPane(
      id = "pane1",
      "Donec at pede.",
      "Pellentesque tristique imperdiet tortor.",
      "Lorem ipsum dolor sit amet, consectetuer adipiscing elit."
    ),
    navPane(
      id = "pane2",
      "Nullam tristique diam non turpis.",
      "Cras placerat accumsan nulla.",
      "Donec at pede."
    ),
    navPane(
      id = "pane3",
      "Phasellus purus.",
      "Etiam laoreet quam sed arcu.",
      "Donec pretium posuere tellus."
    )
  )
)

server <- function(input, output) {
  observeEvent(input$tabs, {
    showPane(input$tabs)
  })

  observeEvent(c(input$action, input$another), {
    if (input$action > 0 || input$another > 0) {
      showPane("pane2")
    }
  })
}

shinyApp(ui, server)
}
}

\section{App with multiple sets of panes}{
\preformatted{ui <- container(
  navInput(
    id = "tabs",
    items = paste("Tab", 1:3),
    values = paste0("pane", 1:3)
  ),
  row(
    column(
      navContent(
        navPane(
          id = "pane1",
          "Aenean eu leo quam. Pellentesque ornare sem lacinia quam ",
          "venenatis vestibulum. Praesent commodo cursus magna, vel ",
          "scelerisque nisl consectetur et. Vivamus sagittis lacus vel ",
          "augue laoreet rutrum faucibus dolor auctor."
        ),
        navPane(
          id = "pane2",
          "Nullam quis risus eget urna mollis ornare vel eu leo. ",
          "Maecenas faucibus mollis interdum. Praesent commodo cursus ",
          "magna, vel scelerisque nisl consectetur et."
        ),
        navPane(
          id = "pane3",
          "Lorem ipsum dolor sit amet, consectetur adipiscing elit. ",
          "Vivamus sagittis lacus vel augue laoreet rutrum faucibus ",
          "dolor auctor. Etiam porta sem malesuada magna mollis euismod."
        )
      )
    ),
    column(
      navContent(
        navPane(
          id = "pane1",
          "Aenean eu leo quam. Pellentesque ornare sem lacinia quam ",
          "venenatis vestibulum. Praesent commodo cursus magna, vel ",
          "scelerisque nisl consectetur et. Vivamus sagittis lacus vel ",
          "augue laoreet rutrum faucibus dolor auctor."
        ),
        navPane(
          id = "pane2",
          "Nullam quis risus eget urna mollis ornare vel eu leo. ",
          "Maecenas faucibus mollis interdum. Praesent commodo cursus ",
          "magna, vel scelerisque nisl consectetur et."
        ),
        navPane(
          id = "pane3",
          "Lorem ipsum dolor sit amet, consectetur adipiscing elit. ",
          "Vivamus sagittis lacus vel augue laoreet rutrum faucibus ",
          "dolor auctor. Etiam porta sem malesuada magna mollis euismod."
        )
      )
    )
  )
)

server <- function(input, output) {
  observeEvent(input$tabs, {
    showPane(input$tabs)
  })
}

shinyApp(ui, server)
}
}

\examples{

### Nav styled as tabs

navInput(
  id = "tabs",
  items = c(
    "Tab 1",
    "Tab 2",
    "Tab 3"
  ),
  appearance = "tabs"
)

### Nav styled as pills

navInput(
  id = "tabs",
  items = paste("Tab", 1:3),
  appearance = "pills"
)

### Nav with dropdown

navInput(
  id = "tabs",
  items = list(
    "Tab 1",
    dropdown(
      label = "Tab 2",
      buttonInput("action", "Action"),
      buttonInput("another", "Another action")
    ),
    "Tab 2"
  )
)

}
\seealso{
Other inputs: \code{\link{addressInput}},
  \code{\link{buttonGroupInput}},
  \code{\link{buttonInput}}, \code{\link{checkboxInput}},
  \code{\link{dateInput}}, \code{\link{fileInput}},
  \code{\link{formInput}}, \code{\link{groupInput}},
  \code{\link{loginInput}}, \code{\link{radioInput}},
  \code{\link{rangeInput}}, \code{\link{selectInput}},
  \code{\link{textInput}}
}
\concept{inputs}
