% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav.R
\name{navInput}
\alias{navInput}
\title{Page navigation}
\usage{
navInput(id, choices, values = choices, selected = NULL, ...,
  appearance = "links", fill = FALSE)
}
\arguments{
\item{id}{A character string specifying the reactive id of the input.}

\item{choices}{A character vector or list of tag elements specifying the
navigation items of the navigation input.}

\item{values}{A character vector specifying custom values for each navigation
item, defaults to \code{choices}.}

\item{selected}{One of \code{values} specifying which choice is selected by
default, defaults to \code{NULL}.}

\item{...}{Additional named arguments passed as HTML attributes to the parent element.}

\item{appearance}{One of \code{"links"}, \code{"pills"}, or \code{"tabs"} specifying the
appearance of the nav input, defaults to \code{"links"}.}

\item{fill}{One of \code{TRUE} or \code{FALSE} specifying if the nav input fills the
width of its parent element. If \code{TRUE}, the space is divided evenly among
the nav items.}
}
\description{
A reactive input styled as a navigation control. The navigation input can be
styled as links, tabs, or pills. A nav input is paired with \code{\link[=navContent]{navContent()}}
and \code{\link[=showPane]{showPane()}} to create tabbed user interfaces.
}
\section{Reactive value}{


A click on a navigation link triggers a single character string.

A click on a dropdown item triggers a character vector of two values, the
value of the nav link and the value of the dropdown item.
}

\examples{

### Nav styled as tabs

navInput(
  id = "tabs1",
  choices = c(
    "Tab 1",
    "Tab 2",
    "Tab 3"
  ),
  appearance = "tabs"
)

### Nav styled as pills

navInput(
  id = "tabs2",
  choices = paste("Tab", 1:3),
  appearance = "pills"
)

### Nav with dropdown

navInput(
  id = "tabs3",
  choices = list(
    "Tab 1",
    menuInput(
      id = NULL,  # <- ignored
      label = "Tab 2",
      choices = c(
        "Action",
        "Another action"
      )
    ),
    "Tab 2"
  ),
  values = c("tab1", "tab2", "tab3")
)

### Full width nav input

navInput(
  id = "tabs4",
  choices = paste("Tab", 1:5),
  values = paste0("tab", 1:5),
  appearance = "pills",
  fill = TRUE
)

### Centering a nav input

navInput(
  id = "tabs5",
  choices = paste("Tab", 1:3)
) \%>\%
  flex(justify = "center")

}
\seealso{
Other inputs: \code{\link{buttonGroupInput}},
  \code{\link{buttonInput}}, \code{\link{checkboxInput}},
  \code{\link{dateInput}}, \code{\link{fileInput}},
  \code{\link{formInput}}, \code{\link{groupInput}},
  \code{\link{listGroupInput}}, \code{\link{menuInput}},
  \code{\link{radioInput}}, \code{\link{rangeInput}},
  \code{\link{selectInput}}, \code{\link{textInput}}
}
\concept{inputs}
