% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav.R
\name{navInput}
\alias{navInput}
\title{Page navigation}
\usage{
navInput(id, items, values = items, ..., fill = FALSE,
  appearance = NULL)
}
\arguments{
\item{id}{A character string specifying the id of a pane.}

\item{values}{A character vector specifying custom values for each navigation
item, defaults to \code{labels}.}

\item{...}{Any number of tag elements or named arguments passed as HTML
attributes to the parent element.}

\item{fill}{One of \code{TRUE} or \code{FALSE} specifying if the nav input fills the
width of its parent element. If \code{TRUE}, the space is divided evenly among
the nav items.}

\item{appearance}{One of \code{"pills"} or \code{"tabs"} specifying the appearance of
the nav input.}

\item{labels}{A character vector specifying the labels of the navigation
items.}
}
\description{
A reactive input styled as a navigation control. The navigation input can be
styled as links, tabs, or pills. A nav input is paired with \code{\link[=navContent]{navContent()}}
and \code{\link[=showPane]{showPane()}} to create tabbed user interfaces.
}
\examples{

### Nav styled as tabs

navInput(
  id = "tabs",
  items = c(
    "Tab 1",
    "Tab 2",
    "Tab 3"
  ),
  appearance = "tabs"
)

### Nav styled as pills

navInput(
  id = "tabs",
  items = paste("Tab", 1:3),
  appearance = "pills"
)

### Nav with dropdown

navInput(
  id = "tabs",
  items = list(
    "Tab 1",
    dropdown(
      label = "Tab 2",
      buttonInput("action", "Action"),
      buttonInput("another", "Another action")
    ),
    "Tab 2"
  )
)

}
\seealso{
Other inputs: \code{\link{addressInput}},
  \code{\link{buttonGroupInput}},
  \code{\link{buttonInput}}, \code{\link{checkboxInput}},
  \code{\link{dateInput}}, \code{\link{fileInput}},
  \code{\link{formInput}}, \code{\link{groupInput}},
  \code{\link{loginInput}}, \code{\link{radioInput}},
  \code{\link{rangeInput}}, \code{\link{selectInput}},
  \code{\link{textInput}}
}
\concept{inputs}
