% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design-spacing.R
\name{padding}
\alias{padding}
\title{Padding}
\usage{
padding(x, all = NULL, top = NULL, right = NULL, bottom = NULL, left = NULL)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{all, top, right, bottom, left}{A \link{responsive} argument. One of \code{0:5} or
\code{"auto"} specifying padding for one or more sides of the tag element. 0
removes all inner space and 5 adds the most space.}
}
\description{
The \code{padding} utility changes the inner spacing of a tag element. The padding
of a tag element is the space between the tag element's border and its
content or child elements. All arguments default to \code{NULL}, in which case
they are ignored.
}
\examples{

div() \%>\% padding(2)
# ->
div() \%>\% cascadess::padding_all(2)

}
\keyword{internal}
