% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{padding}
\alias{padding}
\alias{margin}
\title{Tag element margin and padding}
\usage{
padding(.tag, top = NULL, right = NULL, bottom = NULL, left = NULL)

margin(.tag, top = NULL, right = NULL, bottom = NULL, left = NULL)
}
\arguments{
\item{.tag}{A tag element.}

\item{top}{A \link{responsive} argument. One of \code{0:5} or \code{"auto"}. 0 removes all
space and 5 adds the most space.

If an \strong{unnamed} value is passed as \code{top} \code{margin()} and \code{padding()} will
apply the spcified spacing to \strong{all} sides.}

\item{right, bottom, left}{A \link{responsive} argument. One of \code{0:5} or \code{"auto"}. 0
removes all space and 5 adds the most space.}
}
\description{
Use the \code{margin()} and \code{padding()} utilities to change the margin or padding
of a tag element.  The margin of a tag element is the space outside and
around the tag element, its border, and its content.  The padding of a tag
element is the space between the tag element's border and its content or
child elements.
}
\section{Centering an element}{


In most modern browsers you want to horizontally center a tag element
using the \link{flex} layout. Alternatively, you can horizontally center an element
using \code{margin(<TAG>, right = "auto", left = "auto")}.\preformatted{div("Nam a sapien. Integer placerat tristique nisl.") \%>\%
  width(50) \%>\%
  height(25) \%>\%
  margin(top = 2, r = "auto", b = 2, l = "auto") \%>\%
  padding(3) \%>\%
  background("amber")
)
}
}

\section{Building an inline form}{


Inline form elements automatically use of the flex layout providing you a
means of creating condensed sets of inputs. However you may need to adjust
the spacing of the form's child elements.

Here is an inline form without any additional spacing applied.\preformatted{formInput(
  id = NULL,
  inline = TRUE,
  textInput(id = NULL, placeholder = "Sam Vimes"),
  groupInput(id = NULL, right = "@", placeholder = "Username"),
  checkboxInput(id = NULL, choice = "Remember me")
  )
)
}

Not great. But, with some styling we can make this form sparkle. Notice
we are also adjusting the default submit button added to the form input.\preformatted{formInput(
  id = NULL,
  inline = TRUE,
  textInput(id = NULL, placeholder = "Sam Vimes") \%>\%
    margin(r = c(sm = 2), b = 2),
  groupInput(id = NULL, right = "@", placeholder = "Username") \%>\%
    margin(r = c(sm = 2), b = 2),
  checkboxInput(id = NULL, choice = "Remember me") \%>\%
    margin(r = c(sm = 2), b = 2),
  submit = submitInput() \%>\%
    margin(b = 2)
)
}

Now we're cooking with gas!
}

\examples{

### Getting started

padding(div(), c(xs = 0, sm = 2, lg = 4))

margin(div(), right = c(md = "auto"))

margin(div(), bottom = 3, left = 1)

div(
  div() \%>\%
    margin(4) \%>\%
    padding(4) \%>\%
    background("grey") \%>\%
    border()
)

}
\seealso{
Other design: \code{\link{affix}},
  \code{\link{background}}, \code{\link{border}},
  \code{\link{display}}, \code{\link{float}},
  \code{\link{font}}, \code{\link{rounded}},
  \code{\link{scroll}}, \code{\link{shadow}},
  \code{\link{width}}

Other design: \code{\link{affix}},
  \code{\link{background}}, \code{\link{border}},
  \code{\link{display}}, \code{\link{float}},
  \code{\link{font}}, \code{\link{rounded}},
  \code{\link{scroll}}, \code{\link{shadow}},
  \code{\link{width}}
}
\concept{design}
