% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style-spacing.R
\name{padding}
\alias{padding}
\title{Padding}
\usage{
padding(x, all = NULL, top = NULL, right = NULL, bottom = NULL, left = NULL)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{all, top, right, bottom, left}{A \link{responsive} argument. One of \code{0:5} or
\code{"auto"} specifying padding for one or more sides of the tag element. 0
removes all inner space and 5 adds the most space.}
}
\description{
The \code{padding} utility changes the inner spacing of a tag element. The padding
of a tag element is the space between the tag element's border and its
content or child elements. All arguments default to \code{NULL}, in which case
they are ignored.
}
\seealso{
Other design utilities: 
\code{\link{active}()},
\code{\link{affix}()},
\code{\link{background}()},
\code{\link{border}()},
\code{\link{display}()},
\code{\link{flex}()},
\code{\link{float}()},
\code{\link{font}()},
\code{\link{height}()},
\code{\link{margin}()},
\code{\link{scroll}()},
\code{\link{shadow}()},
\code{\link{width}()}
}
\concept{design utilities}
