% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{padding}
\alias{padding}
\alias{margins}
\title{Element margins and padding}
\usage{
padding(tag, default = NULL, sm = NULL, md = NULL, lg = NULL,
  xl = NULL)

margins(tag, default = NULL, sm = NULL, md = NULL, lg = NULL,
  xl = NULL)
}
\arguments{
\item{default}{One of 0, 1, 2, 3, 4, 5 specifying the default margins or
padding to apply. If the margins and padding remain the same across all
viewports then only \code{default} needs to be specified.}

\item{sm}{Like \code{default}, but the margins or padding are applied once the
viewport is 576 pixels wide, think phone in landscape mode.}

\item{md}{Like \code{default}, but the margins or padding are applied once the
viewport is 768 pixels wide, think tablets.}

\item{lg}{Like \code{default}, but the margins or padding are applied once the
viewport is 992 pixels wide, think desktop.}

\item{xl}{Like \code{default}, but the margins or padding are applied once the
viewport is 1200 pixels wide, think large desktop.}
}
\description{
These functions are used in conjunction with \link{tagReduce} to change the
padding or margins of a tag. Each argument value may be a single value or a
vector of four values. Margins and padding are specified as 0, 1, 2, 3, 4, or
5. 0 removes any margins or padding. 5 is equivalent to \code{3rem} (\code{rem} is a
relative unit of measurement in css).

Specifying a single value changes the margin or padding of all four tag
sides. To apply different margins or padding to each side pass a vector of
four values. In this case, the first value is for the top side, second for
the left side, third for the bottom side, and the fourth value is for the
left side. As a wise help page once said, think "\strong{tr}ou\strong{bl}e" to help
remember the order.
}
\seealso{
Other utilities: \code{\link{affix}}, \code{\link{align}},
  \code{\link{border}}, \code{\link{float}},
  \code{\link{font}}, \code{\link{text}},
  \code{\link{width}}
}
