% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placeholder.R
\name{placeholder}
\alias{placeholder}
\alias{placeholder_button}
\alias{placeholder_block}
\title{Placeholder}
\usage{
placeholder(
  width = NULL,
  height = c("md", "xs", "sm", "lg"),
  container = htmltools::div
)

placeholder_button(width = NULL)

placeholder_block(
  lines = 3,
  width = NULL,
  height = c("md", "xs", "sm", "lg"),
  animate = c("none", "glow", "wave")
)
}
\arguments{
\item{width}{A valid CSS width or number. Numbers are shorthand for
percentages. The width of the placeholder.}

\item{height}{A string. The height of the placeholder.}

\item{container}{An \link[htmltools:builder]{htmltools::tag} function.}

\item{lines}{A number. The number of lines within the placeholder block.}

\item{animate}{A string. An animation for the placeholders within the block.}
}
\description{
A handy placeholder component.
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

placeholder()

placeholder(height = "lg")

card(
  class = "text-bg-info",
  card_body(
    card_title(placeholder()),
    placeholder_block(5)
  ),
  card_footer(
    placeholder_button(),
    placeholder(height = "xs")
  )
)

\dontshow{\}) # examplesIf}
}
