% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{pre}
\alias{pre}
\title{Scrollable code snippets}
\usage{
pre(...)
}
\arguments{
\item{...}{Text, tag elements, or named arguments passed as HTML attributes
to the tag.}
}
\description{
The \code{pre} function adds a maximum height and scroll bar to the standard
\code{<pre>} element.
}
\examples{

### Simple example

pre(
  "shinyApp(",
  "  ui = container(",
  "    row(",
  "      column(",
  "      ",
  "      )",
  "    )",
  "  )",
  "  server = function(input, output) {",
  "  ",
  "  }",
  ")"
)

}
\seealso{
Other content: \code{\link{alert}},
  \code{\link{blockquote}}, \code{\link{card}},
  \code{\link{d1}}, \code{\link{dropdown}},
  \code{\link{icon}}, \code{\link{img}},
  \code{\link{jumbotron}}, \code{\link{modal}},
  \code{\link{popover}}, \code{\link{progressOutlet}},
  \code{\link{tooltip}}
}
\concept{content}
