% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{pre}
\alias{pre}
\title{Scrollable code snippets}
\usage{
pre(...)
}
\arguments{
\item{...}{Text, tag elements, or named arguments passed as HTML attributes
to the tag.}
}
\description{
The \code{pre} function adds a maximum height and scroll bar to the standard
\verb{<pre>} element.
}
\examples{

### Simple example

pre(
  "shinyApp(",
  "  ui = container(",
  "    columns(",
  "      column(",
  "      ",
  "      )",
  "    )",
  "  )",
  "  server = function(input, output) {",
  "  ",
  "  }",
  ")"
)

}
\seealso{
Other components: 
\code{\link{alert}()},
\code{\link{badge}()},
\code{\link{blockquote}()},
\code{\link{card}()},
\code{\link{collapsePane}()},
\code{\link{d1}()},
\code{\link{dropdown}()},
\code{\link{img}()},
\code{\link{jumbotron}()},
\code{\link{modal}()},
\code{\link{navContent}()},
\code{\link{popover}()},
\code{\link{toast}()}
}
\concept{components}
