% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progressOutput}
\alias{progressOutput}
\alias{bar}
\alias{sendBar}
\title{Progress bars}
\usage{
progressOutput(...)

bar(id, value, label = NULL, striped = FALSE, ...)

sendBar(id, value, label = NULL, session = getDefaultReactiveDomain())
}
\arguments{
\item{...}{One or more \code{bar} elements passed to a progress output or named
arguments passed as HTML attributes to the parent element.}

\item{value}{An integer between 0 and 100 specifying the initial value
of a bar.}

\item{label}{A character string speciLabel(s) for a progress bar or components of a progress bar,
defaults to \code{NULL}.}

\item{striped}{If \code{TRUE}, the progress bar has a striped gradient, defaults
to \code{FALSE}.}
}
\description{
Create simple or composite progress bars. To create a composite progress bar
pass multiple calls to \code{bar} to a progress output. Each \code{bar} component has
its own id, value, label, and context.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          buttonInput(id = "inc", "Increment progress")
        ),
        col(
          progressOutput(
            bar(id = "clicks", value = 0) \%>\%
              background("light-blue")
          )
        )
      )
    ),
    server = function(input, output) {
      observeEvent(input$inc, {
        sendBar(
          id = "clicks",
          value = min(input$inc / 20 * 100, 100)
        )
      })
    }
  )
}

if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          progressOutput(
            bar(id = "faster", value = 0) \%>\%
              background("yellow"),
            bar(id = "slower", value = 0)
          )
        )
      )
    ),
    server = function(input, output) {
      observe({
        for (i in seq(from = 0, to = 50, by = 1)) {
          sendBar(
            id = "slower",
            value = i
          )

          sendBar(
            id = "faster",
            value = min(i * 3, 50)
          )

          Sys.sleep(0.1)
        }
      })
    }
  )
}

}
