% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-radiobar.R
\name{radiobarInput}
\alias{radiobarInput}
\alias{updateRadiobarInput}
\title{Radiobar inputs}
\usage{
radiobarInput(..., id, choices, values = choices, selected = values[[1]])

updateRadiobarInput(
  id,
  choices = NULL,
  values = choices,
  selected = NULL,
  enable = NULL,
  disable = NULL,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{choices}{A character vector or list of tag elements specifying the
labels of the input's choices.}

\item{values}{A vector specifying the values of the input's choices,
defaults to \code{choices}.}

\item{selected}{One of \code{values} specifying the input's default selected
choice, defaults to \code{values[[1]]}.}
}
\description{
A stylized group of radio inputs. A radiobar input is similar to a button
group, but with a checked or highlighted stated. Additionally, only one value
at most may be selected at any given time.
}
\seealso{
Other inputs: 
\code{\link{checkbarInput}()},
\code{\link{fileInput}()},
\code{\link{input_button}()},
\code{\link{input_checkbox}()},
\code{\link{input_checkbutton}()},
\code{\link{input_chip}()},
\code{\link{input_form}()},
\code{\link{listGroupInput}()},
\code{\link{menuInput}()},
\code{\link{radioInput}()},
\code{\link{rangeInput}()},
\code{\link{selectInput}()},
\code{\link{textInput}()}
}
\concept{inputs}
