% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-radiobar.R
\name{radiobarInput}
\alias{radiobarInput}
\alias{updateRadiobarInput}
\title{Radiobar inputs}
\usage{
radiobarInput(..., id, choices, values = choices, selected = values[[1]])

updateRadiobarInput(
  id,
  choices = NULL,
  values = choices,
  selected = NULL,
  enable = NULL,
  disable = NULL,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{...}{Additional named arguments passed as HTML attributes to the
parent element or tag elements passed as child elements to the parent
element.}

\item{id}{A character string specifying the id of the reactive input.}

\item{choices}{A character vector or list of tag elements specifying the
labels of the input's choices.}

\item{values}{A vector specifying the values of the input's choices,
defaults to \code{choices}.}

\item{selected}{One of \code{values} specifying the input's default selected
choice, defaults to \code{values[[1]]}.}

\item{enable}{One of \code{values} specifying particular choices to enable or
\code{TRUE} specifying the entire input is enabled, defaults to \code{NULL}.}

\item{disable}{One of \code{values} specifying particular choices to disable or
\code{TRUE} specifying the entire input is disabled, defaults to \code{NULL}.}

\item{session}{A reactive context, defaults to \code{\link[=getDefaultReactiveDomain]{getDefaultReactiveDomain()}}.}
}
\description{
A stylized group of radio inputs. A radiobar input is similar to a button
group, but with a checked or highlighted stated. Additionally, only one value
at most may be selected at any given time.
}
\details{
\subsection{Getting started}{\if{html}{\out{<div class="r">}}\preformatted{radiobarInput(
  .style \%>\%
    background("secondary"),
  id = NULL,
  choices = c("Choice 1", "Choice 2", "Choice 3")
)
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve-->
}}
\if{html}{\out{
<div class="r-output">
}}
\if{html}{\out{
<div class="yonder-radiobar btn-group btn-group-toggle d-flex bg-secondary" data-toggle="buttons">
}}
\if{html}{\out{
<label class="btn active">
<input type="radio" value="Choice 1" checked autocomplete="off"/> Choice
1 </label> <label class="btn">
<input type="radio" value="Choice 2" autocomplete="off"/> Choice 2
</label> <label class="btn">
<input type="radio" value="Choice 3" autocomplete="off"/> Choice 3
</label>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<!--/html_preserve-->
}}

}
}
\seealso{
Other inputs: 
\code{\link{buttonGroupInput}()},
\code{\link{buttonInput}()},
\code{\link{checkbarInput}()},
\code{\link{checkboxInput}()},
\code{\link{chipInput}()},
\code{\link{fileInput}()},
\code{\link{formInput}()},
\code{\link{listGroupInput}()},
\code{\link{menuInput}()},
\code{\link{navInput}()},
\code{\link{radioInput}()},
\code{\link{rangeInput}()},
\code{\link{selectInput}()},
\code{\link{textInput}()}
}
\concept{inputs}
