% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-radio.R
\name{radioInput}
\alias{radioInput}
\alias{updateRadioInput}
\title{Radio inputs}
\usage{
radioInput(
  ...,
  id,
  choices = NULL,
  values = choices,
  selected = values[[1]],
  inline = FALSE
)

updateRadioInput(
  id,
  choices = NULL,
  values = choices,
  selected = NULL,
  inline = FALSE,
  enable = NULL,
  disable = NULL,
  valid = NULL,
  invalid = NULL,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{...}{Additional named arguments passed as HTML attributes to the
parent element or tag elements passed as child elements to the parent
element.}

\item{id}{A character string specifying the id of the reactive input.}

\item{choices}{A character vector or list of tag elements specifying the
input's choices.}

\item{values}{A character vector, list of character strings, vector of values
to coerce to character strings, or list of values to coerce to character
strings specifying the values of the radio input's choices, defaults to
\code{choices}.}

\item{selected}{One of \code{values} indicating the default selected value of the
radio input, defaults to \code{NULL}, in which case the first choice is
selected by default.}

\item{inline}{If \code{TRUE}, the radio input renders inline, defaults to \code{FALSE},
in which case the radio controls render on separate lines.}

\item{enable}{One of \code{values} specifying particular choices to enable or
\code{TRUE} specifying the entire input is enabled, defaults to \code{NULL}.}

\item{disable}{One of \code{values} specifying particular choices to disable or
\code{TRUE} specifying the entire input is disabled, defaults to \code{NULL}.}

\item{valid}{A character string specifying a message to the user indicating
how the input's value is valid, defaults to \code{NULL.}}

\item{invalid}{A character string specifying a message to the user
indicating how the input's value is invalid, defaults to \code{NULL}.}

\item{session}{A reactive context, defaults to \code{\link[=getDefaultReactiveDomain]{getDefaultReactiveDomain()}}.}
}
\description{
A stylized radio input. A reactive input with multiple choices where only one
choice and value at most may be selected.
}
\details{
\subsection{Getting started}{\if{html}{\out{<div class="r">}}\preformatted{radioInput(
  id = NULL,
  choices = c("Choice 1", "Choice 2", "Choice 3")
)
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve-->
}}
\if{html}{\out{
<div class="r-output">
}}
\if{html}{\out{
<div class="yonder-radio">
}}
\if{html}{\out{
<div class="custom-control custom-radio">
}}
\if{html}{\out{
<input class="custom-control-input" type="radio" id="radio-433-345" value="Choice 1" checked autocomplete="off"/>
<label class="custom-control-label" for="radio-433-345">Choice 1</label>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="custom-control custom-radio">
}}
\if{html}{\out{
<input class="custom-control-input" type="radio" id="radio-665-165" value="Choice 2" autocomplete="off"/>
<label class="custom-control-label" for="radio-665-165">Choice 2</label>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="custom-control custom-radio">
}}
\if{html}{\out{
<input class="custom-control-input" type="radio" id="radio-332-323" value="Choice 3" autocomplete="off"/>
<label class="custom-control-label" for="radio-332-323">Choice 3</label>
}}
\if{html}{\out{
<div class="valid-feedback">
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="invalid-feedback">
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<!--/html_preserve-->
}}

}

\subsection{Inline input}{\if{html}{\out{<div class="r">}}\preformatted{radioInput(
  id = NULL,
  choices = c(
    "Choice 1",
    "Choice 2",
    "Choice 3"
  ),
  inline = TRUE  # <-
)
}\if{html}{\out{</div>}}\if{html}{\out{
<!--html_preserve-->
}}
\if{html}{\out{
<div class="r-output">
}}
\if{html}{\out{
<div class="yonder-radio">
}}
\if{html}{\out{
<div class="custom-control custom-radio custom-control-inline">
}}
\if{html}{\out{
<input class="custom-control-input" type="radio" id="radio-274-68" value="Choice 1" checked autocomplete="off"/>
<label class="custom-control-label" for="radio-274-68">Choice 1</label>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="custom-control custom-radio custom-control-inline">
}}
\if{html}{\out{
<input class="custom-control-input" type="radio" id="radio-187-927" value="Choice 2" autocomplete="off"/>
<label class="custom-control-label" for="radio-187-927">Choice 2</label>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="custom-control custom-radio custom-control-inline">
}}
\if{html}{\out{
<input class="custom-control-input" type="radio" id="radio-230-504" value="Choice 3" autocomplete="off"/>
<label class="custom-control-label" for="radio-230-504">Choice 3</label>
}}
\if{html}{\out{
<div class="valid-feedback">
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<div class="invalid-feedback">
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
</div>
}}
\if{html}{\out{
<!--/html_preserve-->
}}

}
}
\seealso{
Other inputs: 
\code{\link{buttonGroupInput}()},
\code{\link{buttonInput}()},
\code{\link{checkbarInput}()},
\code{\link{checkboxInput}()},
\code{\link{chipInput}()},
\code{\link{fileInput}()},
\code{\link{formInput}()},
\code{\link{listGroupInput}()},
\code{\link{menuInput}()},
\code{\link{navInput}()},
\code{\link{radiobarInput}()},
\code{\link{rangeInput}()},
\code{\link{selectInput}()},
\code{\link{textInput}()}
}
\concept{inputs}
