% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-range.R
\name{rangeInput}
\alias{rangeInput}
\alias{updateRangeInput}
\title{Range input}
\usage{
rangeInput(id, min = 0, max = 100, default = min, step = 1, ...)

updateRangeInput(
  id,
  value = NULL,
  enable = NULL,
  disable = NULL,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{min}{A number specifying the minimum value of the input, defaults to
\code{0}.}

\item{max}{A number specifying the maximum value of the input, defaults to
\code{100}.}

\item{default}{A number between \code{min} and \code{max} specifying the default value
of the input, defaults to \code{min}.}

\item{step}{A number specifying the interval step of the input, defaults to
\code{1}.}

\item{value}{A number specifying a new value for the input, defaults to
\code{NULL}.}
}
\description{
\code{rangeInput()} creates a simple numeric range input.
}
\details{
The sophistication of this input will improve as browsers adopt the latest
HTML standards.
}
\seealso{
Other inputs: 
\code{\link{checkbarInput}()},
\code{\link{fileInput}()},
\code{\link{input_button}()},
\code{\link{input_checkbox}()},
\code{\link{input_checkbutton}()},
\code{\link{input_chip}()},
\code{\link{input_form}()},
\code{\link{listGroupInput}()},
\code{\link{menuInput}()},
\code{\link{radioInput}()},
\code{\link{radiobarInput}()},
\code{\link{selectInput}()},
\code{\link{textInput}()}
}
\concept{inputs}
