% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{rounded}
\alias{rounded}
\title{Round tag element borders}
\usage{
rounded(.tag, sides = "all")
}
\arguments{
\item{.tag}{A tag element.}

\item{sides}{One of \code{"top"}, \code{"right"}, \code{"bottom"}, \code{"left"}, \code{"circle"},
\code{"all"} or \code{"none"}, defaults to \code{"all"}, specifying which and how the
the corners of the tag element are rounded.}
}
\description{
The \code{rounded} utility function applies Bootstrap classes to an element. The
styles are applied by sides, e.g. \code{"left"} or \code{"bottom"}. The \code{"circle"}
value heavily rounds all the corners of an element.
}
\examples{

### Different sides

sides <- c("top", "right", "bottom", "left", "circle", "all")

div(
  lapply(
    sides,
    rounded,
    .tag = div() \%>\%
      padding(5) \%>\%
      margin(2) \%>\%
      border("indigo")
  )
) \%>\%
  display("flex") \%>\%
  flex(wrap = TRUE)

}
\seealso{
Other design: \code{\link{affix}},
  \code{\link{background}}, \code{\link{border}},
  \code{\link{display}}, \code{\link{float}},
  \code{\link{font}}, \code{\link{padding}},
  \code{\link{scroll}}, \code{\link{shadow}},
  \code{\link{width}}
}
\concept{design}
