% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{rounded}
\alias{rounded}
\title{Round tag element borders}
\usage{
rounded(.tag, sides = "all")
}
\arguments{
\item{.tag}{A tag element.}
}
\description{
The \code{rounded} utility function applies Bootstrap classes to an element. The
styles are applied by sides, e.g. \code{"left"} or \code{"bottom"}. The \code{"circle"}
value heavily rounds all the corners of an element.
}
\examples{


}
\seealso{
Other design: \code{\link{affix}},
  \code{\link{background}}, \code{\link{border}},
  \code{\link{display}}, \code{\link{float}},
  \code{\link{font}}, \code{\link{height}},
  \code{\link{padding}}, \code{\link{scroll}},
  \code{\link{shadow}}, \code{\link{width}}
}
\concept{design}
