% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rounded}
\alias{rounded}
\title{Round tag element corners}
\usage{
rounded(.tag, sides = "all")
}
\arguments{
\item{.tag}{A tag element.}

\item{sides}{One of \code{"top"}, \code{"right"}, \code{"bottom"}, \code{"left"}, \code{"circle"},
\code{"all"} or \code{"none"}, defaults to \code{"all"}, specifying which and how the
the corners of the tag element are rounded.}
}
\description{
The \code{rounded} utility function applies Bootstrap classes to an element. The
styles are applied by sides, e.g. \code{"left"} or \code{"bottom"}. The \code{"circle"}
value heavily rounds all the corners of an element.
}
\examples{

}
\seealso{
Other utilities: \code{\link{affix}},
  \code{\link{background}}, \code{\link{border}},
  \code{\link{display}}, \code{\link{flex}},
  \code{\link{float}}, \code{\link{font}},
  \code{\link{markInvalid}}, \code{\link{padding}},
  \code{\link{sendModal}}, \code{\link{showPopover}},
  \code{\link{tooltip}}, \code{\link{updateChoices}},
  \code{\link{width}}
}
\concept{utilities}
