% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{scroll}
\alias{scroll}
\title{Vertical and horizontal scroll}
\usage{
scroll(.tag, direction = "y")
}
\arguments{
\item{.tag}{A tag element.}

\item{direction}{One of \code{"x"} or \code{"y"} specifying which direction to scroll
the tag's content, defaults to \code{"y"}, in which case vertical scroll is
applied.}
}
\description{
Many of the applications you build depsite a complex layout will still fit
onto a single page. To help scroll long content along side shorter content
use the \code{scroll()} utility function.
}
\examples{

### A simple scroll

div(
  lapply(
    rep("Integer placerat tristique nisl.", 20),
    p
  )
) \%>\%
  height(20) \%>\%
  border() \%>\%
  scroll()

}
\seealso{
Other design: \code{\link{affix}},
  \code{\link{background}}, \code{\link{border}},
  \code{\link{display}}, \code{\link{float}},
  \code{\link{font}}, \code{\link{height}},
  \code{\link{padding}}, \code{\link{shadow}},
  \code{\link{width}}
}
\concept{design}
