% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design-scroll.R, R/style-scroll.R
\name{scroll}
\alias{scroll}
\title{Vertical and horizontal scroll}
\usage{
scroll(tag, direction = "vertical")

scroll(tag, direction = "vertical")
}
\arguments{
\item{tag}{A tag element.}

\item{direction}{One of \code{"horizontal"} or \code{"vertical"} specifying which
direction to scroll overflowing content, defaults to \code{"vertical"}, in which
case the content may croll up and down.}
}
\description{
Many of the applications you build despite a complex layout will still fit
onto a single page. To help scroll long content alongside shorter content use
the \code{scroll()} utility function.

Many of the applications you build despite a complex layout will still fit
onto a single page. To help scroll long content alongside shorter content use
the \code{scroll()} utility function.
}
\seealso{
Other design utilities: 
\code{\link{active}()},
\code{\link{affix}()},
\code{\link{background}()},
\code{\link{border}()},
\code{\link{display}()},
\code{\link{flex}()},
\code{\link{float}()},
\code{\link{font}()},
\code{\link{height}()},
\code{\link{margin}()},
\code{\link{padding}()},
\code{\link{shadow}()},
\code{\link{width}()}

Other design utilities: 
\code{\link{active}()},
\code{\link{affix}()},
\code{\link{background}()},
\code{\link{border}()},
\code{\link{display}()},
\code{\link{flex}()},
\code{\link{float}()},
\code{\link{font}()},
\code{\link{height}()},
\code{\link{margin}()},
\code{\link{padding}()},
\code{\link{shadow}()},
\code{\link{width}()}
}
\concept{design utilities}
