% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{scroll}
\alias{scroll}
\title{Vertical and horizontal scroll}
\usage{
scroll(.tag, direction = "y")
}
\arguments{
\item{.tag}{A tag element.}

\item{direction}{One of \code{"x"} or \code{"y"} specifying which direction to scroll
the tag's content, defaults to \code{"y"}, in which case horizontal scroll is
applied.}
}
\description{
Many of the applications you build depsite a complex layout will still fit
onto a single page. To help scroll long content along side shorter content
use the \code{scroll()} utility function.
}
\examples{

div(
  lapply(
    rep("Integer placerat tristique nisl.", 20),
    p
  )
) \%>\%
  height(25) \%>\%
  scroll()

if (interactive()) {
  shinyApp(
    ui = container(
      row(
        column(
          width = 3,
          card(
            listGroupThruput(
              id = NULL,
              lapply(
                rep("In id erat non orci commodo lobortis.", 30),
                listGroupItem
              )
            )
          ) \%>\%
            height(100) \%>\%
            scroll()
        ),
        column(

        )
      )
    ),
    server = function(input, output) {

    }
  )
}

}
