% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{selectInput}
\alias{selectInput}
\alias{option}
\title{Select input}
\usage{
selectInput(..., id = NULL)

option(label = NULL, value = NULL, selected = FALSE)
}
\arguments{
\item{id}{A character string specifying an id for the select input, defaults
to \code{NULL}. When specified, a reactive value is available to the shiny
server.}

\item{label}{A character string specifying a label for the select input
option, defaults to \code{NULL}.}

\item{value}{A character string specifying a value for the select input
option, defaults to \code{NULL}.}

\item{selected}{If \code{TRUE}, the select option is selected when the page
renders, defaults to \code{FALSE}.}
}
\description{
A select input.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      formInline(
        label("Preference"),
        selectInput(
          id = "mySelect",
          option("Choose...", selected = TRUE),
          option("One", 1),
          option("Two", 2),
          option("Three", 3)
        ),
        checkboxInput(
          label = "Remember my preference"
        ),
        buttonInput(
          "Go!",
          context = "secondary"
        )
      )
    ),
    server = function(input, output) {

    }
  )

}


}
\seealso{
Other inputs: \code{\link{buttonInput}},
  \code{\link{groupInput}}, \code{\link{radioInput}},
  \code{\link{textInput}}
}
