% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-select.R
\name{selectInput}
\alias{selectInput}
\alias{updateSelectInput}
\alias{groupSelectInput}
\alias{updateGroupSelectInput}
\title{Select inputs}
\usage{
selectInput(
  id,
  choices = NULL,
  values = choices,
  selected = values[[1]],
  ...,
  placeholder = NULL
)

updateSelectInput(
  id,
  choices = NULL,
  values = choices,
  selected = choices[[1]],
  enable = NULL,
  disable = NULL,
  valid = NULL,
  invalid = NULL,
  session = getDefaultReactiveDomain()
)

groupSelectInput(
  id,
  choices,
  values = choices,
  selected = values[[1]],
  ...,
  left = NULL,
  right = NULL
)

updateGroupSelectInput(
  id,
  choices = NULL,
  values = choices,
  selected = NULL,
  enable = NULL,
  disable = NULL,
  valid = NULL,
  invalid = NULL,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{A character string. The id of the reactive input.}

\item{choices}{A character vector specifying the input's choices.}

\item{values}{A character vector specifying the values of the input's
choices, defaults to \code{choices}.}

\item{selected}{One of \code{values} indicating the default value of the input,
defaults to \code{values[[1]]}.}

\item{...}{Optional named arguments specifying HTML attributes for the input
element.}

\item{placeholder}{A character string specifying the placeholder text of
the select input, defaults to \code{NULL}.}

\item{disable}{A boolean. The checkbox starts disabled if \code{TRUE}.}

\item{session}{A shiny session object.}

\item{left, right}{A character vector specifying static addons or
\code{\link[=buttonInput]{buttonInput()}} or \code{\link[=dropdown]{dropdown()}} elements specifying dynamic addons. Addons
affect the reactive value of the group input, see the Details section below
for more information.

\strong{\code{left} is character or \code{right} is character}

If \code{left} or \code{right} are character vectors, then the group input functions
like a text input. The value will update and trigger a reactive event when
the text box is modified. The group input's reactive value is the
concatenation of the static addons specified by \code{left} or \code{right} and the
value of the text input.

\strong{\code{left} is button or \code{right} is button}

The button does not change the value of the group input. However, the input
no longer triggers event when the text box is updated. Instead the value is
updated when a button is clicked. Static addons are still applied to the
group input value.

\strong{\code{left} is a dropdown or \code{right} is a dropdown}

The value of the group input does chance depending on the clicked dropdown
menu item. The value of the input group is the concatenation of the
dropdown input value, the value of the text input, and any static addons.}
}
\description{
Create a select input. Select elements typically appear as a simple menu of
choices and may have one selected choice. A group select input is a select
input with one or two additional components. These addon components are used
to change the reactivity or value of the input, see Details for more
information.
}
\seealso{
Other inputs: 
\code{\link{fileInput}()},
\code{\link{input_button}()},
\code{\link{input_checkbox}()},
\code{\link{input_checkbox_group}()},
\code{\link{input_chip}()},
\code{\link{input_form}()},
\code{\link{input_list_group}()},
\code{\link{menuInput}()},
\code{\link{radioInput}()},
\code{\link{radiobarInput}()},
\code{\link{rangeInput}()},
\code{\link{textInput}()}
}
\concept{inputs}
