% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{selectInput}
\alias{selectInput}
\title{Select input}
\usage{
selectInput(id, choices, values = choices, selected = NULL,
  multiple = FALSE, ...)
}
\arguments{
\item{id}{A character string specifying the id of the select input.}

\item{choices}{A character vector specifying the labels of the select input
options.}

\item{values}{A character vector specifying the values of the select input
options, defaults to \code{chocies}.}

\item{selected}{One of \code{values} indicating the default value of the select
input, defaults to \code{NULL}. If \code{NULL} the first value is selected by
default.}

\item{multiple}{One of \code{TRUE} or \code{FALSE}, if \code{TRUE} multiple values may be
selected, otherwise a single value is selected at a time,
defaults to \code{FALSE}.}

\item{...}{Additional named arguments passed as HTML attributes to the parent
element.}
}
\description{
Create a select input. Select elements often appear as a dropdown menu and
may have one or more selected values, see \code{multiple}.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          selectInput(
            id = "select",
            choices = c("Choose one", "One", "Two", "Three"),
            values = list(NULL, 1, 2, 3),
            multiple = TRUE
          )
        ),
        col(
          d4(
            textOutput("value")
          )
        )
      )
    ),
    server = function(input, output) {
      output$value <- renderText({
        input$select
      })
    }
  )
}


}
\seealso{
Other inputs: \code{\link{buttonInput}},
  \code{\link{checkboxInput}}, \code{\link{formInput}},
  \code{\link{groupInput}}, \code{\link{radioInput}},
  \code{\link{textInput}}
}
