% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{selectInput}
\alias{selectInput}
\title{Select input}
\usage{
selectInput(id, choices, values = choices, selected = NULL,
  multiple = FALSE, ...)
}
\arguments{
\item{id}{A character string specifying the id of the reactive input.}

\item{choices}{A character vector specifying the labels of the select input
options.}

\item{values}{A character vector specifying the values of the select input
options, defaults to \code{chocies}.}

\item{selected}{One of \code{values} indicating the default value of the select
input, defaults to \code{NULL}. If \code{NULL} the first value is selected by
default.}

\item{multiple}{One of \code{TRUE} or \code{FALSE}, if \code{TRUE} multiple values may be
selected, otherwise a single value is selected at a time,
defaults to \code{FALSE}.}

\item{...}{Additional named arguments passed as HTML attributes to the parent
element.}
}
\description{
Create a select input. Select elements often appear as a dropdown menu and
may have one or more selected values, see \code{multiple}.
}
\examples{

### Getting started

selectInput(
  id = NULL,
  choices = c(
    "Choose one",
    "Choice 1",
    "Choice 2",
    "Choice 3"
  ),
  values = list(NULL, 1, 2, 3)
)

}
\seealso{
Other inputs: \code{\link{buttonGroupInput}},
  \code{\link{buttonInput}}, \code{\link{checkboxInput}},
  \code{\link{chipInput}}, \code{\link{fileInput}},
  \code{\link{formInput}}, \code{\link{groupInput}},
  \code{\link{listGroupInput}}, \code{\link{menuInput}},
  \code{\link{navInput}}, \code{\link{radioInput}},
  \code{\link{rangeInput}}, \code{\link{sliderInput}},
  \code{\link{textInput}}
}
\concept{inputs}
