% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{sendAlert}
\alias{sendAlert}
\title{Alerts}
\usage{
sendAlert(id, content, context = "secondary",
  session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{A character vector of id(s). Alerts are inserted above the elements
with these ids.}

\item{content}{A character string specifying the body of the alert or
a custom element to include as the body of the alert.}

\item{context}{One of \code{"primary"}, \code{"secondary"}, \code{"success"}, \code{"info"},
\code{"warning"}, \code{"danger"}, \code{"light"}, or \code{"dark"} specifying the visual
context of the alert, defaults to \code{"secondary"}.}

\item{session}{A \code{session} object passed to the shiny server function,
defaults to \code{\link[=getDefaultReactiveDomain]{getDefaultReactiveDomain()}}.}
}
\description{
Trigger one or more alerts. Use alerts to let the user know of successes
or to call attention to problems.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      buttonInput(id = "button", "A button")
    ),
    server = function(input, output) {
      observeEvent(input$button, {
        sendAlert(
          id = "button",
          content = "You clicked the button!",
          context = "warning"
        )
      })
    }
  )
}

}
\seealso{
Boostrap 4 alert documentation:
\url{https://getbootstrap.com/docs/4.0/components/alerts/}
}
