% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{sendModal}
\alias{sendModal}
\title{Modal dialogs}
\usage{
sendModal(title, body, footer = NULL, ...)
}
\arguments{
\item{body}{A character vectory specifying the main text of the modal or
custom elements to use as the body of the modal, defaults to \code{NULL}.}

\item{footer}{Custom tags to include at the bottom of the modal, defaults to
\code{NULL}.}

\item{...}{Additional named arguments passed as HTML attributes to the parent
element.}

\item{header}{A character vector specifying a title for the modal or a custom
element to use as a header, defaults to \code{NULL}. The default heading tag is
\code{h5}.}
}
\description{
Modals are a flexible alert window, which disable interaction with the page
behind them. Modals may include inputs or buttons or simply include text.
Create a modal that can by shown or hidden by \code{toggleModal}. Unlike shiny
modals, dull modals must be built ahead of time, but may include dynamically
updated components. Modal elements may be included anywhere on a web page.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      buttonInput(id = "toggle", "Click to show modal"),
      modal(
        id = "simple",
        header = "A simple modal",
        body = paste(
          "Cras mattis consectetur purus sit amet fermentum.",
          "Cras justo odio, dapibus ac facilisis in, egestas",
          "eget quam. Morbi leo risus, porta ac consectetur",
          "ac, vestibulum at eros."
        )
      )
    ),
    server = function(input, output) {
      observeEvent(input$toggle, {
        toggleModal("simple")
      })
    }
  )
}

if (interactive()) {
  shinyApp(
    ui = container(
      buttonInput(id = "toggle", "Demo login"),
      modal(
        id = "login",
        header = "Please login",
        body = tagList(
          textInput(
            id = "name",
            label = "Name",
            placeholder = "yourname@email.com"
          ),
          passwordInput(
            id = "password",
            label = "Password"
          )
        ),
        footer = buttonInput(id = "process", "Submit")
      )
    ),
    server = function(input, output) {
      observeEvent(input$toggle, {
        toggleModal("login")
      })

      observeEvent(input$process, {
        print(isolate(input$name))
        print(isolate(input$password))
      })
    }
  )
}

}
