% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{shadow}
\alias{shadow}
\title{Add shadows to tag elements}
\usage{
shadow(.tag, size = "regular")
}
\arguments{
\item{.tag}{A tag element.}

\item{size}{One of \code{"none"}, \code{"small"}, \code{"regular"}, or \code{"large"} specifying
the amount of shadow added, defaults to \code{"regular"}.}
}
\description{
The \code{shadow} utility applies a shadow to a tag element. Elements with a
shadow may appear to pop off the page. The material design set of components,
used on Android and for Google applications, commonly uses shadowing.
Although \code{"none"} is an allowed \code{size}, most elements do not have a shadow by
default.
}
\examples{

if (interactive()) {
  shinyApp(
    ui = tagList(
      navbar(brand = "Navbar") \%>\%
        background("cyan") \%>\%
        shadow("small"),
      container(
        "Cras mattis consectetur purus sit amet fermentum. Donec sed ",
        "odio dui. Lorem ipsum dolor sit amet, consectetur adipiscing ",
        "elit. Aenean eu leo quam. Pellentesque ornare sem lacinia quam ",
        "venenatis vestibulum."
      ) \%>\%
        padding(3)
    ),
    server = function(input, output) {

    }
  )
}

}
\seealso{
Other utilities: \code{\link{affix}},
  \code{\link{background}}, \code{\link{border}},
  \code{\link{collapse}}, \code{\link{display}},
  \code{\link{flex}}, \code{\link{float}},
  \code{\link{font}}, \code{\link{markInvalid}},
  \code{\link{padding}}, \code{\link{rounded}},
  \code{\link{sendModal}}, \code{\link{showPopover}},
  \code{\link{tooltip}}, \code{\link{updateChoices}},
  \code{\link{width}}
}
\concept{utilities}
