% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design-shadow.R
\name{shadow}
\alias{shadow}
\title{Shadows}
\usage{
shadow(x, size = "medium")
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{size}{One of \code{"none"}, \code{"small"}, \code{"medium"}, or \code{"large"} specifying
the amount of shadow added, defaults to \code{"medium"}.}
}
\description{
The \code{shadow} utility applies a shadow to a tag element. By default many
elements include a shadow to help distinguish them. Use \code{"none"} to remove an
element's shadow.
}
\examples{

### Styling a navbar

div(
  navbar(brand = "Navbar") \%>\%
    background("cyan") \%>\%
    shadow("small") \%>\%
    margin(bottom = 3),
  p(
    "Cras mattis consectetur purus sit amet fermentum. Donec sed ",
    "odio dui. Lorem ipsum dolor sit amet, consectetur adipiscing ",
    "elit. Aenean eu leo quam. Pellentesque ornare sem lacinia quam ",
    "venenatis vestibulum."
  )
)

### Different shadows

div(
  buttonInput(id = "b1", label = "Small") \%>\%
    margin(2) \%>\%
    shadow("small"),
  buttonInput(id = "b2", label = "Medium") \%>\%
    margin(2) \%>\%
    shadow("medium"),
  buttonInput(id = "b3", label = "Large") \%>\%
    margin(2) \%>\%
    shadow("large")
)

}
\seealso{
Other design utilities: 
\code{\link{affix}()},
\code{\link{background}()},
\code{\link{border}()},
\code{\link{display}()},
\code{\link{flex}()},
\code{\link{float}()},
\code{\link{font}()},
\code{\link{height}()},
\code{\link{margin}()},
\code{\link{padding}()},
\code{\link{scroll}()},
\code{\link{width}()}
}
\concept{design utilities}
