% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{showAlert}
\alias{showAlert}
\title{Static and actionable alerts}
\usage{
showAlert(text, ..., duration = 4, color = NULL, action = NULL)
}
\arguments{
\item{text}{A character string specifying the message text of the alert.}

\item{...}{Additional named arguments passed as HTML attributes to the
alert element.}

\item{duration}{A positive integer specifying the duration of the alert,
by default the alert is removed after 4 seconds.}

\item{color}{A character string specifying the color of the alert,
for possible colors see \link{background}.}

\item{action}{A character string specifying a reactive id. If specified a
button is added to the alert. If clicked the reactive value
\code{input[[action]]} is set to \code{TRUE}. When the alert is removed
\code{input[[action]]} is reset to \code{NULL}.

[[action]: R:[action
[[action]: R:[action}
}
\description{
Use \code{showAlert} to let the user know of successes or to call attention to
problems. While alerts are static by default they can also be made
actionable. Actionable alerts can be used for undoing or redoing an action
and more.
}
\examples{

if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          buttonInput("show", "Alert!")
        )
      )
    ),
    server = function(input, output) {
      observeEvent(input$show, {
        color <- sample(c("teal", "red", "orange", "blue"), 1)
        showAlert("Alert", color = color)
      })
    }
  )
}

if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          buttonInput("show", "Alert!")
        )
      )
    ),
    server = function(input, output) {
      observeEvent(input$show, {
        color <- sample(c("teal", "red", "orange", "blue"), 1)
        showAlert("Alert", color = color, action = "undo")
      })

      observeEvent(input$undo, {
        print("OOOOOOHHH yes!")
      })
    }
  )
}

}
\seealso{
Boostrap 4 alert documentation:
\url{https://getbootstrap.com/docs/4.0/components/alerts/}
}
