% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{showAlert}
\alias{showAlert}
\alias{closeAlert}
\title{Static and actionable alerts}
\usage{
showAlert(text, ..., duration = 4, color = NULL, action = NULL)

closeAlert(...)
}
\arguments{
\item{text}{A character string specifying the message text of the alert.}

\item{...}{Additional named arguments passed as HTML attributes to the
alert element.}

\item{duration}{A positive integer or \code{NULL} specifying the duration of the
alert, by default the alert is removed after 4 seconds. If \code{NULL} the
alert is not automatically removed.}

\item{color}{A character string specifying the color of the alert,
for possible colors see \link{background}.}

\item{action}{A character string specifying a reactive id. If specified a
button is added to the alert. If clicked the reactive value
\code{input[[action]]} is set to \code{TRUE}. When the alert is removed
\code{input[[action]]} is reset to \code{NULL}.

[[action]: R:[action
[[action]: R:[action}
}
\description{
Use \code{showAlert} to let the user know of successes or to call attention to
problems. While alerts are static by default they can also be made
actionable. Actionable alerts can be used for undoing or redoing an action
and more.
}
\examples{

if (interactive()) {
  shinyApp(
    ui = container(
      buttonInput("show", "Alert!") \%>\%
        margin(3)
    ),
    server = function(input, output) {
      observeEvent(input$show, {
        color <- sample(c("teal", "red", "orange", "blue"), 1)
        showAlert("Alert", color = color)
      })
    }
  )
}


if (interactive()) {
  shinyApp(
    ui = container(
      row(
        column(
          groupInput(
            id = "text",
            right = buttonInput("clear", icon("times")) \%>\%
              background("red")
          )
        ),
        column(
          verbatimTextOutput("value")
        )
      ) \%>\%
        margin(3)
    ),
    server = function(input, output) {
      oldValue <- NULL

      output$value <- renderPrint(input$text)

      observeEvent(input$clear, {
        oldValue <<- input$text
        updateValues("text", "")
        showAlert("Undo clear.", color = "yellow", action = "undo")
      })

      observeEvent(input$undo, {
        updateValues("text", oldValue)
      })
    }
  )
}


if (interactive()) {
  shinyApp(
    ui = container(
      buttonInput("add", "Alert") \%>\%
        margin(3),
      buttonInput("first", "Remove first alert"),
      buttonInput(
        id = "reds",
        label = "Remove red alerts",
        alt = "the red ones offend the aesthetic"
      ),
      buttonInput("alert", "Remove 'Alert' alerts")
    ),
    server = function(input, output) {
      observeEvent(input$add, {
        color <- sample(c("grey", "teal", "purple", "yellow", "red"), 1)
        showAlert("Alert", duration = NULL, color = color)
      })

      observeEvent(input$first, {
        closeAlert(1)
      })

      observeEvent(input$reds, {
        closeAlert(class = "alert-red")
      })

      observeEvent(input$alert, {
        closeAlert("Alert")
      })
    }
  )
}


# this is a variation of the second example
if (interactive()) {
  shinyApp(
    ui = container(
      row(
        column(
          groupInput(
            id = "text",
            right = buttonInput("clear", icon("times")) \%>\%
              background("red")
          )
        ),
        column(
          verbatimTextOutput("value")
        )
      ) \%>\%
        margin(3)
    ),
    server = function(input, output) {
      oldValue <- NULL

      output$value <- renderPrint(input$text)

      observeEvent(input$clear, {
        oldValue <<- input$text
        updateValues("text", "")
        showAlert(
          text = "Undo clear.",
          color = "yellow",
          action = "undo",
          duration = NULL
        )
      })

      observeEvent(input$undo, {
        updateValues("text", oldValue)
        closeAlert(1)
      })
    }
  )
}

}
\seealso{
Boostrap 4 alert documentation:
\url{https://getbootstrap.com/docs/4.0/components/alerts/}
}
