% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkline.R
\name{sparklineOutput}
\alias{sparklineOutput}
\alias{renderSparkline}
\title{Sparkline output}
\usage{
sparklineOutput(id, type = "bar", labels = TRUE, ...)

renderSparkline(values, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{id}{A character string specifying the id of the sparkline output.}

\item{type}{One of \code{"bar"}, \code{"dot"}, or \code{"line"} specifying the type of
chart to render, defaults to \code{"bar"}.}

\item{labels}{If \code{TRUE}, the first and last value of the sparkline data
are used as labels, defaults to \code{TRUE}.}

\item{values}{A numeric vector of values specifying the sparkline data
points. Bar and dot sparklines expect values to be between 0 and 100 and
line sparklines expect values to be between 0 and 10.}
}
\description{
Display concise, reactive inline bar, line, or dot charts.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          listGroupInput(
            id = "toc",
            items = list(
              tags$div(
                class = "d-flex w-100 justify-content-between",
                "Bars",
                sparklineOutput("bar", type = "bar")
              ),
              tags$div(
                class = "d-flex w-100 justify-content-between",
                "Dots",
                sparklineOutput("dot", type = "dot")
              ),
              tags$div(
                class = "d-flex w-100 justify-content-between",
                "Line",
                sparklineOutput("line", type = "line")
              )
            ),
            values = c("bar", "dot", "line")
          )
        ),
        col(

        )
      )
    ),
    server = function(input, output) {
      gen10 <- function() {
        sample(10:100, 10, replace = TRUE) \%/\% 10 * 10
      }

      output$bar <- renderSparkline({
        gen10()
      })

      output$dot <- renderSparkline({
        gen10()
      })

      output$line <- renderSparkline({
        gen10()
      })
    }
  )
}

}
