% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkline.R
\name{sparklineOutput}
\alias{sparklineOutput}
\alias{renderSparkline}
\title{Sparkline output}
\usage{
sparklineOutput(id, type = "bar", labels = TRUE, ...)

renderSparkline(values, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{id}{A character string specifying the id of the sparkline output.}

\item{type}{One of \code{"bar"}, \code{"dot"}, or \code{"line"} specifying the type of
chart to render, defaults to \code{"bar"}.}

\item{labels}{If \code{TRUE}, the first and last value of the sparkline data
are used as labels, defaults to \code{TRUE}.}

\item{...}{Additional named arguments passed as HTML attributes to the parent
element.}

\item{values}{A numeric vector of values specifying the sparkline data
points. Bar and dot sparklines expect values to be between 0 and 100 and
line sparklines expect values to be between 0 and 10.}

\item{env}{The environment in which to evaluate \code{values}.}

\item{quoted}{One of \code{TRUE} or \code{FALSE} specifying if \code{values} is a quoted
expression.}
}
\description{
Display concise, reactive inline bar, line, or dot charts.
}
\examples{

}
\seealso{
Other outputs: \code{\link{progressOutput}},
  \code{\link{streamOutput}}

Other outputs: \code{\link{progressOutput}},
  \code{\link{streamOutput}}
}
\concept{outputs}
