% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{streamOutput}
\alias{streamOutput}
\alias{updateStream}
\title{Stream notifications}
\usage{
streamOutput(id, ...)

updateStream(id, message, template = listGroupItem(),
  session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{A character string specifying the id of the stream output.}

\item{message}{A character string specifying an update message.}

\item{template}{A call \code{listGroupItem} which is used as a template for the
message sent, any body element(s) or text of the item are replaced by
\code{message}, defaults to \code{listGroupItem()}.}

\item{session}{The shiny session, defaults to the default reactive domain.}
}
\description{
These functionalities allow an application to convey text progress updates
to the user during long running processes.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      button(id = "trigger", "Go!"),
      tags$div(
        id = "myStream"
      )
    ),
    server = function(input, output, session) {
      observeEvent(input$trigger, {
        for (i in seq_len(5)) {
          updateStream(
            id = "myStream",
            message = paste("Update:", i),
            template = listGroupItem(
              context = if (i \%\% 2) "warning" else "info"
            )
          )
          Sys.sleep(1);
        }
      })
    }
  )
}

}
