% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{tabContent}
\alias{tabContent}
\alias{tabPane}
\title{Tab content, panes}
\usage{
tabContent(list, ...)

tabPane(..., fade = TRUE)
}
\arguments{
\item{list}{A character string specifying the id of a tab list, see \code{id}
in \code{\link[=tabList]{tabList()}}.}

\item{...}{For \strong{tabContent}, calls to \code{tabPane} or named arguments passed
as HTML attributes to the parent element.

For \strong{tabPane}, any number of tag elements or named arguments passed as
HTML attributes to the parent element.}
}
\description{
Create tabbed content. You will need to create a tab pane for each of the
tabs in the corresponding tab list. A tab list and tab content are linked
through the \code{id} argument and \code{list} argument respectively.
}
\examples{

tabContent(
  list = "tabListId",
  tabPane(
    tags$h1("Pane 1"),
    tags$p("Lorem Dapibus Malesuada Cras Cursus")
  ),
  tabPane(
    tags$p("Pane 2"),
    tags$p("Magna Aenean Mattis Ultricies Ridiculus")
  )
)

if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          default = 4,
          card(
            p("Nullam quis risus eget urna mollis ornare vel eu leo. Nullam
            id dolor id nibh ultricies vehicula ut id elit.")
          ) \%>\%
            background("grey", +2)
        ),
        col(
          tabList(
            id = "myTabs",
            labels = c("Home", "About", "Posts")
          ) \%>\%
            margins(c(0, 0, 5, 0)),
          tabContent(
            list = "myTabs",
            tabPane(
              h4("This is home."),
              p("Pellentesque dapibus suscipit ligula.  Donec ",
                "posuere augue in quam.  Etiam vel tortor ",
                "sodales tellus ultricies commodo.  Suspendisse ",
                "potenti.  Aenean in sem ac leo mollis blandit."),
              p("Donec neque quam, dignissim in, mollis nec, ",
                "sagittis eu, wisi.  Phasellus lacus.  Etiam ",
                "laoreet quam sed arcu.  Phasellus at dui in ligula",
                "mollis ultricies.  Integer placerat tristique nisl.")
            ),
            tabPane(
              h4("All about."),
              p("Pellentesque dapibus suscipit ligula. ",
                "Phasellus neque orci, porta a, aliquet quis, semper a, massa. ",
                "Nullam tristique diam non turpis. ",
                "Lorem ipsum dolor sit amet, consectetuer adipiscing elit.")
            ),
            tabPane(
              h4("Blog items."),
              p("Nullam tempus.  Mauris ac felis vel velit ",
                "tristique imperdiet.  Donec at pede.  Etiam vel ",
                "neque nec dui dignissim bibendum.  Vivamus id ",
                "enim.  Phasellus neque orci, porta a, aliquet ",
                "quis, semper a, massa.  Phasellus purus.")
            )
          )
        )
      ) \%>\%
        padding(3)
    ),
    server = function(input, output) {
    }
  )
}

}
\seealso{
Other tabs: \code{\link{tabList}}

Other tabs: \code{\link{tabList}}
}
