% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{tableThruput}
\alias{tableThruput}
\alias{renderTable}
\title{Table output}
\usage{
tableThruput(id, borders = FALSE, compact = FALSE, invert = FALSE,
  striped = FALSE, hoverable = FALSE, context = NULL, responsive = TRUE,
  ...)

renderTable(expr, numbered = FALSE, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{id}{A character string specifying the id of the table output.}

\item{borders}{If \code{TRUE}, the table renders with borders, defaults to
\code{FALSE}.}

\item{compact}{If \code{TRUE}, table cell padding is cut in half to reduce the
size of the table, defaults to \code{FALSE}.}

\item{invert}{If \code{TRUE}, the table renders with a dark background, light
text, and any borders are light, defaults to \code{FALSE}.}

\item{striped}{If \code{TRUE}, the table renders with a striped pattern, defaults
to \code{FALSE}.}

\item{hoverable}{If \code{TRUE}, each table row has a hover state, defaults to
\code{FALSE}.}

\item{context}{One \code{"success"}, \code{"info"}, \code{"warning"}, \code{"danger"}, specifying
the context of selected table rows, defaults to \code{NULL}, in which case
selected rows are highlighted in grey.}

\item{responsive}{If \code{TRUE}, the table will scroll horizontally on small
viewports, viewports under 768 pixels, defaults to \code{TRUE}. There is no
effect when the viewport height is greater than 768 pixels.}

\item{...}{Additional named arguments passed as HTML attributes to the parent
element.}
}
\description{
Render a table.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          tableThruput(
            id = "table",
            context = "danger"
          )
        ),
        col(
          verbatimTextOutput("value")
        )
      )
    ),
    server = function(input, output) {
      output$table <- renderTable(
        numbered = TRUE,
        iris[1:10, ]
      )

      output$value <- renderPrint({
        input$table
      })
    }
  )
}

if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          tableThruput(
            id = "tbl",
            hoverable = TRUE,
            context = "warning",
            borders = TRUE
          )
        ),
        col(
          tableThruput(
            id = "subset",
            borders = TRUE
          )
        )
      )
    ),
    server = function(input, output) {
      output$tbl <- renderTable(
        numbered = TRUE,
        iris[1:10, ]
      )

      output$subset <- renderTable(
        numbered = TRUE,
        input$tbl
      )
    }
  )
}


}
