% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{tabList}
\alias{tabList}
\title{Tab list}
\usage{
tabList(labels, panes, values = labels, active = values[1], ...,
  id = NULL)
}
\arguments{
\item{labels}{A character vector specifying the labels of the tabs.}

\item{panes}{A character vector specifying the ids of the tab panes
controlled by the tab list, see \code{id} in \link{tabPane}.}

\item{values}{A character vector specifying a reactive value for each tab,
defaults to \code{labels}.}

\item{active}{One of \code{values} specifying which tab is initially shown,
defaults to \code{values[1]}.`}

\item{...}{Additional named arguments passed as HTML attributes to the
parent element.}

\item{id}{A character string specifying a reactive id for the tab list,
defaults to \code{NULL}, in which case a reactive value is not created. If \emph{not}
\code{NULL} the tab lists's reactive value is the value of the active tab.}
}
\description{
Create a tab list for controlling tab content.
}
\examples{

if (interactive()) {
  shinyApp(
    ui = container(
      card(
        header = tabList(
          labels = c("Home", "Profile", "Contact"),
          panes = c("homePane", "profPane", "contactPane"),
        ),
        tabContent(
          tabPane(
            id = "homePane",
            class = "show active",
            "Vestibulum id ligula porta felis euismod semper. Cras justo
             odio, dapibus ac facilisis in, egestas eget quam."
          ),
          tabPane(
            id = "profPane",
            "Duis mollis, est non commodo luctus, nisi erat porttitor ligula,
             eget lacinia odio sem nec elit. Aenean eu leo quam. Pellentesque
             ornare sem lacinia quam venenatis vestibulum."
          ),
          tabPane(
            id = "contactPane",
            "Donec ullamcorper nulla non metus auctor fringilla. Nullam id
             dolor id nibh ultricies vehicula ut id elit."
          )
        )
      ) \%>\%
        margins(4) \%>\%
        width(50)
    ),
    server = function(input, output) {

    }
  )
}

}
\seealso{
Other tabs: \code{\link{tabContent}}
}
