% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{text}
\alias{text}
\alias{search}
\alias{email}
\alias{url}
\alias{telephone}
\alias{password}
\alias{number}
\alias{datetime}
\alias{date}
\alias{month}
\alias{week}
\alias{time}
\alias{color}
\title{Textual Inputs}
\usage{
inputs$text(label = NULL, ...)
}
\arguments{
\item{label}{Input label.}

\item{value}{Input value.}

\item{placeholder}{Input placeholder text.}

\item{toggleable}{If \code{TRUE}, a checkbox is included as part of the input,
when unchecked the value of the input is always \code{NULL}, defaults to
\code{FALSE}.}
}
\description{
description
}
