% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{text}
\alias{text}
\alias{textInput}
\alias{search}
\alias{email}
\alias{url}
\alias{telephone}
\alias{password}
\alias{number}
\alias{datetime}
\alias{date}
\alias{month}
\alias{week}
\alias{time}
\alias{color}
\title{Inputs}
\usage{
textInput(value = NULL, placeholder = NULL, ..., id = NULL)
}
\arguments{
\item{placeholder}{A character string specifying placeholder text for the
input, defaults to \code{NULL}.}

\item{...}{Named arguments passed as HTML attributes to the input. Specify
\code{value} to give the input a default value. Specify \code{placeholder} to give
the input hint text about the expected value of the input.}

\item{id}{A character string specifying the id of the textual input, defaults
to \code{NULL}. If specified, a reactive value is available to the shiny server
function.}

\item{label}{A character vector specifying a label for the input, defaults to
\code{NULL}. If a label is specified it is advised, though not necessary, to
also specify an \code{id} for the input.}

\item{type}{A character string specifying the type of the text input, a
specific type may be used when appropriate, e.g. when \code{type = "password"}
the input's value is obscured, displays as asterisks. See \code{Details} for
possible \code{type} values.}
}
\description{
Form control inputs.
}
\details{
** \code{text} values **
\itemize{
\item text - Single-line text input, linebreaks are removed.
\item search - Single-line text input for search strings, linebreaks are removed.
\item email - Single-line text input for email addresses, validated to contain
the empty string or a valid email address.
\item url - Single-line text input for URLs, validated to contain the empty
string or valid absolute URL.
\item tel - Single-line text input for telephone numbers, syntax is not enforced.
\item password - Single-line text input for password, value is obscured.
\item number - Number picker.
\item datetime - Date and time picker, no time zone.
\item date - Year, month, and day picker, no time.
\item month - Month and year picker, no time zone.
\item week - Week of year picker.
\item time - Time picker.
\item color - Color picker.
}
}
\seealso{
Other inputs: \code{\link{button}},
  \code{\link{groupInput}}, \code{\link{radiosInput}},
  \code{\link{selectInput}}
}
