% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{text}
\alias{text}
\title{Element text color, font case, font weight, and more}
\usage{
text(tag, context = NULL, align = NULL, truncate = FALSE, case = NULL,
  weight = NULL, style = NULL)
}
\arguments{
\item{context}{One of \code{"primary"}, \code{"secondary"}, \code{"success"}, \code{"info"},
\code{"warning"}, \code{"danger"}, \code{"light"}, \code{"dark"}, \code{"muted"}, or \code{"white"}
specifying the visual context of the text. \code{"white"} is best suited for
elements with dark backgrounds. Defaults to \code{NULL}, the element text color
is left as is.}

\item{align}{One of \code{"justify"}, \code{"left"}, \code{"right"}, or \code{"center"}
specifying the alignment of an element's text. Defaults to \code{NULL}, no
alignment is applied, however the CSS default is to left align element
text.}

\item{truncate}{One of \code{TRUE} or \code{FALSE} specifying whether or not an
element's text is truncated. Defaults to \code{FALSE}, the element text is left
as is.}

\item{case}{One of \code{"upper"}, \code{"lower"}, or \code{"capitalize"} specifying a
transformation of an element's text. Defaults to \code{NULL}, element case is
left as is. \code{"capitalize"} changes only the first letter to upper case,
all other characters are left as is.}
}
\description{
The \code{text} helper function applies text related Boostrap classes to a tag
element. These classes apply CSS to change text color, alignment, or case.
}
\examples{
tags$span("hello, world!") \%>\%
  text(context = "warning", case = "upper")

tags$div("Goodnight, moon") \%>\%
  text(truncate = TRUE, weight = "bold") \%>\%
  text("success")

tags$div("Howdy") \%>\%
  text("white") \%>\%
  background("dark")

}
\seealso{
Other utilities: \code{\link{affix}}, \code{\link{align}},
  \code{\link{border}}, \code{\link{float}},
  \code{\link{font}}, \code{\link{padding}},
  \code{\link{width}}
}
