% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textual.R
\name{textInput}
\alias{textInput}
\alias{searchInput}
\alias{emailInput}
\alias{urlInput}
\alias{telephoneInput}
\alias{passwordInput}
\alias{numberInput}
\title{Textual inputs}
\usage{
textInput(id, value = NULL, placeholder = NULL, size = NULL,
  readonly = FALSE, help = NULL, ...)

searchInput(id, value = NULL, placeholder = NULL, size = NULL,
  readonly = FALSE, help = NULL, ...)

emailInput(id, value = NULL, placeholder = NULL, size = NULL,
  readonly = FALSE, help = NULL, ...)

urlInput(id, value = NULL, placeholder = NULL, size = NULL,
  readonly = FALSE, help = NULL, ...)

telephoneInput(id, value = NULL, placeholder = NULL, size = NULL,
  readonly = FALSE, help = NULL, ...)

passwordInput(id, value = NULL, placeholder = NULL, size = NULL,
  readonly = FALSE, help = NULL, ...)

numberInput(id, value = NULL, placeholder = NULL, size = NULL,
  readonly = FALSE, help = NULL, ...)
}
\arguments{
\item{id}{A character string specifying the id of the textual input, defaults
to \code{NULL}.}

\item{value}{A character string or a value coerced to a character string
specifying the default value of the textual input.}

\item{placeholder}{A character string specifying placeholder text for the
input, defaults to \code{NULL}, in which case there is no placeholder text.}

\item{size}{One of \code{"small"} or \code{"large"} specifying the size transformation
of the input, defaults to \code{NULL}, in which case the input's size is
unchanged.}

\item{readonly}{If \code{TRUE}, the textual input is read-only preventing
modification of the value, defaults \code{FALSE}.}

\item{help}{A character string specifying the help text of the textual input,
defaults to \code{NULL}, in which case no help text is added.}

\item{...}{Additional named arguments passed as HTML attributes to the parent
element.}
}
\description{
Different types of textual inputs are provided to best support mobile
keyboards and assistive technologies. A password input will mask its
contents. Email inputs offer client-side validation depending on the browser.
}
\examples{

### Basic text

textInput(id = "text")

### Search

searchInput(id = "search")

### Email

emailInput(id = "email")

### URLs

urlInput(id = "url")

### Telephone numbers

telephoneInput(id = "tele")

### Passwords

passwordInput(id = "password")

### Numbers

numberInput(id = "num")

}
\seealso{
Other inputs: \code{\link{addressInput}},
  \code{\link{buttonGroupInput}},
  \code{\link{buttonInput}}, \code{\link{checkboxInput}},
  \code{\link{dateInput}}, \code{\link{fileInput}},
  \code{\link{formInput}}, \code{\link{groupInput}},
  \code{\link{loginInput}}, \code{\link{radioInput}},
  \code{\link{rangeInput}}, \code{\link{selectInput}}
}
\concept{inputs}
