% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-textual.R
\name{textInput}
\alias{textInput}
\alias{numberInput}
\alias{updateTextInput}
\alias{groupTextInput}
\alias{updateGroupTextInput}
\title{Text inputs}
\usage{
textInput(..., id, value = NULL, placeholder = NULL, type = "text")

numberInput(
  ...,
  id,
  value = NULL,
  placeholder = NULL,
  min = NULL,
  max = NULL,
  step = 1
)

updateTextInput(
  id,
  value = NULL,
  enable = NULL,
  disable = NULL,
  valid = NULL,
  invalid = NULL,
  session = getDefaultReactiveDomain()
)

groupTextInput(
  ...,
  id,
  value = NULL,
  placeholder = NULL,
  type = "text",
  left = NULL,
  right = NULL
)

updateGroupTextInput(
  id,
  value = NULL,
  enable = NULL,
  disable = NULL,
  valid = NULL,
  invalid = NULL,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{...}{Optional named arguments specifying HTML attributes for the input
element.}

\item{id}{A character string. The id of the reactive input.}

\item{value}{A character string or a value coerced to a character string
specifying the default value of the textual input.}

\item{placeholder}{A character string specifying placeholder text for the
input, defaults to \code{NULL}, in which case there is no placeholder text.}

\item{type}{One of "\code{color}", "\code{date}", "\code{datetime-local}", "\code{email}", "\code{month}", "\code{number}", "\code{password}", "\code{search}", "\code{tel}", "\code{text}", "\code{time}", "\code{url}" or "\code{week}" specifying the type of text input, defaults to \code{"text"}.

For details on a particular type please see \url{https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input}.}

\item{min}{A number specifying the minimum allowed value of the number input,
defaults to \code{NULL}.}

\item{max}{A number specifying the maximum allowed value of the number input,
defaults to \code{NULL}.}

\item{step}{A number specifying the increment step of the number input,
defaults to 1.}

\item{disable}{A boolean. The checkbox starts disabled if \code{TRUE}.}

\item{session}{A shiny session object.}

\item{left, right}{A character vector specifying static addons or
\code{\link[=buttonInput]{buttonInput()}} or \code{\link[=dropdown]{dropdown()}} elements specifying dynamic addons. Addons
affect the reactive value of the group input, see the Details section below
for more information.

\strong{\code{left} is character or \code{right} is character}

If \code{left} or \code{right} are character vectors, then the group input functions
like a text input. The value will update and trigger a reactive event when
the text box is modified. The group input's reactive value is the
concatenation of the static addons specified by \code{left} or \code{right} and the
value of the text input.

\strong{\code{left} is button or \code{right} is button}

The button does not change the value of the group input. However, the input
no longer triggers event when the text box is updated. Instead the value is
updated when a button is clicked. Static addons are still applied to the
group input value.

\strong{\code{left} is a dropdown or \code{right} is a dropdown}

The value of the group input does chance depending on the clicked dropdown
menu item. The value of the input group is the concatenation of the
dropdown input value, the value of the text input, and any static addons.}
}
\description{
A text input. A group text input is an alternative text input. The group text
input allows you to include static prefixes or buttons with a standard text
input.

\code{numberInput()} is a simple wrapper around \code{textInput()} with \code{type} set to
\code{"number"} and explicit arguments for specifying a min value, max value, and
the step amount. Use \code{updateTextInput()} to update a number input.
}
\seealso{
Other inputs: 
\code{\link{fileInput}()},
\code{\link{input_button}()},
\code{\link{input_checkbox}()},
\code{\link{input_checkbox_group}()},
\code{\link{input_chip}()},
\code{\link{input_form}()},
\code{\link{input_list_group}()},
\code{\link{input_menu}()},
\code{\link{input_radio_group}()},
\code{\link{radiobarInput}()},
\code{\link{rangeInput}()},
\code{\link{selectInput}()}
}
\concept{inputs}
