% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{textInput}
\alias{textInput}
\alias{searchInput}
\alias{emailInput}
\alias{urlInput}
\alias{telephoneInput}
\alias{passwordInput}
\alias{numberInput}
\alias{colorInput}
\title{Inputs}
\usage{
textInput(id, label, value = NULL, placeholder = NULL, ...)

searchInput(id, label, value = NULL, placeholder = NULL, ...)

emailInput(id, label, value = NULL, placeholder = NULL, ...)

urlInput(id, label, value = NULL, placeholder = NULL, ...)

telephoneInput(id, label, value = NULL, placeholder = NULL, ...)

passwordInput(id, label, value = NULL, placeholder = NULL, ...)

numberInput(id, label, value = NULL, placeholder = NULL, ...)

colorInput(id, label, value = NULL, placeholder = NULL, ...)
}
\arguments{
\item{id}{A character string specifying the id of the textual input, defaults
to \code{NULL}. If specified, a reactive value is available to the shiny server
function.}

\item{label}{A character string specifying a label for the input, defaults to
\code{NULL}. If a label is specified it is advised, though not necessary, to
also specify an \code{id} for the input.}

\item{placeholder}{A character string specifying placeholder text for the
input, defaults to \code{NULL}, in which case there is no placeholder text.}

\item{...}{Additional named arguments passed as HTML attributes to the input.}
}
\description{
Form control inputs.
}
\examples{

stub

}
\seealso{
Other inputs: \code{\link{buttonInput}},
  \code{\link{checkboxInput}}, \code{\link{formInput}},
  \code{\link{groupInput}}, \code{\link{radioInput}},
  \code{\link{selectInput}}
}
