% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textual.R
\name{textInput}
\alias{textInput}
\alias{searchInput}
\alias{emailInput}
\alias{urlInput}
\alias{telephoneInput}
\alias{passwordInput}
\alias{numberInput}
\alias{colorInput}
\title{Textual inputs}
\usage{
textInput(id, value = NULL, placeholder = NULL, size = NULL,
  readonly = FALSE, help = NULL, ...)

searchInput(id, value = NULL, placeholder = NULL, size = NULL,
  readonly = FALSE, help = NULL, ...)

emailInput(id, value = NULL, placeholder = NULL, size = NULL,
  readonly = FALSE, help = NULL, ...)

urlInput(id, value = NULL, placeholder = NULL, size = NULL,
  readonly = FALSE, help = NULL, ...)

telephoneInput(id, value = NULL, placeholder = NULL, size = NULL,
  readonly = FALSE, help = NULL, ...)

passwordInput(id, value = NULL, placeholder = NULL, size = NULL,
  readonly = FALSE, help = NULL, ...)

numberInput(id, value = NULL, placeholder = NULL, size = NULL,
  readonly = FALSE, help = NULL, ...)

colorInput(id, value = NULL, placeholder = NULL, size = NULL,
  readonly = FALSE, help = NULL, ...)
}
\arguments{
\item{id}{A character string specifying the id of the textual input, defaults
to \code{NULL}.}

\item{value}{A character string or a value coerced to a character string
specifying the default value of the textual input.}

\item{placeholder}{A character string specifying placeholder text for the
input, defaults to \code{NULL}, in which case there is no placeholder text.}

\item{size}{One of \code{"small"} or \code{"large"} specifying the size transformation
of the input, defaults to \code{NULL}, in which case the input's size is
unchanged.}

\item{readonly}{If \code{TRUE}, the textual input is read-only preventing
modification of the value, defaults \code{FALSE}.}

\item{help}{A character string specifying the help text of the textual input,
defaults to \code{NULL}, in which case no help text is added.}

\item{...}{Additional named arguments passed as HTML attributes to the parent
element.}
}
\description{
Textual inputs.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          p("For best results open in a browser") \%>\%
            font(weight = "bold")
        )
      ),
      row(
        col(
          h6("Text input:"),
          textInput(id = "text"),
          h6("Search input:"),
          searchInput(id = "search"),
          h6("Email input:"),
          emailInput(id = "email"),
          h6("URL input:"),
          urlInput(id = "url"),
          h6("Telephone input:"),
          telephoneInput(id = "tel"),
          h6("Password input:"),
          passwordInput(id = "pass"),
          h6("Number input:"),
          numberInput(id = "num") \%>\%
            background("lime", +1)
        ),
        col(
          verbatimTextOutput("values")
        )
      )
    ),
    server = function(input, output) {
      output$values <- renderPrint({
        list(
          text = input$text, search = input$search, email = input$email,
          url = input$url, telephone = input$tel, password = input$pass,
          number = input$num
         )
      })
    }
  )
}


}
\seealso{
Other inputs: \code{\link{buttonGroupInput}},
  \code{\link{buttonInput}}, \code{\link{checkboxInput}},
  \code{\link{formInput}}, \code{\link{groupInput}},
  \code{\link{radioInput}}, \code{\link{selectInput}}
}
\concept{inputs}
