% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{textInput}
\alias{textInput}
\alias{searchInput}
\alias{emailInput}
\alias{urlInput}
\alias{telephoneInput}
\alias{passwordInput}
\alias{numberInput}
\alias{colorInput}
\title{Textual nputs}
\usage{
textInput(id, label, value = NULL, placeholder = NULL, readonly = FALSE,
  help = NULL, ...)

searchInput(id, label, value = NULL, placeholder = NULL, readonly = FALSE,
  help = NULL, ...)

emailInput(id, label, value = NULL, placeholder = NULL, readonly = FALSE,
  help = NULL, ...)

urlInput(id, label, value = NULL, placeholder = NULL, readonly = FALSE,
  help = NULL, ...)

telephoneInput(id, label, value = NULL, placeholder = NULL,
  readonly = FALSE, help = NULL, ...)

passwordInput(id, label, value = NULL, placeholder = NULL,
  readonly = FALSE, help = NULL, ...)

numberInput(id, label, value = NULL, placeholder = NULL, readonly = FALSE,
  help = NULL, ...)

colorInput(id, label, value = NULL, placeholder = NULL, readonly = FALSE,
  help = NULL, ...)
}
\arguments{
\item{id}{A character string specifying the id of the textual input, defaults
to \code{NULL}.}

\item{label}{A character string specifying a label for the input, defaults to
\code{NULL}.}

\item{value}{A character string or a value coerced to a character string
specifying the default value of the textual input.}

\item{placeholder}{A character string specifying placeholder text for the
input, defaults to \code{NULL}, in which case there is no placeholder text.}

\item{readonly}{If \code{TRUE}, the textual input is read-only preventing
modification of the value, defaults \code{FALSE}.}

\item{help}{A character string specifying the help text of the textual input,
defaults to \code{NULL}, in which case no help text is added.}

\item{...}{Additional named arguments passed as HTML attributes to the parent
element.}
}
\description{
Textual inputs.
}
\examples{
if (interactive()) {

}


}
\seealso{
Other inputs: \code{\link{buttonInput}},
  \code{\link{checkboxInput}}, \code{\link{formInput}},
  \code{\link{groupInput}}, \code{\link{radioInput}},
  \code{\link{selectInput}}
}
