% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textual.R
\name{textInput}
\alias{textInput}
\alias{searchInput}
\alias{emailInput}
\alias{urlInput}
\alias{telephoneInput}
\alias{passwordInput}
\alias{numberInput}
\title{Textual inputs}
\usage{
textInput(id, value = NULL, placeholder = NULL, ...)

searchInput(id, value = NULL, placeholder = NULL, ...)

emailInput(id, value = NULL, placeholder = NULL, ...)

urlInput(id, value = NULL, placeholder = NULL, ...)

telephoneInput(id, value = NULL, placeholder = NULL, ...)

passwordInput(id, value = NULL, placeholder = NULL, ...)

numberInput(id, value = NULL, placeholder = NULL, ...)
}
\arguments{
\item{id}{A character string specifying the reactive id of the input.}

\item{value}{A character string or a value coerced to a character string
specifying the default value of the textual input.}

\item{placeholder}{A character string specifying placeholder text for the
input, defaults to \code{NULL}, in which case there is no placeholder text.}

\item{...}{Additional named arguments passed as HTML attributes to the parent element.}
}
\description{
Different types of textual inputs are provided to best support mobile
keyboards and assistive technologies. A password input will mask its
contents. Email inputs offer client-side validation depending on the browser.
}
\examples{

### Basic text

textInput(id = "text")

### Search

searchInput(id = "search")

### Email

emailInput(id = "email")

### URLs

urlInput(id = "url")

### Telephone numbers

telephoneInput(id = "tele")

### Passwords

passwordInput(id = "password")

### Numbers

numberInput(id = "num")

}
\seealso{
Other inputs: \code{\link{buttonGroupInput}},
  \code{\link{buttonInput}}, \code{\link{checkboxInput}},
  \code{\link{fileInput}}, \code{\link{formInput}},
  \code{\link{groupInput}}, \code{\link{listGroupInput}},
  \code{\link{menuInput}}, \code{\link{navInput}},
  \code{\link{radioInput}}, \code{\link{rangeInput}},
  \code{\link{selectInput}}
}
\concept{inputs}
