% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toast.R
\name{toast}
\alias{toast}
\alias{toast_container}
\title{Toasts}
\usage{
toast(..., id = NULL, visibility = c("show", "hide"), wrapper = toast_body)

toast_container(..., position = c("top", "bottom"), padding = 3)
}
\arguments{
\item{...}{Components to include inside a toast. Named arguments are added as
HTML attributes to the parent element.}

\item{id}{A character string. An optional reactive id to track the state of
the toast via \verb{input$<id>}. When visible this value is \code{"shown"} and when
the toast is dismissed the value is \code{"hidden"}.}
}
\description{
Toasts relay notification-level information to the user.
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

library(shiny)
library(bslib)

ui <- page_fluid(
  toast_container(
    toast()
  ),
  card(
    input_button(
      id = "begin",
      label = "Show notification"
    )
  )
)

server <- function(input, output) {
  observeEvent(input$begin, {
    showToast(
      toast(
        list(
          span("Notification") \%>\%
            margin(right = "4"),
          span(strftime(Sys.time(), "\%H:\%M")) \%>\%
            margin(right = 1)
        ),
        "This is notification ", input$show
      ) \%>\%
        margin(right = 2, top = 2)
    )
  })
}

shinyApp(ui, server)

\dontshow{\}) # examplesIf}
}
\seealso{
Other components: 
\code{\link{alert}()},
\code{\link{badge}()},
\code{\link{blockquote}()},
\code{\link{collapsePane}()},
\code{\link{d1}()},
\code{\link{dropdown}()},
\code{\link{img}()},
\code{\link{jumbotron}()},
\code{\link{modal}()},
\code{\link{pre}()}
}
\concept{components}
