% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toast.R
\name{toast}
\alias{toast}
\alias{toast_container}
\title{Toasts}
\usage{
toast(..., id = NULL, visibility = c("show", "hide"), wrapper = toast_body)

toast_container(..., position = c("top", "bottom"), padding = 3)
}
\arguments{
\item{...}{Components to include inside a toast. Named arguments are added as
HTML attributes to the parent element.}

\item{id}{A character string. An optional reactive id to track the state of
the toast via \verb{input$<id>}. When visible this value is \code{"shown"} and when
the toast is dismissed the value is \code{"hidden"}.}
}
\description{
Toasts relay notification-level information to the user.
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

library(shiny)
library(bslib)

ui <- page_fluid(
  toast_container(
    id = "notifications",
    toast()
  ),
  card(
    input_button(
      id = "begin",
      text = "Begin task"
    )
  )
)

server <- function(input, output) {
  observeEvent(input$begin, {
    toast_add(
      target = "notifications",
      toast(
        id = sprintf("task\%s", input$begin),
        toast_header("Task beginning"),
        "This may take a while"
      )
    )
  })

  observe({
    print(input$task1)
  })
}

shinyApp(ui, server)

\dontshow{\}) # examplesIf}
}
\seealso{
Other components: 
\code{\link{alert}()},
\code{\link{badge}()},
\code{\link{blockquote}()},
\code{\link{collapsePane}()},
\code{\link{d1}()},
\code{\link{dropdown}()},
\code{\link{img}()},
\code{\link{jumbotron}()},
\code{\link{modal}()},
\code{\link{pre}()}
}
\concept{components}
