% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toast.R
\name{toast_body}
\alias{toast_body}
\alias{toast_header}
\alias{toast_button}
\title{Toast components}
\usage{
toast_body(...)

toast_header(title, ..., icon = NULL, dismiss = toast_button())

toast_button()
}
\value{
A \code{shiny.tag} object.
}
\description{
Components to include inside a \link{toast}.
}
\seealso{
\code{\link[=toast]{toast()}} for creating toasts.
}
