% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip.R
\name{tooltip}
\alias{tooltip}
\alias{popover}
\title{Add Tooltips and Popovers}
\usage{
tooltip(content, text, placement = "top")

popover(content, text, placement = "top")
}
\arguments{
\item{content}{The tag to add a tooltip or popover to.}

\item{text}{The tooltip or popover text.}

\item{placement}{A character vector specifying where the tooltip or popover
is placed, defaults to top, one of \code{"left"}, \code{"top"}, \code{"right"}, or
\code{"bottom"}.}
}
\description{
Both functions are a means of adding help text to HTML components. Tooltips
appear on hover and popovers appear when the element is clicked.
}
\examples{

popover(
  buttonInput("Click me!"),
  "This text appears when the button is clicked"
)

}
