% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip.R
\name{tooltip}
\alias{tooltip}
\title{Tooltips}
\usage{
tooltip(.tag, text, placement = "top")
}
\arguments{
\item{.tag}{A tag element.}

\item{text}{The tooltip text.}

\item{placement}{One of \code{"top"}, \code{"right"}, \code{"bottom"}, or \code{"left"}
specifying what side of the tag element the tooltip appears on.}
}
\description{
Both functions are a means of adding help text to HTML components. Tooltips
appear on hover. To add tooltips to text first wrap the text in a \code{span()} tag
element.
}
\examples{

div(
  "The island of ",
  span("Yll") \%>\%
    tooltip("An island of south of the Commonwealth")
)

}
\seealso{
Other utilities: \code{\link{affix}},
  \code{\link{background}}, \code{\link{border}},
  \code{\link{display}}, \code{\link{flex}},
  \code{\link{float}}, \code{\link{font}},
  \code{\link{markInvalid}}, \code{\link{padding}},
  \code{\link{rounded}}, \code{\link{sendModal}},
  \code{\link{showPopover}}, \code{\link{updateChoices}},
  \code{\link{width}}
}
