% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip.R
\name{tooltip}
\alias{tooltip}
\alias{popover}
\title{Add Tooltips and Popovers}
\usage{
tooltip(content, text, placement = "top")

popover(content, text, placement = "top")
}
\arguments{
\item{content}{The tag to add a tooltip or popover to.}

\item{text}{The tooltip or popover text.}

\item{placement}{A character vector specifying where the tooltip or popover
is placed, defaults to top, one of \code{"left"}, \code{"top"}, \code{"right"}, or
\code{"bottom"}.}
}
\description{
Both functions are a means of adding help text to HTML components. Tooltips
appear on hover and popovers appear when the element is clicked.
}
\examples{
tooltip(
  button("My button!"),
  text = "This is my button",
)

tooltip(
  button("Another button"),
  text = "I won't do much when clicked, sorry",
  placement = "bottom"
)

listGroup(
  tooltip(
    listItem("Item 1"),
    "This is the first item"
  ),
  tooltip(
    listItem("Item 2"),
    "This is the second item"
  )
)

popover(
  button("Click me!"),
  "This text appears on click"
)

# list items with popovers
listGroup(
  popover(
    listItem("Some text"),
    "The first list item"
  ),
  popover(
    listItem("More text"),
    "The second list item"
  ),
  popover(
    listItem("Final text"),
    "Did you find this text?"
  )
)

}
