% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip.R
\name{tooltip}
\alias{tooltip}
\title{Tooltips}
\usage{
tooltip(tag, text, placement = "top")
}
\arguments{
\item{tag}{A tag element.}

\item{text}{The tooltip text.}

\item{placement}{One of \code{"top"}, \code{"right"}, \code{"bottom"}, or \code{"left"}
specifying what side of the tag element the tooltip appears on.}
}
\description{
Add a tooltip to a tag element. Tooltips may be placed above, below, left, or
right of an element.
}
\examples{

### Tooltips galore

formGroup(
  label = tags$label(
    "An exciting input",
    tooltip(span(icon("info-circle")), "What is exciting here?")
  ),
  radioInput(
    id = "radios",
    choices = c("Ready", "Set", "Go")
  )
)

### Describing links (link inputs)

div(
  p("Nunc rutrum turpis sed pede."),
  p(
    "Donec posuere augue in ",
    linkInput(NULL, "quam.") \%>\%
      tooltip("This is bound to do something")
  ),
  p(
    "Etiam vel tortor sodales ",
    linkInput(NULL, "tellus") \%>\%
      tooltip("Tell us more?"),
    " ultricies commodo."
  )
)

}
\seealso{
Other content: \code{\link{alert}}, \code{\link{badge}},
  \code{\link{blockquote}}, \code{\link{card}},
  \code{\link{collapsiblePane}}, \code{\link{d1}},
  \code{\link{dropdown}}, \code{\link{img}},
  \code{\link{jumbotron}}, \code{\link{modal}},
  \code{\link{navContent}}, \code{\link{popover}},
  \code{\link{pre}}, \code{\link{progressOutlet}},
  \code{\link{toast}}
}
\concept{content}
