% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip.R
\name{tooltip}
\alias{tooltip}
\alias{popover}
\title{Add Tooltips and Popovers}
\usage{
tooltip(content, text, placement = "top")

popover(content, text, placement = "top")
}
\arguments{
\item{content}{The tag to add a tooltip or popover to.}

\item{text}{The tooltip or popover text.}

\item{placement}{A character vector specifying where the tooltip or popover
is placed, defaults to top, one of \code{"left"}, \code{"top"}, \code{"right"}, or
\code{"bottom"}.}
}
\description{
Both functions are a means of adding help text to HTML components. Tooltips
appear on hover and popovers appear when the element is clicked.
}
\examples{
if (interactive()) {
  shinyApp(
    ui = container(
      listGroupInput(
        tooltip(
          listGroupItem("Item 1"),
          "This is the first item",
          placement = "bottom"
        ),
        tooltip(
          listGroupItem("Item 2"),
          "This is the second item"
        )
      )
    ),
    server = function(input, output) {

    }
  )
}

popover(
  buttonInput("Click me!"),
  "This text appears on click"
)

# list items with popovers
listGroup(
  popover(
    listGroupItem("Some text"),
    "The first list item"
  ),
  popover(
    listGroupItem("More text"),
    "The second list item"
  ),
  popover(
    listGroupItem("Final text"),
    "Did you find this text?"
  )
)

}
