% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{typography}
\alias{typography}
\alias{display1}
\alias{display2}
\alias{display3}
\alias{display4}
\alias{display5}
\title{Text and Heading Styles}
\usage{
display1(...)

display2(...)

display3(...)

display4(...)

display5(...)
}
\arguments{
\item{...}{Child elements, additional HTML classes, or HTML attributes
passed to the parent element.}
}
\description{
These function encapsulate Bootstrap classes for styling text and headings.
Display headings are not meant to replace the standard HTML heading tags,
they are a stand out alternative for eye-catching titles. Use \code{lead} to
emphasize a paragraph or snippet of text.

\code{mark} and \code{small} are HTML class-based alternatives to \code{<mark>} and
\code{<small>}. Typically, \code{<mark>} is used to indicate relevance. By using
Bootstrap's class alternatives one can utilize the highlighting aspect
of \code{<mark>} while avoiding the usage and purpose restrictions.
}
\examples{
display1("Eye-catching!")

}
