% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{updateChoices}
\alias{updateChoices}
\alias{updateValues}
\alias{updateSelected}
\title{Update choices, values, selected choices}
\usage{
updateChoices(id, ...)

updateValues(id, ...)

updateSelected(id, ...)
}
\arguments{
\item{id}{A character string specifying the id of an input to update.}

\item{...}{Named values or unnamed values specifying which and how existing
or new input choices, values, or which values are selected are updated,
values are \code{\link[=dots_list]{dots_list()}} allowing the use of bangs and splices.}
}
\description{
Functions to update choices, values, and selected choices.
}
\seealso{
Other utilities: \code{\link{affix}},
  \code{\link{background}}, \code{\link{border}},
  \code{\link{display}}, \code{\link{flex}},
  \code{\link{float}}, \code{\link{font}},
  \code{\link{markInvalid}}, \code{\link{padding}},
  \code{\link{rounded}}, \code{\link{sendModal}},
  \code{\link{showPopover}}, \code{\link{tooltip}},
  \code{\link{width}}

Other utilities: \code{\link{affix}},
  \code{\link{background}}, \code{\link{border}},
  \code{\link{display}}, \code{\link{flex}},
  \code{\link{float}}, \code{\link{font}},
  \code{\link{markInvalid}}, \code{\link{padding}},
  \code{\link{rounded}}, \code{\link{sendModal}},
  \code{\link{showPopover}}, \code{\link{tooltip}},
  \code{\link{width}}

Other utilities: \code{\link{affix}},
  \code{\link{background}}, \code{\link{border}},
  \code{\link{display}}, \code{\link{flex}},
  \code{\link{float}}, \code{\link{font}},
  \code{\link{markInvalid}}, \code{\link{padding}},
  \code{\link{rounded}}, \code{\link{sendModal}},
  \code{\link{showPopover}}, \code{\link{tooltip}},
  \code{\link{width}}
}
