% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{updateInput}
\alias{updateInput}
\title{Update input}
\usage{
updateInput(id, choices, values = choices, selected = NULL,
  session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{A character string.}

\item{choices}{A character vector, tag element, or list specifying choices
for the input.}

\item{values}{An object specifying values for the input, this object is
coerced to a character vector.}

\item{selected}{One of \code{values} specifying which choice to select, defaults
to \code{NULL}, in which case a choice is not selected. Note that browsers may
automatically select a choice if not specified.}

\item{session}{A reactive context, defaults to \code{\link[=getDefaultReactiveDomain]{getDefaultReactiveDomain()}}.}
}
\description{
Input utilities.
}
