% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{updateInput}
\alias{updateInput}
\title{Update input choices and values}
\usage{
updateInput(id, choices, values = choices, selected = NULL,
  session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{A character string specifying the id of an input.}

\item{choices}{A character vector or list of tag elements specifying new
choices for the input.}

\item{values}{An atomic vector specifying new values for the input.}

\item{selected}{One of \code{values} specifying which choice to select, defaults
to \code{NULL}, in which case a choice is not selected. Note that browsers may
automatically select a choice if not specified.}

\item{session}{A reactive context, defaults to \code{\link[=getDefaultReactiveDomain]{getDefaultReactiveDomain()}}.}
}
\description{
Use \code{updateInput()} to replace an input's choices and values. Optionally,
a new choices may be selected.
}
\section{Button inputs}{


When updating a button input if \code{values} equals \code{choices}, the default value
for \code{values}, the value of the button input is left as is.

Passing a non-numeric value will reset the button input's value.
}

\section{File inputs}{


Files inputs do not currently support \code{updateInput()}.
}

\section{Form inputs}{


Form inputs do not support \code{updateInput()}, instead update the specific
inputs within the form.
}

\seealso{
Other utilities: \code{\link{downloadEvent}}
}
\concept{utilities}
