% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{width}
\alias{width}
\alias{height}
\title{Tag width and height}
\usage{
width(tag, percentage = NULL, max = NULL)

height(percentage = NULL, max = NULL)
}
\arguments{
\item{percentage}{One of 25, 50, 75, or 100 specifying width or height as a
percentage of a parent element's width or height.}

\item{max}{One of 25, 50, 75, or 100 specifying max width or max height as a
percentage of a parent element's width or height.}
}
\description{
Used in conjunction with \link{tagReduce} to change a tag element's width or
height. Widths and heights are specified as percentages of the parent
object's width or height.
}
\examples{
tags$div() \%>\%
  width(25) \%>\%
  height(100)

tags$div() \%>\%
  width(max = 75)

}
\seealso{
Other utilities: \code{\link{affix}}, \code{\link{align}},
  \code{\link{background}}, \code{\link{border}},
  \code{\link{display}}, \code{\link{float}},
  \code{\link{font}}, \code{\link{padding}},
  \code{\link{text}}

Other utilities: \code{\link{affix}}, \code{\link{align}},
  \code{\link{background}}, \code{\link{border}},
  \code{\link{display}}, \code{\link{float}},
  \code{\link{font}}, \code{\link{padding}},
  \code{\link{text}}
}
