% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{width}
\alias{width}
\alias{height}
\title{Tag element width and height}
\usage{
width(.tag, percentage = NULL, max = NULL)

height(.tag, percentage = NULL, max = NULL)
}
\arguments{
\item{.tag}{A tag element.}

\item{percentage}{One of 25, 50, 75, or 100 specifying width or height as a
percentage of a parent element's width or height.}

\item{max}{One of 25, 50, 75, or 100 specifying max width or max height as a
percentage of a parent element's width or height.}
}
\description{
Utility functions to change a tag element's width or height. Widths and
heights are specified as percentages of the parent object's width or height.
}
\examples{

tags$div() \%>\%
  width(25) \%>\%
  height(100)

tags$div() \%>\%
  width(max = 75)

}
\seealso{
Other utilities: \code{\link{affix}},
  \code{\link{background}}, \code{\link{border}},
  \code{\link{display}}, \code{\link{float}},
  \code{\link{font}}, \code{\link{padding}},
  \code{\link{rounded}}, \code{\link{text}}

Other utilities: \code{\link{affix}},
  \code{\link{background}}, \code{\link{border}},
  \code{\link{display}}, \code{\link{float}},
  \code{\link{font}}, \code{\link{padding}},
  \code{\link{rounded}}, \code{\link{text}}
}
\concept{utilities}
