% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design-sizing.R
\name{width}
\alias{width}
\title{Width}
\usage{
width(x, size)
}
\arguments{
\item{x}{A tag element or \link{.style} pronoun.}

\item{size}{A character string or number specifying the width of the tag
element. Possible values:

One of 25, 50, 75, or 100 specifying the element's width is a percentage of
its parent's width. The width of the parent element must be
specified. Percentages do not account for margins or padding and may cause
an element to extend beyond its parent element.

\code{"auto"}, in which case the element's width is determined by the browser.
The browser will take into account the width, padding, margins, and border
of the tag element's parent to keep the element from extending beyond its
parent.

\code{"viewport"}, in which case the element's width is determined by the size
of the browser window.}
}
\description{
Utility function to change a tag element's width. Widths are specified
relative the width of a parent element, an element's content, or the size of
the browser window.
}
\seealso{
Other design utilities: 
\code{\link{active}()},
\code{\link{affix}()},
\code{\link{background}()},
\code{\link{border}()},
\code{\link{display}()},
\code{\link{flex}()},
\code{\link{float}()},
\code{\link{font}()},
\code{\link{height}()},
\code{\link{margin}()},
\code{\link{padding}()},
\code{\link{scroll}()},
\code{\link{shadow}()}
}
\concept{design utilities}
