% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex.R
\name{wrap}
\alias{wrap}
\title{Wrapping flex items}
\usage{
wrap(tag, default = NULL, sm = NULL, md = NULL, lg = NULL, xl = NULL)
}
\arguments{
\item{tag}{A tag element.}

\item{default}{One of \code{"nowrap"}, \code{"wrap"}, or \code{"reverse"} specifying how the
flex items of an element wrap.}

\item{sm}{Like \code{default}, but the wrapping behavior is applied once the
viewport is 576 pixels wide, think phone in landscape mode.}

\item{md}{Like \code{default}, but the wrapping behavior is applied once the
viewport is 768 pixels wide, think tablets.}

\item{lg}{Like \code{default}, but the wrapping behavior is applied once the
viewport is 992 pixels wide, think desktop.}

\item{xl}{Like \code{default}, but the wrapping behavior is applied once the
viewport is 1200 pixels wide, think large desktop.}
}
\description{
This function applies bootstrap classes to a tag element to change how the
element's flex items wrap or do not wrap. By default items will not wrap
onto new lines. See the \emph{Wrapping} section below for more information on
the possible wrapping behaviors.
}
\section{Wrapping}{


\strong{\code{"nowrap"}}, flex items do not wrap onto a new line and will extend
beyond the boundaries of the parent element. This is the browser default.\preformatted{| Item | Item | Item | Item | Item | Item |
| 1    | 2    | 3    | 4    | 5    | 6    |
}

\strong{\code{"wrap"}}, flex items will wrap onto a new line.\preformatted{| Item 1 | Item 2 | Item 3 | Item 4 | === |
| Item 5 | Item 6 | ===================== |
}

\strong{\code{"reverse"}}, rows of flex items appear in reverse order wrapping from the
bottom of the parent element up.\preformatted{| Item 5 | Item 6 | ===================== |
| Item 1 | Item 2 | Item 3 | Item 4 | === |
}
}

\examples{
# Make sure to try resizing the browser or viewer window after running
# the following examples.

if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          tags$p("No wrap") \%>\%
            font(style = "italics"),
          lapply(
            1:15,
            . \%>\%
              paste("Flex item", .) \%>\%
              tags$div() \%>\%
              width(25) \%>\%
              border()
          ) \%>\%
            tags$div() \%>\%
            width(100) \%>\%
            display("flex")
        )
      )
    ),
    server = function(input, output) {

    }
  )
}

if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          tags$p("wrap") \%>\%
            font(style = "italics"),
          lapply(
            1:15,
            . \%>\%
              paste("Flex item", .) \%>\%
              tags$div() \%>\%
              width(25) \%>\%
              border()
          ) \%>\%
            tags$div() \%>\%
            width(100) \%>\%
            display("flex") \%>\%
            wrap("wrap")
        )
      )
    ),
    server = function(input, output) {

    }
  )
}

}
\seealso{
Other flex: \code{\link{content}}, \code{\link{direction}},
  \code{\link{items}}
}
\concept{flex}
