% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yonder.R
\docType{package}
\name{yonder}
\alias{yonder-package}
\alias{yonder}
\title{A shiny framework}
\description{
Yonder is a set of tools for flexible and creative shiny application
construction and design.
}
\section{Getting started}{

\subsection{Inputs}{

Yonder provides many familiar inputs like \code{\link[=input_select]{input_select()}} or
\code{\link[=input_radio_group]{input_radio_group()}}. There are also new inputs like \code{\link[=input_text_group]{input_text_group()}} or
\code{\link[=input_form]{input_form()}}.
\itemize{
\item Input functions have an \code{id} argument instead of \code{inputId}.
\item Input functions include a \code{label} argument for adding a label above the
input. Pass a \code{\link[=floating_label]{floating_label()}} to a text, numeric, select, or text
group input to render a floating label instead.
}
\subsection{Looking for ... ?}{
\itemize{
\item \code{actionButton()} or \code{actionLink()} use \code{\link[=input_button]{input_button()}} or \code{\link[=input_link]{input_link()}}
\item \code{radioButtons()} use \code{\link[=input_radio_group]{input_radio_group()}}
\item \code{checkboxGroupInput()} use \code{\link[=input_checkbox_group]{input_checkbox_group()}} or \code{\link[=input_checkbox]{input_checkbox()}}
\item \code{numericInput()} use \code{\link[=input_numeric]{input_numeric()}}
\item \code{selectizeInput()} use \code{\link[=input_select]{input_select()}} or \code{\link[=input_multi_select]{input_multi_select()}}
\item \code{submitButton()} use \code{\link[=input_form]{input_form()}} and \code{\link[=form_submit_button]{form_submit_button()}}
}
}

}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://nteetor.github.io/yonder}
  \item Report bugs at \url{https://github.com/nteetor/yonder/issues}
}

}
\author{
\strong{Maintainer}: Nathan Teetor \email{nathanteetor@gmail.com}

Authors:
\itemize{
  \item Nathan Teetor \email{nathanteetor@gmail.com}
}

Other contributors:
\itemize{
  \item  The Bootstrap Authors (Bootstrap library) [copyright holder]
  \item  Twitter, Inc (Bootstrap library) [copyright holder]
  \item  JS Foundation (jQuery library) [copyright holder]
  \item Federico Zivolo (popper.js library) [contributor, copyright holder]
  \item Johann Servoire (bs-custom-file-input library) [contributor, copyright holder]
}

}
