% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yonder.R
\docType{package}
\name{yonder}
\alias{yonder}
\alias{yonder-package}
\title{A new approach to shiny applications}
\description{
More freedom to the programmer, more to come.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nteetor/yonder}
}

}
\author{
\strong{Maintainer}: Nathan Teetor \email{nathanteetor@gmail.com}

}
