% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yonder.R
\docType{package}
\name{yonder}
\alias{yonder}
\alias{yonder-package}
\title{A new approach to shiny applications}
\description{
More freedom to the programmer, more to come.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nteetor/yonder}
  \item Report bugs at \url{https://github.com/nteetor/yonder/issues}
}

}
\author{
\strong{Maintainer}: Nathan Teetor \email{nathanteetor@gmail.com}

Other contributors:
\itemize{
  \item  The Bootstrap Authors (Bootstrap library) [copyright holder]
  \item  Twitter, Inc (Bootstrap library) [copyright holder]
  \item Eli Grey (FileSaver.js library) [contributor, copyright holder]
  \item Denis Ineshin (ion.rangeSlider library) [contributor, copyright holder]
  \item  JS Foundation (jQuery library) [copyright holder]
  \item Gregory Petrosyan (flatpicker library) [contributor, copyright holder]
  \item Federico Zivolo (popper.js library) [contributor, copyright holder]
  \item Julian Shapiro (Velocity library) [contributor, copyright holder]
}

}
